aws-elixir v0.6.0 AWS.APIGateway

Amazon API Gateway

Amazon API Gateway helps developers deliver robust, secure, and scalable mobile and web application back ends.

API Gateway allows developers to securely connect mobile and web applications to APIs that run on AWS Lambda, Amazon EC2, or other publicly addressable web services that are hosted outside of AWS.

Functions

create_api_key(client, input, options \\ [])

Create an ApiKey resource.

See also: AWS CLI

create_authorizer(client, rest_api_id, input, options \\ [])

Adds a new Authorizer resource to an existing RestApi resource.

See also: AWS CLI

create_base_path_mapping(client, domain_name, input, options \\ [])

Creates a new BasePathMapping resource.

create_deployment(client, rest_api_id, input, options \\ [])

Creates a Deployment resource, which makes a specified RestApi callable over the internet.

create_documentation_part(client, rest_api_id, input, options \\ [])

create_documentation_version(client, rest_api_id, input, options \\ [])

create_domain_name(client, input, options \\ [])

Creates a new domain name.

create_model(client, rest_api_id, input, options \\ [])

Adds a new Model resource to an existing RestApi resource.

create_request_validator(client, rest_api_id, input, options \\ [])

Creates a ReqeustValidator of a given RestApi.

create_resource(client, parent_id, rest_api_id, input, options \\ [])

Creates a Resource resource.

create_rest_api(client, input, options \\ [])

Creates a new RestApi resource.

create_stage(client, rest_api_id, input, options \\ [])

Creates a new Stage resource that references a pre-existing Deployment for the API.

create_usage_plan(client, input, options \\ [])

Creates a usage plan with the throttle and quota limits, as well as the associated API stages, specified in the payload.

create_usage_plan_key(client, usage_plan_id, input, options \\ [])

Creates a usage plan key for adding an existing API key to a usage plan.

create_vpc_link(client, input, options \\ [])

Creates a VPC link, under the caller's account in a selected region, in an asynchronous operation that typically takes 2-4 minutes to complete and become operational.

The caller must have permissions to create and update VPC Endpoint services.

delete_api_key(client, api_key, input, options \\ [])

Deletes the ApiKey resource.

delete_authorizer(client, authorizer_id, rest_api_id, input, options \\ [])

Deletes an existing Authorizer resource.

See also: AWS CLI

delete_base_path_mapping(client, base_path, domain_name, input, options \\ [])

Deletes the BasePathMapping resource.

delete_client_certificate(client, client_certificate_id, input, options \\ [])

Deletes the ClientCertificate resource.

delete_deployment(client, deployment_id, rest_api_id, input, options \\ [])

Deletes a Deployment resource.

Deleting a deployment will only succeed if there are no Stage resources associated with it.

delete_documentation_part(client, documentation_part_id, rest_api_id, input, options \\ [])

delete_documentation_version(client, documentation_version, rest_api_id, input, options \\ [])

delete_domain_name(client, domain_name, input, options \\ [])

Deletes the DomainName resource.

delete_gateway_response(client, response_type, rest_api_id, input, options \\ [])

Clears any customization of a GatewayResponse of a specified response type on the given RestApi and resets it with the default settings.

delete_integration(client, http_method, resource_id, rest_api_id, input, options \\ [])

Represents a delete integration.

delete_integration_response(client, http_method, resource_id, rest_api_id, status_code, input, options \\ [])

Represents a delete integration response.

delete_method(client, http_method, resource_id, rest_api_id, input, options \\ [])

Deletes an existing Method resource.

delete_method_response(client, http_method, resource_id, rest_api_id, status_code, input, options \\ [])

Deletes an existing MethodResponse resource.

delete_model(client, model_name, rest_api_id, input, options \\ [])

Deletes a model.

delete_request_validator(client, request_validator_id, rest_api_id, input, options \\ [])

Deletes a RequestValidator of a given RestApi.

delete_resource(client, resource_id, rest_api_id, input, options \\ [])

Deletes a Resource resource.

delete_rest_api(client, rest_api_id, input, options \\ [])

Deletes the specified API.

delete_stage(client, rest_api_id, stage_name, input, options \\ [])

Deletes a Stage resource.

delete_usage_plan(client, usage_plan_id, input, options \\ [])

Deletes a usage plan of a given plan Id.

delete_usage_plan_key(client, key_id, usage_plan_id, input, options \\ [])

Deletes a usage plan key and remove the underlying API key from the associated usage plan.

delete_vpc_link(client, vpc_link_id, input, options \\ [])

Deletes an existing VpcLink of a specified identifier.

flush_stage_authorizers_cache(client, rest_api_id, stage_name, input, options \\ [])

Flushes all authorizer cache entries on a stage.

flush_stage_cache(client, rest_api_id, stage_name, input, options \\ [])

Flushes a stage's cache.

generate_client_certificate(client, input, options \\ [])

Generates a ClientCertificate resource.

get_account(client, options \\ [])

Gets information about the current Account resource.

get_api_key(client, api_key, include_value \\ nil, options \\ [])

Gets information about the current ApiKey resource.

get_api_keys(client, customer_id \\ nil, include_values \\ nil, limit \\ nil, name_query \\ nil, position \\ nil, options \\ [])

Gets information about the current ApiKeys resource.

get_authorizer(client, authorizer_id, rest_api_id, options \\ [])

Describe an existing Authorizer resource.

See also: AWS CLI

get_authorizers(client, rest_api_id, limit \\ nil, position \\ nil, options \\ [])

Describe an existing Authorizers resource.

See also: AWS CLI

get_base_path_mapping(client, base_path, domain_name, options \\ [])

Describe a BasePathMapping resource.

get_base_path_mappings(client, domain_name, limit \\ nil, position \\ nil, options \\ [])

Represents a collection of BasePathMapping resources.

get_client_certificate(client, client_certificate_id, options \\ [])

Gets information about the current ClientCertificate resource.

get_client_certificates(client, limit \\ nil, position \\ nil, options \\ [])

Gets a collection of ClientCertificate resources.

get_deployment(client, deployment_id, rest_api_id, embed \\ nil, options \\ [])

Gets information about a Deployment resource.

get_deployments(client, rest_api_id, limit \\ nil, position \\ nil, options \\ [])

Gets information about a Deployments collection.

get_documentation_part(client, documentation_part_id, rest_api_id, options \\ [])

get_documentation_parts(client, rest_api_id, limit \\ nil, location_status \\ nil, name_query \\ nil, path \\ nil, position \\ nil, type \\ nil, options \\ [])

get_documentation_version(client, documentation_version, rest_api_id, options \\ [])

get_documentation_versions(client, rest_api_id, limit \\ nil, position \\ nil, options \\ [])

get_domain_name(client, domain_name, options \\ [])

Represents a domain name that is contained in a simpler, more intuitive URL that can be called.

get_domain_names(client, limit \\ nil, position \\ nil, options \\ [])

Represents a collection of DomainName resources.

get_export(client, export_type, rest_api_id, stage_name, parameters \\ nil, accepts \\ nil, options \\ [])

Exports a deployed version of a RestApi in a specified format.

get_gateway_response(client, response_type, rest_api_id, options \\ [])

Gets a GatewayResponse of a specified response type on the given RestApi.

get_gateway_responses(client, rest_api_id, limit \\ nil, position \\ nil, options \\ [])

Gets the GatewayResponses collection on the given RestApi.

If an API developer has not added any definitions for gateway responses, the result will be the API Gateway-generated default GatewayResponses collection for the supported response types.

get_integration(client, http_method, resource_id, rest_api_id, options \\ [])

Get the integration settings.

get_integration_response(client, http_method, resource_id, rest_api_id, status_code, options \\ [])

Represents a get integration response.

get_method(client, http_method, resource_id, rest_api_id, options \\ [])

Describe an existing Method resource.

get_method_response(client, http_method, resource_id, rest_api_id, status_code, options \\ [])

Describes a MethodResponse resource.

get_model(client, model_name, rest_api_id, flatten \\ nil, options \\ [])

Describes an existing model defined for a RestApi resource.

get_model_template(client, model_name, rest_api_id, options \\ [])

Generates a sample mapping template that can be used to transform a payload into the structure of a model.

get_models(client, rest_api_id, limit \\ nil, position \\ nil, options \\ [])

Describes existing Models defined for a RestApi resource.

get_request_validator(client, request_validator_id, rest_api_id, options \\ [])

Gets a RequestValidator of a given RestApi.

get_request_validators(client, rest_api_id, limit \\ nil, position \\ nil, options \\ [])

Gets the RequestValidators collection of a given RestApi.

get_resource(client, resource_id, rest_api_id, embed \\ nil, options \\ [])

Lists information about a resource.

get_resources(client, rest_api_id, embed \\ nil, limit \\ nil, position \\ nil, options \\ [])

Lists information about a collection of Resource resources.

get_rest_api(client, rest_api_id, options \\ [])

Lists the RestApi resource in the collection.

get_rest_apis(client, limit \\ nil, position \\ nil, options \\ [])

Lists the RestApis resources for your collection.

get_sdk(client, rest_api_id, sdk_type, stage_name, parameters \\ nil, options \\ [])

Generates a client SDK for a RestApi and Stage.

get_sdk_type(client, id, options \\ [])

get_sdk_types(client, limit \\ nil, position \\ nil, options \\ [])

get_stage(client, rest_api_id, stage_name, options \\ [])

Gets information about a Stage resource.

get_stages(client, rest_api_id, deployment_id \\ nil, options \\ [])

Gets information about one or more Stage resources.

get_tags(client, resource_arn, limit \\ nil, position \\ nil, options \\ [])

Gets the Tags collection for a given resource.

get_usage(client, usage_plan_id, end_date, key_id \\ nil, limit \\ nil, position \\ nil, start_date, options \\ [])

Gets the usage data of a usage plan in a specified time interval.

get_usage_plan(client, usage_plan_id, options \\ [])

Gets a usage plan of a given plan identifier.

get_usage_plan_key(client, key_id, usage_plan_id, options \\ [])

Gets a usage plan key of a given key identifier.

get_usage_plan_keys(client, usage_plan_id, limit \\ nil, name_query \\ nil, position \\ nil, options \\ [])

Gets all the usage plan keys representing the API keys added to a specified usage plan.

get_usage_plans(client, key_id \\ nil, limit \\ nil, position \\ nil, options \\ [])

Gets all the usage plans of the caller's account.

get_vpc_link(client, vpc_link_id, options \\ [])

Gets a specified VPC link under the caller's account in a region.

get_vpc_links(client, limit \\ nil, position \\ nil, options \\ [])

Gets the VpcLinks collection under the caller's account in a selected region.

import_api_keys(client, input, options \\ [])

Import API keys from an external source, such as a CSV-formatted file.

import_documentation_parts(client, rest_api_id, input, options \\ [])

import_rest_api(client, input, options \\ [])

A feature of the API Gateway control service for creating a new API from an external API definition file.

put_gateway_response(client, response_type, rest_api_id, input, options \\ [])

Creates a customization of a GatewayResponse of a specified response type and status code on the given RestApi.

put_integration(client, http_method, resource_id, rest_api_id, input, options \\ [])

Sets up a method's integration.

put_integration_response(client, http_method, resource_id, rest_api_id, status_code, input, options \\ [])

Represents a put integration.

put_method(client, http_method, resource_id, rest_api_id, input, options \\ [])

Add a method to an existing Resource resource.

put_method_response(client, http_method, resource_id, rest_api_id, status_code, input, options \\ [])

Adds a MethodResponse to an existing Method resource.

put_rest_api(client, rest_api_id, input, options \\ [])

A feature of the API Gateway control service for updating an existing API with an input of external API definitions.

The update can take the form of merging the supplied definition into the existing API or overwriting the existing API.

tag_resource(client, resource_arn, input, options \\ [])

Adds or updates a tag on a given resource.

test_invoke_authorizer(client, authorizer_id, rest_api_id, input, options \\ [])

Simulate the execution of an Authorizer in your RestApi with headers, parameters, and an incoming request body.

See also: Use Lambda Function as Authorizer Use Cognito User Pool as Authorizer

test_invoke_method(client, http_method, resource_id, rest_api_id, input, options \\ [])

Simulate the execution of a Method in your RestApi with headers, parameters, and an incoming request body.

untag_resource(client, resource_arn, input, options \\ [])

Removes a tag from a given resource.

update_account(client, input, options \\ [])

Changes information about the current Account resource.

update_api_key(client, api_key, input, options \\ [])

Changes information about an ApiKey resource.

update_authorizer(client, authorizer_id, rest_api_id, input, options \\ [])

Updates an existing Authorizer resource.

See also: AWS CLI

update_base_path_mapping(client, base_path, domain_name, input, options \\ [])

Changes information about the BasePathMapping resource.

update_client_certificate(client, client_certificate_id, input, options \\ [])

Changes information about an ClientCertificate resource.

update_deployment(client, deployment_id, rest_api_id, input, options \\ [])

Changes information about a Deployment resource.

update_documentation_part(client, documentation_part_id, rest_api_id, input, options \\ [])

update_documentation_version(client, documentation_version, rest_api_id, input, options \\ [])

update_domain_name(client, domain_name, input, options \\ [])

Changes information about the DomainName resource.

update_gateway_response(client, response_type, rest_api_id, input, options \\ [])

Updates a GatewayResponse of a specified response type on the given RestApi.

update_integration(client, http_method, resource_id, rest_api_id, input, options \\ [])

Represents an update integration.

update_integration_response(client, http_method, resource_id, rest_api_id, status_code, input, options \\ [])

Represents an update integration response.

update_method(client, http_method, resource_id, rest_api_id, input, options \\ [])

Updates an existing Method resource.

update_method_response(client, http_method, resource_id, rest_api_id, status_code, input, options \\ [])

Updates an existing MethodResponse resource.

update_model(client, model_name, rest_api_id, input, options \\ [])

Changes information about a model.

update_request_validator(client, request_validator_id, rest_api_id, input, options \\ [])

Updates a RequestValidator of a given RestApi.

update_resource(client, resource_id, rest_api_id, input, options \\ [])

Changes information about a Resource resource.

update_rest_api(client, rest_api_id, input, options \\ [])

Changes information about the specified API.

update_stage(client, rest_api_id, stage_name, input, options \\ [])

Changes information about a Stage resource.

update_usage(client, key_id, usage_plan_id, input, options \\ [])

Grants a temporary extension to the remaining quota of a usage plan associated with a specified API key.

update_usage_plan(client, usage_plan_id, input, options \\ [])

Updates a usage plan of a given plan Id.

update_vpc_link(client, vpc_link_id, input, options \\ [])

Updates an existing VpcLink of a specified identifier.