aws-elixir v0.6.0 AWS.AccessAnalyzer

AWS IAM Access Analyzer helps identify potential resource-access risks by enabling you to identify any policies that grant access to an external principal.

It does this by using logic-based reasoning to analyze resource-based policies in your AWS environment. An external principal can be another AWS account, a root user, an IAM user or role, a federated user, an AWS service, or an anonymous user. This guide describes the AWS IAM Access Analyzer operations that you can call programmatically. For general information about Access Analyzer, see the AWS IAM Access Analyzer section of the IAM User Guide.

To start using Access Analyzer, you first need to create an analyzer.

Functions

create_analyzer(client, input, options \\ [])

Creates an analyzer for your account.

create_archive_rule(client, analyzer_name, input, options \\ [])

Creates an archive rule for the specified analyzer.

Archive rules automatically archive findings that meet the criteria you define when you create the rule.

delete_analyzer(client, analyzer_name, input, options \\ [])

Deletes the specified analyzer.

When you delete an analyzer, Access Analyzer is disabled for the account in the current or specific Region. All findings that were generated by the analyzer are deleted. You cannot undo this action.

delete_archive_rule(client, analyzer_name, rule_name, input, options \\ [])

Deletes the specified archive rule.

get_analyzed_resource(client, analyzer_arn, resource_arn, options \\ [])

Retrieves information about a resource that was analyzed.

get_analyzer(client, analyzer_name, options \\ [])

Retrieves information about the specified analyzer.

get_archive_rule(client, analyzer_name, rule_name, options \\ [])

Retrieves information about an archive rule.

get_finding(client, id, analyzer_arn, options \\ [])

Retrieves information about the specified finding.

list_analyzed_resources(client, input, options \\ [])

Retrieves a list of resources of the specified type that have been analyzed by the specified analyzer..

list_analyzers(client, max_results \\ nil, next_token \\ nil, type \\ nil, options \\ [])

Retrieves a list of analyzers.

list_archive_rules(client, analyzer_name, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of archive rules created for the specified analyzer.

list_findings(client, input, options \\ [])

Retrieves a list of findings generated by the specified analyzer.

list_tags_for_resource(client, resource_arn, options \\ [])

Retrieves a list of tags applied to the specified resource.

start_resource_scan(client, input, options \\ [])

Immediately starts a scan of the policies applied to the specified resource.

tag_resource(client, resource_arn, input, options \\ [])

Adds a tag to the specified resource.

untag_resource(client, resource_arn, input, options \\ [])

Removes a tag from the specified resource.

update_archive_rule(client, analyzer_name, rule_name, input, options \\ [])

Updates the criteria and values for the specified archive rule.

update_findings(client, input, options \\ [])

Updates the status for the specified findings.