aws-elixir v0.6.0 AWS.AlexaForBusiness

Alexa for Business helps you use Alexa in your organization.

Alexa for Business provides you with the tools to manage Alexa devices, enroll your users, and assign skills, at scale. You can build your own context-aware voice skills using the Alexa Skills Kit and the Alexa for Business API operations. You can also make these available as private skills for your organization. Alexa for Business makes it efficient to voice-enable your products and services, thus providing context-aware voice experiences for your customers. Device makers building with the Alexa Voice Service (AVS) can create fully integrated solutions, register their products with Alexa for Business, and manage them as shared devices in their organization.

Functions

approve_skill(client, input, options \\ [])

Associates a skill with the organization under the customer's AWS account.

If a skill is private, the user implicitly accepts access to this skill during enablement.

associate_contact_with_address_book(client, input, options \\ [])

Associates a contact with a given address book.

associate_device_with_network_profile(client, input, options \\ [])

Associates a device with the specified network profile.

associate_device_with_room(client, input, options \\ [])

Associates a device with a given room.

This applies all the settings from the room profile to the device, and all the skills in any skill groups added to that room. This operation requires the device to be online, or else a manual sync is required.

associate_skill_group_with_room(client, input, options \\ [])

Associates a skill group with a given room.

This enables all skills in the associated skill group on all devices in the room.

associate_skill_with_skill_group(client, input, options \\ [])

Associates a skill with a skill group.

associate_skill_with_users(client, input, options \\ [])

Makes a private skill available for enrolled users to enable on their devices.

create_address_book(client, input, options \\ [])

Creates an address book with the specified details.

create_business_report_schedule(client, input, options \\ [])

Creates a recurring schedule for usage reports to deliver to the specified S3 location with a specified daily or weekly interval.

create_conference_provider(client, input, options \\ [])

Adds a new conference provider under the user's AWS account.

create_contact(client, input, options \\ [])

Creates a contact with the specified details.

create_gateway_group(client, input, options \\ [])

Creates a gateway group with the specified details.

create_network_profile(client, input, options \\ [])

Creates a network profile with the specified details.

create_profile(client, input, options \\ [])

Creates a new room profile with the specified details.

create_room(client, input, options \\ [])

Creates a room with the specified details.

create_skill_group(client, input, options \\ [])

Creates a skill group with a specified name and description.

create_user(client, input, options \\ [])

Creates a user.

delete_address_book(client, input, options \\ [])

Deletes an address book by the address book ARN.

delete_business_report_schedule(client, input, options \\ [])

Deletes the recurring report delivery schedule with the specified schedule ARN.

delete_conference_provider(client, input, options \\ [])

Deletes a conference provider.

delete_contact(client, input, options \\ [])

Deletes a contact by the contact ARN.

delete_device(client, input, options \\ [])

Removes a device from Alexa For Business.

delete_device_usage_data(client, input, options \\ [])

When this action is called for a specified shared device, it allows authorized users to delete the device's entire previous history of voice input data and associated response data.

This action can be called once every 24 hours for a specific shared device.

delete_gateway_group(client, input, options \\ [])

Deletes a gateway group.

delete_network_profile(client, input, options \\ [])

Deletes a network profile by the network profile ARN.

delete_profile(client, input, options \\ [])

Deletes a room profile by the profile ARN.

delete_room(client, input, options \\ [])

Deletes a room by the room ARN.

delete_room_skill_parameter(client, input, options \\ [])

Deletes room skill parameter details by room, skill, and parameter key ID.

delete_skill_authorization(client, input, options \\ [])

Unlinks a third-party account from a skill.

delete_skill_group(client, input, options \\ [])

Deletes a skill group by skill group ARN.

delete_user(client, input, options \\ [])

Deletes a specified user by user ARN and enrollment ARN.

disassociate_contact_from_address_book(client, input, options \\ [])

Disassociates a contact from a given address book.

disassociate_device_from_room(client, input, options \\ [])

Disassociates a device from its current room.

The device continues to be connected to the Wi-Fi network and is still registered to the account. The device settings and skills are removed from the room.

disassociate_skill_from_skill_group(client, input, options \\ [])

Disassociates a skill from a skill group.

disassociate_skill_from_users(client, input, options \\ [])

Makes a private skill unavailable for enrolled users and prevents them from enabling it on their devices.

disassociate_skill_group_from_room(client, input, options \\ [])

Disassociates a skill group from a specified room.

This disables all skills in the skill group on all devices in the room.

forget_smart_home_appliances(client, input, options \\ [])

Forgets smart home appliances associated to a room.

get_address_book(client, input, options \\ [])

Gets address the book details by the address book ARN.

get_conference_preference(client, input, options \\ [])

Retrieves the existing conference preferences.

get_conference_provider(client, input, options \\ [])

Gets details about a specific conference provider.

get_contact(client, input, options \\ [])

Gets the contact details by the contact ARN.

get_device(client, input, options \\ [])

Gets the details of a device by device ARN.

get_gateway(client, input, options \\ [])

Retrieves the details of a gateway.

get_gateway_group(client, input, options \\ [])

Retrieves the details of a gateway group.

get_invitation_configuration(client, input, options \\ [])

Retrieves the configured values for the user enrollment invitation email template.

get_network_profile(client, input, options \\ [])

Gets the network profile details by the network profile ARN.

get_profile(client, input, options \\ [])

Gets the details of a room profile by profile ARN.

get_room(client, input, options \\ [])

Gets room details by room ARN.

get_room_skill_parameter(client, input, options \\ [])

Gets room skill parameter details by room, skill, and parameter key ARN.

get_skill_group(client, input, options \\ [])

Gets skill group details by skill group ARN.

list_business_report_schedules(client, input, options \\ [])

Lists the details of the schedules that a user configured.

A download URL of the report associated with each schedule is returned every time this action is called. A new download URL is returned each time, and is valid for 24 hours.

list_conference_providers(client, input, options \\ [])

Lists conference providers under a specific AWS account.

list_device_events(client, input, options \\ [])

Lists the device event history, including device connection status, for up to 30 days.

list_gateway_groups(client, input, options \\ [])

Retrieves a list of gateway group summaries.

Use GetGatewayGroup to retrieve details of a specific gateway group.

list_gateways(client, input, options \\ [])

Retrieves a list of gateway summaries.

Use GetGateway to retrieve details of a specific gateway. An optional gateway group ARN can be provided to only retrieve gateway summaries of gateways that are associated with that gateway group ARN.

list_skills(client, input, options \\ [])

Lists all enabled skills in a specific skill group.

list_skills_store_categories(client, input, options \\ [])

Lists all categories in the Alexa skill store.

list_skills_store_skills_by_category(client, input, options \\ [])

Lists all skills in the Alexa skill store by category.

list_smart_home_appliances(client, input, options \\ [])

Lists all of the smart home appliances associated with a room.

list_tags(client, input, options \\ [])

Lists all tags for the specified resource.

put_conference_preference(client, input, options \\ [])

Sets the conference preferences on a specific conference provider at the account level.

put_invitation_configuration(client, input, options \\ [])

Configures the email template for the user enrollment invitation with the specified attributes.

put_room_skill_parameter(client, input, options \\ [])

Updates room skill parameter details by room, skill, and parameter key ID.

Not all skills have a room skill parameter.

put_skill_authorization(client, input, options \\ [])

Links a user's account to a third-party skill provider.

If this API operation is called by an assumed IAM role, the skill being linked must be a private skill. Also, the skill must be owned by the AWS account that assumed the IAM role.

register_a_v_s_device(client, input, options \\ [])

Registers an Alexa-enabled device built by an Original Equipment Manufacturer (OEM) using Alexa Voice Service (AVS).

reject_skill(client, input, options \\ [])

Disassociates a skill from the organization under a user's AWS account.

If the skill is a private skill, it moves to an AcceptStatus of PENDING. Any private or public skill that is rejected can be added later by calling the ApproveSkill API.

resolve_room(client, input, options \\ [])

Determines the details for the room from which a skill request was invoked.

This operation is used by skill developers.

revoke_invitation(client, input, options \\ [])

Revokes an invitation and invalidates the enrollment URL.

search_address_books(client, input, options \\ [])

Searches address books and lists the ones that meet a set of filter and sort criteria.

search_contacts(client, input, options \\ [])

Searches contacts and lists the ones that meet a set of filter and sort criteria.

search_devices(client, input, options \\ [])

Searches devices and lists the ones that meet a set of filter criteria.

search_network_profiles(client, input, options \\ [])

Searches network profiles and lists the ones that meet a set of filter and sort criteria.

search_profiles(client, input, options \\ [])

Searches room profiles and lists the ones that meet a set of filter criteria.

search_rooms(client, input, options \\ [])

Searches rooms and lists the ones that meet a set of filter and sort criteria.

search_skill_groups(client, input, options \\ [])

Searches skill groups and lists the ones that meet a set of filter and sort criteria.

search_users(client, input, options \\ [])

Searches users and lists the ones that meet a set of filter and sort criteria.

send_announcement(client, input, options \\ [])

Triggers an asynchronous flow to send text, SSML, or audio announcements to rooms that are identified by a search or filter.

send_invitation(client, input, options \\ [])

Sends an enrollment invitation email with a URL to a user.

The URL is valid for 30 days or until you call this operation again, whichever comes first.

start_device_sync(client, input, options \\ [])

Resets a device and its account to the known default settings.

This clears all information and settings set by previous users in the following ways:

Bluetooth - This unpairs all bluetooth devices paired with your echo device.

Volume - This resets the echo device's volume to the default value.

Notifications - This clears all notifications from your echo device.

Lists - This clears all to-do items from your echo device.

Settings - This internally syncs the room's profile (if the device is assigned to a room), contacts, address books, delegation access for account linking, and communications (if enabled on the room profile).

start_smart_home_appliance_discovery(client, input, options \\ [])

Initiates the discovery of any smart home appliances associated with the room.

tag_resource(client, input, options \\ [])

Adds metadata tags to a specified resource.

untag_resource(client, input, options \\ [])

Removes metadata tags from a specified resource.

update_address_book(client, input, options \\ [])

Updates address book details by the address book ARN.

update_business_report_schedule(client, input, options \\ [])

Updates the configuration of the report delivery schedule with the specified schedule ARN.

update_conference_provider(client, input, options \\ [])

Updates an existing conference provider's settings.

update_contact(client, input, options \\ [])

Updates the contact details by the contact ARN.

update_device(client, input, options \\ [])

Updates the device name by device ARN.

update_gateway(client, input, options \\ [])

Updates the details of a gateway.

If any optional field is not provided, the existing corresponding value is left unmodified.

update_gateway_group(client, input, options \\ [])

Updates the details of a gateway group.

If any optional field is not provided, the existing corresponding value is left unmodified.

update_network_profile(client, input, options \\ [])

Updates a network profile by the network profile ARN.

update_profile(client, input, options \\ [])

Updates an existing room profile by room profile ARN.

update_room(client, input, options \\ [])

Updates room details by room ARN.

update_skill_group(client, input, options \\ [])

Updates skill group details by skill group ARN.