aws-elixir v0.6.0 AWS.Amplify

Amplify enables developers to develop and deploy cloud-powered mobile and web apps.

The Amplify Console provides a continuous delivery and hosting service for web applications. For more information, see the Amplify Console User Guide. The Amplify Framework is a comprehensive set of SDKs, libraries, tools, and documentation for client app development. For more information, see the Amplify Framework.

Functions

create_app(client, input, options \\ [])

Creates a new Amplify app.

create_backend_environment(client, app_id, input, options \\ [])

Creates a new backend environment for an Amplify app.

create_branch(client, app_id, input, options \\ [])

Creates a new branch for an Amplify app.

create_deployment(client, app_id, branch_name, input, options \\ [])

Creates a deployment for a manually deployed Amplify app.

Manually deployed apps are not connected to a repository.

create_domain_association(client, app_id, input, options \\ [])

Creates a new domain association for an Amplify app.

This action associates a custom domain with the Amplify app

create_webhook(client, app_id, input, options \\ [])

Creates a new webhook on an Amplify app.

delete_app(client, app_id, input, options \\ [])

Deletes an existing Amplify app specified by an app ID.

delete_backend_environment(client, app_id, environment_name, input, options \\ [])

Deletes a backend environment for an Amplify app.

delete_branch(client, app_id, branch_name, input, options \\ [])

Deletes a branch for an Amplify app.

delete_domain_association(client, app_id, domain_name, input, options \\ [])

Deletes a domain association for an Amplify app.

delete_job(client, app_id, branch_name, job_id, input, options \\ [])

Deletes a job for a branch of an Amplify app.

delete_webhook(client, webhook_id, input, options \\ [])

Deletes a webhook.

generate_access_logs(client, app_id, input, options \\ [])

Returns the website access logs for a specific time range using a presigned URL.

get_app(client, app_id, options \\ [])

Returns an existing Amplify app by appID.

get_artifact_url(client, artifact_id, options \\ [])

Returns the artifact info that corresponds to an artifact id.

get_backend_environment(client, app_id, environment_name, options \\ [])

Returns a backend environment for an Amplify app.

get_branch(client, app_id, branch_name, options \\ [])

Returns a branch for an Amplify app.

get_domain_association(client, app_id, domain_name, options \\ [])

Returns the domain information for an Amplify app.

get_job(client, app_id, branch_name, job_id, options \\ [])

Returns a job for a branch of an Amplify app.

get_webhook(client, webhook_id, options \\ [])

Returns the webhook information that corresponds to a specified webhook ID.

list_apps(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of the existing Amplify apps.

list_artifacts(client, app_id, branch_name, job_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of artifacts for a specified app, branch, and job.

list_backend_environments(client, app_id, environment_name \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the backend environments for an Amplify app.

list_branches(client, app_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the branches of an Amplify app.

list_domain_associations(client, app_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns the domain associations for an Amplify app.

list_jobs(client, app_id, branch_name, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the jobs for a branch of an Amplify app.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns a list of tags for a specified Amazon Resource Name (ARN).

list_webhooks(client, app_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of webhooks for an Amplify app.

start_deployment(client, app_id, branch_name, input, options \\ [])

Starts a deployment for a manually deployed app.

Manually deployed apps are not connected to a repository.

start_job(client, app_id, branch_name, input, options \\ [])

Starts a new job for a branch of an Amplify app.

stop_job(client, app_id, branch_name, job_id, input, options \\ [])

Stops a job that is in progress for a branch of an Amplify app.

tag_resource(client, resource_arn, input, options \\ [])

Tags the resource with a tag key and value.

untag_resource(client, resource_arn, input, options \\ [])

Untags a resource with a specified Amazon Resource Name (ARN).

update_app(client, app_id, input, options \\ [])

Updates an existing Amplify app.

update_branch(client, app_id, branch_name, input, options \\ [])

Updates a branch for an Amplify app.

update_domain_association(client, app_id, domain_name, input, options \\ [])

Creates a new domain association for an Amplify app.

update_webhook(client, webhook_id, input, options \\ [])

Updates a webhook.