aws-elixir v0.6.0 AWS.ApiGatewayManagementApi

The Amazon API Gateway Management API allows you to directly manage runtime aspects of your deployed APIs.

To use it, you must explicitly set the SDK's endpoint to point to the endpoint of your deployed API. The endpoint will be of the form https://{api-id}.execute-api.{region}.amazonaws.com/{stage}, or will be the endpoint corresponding to your API's custom domain and base path, if applicable.

Functions

delete_connection(client, connection_id, input, options \\ [])

Delete the connection with the provided id.

get_connection(client, connection_id, options \\ [])

Get information about the connection with the provided id.

post_to_connection(client, connection_id, input, options \\ [])

Sends the provided data to the specified connection.