aws-elixir v0.6.0 AWS.ApiGatewayV2

Amazon API Gateway V2

Functions

create_api(client, input, options \\ [])

Creates an Api resource.

create_api_mapping(client, domain_name, input, options \\ [])

Creates an API mapping.

create_authorizer(client, api_id, input, options \\ [])

Creates an Authorizer for an API.

create_deployment(client, api_id, input, options \\ [])

Creates a Deployment for an API.

create_domain_name(client, input, options \\ [])

Creates a domain name.

create_integration(client, api_id, input, options \\ [])

Creates an Integration.

create_integration_response(client, api_id, integration_id, input, options \\ [])

Creates an IntegrationResponses.

create_model(client, api_id, input, options \\ [])

Creates a Model for an API.

create_route(client, api_id, input, options \\ [])

Creates a Route for an API.

create_route_response(client, api_id, route_id, input, options \\ [])

Creates a RouteResponse for a Route.

create_stage(client, api_id, input, options \\ [])

Creates a Stage for an API.

create_vpc_link(client, input, options \\ [])

Creates a VPC link.

delete_access_log_settings(client, api_id, stage_name, input, options \\ [])

Deletes the AccessLogSettings for a Stage.

To disable access logging for a Stage, delete its AccessLogSettings.

delete_api(client, api_id, input, options \\ [])

Deletes an Api resource.

delete_api_mapping(client, api_mapping_id, domain_name, input, options \\ [])

Deletes an API mapping.

delete_authorizer(client, api_id, authorizer_id, input, options \\ [])

Deletes an Authorizer.

delete_cors_configuration(client, api_id, input, options \\ [])

Deletes a CORS configuration.

delete_deployment(client, api_id, deployment_id, input, options \\ [])

Deletes a Deployment.

delete_domain_name(client, domain_name, input, options \\ [])

Deletes a domain name.

delete_integration(client, api_id, integration_id, input, options \\ [])

Deletes an Integration.

delete_integration_response(client, api_id, integration_id, integration_response_id, input, options \\ [])

Deletes an IntegrationResponses.

delete_model(client, api_id, model_id, input, options \\ [])

Deletes a Model.

delete_route(client, api_id, route_id, input, options \\ [])

Deletes a Route.

delete_route_request_parameter(client, api_id, request_parameter_key, route_id, input, options \\ [])

Deletes a route request parameter.

delete_route_response(client, api_id, route_id, route_response_id, input, options \\ [])

Deletes a RouteResponse.

delete_route_settings(client, api_id, route_key, stage_name, input, options \\ [])

Deletes the RouteSettings for a stage.

delete_stage(client, api_id, stage_name, input, options \\ [])

Deletes a Stage.

delete_vpc_link(client, vpc_link_id, input, options \\ [])

Deletes a VPC link.

export_api(client, api_id, specification, export_version \\ nil, include_extensions \\ nil, output_type, stage_name \\ nil, options \\ [])

Exports a definition of an API in a particular output format and specification.

get_api(client, api_id, options \\ [])

Gets an Api resource.

get_api_mapping(client, api_mapping_id, domain_name, options \\ [])

Gets an API mapping.

get_api_mappings(client, domain_name, max_results \\ nil, next_token \\ nil, options \\ [])

Gets API mappings.

get_apis(client, max_results \\ nil, next_token \\ nil, options \\ [])

Gets a collection of Api resources.

get_authorizer(client, api_id, authorizer_id, options \\ [])

Gets an Authorizer.

get_authorizers(client, api_id, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the Authorizers for an API.

get_deployment(client, api_id, deployment_id, options \\ [])

Gets a Deployment.

get_deployments(client, api_id, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the Deployments for an API.

get_domain_name(client, domain_name, options \\ [])

Gets a domain name.

get_domain_names(client, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the domain names for an AWS account.

get_integration(client, api_id, integration_id, options \\ [])

Gets an Integration.

get_integration_response(client, api_id, integration_id, integration_response_id, options \\ [])

Gets an IntegrationResponses.

get_integration_responses(client, api_id, integration_id, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the IntegrationResponses for an Integration.

get_integrations(client, api_id, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the Integrations for an API.

get_model(client, api_id, model_id, options \\ [])

Gets a Model.

get_model_template(client, api_id, model_id, options \\ [])

Gets a model template.

get_models(client, api_id, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the Models for an API.

get_route(client, api_id, route_id, options \\ [])

Gets a Route.

get_route_response(client, api_id, route_id, route_response_id, options \\ [])

Gets a RouteResponse.

get_route_responses(client, api_id, route_id, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the RouteResponses for a Route.

get_routes(client, api_id, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the Routes for an API.

get_stage(client, api_id, stage_name, options \\ [])

Gets a Stage.

get_stages(client, api_id, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the Stages for an API.

get_tags(client, resource_arn, options \\ [])

Gets a collection of Tag resources.

get_vpc_link(client, vpc_link_id, options \\ [])

Gets a VPC link.

get_vpc_links(client, max_results \\ nil, next_token \\ nil, options \\ [])

Gets a collection of VPC links.

import_api(client, input, options \\ [])

Imports an API.

reimport_api(client, api_id, input, options \\ [])

Puts an Api resource.

reset_authorizers_cache(client, api_id, stage_name, input, options \\ [])

Resets all authorizer cache entries for the specified stage.

Supported only for HTTP API Lambda authorizers.

tag_resource(client, resource_arn, input, options \\ [])

Creates a new Tag resource to represent a tag.

untag_resource(client, resource_arn, input, options \\ [])

Deletes a Tag.

update_api(client, api_id, input, options \\ [])

Updates an Api resource.

update_api_mapping(client, api_mapping_id, domain_name, input, options \\ [])

The API mapping.

update_authorizer(client, api_id, authorizer_id, input, options \\ [])

Updates an Authorizer.

update_deployment(client, api_id, deployment_id, input, options \\ [])

Updates a Deployment.

update_domain_name(client, domain_name, input, options \\ [])

Updates a domain name.

update_integration(client, api_id, integration_id, input, options \\ [])

Updates an Integration.

update_integration_response(client, api_id, integration_id, integration_response_id, input, options \\ [])

Updates an IntegrationResponses.

update_model(client, api_id, model_id, input, options \\ [])

Updates a Model.

update_route(client, api_id, route_id, input, options \\ [])

Updates a Route.

update_route_response(client, api_id, route_id, route_response_id, input, options \\ [])

Updates a RouteResponse.

update_stage(client, api_id, stage_name, input, options \\ [])

Updates a Stage.

update_vpc_link(client, vpc_link_id, input, options \\ [])

Updates a VPC link.