aws-elixir v0.6.0 AWS.AppConfig

AWS AppConfig

Use AWS AppConfig, a capability of AWS Systems Manager, to create, manage, and quickly deploy application configurations.

AppConfig supports controlled deployments to applications of any size and includes built-in validation checks and monitoring. You can use AppConfig with applications hosted on Amazon EC2 instances, AWS Lambda, containers, mobile applications, or IoT devices.

To prevent errors when deploying application configurations, especially for production systems where a simple typo could cause an unexpected outage, AppConfig includes validators. A validator provides a syntactic or semantic check to ensure that the configuration you want to deploy works as intended. To validate your application configuration data, you provide a schema or a Lambda function that runs against the configuration. The configuration deployment or update can only proceed when the configuration data is valid.

During a configuration deployment, AppConfig monitors the application to ensure that the deployment is successful. If the system encounters an error, AppConfig rolls back the change to minimize impact for your application users. You can configure a deployment strategy for each application or environment that includes deployment criteria, including velocity, bake time, and alarms to monitor. Similar to error monitoring, if a deployment triggers an alarm, AppConfig automatically rolls back to the previous version.

AppConfig supports multiple use cases. Here are some examples.

Application tuning: Use AppConfig to carefully introduce changes to your application that can only be tested with production traffic.

Feature toggle: Use AppConfig to turn on new features that require a timely deployment, such as a product launch or announcement.

Allow list: Use AppConfig to allow premium subscribers to access paid content.

Operational issues: Use AppConfig to reduce stress on your application when a dependency or other external factor impacts the system.

This reference is intended to be used with the AWS AppConfig User Guide.

Functions

create_application(client, input, options \\ [])

An application in AppConfig is a logical unit of code that provides capabilities for your customers.

For example, an application can be a microservice that runs on Amazon EC2 instances, a mobile application installed by your users, a serverless application using Amazon API Gateway and AWS Lambda, or any system you run on behalf of others.

create_configuration_profile(client, application_id, input, options \\ [])

Information that enables AppConfig to access the configuration source.

Valid configuration sources include Systems Manager (SSM) documents, SSM Parameter Store parameters, and Amazon S3 objects. A configuration profile includes the following information.

The Uri location of the configuration data.

The AWS Identity and Access Management (IAM) role that provides access to the configuration data.

A validator for the configuration data. Available validators include either a JSON Schema or an AWS Lambda function.

For more information, see Create a Configuration and a Configuration Profile in the AWS AppConfig User Guide.

create_deployment_strategy(client, input, options \\ [])

A deployment strategy defines important criteria for rolling out your configuration to the designated targets.

A deployment strategy includes: the overall duration required, a percentage of targets to receive the deployment during each interval, an algorithm that defines how percentage grows, and bake time.

create_environment(client, application_id, input, options \\ [])

For each application, you define one or more environments.

An environment is a logical deployment group of AppConfig targets, such as applications in a Beta or Production environment. You can also define environments for application subcomponents such as the Web, Mobile and Back-end components for your application. You can configure Amazon CloudWatch alarms for each environment. The system monitors alarms during a configuration deployment. If an alarm is triggered, the system rolls back the configuration.

create_hosted_configuration_version(client, application_id, configuration_profile_id, input, options \\ [])

Create a new configuration in the AppConfig configuration store.

delete_application(client, application_id, input, options \\ [])

Delete an application.

Deleting an application does not delete a configuration from a host.

delete_configuration_profile(client, application_id, configuration_profile_id, input, options \\ [])

Delete a configuration profile.

Deleting a configuration profile does not delete a configuration from a host.

delete_deployment_strategy(client, deployment_strategy_id, input, options \\ [])

Delete a deployment strategy.

Deleting a deployment strategy does not delete a configuration from a host.

delete_environment(client, application_id, environment_id, input, options \\ [])

Delete an environment.

Deleting an environment does not delete a configuration from a host.

delete_hosted_configuration_version(client, application_id, configuration_profile_id, version_number, input, options \\ [])

Delete a version of a configuration from the AppConfig configuration store.

get_application(client, application_id, options \\ [])

Retrieve information about an application.

get_configuration(client, application, configuration, environment, client_configuration_version \\ nil, client_id, options \\ [])

Receive information about a configuration.

AWS AppConfig uses the value of the ClientConfigurationVersion parameter to identify the configuration version on your clients. If you don’t send ClientConfigurationVersion with each call to GetConfiguration, your clients receive the current configuration. You are charged each time your clients receive a configuration.

To avoid excess charges, we recommend that you include the ClientConfigurationVersion value with every call to GetConfiguration. This value must be saved on your client. Subsequent calls to GetConfiguration must pass this value by using the ClientConfigurationVersion parameter.

get_configuration_profile(client, application_id, configuration_profile_id, options \\ [])

Retrieve information about a configuration profile.

get_deployment(client, application_id, deployment_number, environment_id, options \\ [])

Retrieve information about a configuration deployment.

get_deployment_strategy(client, deployment_strategy_id, options \\ [])

Retrieve information about a deployment strategy.

A deployment strategy defines important criteria for rolling out your configuration to the designated targets. A deployment strategy includes: the overall duration required, a percentage of targets to receive the deployment during each interval, an algorithm that defines how percentage grows, and bake time.

get_environment(client, application_id, environment_id, options \\ [])

Retrieve information about an environment.

An environment is a logical deployment group of AppConfig applications, such as applications in a Production environment or in an EU_Region environment. Each configuration deployment targets an environment. You can enable one or more Amazon CloudWatch alarms for an environment. If an alarm is triggered during a deployment, AppConfig roles back the configuration.

get_hosted_configuration_version(client, application_id, configuration_profile_id, version_number, options \\ [])

Get information about a specific configuration version.

list_applications(client, max_results \\ nil, next_token \\ nil, options \\ [])

List all applications in your AWS account.

list_configuration_profiles(client, application_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the configuration profiles for an application.

list_deployment_strategies(client, max_results \\ nil, next_token \\ nil, options \\ [])

List deployment strategies.

list_deployments(client, application_id, environment_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the deployments for an environment.

list_environments(client, application_id, max_results \\ nil, next_token \\ nil, options \\ [])

List the environments for an application.

list_hosted_configuration_versions(client, application_id, configuration_profile_id, max_results \\ nil, next_token \\ nil, options \\ [])

View a list of configurations stored in the AppConfig configuration store by version.

list_tags_for_resource(client, resource_arn, options \\ [])

Retrieves the list of key-value tags assigned to the resource.

start_deployment(client, application_id, environment_id, input, options \\ [])

Starts a deployment.

stop_deployment(client, application_id, deployment_number, environment_id, input, options \\ [])

Stops a deployment.

This API action works only on deployments that have a status of DEPLOYING. This action moves the deployment to a status of ROLLED_BACK.

tag_resource(client, resource_arn, input, options \\ [])

Metadata to assign to an AppConfig resource.

Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define. You can specify a maximum of 50 tags for a resource.

untag_resource(client, resource_arn, input, options \\ [])

Deletes a tag key and value from an AppConfig resource.

update_application(client, application_id, input, options \\ [])

Updates an application.

update_configuration_profile(client, application_id, configuration_profile_id, input, options \\ [])

Updates a configuration profile.

update_deployment_strategy(client, deployment_strategy_id, input, options \\ [])

Updates a deployment strategy.

update_environment(client, application_id, environment_id, input, options \\ [])

Updates an environment.

validate_configuration(client, application_id, configuration_profile_id, input, options \\ [])

Uses the validators in a configuration profile to validate a configuration.