aws-elixir v0.6.0 AWS.AppSync

AWS AppSync provides API actions for creating and interacting with data sources using GraphQL from your application.

Functions

create_api_cache(client, api_id, input, options \\ [])

Creates a cache for the GraphQL API.

create_api_key(client, api_id, input, options \\ [])

Creates a unique key that you can distribute to clients who are executing your API.

create_data_source(client, api_id, input, options \\ [])

Creates a DataSource object.

create_function(client, api_id, input, options \\ [])

Creates a Function object.

A function is a reusable entity. Multiple functions can be used to compose the resolver logic.

create_graphql_api(client, input, options \\ [])

Creates a GraphqlApi object.

create_resolver(client, api_id, type_name, input, options \\ [])

Creates a Resolver object.

A resolver converts incoming requests into a format that a data source can understand and converts the data source's responses into GraphQL.

create_type(client, api_id, input, options \\ [])

Creates a Type object.

delete_api_cache(client, api_id, input, options \\ [])

Deletes an ApiCache object.

delete_api_key(client, api_id, id, input, options \\ [])

Deletes an API key.

delete_data_source(client, api_id, name, input, options \\ [])

Deletes a DataSource object.

delete_function(client, api_id, function_id, input, options \\ [])

Deletes a Function.

delete_graphql_api(client, api_id, input, options \\ [])

Deletes a GraphqlApi object.

delete_resolver(client, api_id, field_name, type_name, input, options \\ [])

Deletes a Resolver object.

delete_type(client, api_id, type_name, input, options \\ [])

Deletes a Type object.

flush_api_cache(client, api_id, input, options \\ [])

Flushes an ApiCache object.

get_api_cache(client, api_id, options \\ [])

Retrieves an ApiCache object.

get_data_source(client, api_id, name, options \\ [])

Retrieves a DataSource object.

get_function(client, api_id, function_id, options \\ [])

Get a Function.

get_graphql_api(client, api_id, options \\ [])

Retrieves a GraphqlApi object.

get_introspection_schema(client, api_id, format, include_directives \\ nil, options \\ [])

Retrieves the introspection schema for a GraphQL API.

get_resolver(client, api_id, field_name, type_name, options \\ [])

Retrieves a Resolver object.

get_schema_creation_status(client, api_id, options \\ [])

Retrieves the current status of a schema creation operation.

get_type(client, api_id, type_name, format, options \\ [])

Retrieves a Type object.

list_api_keys(client, api_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the API keys for a given API.

API keys are deleted automatically 60 days after they expire. However, they may still be included in the response until they have actually been deleted. You can safely call DeleteApiKey to manually delete a key before it's automatically deleted.

list_data_sources(client, api_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the data sources for a given API.

list_functions(client, api_id, max_results \\ nil, next_token \\ nil, options \\ [])

List multiple functions.

list_graphql_apis(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists your GraphQL APIs.

list_resolvers(client, api_id, type_name, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the resolvers for a given API and type.

list_resolvers_by_function(client, api_id, function_id, max_results \\ nil, next_token \\ nil, options \\ [])

List the resolvers that are associated with a specific function.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists the tags for a resource.

list_types(client, api_id, format, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the types for a given API.

start_schema_creation(client, api_id, input, options \\ [])

Adds a new schema to your GraphQL API.

This operation is asynchronous. Use to determine when it has completed.

tag_resource(client, resource_arn, input, options \\ [])

Tags a resource with user-supplied tags.

untag_resource(client, resource_arn, input, options \\ [])

Untags a resource.

update_api_cache(client, api_id, input, options \\ [])

Updates the cache for the GraphQL API.

update_api_key(client, api_id, id, input, options \\ [])

Updates an API key.

The key can be updated while it is not deleted.

update_data_source(client, api_id, name, input, options \\ [])

Updates a DataSource object.

update_function(client, api_id, function_id, input, options \\ [])

Updates a Function object.

update_graphql_api(client, api_id, input, options \\ [])

Updates a GraphqlApi object.

update_resolver(client, api_id, field_name, type_name, input, options \\ [])

Updates a Resolver object.

update_type(client, api_id, type_name, input, options \\ [])

Updates a Type object.