aws-elixir v0.6.0 AWS.Appflow

Welcome to the Amazon AppFlow API reference.

This guide is for developers who need detailed information about the Amazon AppFlow API operations, data types, and errors.

Amazon AppFlow is a fully managed integration service that enables you to securely transfer data between software as a service (SaaS) applications like Salesforce, Marketo, Slack, and ServiceNow, and AWS services like Amazon S3 and Amazon Redshift.

Use the following links to get started on the Amazon AppFlow API:

* Actions: An alphabetical list of all Amazon AppFlow API operations.

Data types: An alphabetical list of all Amazon AppFlow data types.

Common parameters: Parameters that all Query operations can use.

Common errors: Client and server errors that all operations can return.

If you're new to Amazon AppFlow, we recommend that you review the Amazon AppFlow User Guide.

Amazon AppFlow API users can use vendor-specific mechanisms for OAuth, and include applicable OAuth attributes (such as auth-code and redirecturi) with the connector-specific ConnectorProfileProperties when creating a new connector profile using Amazon AppFlow API operations. For example, Salesforce users can refer to the Authorize Apps with OAuth documentation.

Functions

create_connector_profile(client, input, options \\ [])

Creates a new connector profile associated with your AWS account.

There is a soft quota of 100 connector profiles per AWS account. If you need more connector profiles than this quota allows, you can submit a request to the Amazon AppFlow team through the Amazon AppFlow support channel.

create_flow(client, input, options \\ [])

Enables your application to create a new flow using Amazon AppFlow.

You must create a connector profile before calling this API. Please note that the Request Syntax below shows syntax for multiple destinations, however, you can only transfer data to one item in this list at a time. Amazon AppFlow does not currently support flows to multiple destinations at once.

delete_connector_profile(client, input, options \\ [])

Enables you to delete an existing connector profile.

delete_flow(client, input, options \\ [])

Enables your application to delete an existing flow.

Before deleting the flow, Amazon AppFlow validates the request by checking the flow configuration and status. You can delete flows one at a time.

describe_connector_entity(client, input, options \\ [])

Provides details regarding the entity used with the connector, with a description of the data model for each entity.

describe_connector_profiles(client, input, options \\ [])

Returns a list of connector-profile details matching the provided connector-profile names and connector-types.

Both input lists are optional, and you can use them to filter the result.

If no names or connector-types are provided, returns all connector profiles in a paginated form. If there is no match, this operation returns an empty list.

describe_connectors(client, input, options \\ [])

Describes the connectors vended by Amazon AppFlow for specified connector types.

If you don't specify a connector type, this operation describes all connectors vended by Amazon AppFlow. If there are more connectors than can be returned in one page, the response contains a nextToken object, which can be be passed in to the next call to the DescribeConnectors API operation to retrieve the next page.

describe_flow(client, input, options \\ [])

Provides a description of the specified flow.

describe_flow_execution_records(client, input, options \\ [])

Fetches the execution history of the flow.

list_connector_entities(client, input, options \\ [])

Returns the list of available connector entities supported by Amazon AppFlow.

For example, you can query Salesforce for Account and Opportunity entities, or query ServiceNow for the Incident entity.

list_flows(client, input, options \\ [])

Lists all of the flows associated with your account.

list_tags_for_resource(client, resource_arn, options \\ [])

Retrieves the tags that are associated with a specified flow.

start_flow(client, input, options \\ [])

Activates an existing flow.

For on-demand flows, this operation runs the flow immediately. For schedule and event-triggered flows, this operation activates the flow.

stop_flow(client, input, options \\ [])

Deactivates the existing flow.

For on-demand flows, this operation returns an unsupportedOperationException error message. For schedule and event-triggered flows, this operation deactivates the flow.

tag_resource(client, resource_arn, input, options \\ [])

Applies a tag to the specified flow.

untag_resource(client, resource_arn, input, options \\ [])

Removes a tag from the specified flow.

update_connector_profile(client, input, options \\ [])

Updates a given connector profile associated with your account.

update_flow(client, input, options \\ [])

Updates an existing flow.