aws-elixir v0.6.0 AWS.ApplicationInsights

Amazon CloudWatch Application Insights for .NET and SQL Server

Amazon CloudWatch Application Insights for .NET and SQL Server is a service that helps you detect common problems with your .NET and SQL Server-based applications.

It enables you to pinpoint the source of issues in your applications (built with technologies such as Microsoft IIS, .NET, and Microsoft SQL Server), by providing key insights into detected problems.

After you onboard your application, CloudWatch Application Insights for .NET and SQL Server identifies, recommends, and sets up metrics and logs. It continuously analyzes and correlates your metrics and logs for unusual behavior to surface actionable problems with your application. For example, if your application is slow and unresponsive and leading to HTTP 500 errors in your Application Load Balancer (ALB), Application Insights informs you that a memory pressure problem with your SQL Server database is occurring. It bases this analysis on impactful metrics and log errors.

Functions

create_application(client, input, options \\ [])

Adds an application that is created from a resource group.

create_component(client, input, options \\ [])

Creates a custom component by grouping similar standalone instances to monitor.

create_log_pattern(client, input, options \\ [])

Adds an log pattern to a LogPatternSet.

delete_application(client, input, options \\ [])

Removes the specified application from monitoring.

Does not delete the application.

delete_component(client, input, options \\ [])

Ungroups a custom component.

When you ungroup custom components, all applicable monitors that are set up for the component are removed and the instances revert to their standalone status.

delete_log_pattern(client, input, options \\ [])

Removes the specified log pattern from a LogPatternSet.

describe_application(client, input, options \\ [])

Describes the application.

describe_component(client, input, options \\ [])

Describes a component and lists the resources that are grouped together in a component.

describe_component_configuration(client, input, options \\ [])

Describes the monitoring configuration of the component.

describe_component_configuration_recommendation(client, input, options \\ [])

Describes the recommended monitoring configuration of the component.

describe_log_pattern(client, input, options \\ [])

Describe a specific log pattern from a LogPatternSet.

describe_observation(client, input, options \\ [])

Describes an anomaly or error with the application.

describe_problem(client, input, options \\ [])

Describes an application problem.

describe_problem_observations(client, input, options \\ [])

Describes the anomalies or errors associated with the problem.

list_applications(client, input, options \\ [])

Lists the IDs of the applications that you are monitoring.

list_components(client, input, options \\ [])

Lists the auto-grouped, standalone, and custom components of the application.

list_configuration_history(client, input, options \\ [])

Lists the INFO, WARN, and ERROR events for periodic configuration updates performed by Application Insights.

Examples of events represented are:

INFO: creating a new alarm or updating an alarm threshold.

WARN: alarm not created due to insufficient data points used to predict thresholds.

ERROR: alarm not created due to permission errors or exceeding quotas.

list_log_pattern_sets(client, input, options \\ [])

Lists the log pattern sets in the specific application.

list_log_patterns(client, input, options \\ [])

Lists the log patterns in the specific log LogPatternSet.

list_problems(client, input, options \\ [])

Lists the problems with your application.

list_tags_for_resource(client, input, options \\ [])

Retrieve a list of the tags (keys and values) that are associated with a specified application.

A tag is a label that you optionally define and associate with an application. Each tag consists of a required tag key and an optional associated tag value. A tag key is a general label that acts as a category for more specific tag values. A tag value acts as a descriptor within a tag key.

tag_resource(client, input, options \\ [])

Add one or more tags (keys and values) to a specified application.

A tag is a label that you optionally define and associate with an application. Tags can help you categorize and manage application in different ways, such as by purpose, owner, environment, or other criteria.

Each tag consists of a required tag key and an associated tag value, both of which you define. A tag key is a general label that acts as a category for more specific tag values. A tag value acts as a descriptor within a tag key.

untag_resource(client, input, options \\ [])

Remove one or more tags (keys and values) from a specified application.

update_application(client, input, options \\ [])

Updates the application.

update_component(client, input, options \\ [])

Updates the custom component name and/or the list of resources that make up the component.

update_component_configuration(client, input, options \\ [])

Updates the monitoring configurations for the component.

The configuration input parameter is an escaped JSON of the configuration and should match the schema of what is returned by DescribeComponentConfigurationRecommendation.

update_log_pattern(client, input, options \\ [])

Adds a log pattern to a LogPatternSet.