aws-elixir v0.6.0 AWS.Athena

Amazon Athena is an interactive query service that lets you use standard SQL to analyze data directly in Amazon S3.

You can point Athena at your data in Amazon S3 and run ad-hoc queries and get results in seconds. Athena is serverless, so there is no infrastructure to set up or manage. You pay only for the queries you run. Athena scales automatically—executing queries in parallel—so results are fast, even with large datasets and complex queries. For more information, see What is Amazon Athena in the Amazon Athena User Guide.

If you connect to Athena using the JDBC driver, use version 1.1.0 of the driver or later with the Amazon Athena API. Earlier version drivers do not support the API. For more information and to download the driver, see Accessing Amazon Athena with JDBC.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

Functions

batch_get_named_query(client, input, options \\ [])

Returns the details of a single named query or a list of up to 50 queries, which you provide as an array of query ID strings.

Requires you to have access to the workgroup in which the queries were saved. Use ListNamedQueriesInput to get the list of named query IDs in the specified workgroup. If information could not be retrieved for a submitted query ID, information about the query ID submitted is listed under UnprocessedNamedQueryId. Named queries differ from executed queries. Use BatchGetQueryExecutionInput to get details about each unique query execution, and ListQueryExecutionsInput to get a list of query execution IDs.

batch_get_query_execution(client, input, options \\ [])

Returns the details of a single query execution or a list of up to 50 query executions, which you provide as an array of query execution ID strings.

Requires you to have access to the workgroup in which the queries ran. To get a list of query execution IDs, use ListQueryExecutionsInput$WorkGroup. Query executions differ from named (saved) queries. Use BatchGetNamedQueryInput to get details about named queries.

create_data_catalog(client, input, options \\ [])

Creates (registers) a data catalog with the specified name and properties.

Catalogs created are visible to all users of the same AWS account.

create_named_query(client, input, options \\ [])

Creates a named query in the specified workgroup.

Requires that you have access to the workgroup.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

create_work_group(client, input, options \\ [])

Creates a workgroup with the specified name.

delete_data_catalog(client, input, options \\ [])

Deletes a data catalog.

delete_named_query(client, input, options \\ [])

Deletes the named query if you have access to the workgroup in which the query was saved.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

delete_work_group(client, input, options \\ [])

Deletes the workgroup with the specified name.

The primary workgroup cannot be deleted.

get_data_catalog(client, input, options \\ [])

Returns the specified data catalog.

get_database(client, input, options \\ [])

Returns a database object for the specfied database and data catalog.

get_named_query(client, input, options \\ [])

Returns information about a single query.

Requires that you have access to the workgroup in which the query was saved.

get_query_execution(client, input, options \\ [])

Returns information about a single execution of a query if you have access to the workgroup in which the query ran.

Each time a query executes, information about the query execution is saved with a unique ID.

get_query_results(client, input, options \\ [])

Streams the results of a single query execution specified by QueryExecutionId from the Athena query results location in Amazon S3.

For more information, see Query Results in the Amazon Athena User Guide. This request does not execute the query but returns results. Use StartQueryExecution to run a query.

To stream query results successfully, the IAM principal with permission to call GetQueryResults also must have permissions to the Amazon S3 GetObject action for the Athena query results location.

IAM principals with permission to the Amazon S3 GetObject action for the query results location are able to retrieve query results from Amazon S3 even if permission to the GetQueryResults action is denied. To restrict user or role access, ensure that Amazon S3 permissions to the Athena query location are denied.

get_table_metadata(client, input, options \\ [])

Returns table metadata for the specified catalog, database, and table.

get_work_group(client, input, options \\ [])

Returns information about the workgroup with the specified name.

list_data_catalogs(client, input, options \\ [])

Lists the data catalogs in the current AWS account.

list_databases(client, input, options \\ [])

Lists the databases in the specified data catalog.

list_named_queries(client, input, options \\ [])

Provides a list of available query IDs only for queries saved in the specified workgroup.

Requires that you have access to the specified workgroup. If a workgroup is not specified, lists the saved queries for the primary workgroup.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

list_query_executions(client, input, options \\ [])

Provides a list of available query execution IDs for the queries in the specified workgroup.

If a workgroup is not specified, returns a list of query execution IDs for the primary workgroup. Requires you to have access to the workgroup in which the queries ran.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

list_table_metadata(client, input, options \\ [])

Lists the metadata for the tables in the specified data catalog database.

list_tags_for_resource(client, input, options \\ [])

Lists the tags associated with an Athena workgroup or data catalog resource.

list_work_groups(client, input, options \\ [])

Lists available workgroups for the account.

start_query_execution(client, input, options \\ [])

Runs the SQL query statements contained in the Query.

Requires you to have access to the workgroup in which the query ran. Running queries against an external catalog requires GetDataCatalog permission to the catalog. For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

stop_query_execution(client, input, options \\ [])

Stops a query execution.

Requires you to have access to the workgroup in which the query ran.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

tag_resource(client, input, options \\ [])

Adds one or more tags to an Athena resource.

A tag is a label that you assign to a resource. In Athena, a resource can be a workgroup or data catalog. Each tag consists of a key and an optional value, both of which you define. For example, you can use tags to categorize Athena workgroups or data catalogs by purpose, owner, or environment. Use a consistent set of tag keys to make it easier to search and filter workgroups or data catalogs in your account. For best practices, see Tagging Best Practices. Tag keys can be from 1 to 128 UTF-8 Unicode characters, and tag values can be from 0 to 256 UTF-8 Unicode characters. Tags can use letters and numbers representable in UTF-8, and the following characters: + - = . _ : / @. Tag keys and values are case-sensitive. Tag keys must be unique per resource. If you specify more than one tag, separate them by commas.

untag_resource(client, input, options \\ [])

Removes one or more tags from a data catalog or workgroup resource.

update_data_catalog(client, input, options \\ [])

Updates the data catalog that has the specified name.

update_work_group(client, input, options \\ [])

Updates the workgroup with the specified name.

The workgroup's name cannot be changed.