aws-elixir v0.6.0 AWS.AutoScalingPlans

AWS Auto Scaling

Use AWS Auto Scaling to quickly discover all the scalable AWS resources for your application and configure dynamic scaling and predictive scaling for your resources using scaling plans.

Use this service in conjunction with the Amazon EC2 Auto Scaling, Application Auto Scaling, Amazon CloudWatch, and AWS CloudFormation services.

Currently, predictive scaling is only available for Amazon EC2 Auto Scaling groups.

For more information about AWS Auto Scaling, including information about granting IAM users required permissions for AWS Auto Scaling actions, see the AWS Auto Scaling User Guide.

Functions

create_scaling_plan(client, input, options \\ [])

Creates a scaling plan.

delete_scaling_plan(client, input, options \\ [])

Deletes the specified scaling plan.

Deleting a scaling plan deletes the underlying ScalingInstruction for all of the scalable resources that are covered by the plan.

If the plan has launched resources or has scaling activities in progress, you must delete those resources separately.

describe_scaling_plan_resources(client, input, options \\ [])

Describes the scalable resources in the specified scaling plan.

describe_scaling_plans(client, input, options \\ [])

Describes one or more of your scaling plans.

get_scaling_plan_resource_forecast_data(client, input, options \\ [])

Retrieves the forecast data for a scalable resource.

Capacity forecasts are represented as predicted values, or data points, that are calculated using historical data points from a specified CloudWatch load metric. Data points are available for up to 56 days.

update_scaling_plan(client, input, options \\ [])

Updates the specified scaling plan.

You cannot update a scaling plan if it is in the process of being created, updated, or deleted.