aws-elixir v0.6.0 AWS.Braket

The Amazon Braket API Reference provides information about the operations and structures supported in Amazon Braket.

Functions

cancel_quantum_task(client, quantum_task_arn, input, options \\ [])

Cancels the specified task.

create_quantum_task(client, input, options \\ [])

Creates a quantum task.

get_device(client, device_arn, options \\ [])

Retrieves the devices available in Amazon Braket.

get_quantum_task(client, quantum_task_arn, options \\ [])

Retrieves the specified quantum task.

search_devices(client, input, options \\ [])

Searches for devices using the specified filters.

search_quantum_tasks(client, input, options \\ [])

Searches for tasks that match the specified filter values.