aws-elixir v0.6.0 AWS.Budgets

The AWS Budgets API enables you to use AWS Budgets to plan your service usage, service costs, and instance reservations.

The API reference provides descriptions, syntax, and usage examples for each of the actions and data types for AWS Budgets.

Budgets provide you with a way to see the following information:

How close your plan is to your budgeted amount or to the free tier limits

Your usage-to-date, including how much you've used of your Reserved Instances (RIs)

Your current estimated charges from AWS, and how much your predicted usage will accrue in charges by the end of the month

How much of your budget has been used

AWS updates your budget status several times a day. Budgets track your unblended costs, subscriptions, refunds, and RIs. You can create the following types of budgets:

Cost budgets - Plan how much you want to spend on a service.

Usage budgets - Plan how much you want to use one or more services.

RI utilization budgets - Define a utilization threshold, and receive alerts when your RI usage falls below that threshold. This lets you see if your RIs are unused or under-utilized.

RI coverage budgets - Define a coverage threshold, and receive alerts when the number of your instance hours that are covered by RIs fall below that threshold. This lets you see how much of your instance usage is covered by a reservation.

Service Endpoint

The AWS Budgets API provides the following endpoint:

https://budgets.amazonaws.com

For information about costs that are associated with the AWS Budgets API, see AWS Cost Management Pricing.

Functions

create_budget(client, input, options \\ [])

Creates a budget and, if included, notifications and subscribers.

Only one of BudgetLimit or PlannedBudgetLimits can be present in the syntax at one time. Use the syntax that matches your case. The Request Syntax section shows the BudgetLimit syntax. For PlannedBudgetLimits, see the Examples section.

create_notification(client, input, options \\ [])

Creates a notification.

You must create the budget before you create the associated notification.

create_subscriber(client, input, options \\ [])

Creates a subscriber.

You must create the associated budget and notification before you create the subscriber.

delete_budget(client, input, options \\ [])

Deletes a budget.

You can delete your budget at any time.

Deleting a budget also deletes the notifications and subscribers that are associated with that budget.

delete_notification(client, input, options \\ [])

Deletes a notification.

Deleting a notification also deletes the subscribers that are associated with the notification.

delete_subscriber(client, input, options \\ [])

Deletes a subscriber.

Deleting the last subscriber to a notification also deletes the notification.

describe_budget(client, input, options \\ [])

Describes a budget.

The Request Syntax section shows the BudgetLimit syntax. For PlannedBudgetLimits, see the Examples section.

describe_budget_performance_history(client, input, options \\ [])

Describes the history for DAILY, MONTHLY, and QUARTERLY budgets.

Budget history isn't available for ANNUAL budgets.

describe_budgets(client, input, options \\ [])

Lists the budgets that are associated with an account.

The Request Syntax section shows the BudgetLimit syntax. For PlannedBudgetLimits, see the Examples section.

describe_notifications_for_budget(client, input, options \\ [])

Lists the notifications that are associated with a budget.

describe_subscribers_for_notification(client, input, options \\ [])

Lists the subscribers that are associated with a notification.

update_budget(client, input, options \\ [])

Updates a budget.

You can change every part of a budget except for the budgetName and the calculatedSpend. When you modify a budget, the calculatedSpend drops to zero until AWS has new usage data to use for forecasting.

Only one of BudgetLimit or PlannedBudgetLimits can be present in the syntax at one time. Use the syntax that matches your case. The Request Syntax section shows the BudgetLimit syntax. For PlannedBudgetLimits, see the Examples section.

update_notification(client, input, options \\ [])

Updates a notification.

update_subscriber(client, input, options \\ [])

Updates a subscriber.