aws-elixir v0.6.0 AWS.Chime

The Amazon Chime API (application programming interface) is designed for developers to perform key tasks, such as creating and managing Amazon Chime accounts, users, and Voice Connectors.

This guide provides detailed information about the Amazon Chime API, including operations, types, inputs and outputs, and error codes. It also includes some server-side API actions to use with the Amazon Chime SDK. For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

You can use an AWS SDK, the AWS Command Line Interface (AWS CLI), or the REST API to make API calls. We recommend using an AWS SDK or the AWS CLI. Each API operation includes links to information about using it with a language-specific AWS SDK or the AWS CLI.

Definitions

Using an AWS SDK

You don't need to write code to calculate a signature for request authentication. The SDK clients authenticate your requests by using access keys that you provide. For more information about AWS SDKs, see the AWS Developer Center.

Using the AWS CLI

Use your access keys with the AWS CLI to make API calls. For information about setting up the AWS CLI, see Installing the AWS Command Line Interface in the AWS Command Line Interface User Guide. For a list of available Amazon Chime commands, see the Amazon Chime commands in the AWS CLI Command Reference.

Using REST API

If you use REST to make API calls, you must authenticate your request by providing a signature. Amazon Chime supports signature version 4. For more information, see Signature Version 4 Signing Process in the Amazon Web Services General Reference.

When making REST API calls, use the service name chime and REST endpoint https://service.chime.aws.amazon.com.

Administrative permissions are controlled using AWS Identity and Access Management (IAM). For more information, see Identity and Access Management for Amazon Chime in the Amazon Chime Administration Guide.

Functions

associate_phone_number_with_user(client, account_id, user_id, input, options \\ [])

Associates a phone number with the specified Amazon Chime user.

associate_phone_numbers_with_voice_connector(client, voice_connector_id, input, options \\ [])

Associates phone numbers with the specified Amazon Chime Voice Connector.

associate_phone_numbers_with_voice_connector_group(client, voice_connector_group_id, input, options \\ [])

Associates phone numbers with the specified Amazon Chime Voice Connector group.

associate_signin_delegate_groups_with_account(client, account_id, input, options \\ [])

Associates the specified sign-in delegate groups with the specified Amazon Chime account.

batch_create_attendee(client, meeting_id, input, options \\ [])

Creates up to 100 new attendees for an active Amazon Chime SDK meeting.

For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

batch_create_room_membership(client, account_id, room_id, input, options \\ [])

Adds up to 50 members to a chat room in an Amazon Chime Enterprise account.

Members can be either users or bots. The member role designates whether the member is a chat room administrator or a general chat room member.

batch_delete_phone_number(client, input, options \\ [])

Moves phone numbers into the Deletion queue.

Phone numbers must be disassociated from any users or Amazon Chime Voice Connectors before they can be deleted.

Phone numbers remain in the Deletion queue for 7 days before they are deleted permanently.

batch_suspend_user(client, account_id, input, options \\ [])

Suspends up to 50 users from a Team or EnterpriseLWA Amazon Chime account.

For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

Users suspended from a Team account are disassociated from the account, but they can continue to use Amazon Chime as free users. To remove the suspension from suspended Team account users, invite them to the Team account again. You can use the InviteUsers action to do so.

Users suspended from an EnterpriseLWA account are immediately signed out of Amazon Chime and can no longer sign in. To remove the suspension from suspended EnterpriseLWA account users, use the BatchUnsuspendUser action.

To sign out users without suspending them, use the LogoutUser action.

batch_unsuspend_user(client, account_id, input, options \\ [])

Removes the suspension from up to 50 previously suspended users for the specified Amazon Chime EnterpriseLWA account.

Only users on EnterpriseLWA accounts can be unsuspended using this action. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

Previously suspended users who are unsuspended using this action are returned to Registered status. Users who are not previously suspended are ignored.

batch_update_phone_number(client, input, options \\ [])

Updates phone number product types or calling names.

You can update one attribute at a time for each UpdatePhoneNumberRequestItem. For example, you can update either the product type or the calling name.

For product types, choose from Amazon Chime Business Calling and Amazon Chime Voice Connector. For toll-free numbers, you must use the Amazon Chime Voice Connector product type.

Updates to outbound calling names can take up to 72 hours to complete. Pending updates to outbound calling names must be complete before you can request another update.

batch_update_user(client, account_id, input, options \\ [])

Updates user details within the UpdateUserRequestItem object for up to 20 users for the specified Amazon Chime account.

Currently, only LicenseType updates are supported for this action.

create_account(client, input, options \\ [])

Creates an Amazon Chime account under the administrator's AWS account.

Only Team account types are currently supported for this action. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

create_attendee(client, meeting_id, input, options \\ [])

Creates a new attendee for an active Amazon Chime SDK meeting.

For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

create_bot(client, account_id, input, options \\ [])

Creates a bot for an Amazon Chime Enterprise account.

create_meeting(client, input, options \\ [])

Creates a new Amazon Chime SDK meeting in the specified media Region with no initial attendees.

For more information about specifying media Regions, see Amazon Chime SDK Media Regions in the Amazon Chime Developer Guide. For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

create_meeting_with_attendees(client, input, options \\ [])

Creates a new Amazon Chime SDK meeting in the specified media Region, with attendees.

For more information about specifying media Regions, see Amazon Chime SDK Media Regions in the Amazon Chime Developer Guide. For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

create_phone_number_order(client, input, options \\ [])

Creates an order for phone numbers to be provisioned.

Choose from Amazon Chime Business Calling and Amazon Chime Voice Connector product types. For toll-free numbers, you must use the Amazon Chime Voice Connector product type.

create_proxy_session(client, voice_connector_id, input, options \\ [])

Creates a proxy session on the specified Amazon Chime Voice Connector for the specified participant phone numbers.

create_room(client, account_id, input, options \\ [])

Creates a chat room for the specified Amazon Chime Enterprise account.

create_room_membership(client, account_id, room_id, input, options \\ [])

Adds a member to a chat room in an Amazon Chime Enterprise account.

A member can be either a user or a bot. The member role designates whether the member is a chat room administrator or a general chat room member.

create_user(client, account_id, input, options \\ [])

Creates a user under the specified Amazon Chime account.

create_voice_connector(client, input, options \\ [])

Creates an Amazon Chime Voice Connector under the administrator's AWS account.

You can choose to create an Amazon Chime Voice Connector in a specific AWS Region.

Enabling CreateVoiceConnectorRequest$RequireEncryption configures your Amazon Chime Voice Connector to use TLS transport for SIP signaling and Secure RTP (SRTP) for media. Inbound calls use TLS transport, and unencrypted outbound calls are blocked.

create_voice_connector_group(client, input, options \\ [])

Creates an Amazon Chime Voice Connector group under the administrator's AWS account.

You can associate Amazon Chime Voice Connectors with the Amazon Chime Voice Connector group by including VoiceConnectorItems in the request.

You can include Amazon Chime Voice Connectors from different AWS Regions in your group. This creates a fault tolerant mechanism for fallback in case of availability events.

delete_account(client, account_id, input, options \\ [])

Deletes the specified Amazon Chime account.

You must suspend all users before deleting a Team account. You can use the BatchSuspendUser action to do so.

For EnterpriseLWA and EnterpriseAD accounts, you must release the claimed domains for your Amazon Chime account before deletion. As soon as you release the domain, all users under that account are suspended.

Deleted accounts appear in your Disabled accounts list for 90 days. To restore a deleted account from your Disabled accounts list, you must contact AWS Support.

After 90 days, deleted accounts are permanently removed from your Disabled accounts list.

delete_attendee(client, attendee_id, meeting_id, input, options \\ [])

Deletes an attendee from the specified Amazon Chime SDK meeting and deletes their JoinToken.

Attendees are automatically deleted when a Amazon Chime SDK meeting is deleted. For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

delete_events_configuration(client, account_id, bot_id, input, options \\ [])

Deletes the events configuration that allows a bot to receive outgoing events.

delete_meeting(client, meeting_id, input, options \\ [])

Deletes the specified Amazon Chime SDK meeting.

When a meeting is deleted, its attendees are also deleted and clients can no longer join it. For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

delete_phone_number(client, phone_number_id, input, options \\ [])

Moves the specified phone number into the Deletion queue.

A phone number must be disassociated from any users or Amazon Chime Voice Connectors before it can be deleted.

Deleted phone numbers remain in the Deletion queue for 7 days before they are deleted permanently.

delete_proxy_session(client, proxy_session_id, voice_connector_id, input, options \\ [])

Deletes the specified proxy session from the specified Amazon Chime Voice Connector.

delete_room(client, account_id, room_id, input, options \\ [])

Deletes a chat room in an Amazon Chime Enterprise account.

delete_room_membership(client, account_id, member_id, room_id, input, options \\ [])

Removes a member from a chat room in an Amazon Chime Enterprise account.

delete_voice_connector(client, voice_connector_id, input, options \\ [])

Deletes the specified Amazon Chime Voice Connector.

Any phone numbers associated with the Amazon Chime Voice Connector must be disassociated from it before it can be deleted.

delete_voice_connector_emergency_calling_configuration(client, voice_connector_id, input, options \\ [])

Deletes the emergency calling configuration details from the specified Amazon Chime Voice Connector.

delete_voice_connector_group(client, voice_connector_group_id, input, options \\ [])

Deletes the specified Amazon Chime Voice Connector group.

Any VoiceConnectorItems and phone numbers associated with the group must be removed before it can be deleted.

delete_voice_connector_origination(client, voice_connector_id, input, options \\ [])

Deletes the origination settings for the specified Amazon Chime Voice Connector.

If emergency calling is configured for the Amazon Chime Voice Connector, it must be deleted prior to deleting the origination settings.

delete_voice_connector_proxy(client, voice_connector_id, input, options \\ [])

Deletes the proxy configuration from the specified Amazon Chime Voice Connector.

delete_voice_connector_streaming_configuration(client, voice_connector_id, input, options \\ [])

Deletes the streaming configuration for the specified Amazon Chime Voice Connector.

delete_voice_connector_termination(client, voice_connector_id, input, options \\ [])

Deletes the termination settings for the specified Amazon Chime Voice Connector.

If emergency calling is configured for the Amazon Chime Voice Connector, it must be deleted prior to deleting the termination settings.

delete_voice_connector_termination_credentials(client, voice_connector_id, input, options \\ [])

Deletes the specified SIP credentials used by your equipment to authenticate during call termination.

disassociate_phone_number_from_user(client, account_id, user_id, input, options \\ [])

Disassociates the primary provisioned phone number from the specified Amazon Chime user.

disassociate_phone_numbers_from_voice_connector(client, voice_connector_id, input, options \\ [])

Disassociates the specified phone numbers from the specified Amazon Chime Voice Connector.

disassociate_phone_numbers_from_voice_connector_group(client, voice_connector_group_id, input, options \\ [])

Disassociates the specified phone numbers from the specified Amazon Chime Voice Connector group.

disassociate_signin_delegate_groups_from_account(client, account_id, input, options \\ [])

Disassociates the specified sign-in delegate groups from the specified Amazon Chime account.

get_account(client, account_id, options \\ [])

Retrieves details for the specified Amazon Chime account, such as account type and supported licenses.

get_account_settings(client, account_id, options \\ [])

Retrieves account settings for the specified Amazon Chime account ID, such as remote control and dial out settings.

For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

get_attendee(client, attendee_id, meeting_id, options \\ [])

Gets the Amazon Chime SDK attendee details for a specified meeting ID and attendee ID.

For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

get_bot(client, account_id, bot_id, options \\ [])

Retrieves details for the specified bot, such as bot email address, bot type, status, and display name.

get_events_configuration(client, account_id, bot_id, options \\ [])

Gets details for an events configuration that allows a bot to receive outgoing events, such as an HTTPS endpoint or Lambda function ARN.

get_global_settings(client, options \\ [])

Retrieves global settings for the administrator's AWS account, such as Amazon Chime Business Calling and Amazon Chime Voice Connector settings.

get_meeting(client, meeting_id, options \\ [])

Gets the Amazon Chime SDK meeting details for the specified meeting ID.

For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

get_phone_number(client, phone_number_id, options \\ [])

Retrieves details for the specified phone number ID, such as associations, capabilities, and product type.

get_phone_number_order(client, phone_number_order_id, options \\ [])

Retrieves details for the specified phone number order, such as order creation timestamp, phone numbers in E.164 format, product type, and order status.

get_phone_number_settings(client, options \\ [])

Retrieves the phone number settings for the administrator's AWS account, such as the default outbound calling name.

get_proxy_session(client, proxy_session_id, voice_connector_id, options \\ [])

Gets the specified proxy session details for the specified Amazon Chime Voice Connector.

get_retention_settings(client, account_id, options \\ [])

Gets the retention settings for the specified Amazon Chime Enterprise account.

For more information about retention settings, see Managing Chat Retention Policies in the Amazon Chime Administration Guide.

get_room(client, account_id, room_id, options \\ [])

Retrieves room details, such as the room name, for a room in an Amazon Chime Enterprise account.

get_user(client, account_id, user_id, options \\ [])

Retrieves details for the specified user ID, such as primary email address, license type, and personal meeting PIN.

To retrieve user details with an email address instead of a user ID, use the ListUsers action, and then filter by email address.

get_user_settings(client, account_id, user_id, options \\ [])

Retrieves settings for the specified user ID, such as any associated phone number settings.

get_voice_connector(client, voice_connector_id, options \\ [])

Retrieves details for the specified Amazon Chime Voice Connector, such as timestamps, name, outbound host, and encryption requirements.

get_voice_connector_emergency_calling_configuration(client, voice_connector_id, options \\ [])

Gets the emergency calling configuration details for the specified Amazon Chime Voice Connector.

get_voice_connector_group(client, voice_connector_group_id, options \\ [])

Retrieves details for the specified Amazon Chime Voice Connector group, such as timestamps, name, and associated VoiceConnectorItems.

get_voice_connector_logging_configuration(client, voice_connector_id, options \\ [])

Retrieves the logging configuration details for the specified Amazon Chime Voice Connector.

Shows whether SIP message logs are enabled for sending to Amazon CloudWatch Logs.

get_voice_connector_origination(client, voice_connector_id, options \\ [])

Retrieves origination setting details for the specified Amazon Chime Voice Connector.

get_voice_connector_proxy(client, voice_connector_id, options \\ [])

Gets the proxy configuration details for the specified Amazon Chime Voice Connector.

get_voice_connector_streaming_configuration(client, voice_connector_id, options \\ [])

Retrieves the streaming configuration details for the specified Amazon Chime Voice Connector.

Shows whether media streaming is enabled for sending to Amazon Kinesis. It also shows the retention period, in hours, for the Amazon Kinesis data.

get_voice_connector_termination(client, voice_connector_id, options \\ [])

Retrieves termination setting details for the specified Amazon Chime Voice Connector.

get_voice_connector_termination_health(client, voice_connector_id, options \\ [])

Retrieves information about the last time a SIP OPTIONS ping was received from your SIP infrastructure for the specified Amazon Chime Voice Connector.

invite_users(client, account_id, input, options \\ [])

Sends email to a maximum of 50 users, inviting them to the specified Amazon Chime Team account.

Only Team account types are currently supported for this action.

list_accounts(client, max_results \\ nil, name \\ nil, next_token \\ nil, user_email \\ nil, options \\ [])

Lists the Amazon Chime accounts under the administrator's AWS account.

You can filter accounts by account name prefix. To find out which Amazon Chime account a user belongs to, you can filter by the user's email address, which returns one account result.

list_attendee_tags(client, attendee_id, meeting_id, options \\ [])

Lists the tags applied to an Amazon Chime SDK attendee resource.

list_attendees(client, meeting_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the attendees for the specified Amazon Chime SDK meeting.

For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

list_bots(client, account_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the bots associated with the administrator's Amazon Chime Enterprise account ID.

list_meeting_tags(client, meeting_id, options \\ [])

Lists the tags applied to an Amazon Chime SDK meeting resource.

list_meetings(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists up to 100 active Amazon Chime SDK meetings.

For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

list_phone_number_orders(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the phone number orders for the administrator's Amazon Chime account.

list_phone_numbers(client, filter_name \\ nil, filter_value \\ nil, max_results \\ nil, next_token \\ nil, product_type \\ nil, status \\ nil, options \\ [])

Lists the phone numbers for the specified Amazon Chime account, Amazon Chime user, Amazon Chime Voice Connector, or Amazon Chime Voice Connector group.

list_proxy_sessions(client, voice_connector_id, max_results \\ nil, next_token \\ nil, status \\ nil, options \\ [])

Lists the proxy sessions for the specified Amazon Chime Voice Connector.

list_room_memberships(client, account_id, room_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the membership details for the specified room in an Amazon Chime Enterprise account, such as the members' IDs, email addresses, and names.

list_rooms(client, account_id, max_results \\ nil, member_id \\ nil, next_token \\ nil, options \\ [])

Lists the room details for the specified Amazon Chime Enterprise account.

Optionally, filter the results by a member ID (user ID or bot ID) to see a list of rooms that the member belongs to.

list_tags_for_resource(client, resource_a_r_n, options \\ [])

Lists the tags applied to an Amazon Chime SDK meeting resource.

list_users(client, account_id, max_results \\ nil, next_token \\ nil, user_email \\ nil, user_type \\ nil, options \\ [])

Lists the users that belong to the specified Amazon Chime account.

You can specify an email address to list only the user that the email address belongs to.

list_voice_connector_groups(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the Amazon Chime Voice Connector groups for the administrator's AWS account.

list_voice_connector_termination_credentials(client, voice_connector_id, options \\ [])

Lists the SIP credentials for the specified Amazon Chime Voice Connector.

list_voice_connectors(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the Amazon Chime Voice Connectors for the administrator's AWS account.

logout_user(client, account_id, user_id, input, options \\ [])

Logs out the specified user from all of the devices they are currently logged into.

put_events_configuration(client, account_id, bot_id, input, options \\ [])

Creates an events configuration that allows a bot to receive outgoing events sent by Amazon Chime.

Choose either an HTTPS endpoint or a Lambda function ARN. For more information, see Bot.

put_retention_settings(client, account_id, input, options \\ [])

Puts retention settings for the specified Amazon Chime Enterprise account.

We recommend using AWS CloudTrail to monitor usage of this API for your account. For more information, see Logging Amazon Chime API Calls with AWS CloudTrail in the Amazon Chime Administration Guide.

To turn off existing retention settings, remove the number of days from the corresponding RetentionDays field in the RetentionSettings object. For more information about retention settings, see Managing Chat Retention Policies in the Amazon Chime Administration Guide.

put_voice_connector_emergency_calling_configuration(client, voice_connector_id, input, options \\ [])

Puts emergency calling configuration details to the specified Amazon Chime Voice Connector, such as emergency phone numbers and calling countries.

Origination and termination settings must be enabled for the Amazon Chime Voice Connector before emergency calling can be configured.

put_voice_connector_logging_configuration(client, voice_connector_id, input, options \\ [])

Adds a logging configuration for the specified Amazon Chime Voice Connector.

The logging configuration specifies whether SIP message logs are enabled for sending to Amazon CloudWatch Logs.

put_voice_connector_origination(client, voice_connector_id, input, options \\ [])

Adds origination settings for the specified Amazon Chime Voice Connector.

If emergency calling is configured for the Amazon Chime Voice Connector, it must be deleted prior to turning off origination settings.

put_voice_connector_proxy(client, voice_connector_id, input, options \\ [])

Puts the specified proxy configuration to the specified Amazon Chime Voice Connector.

put_voice_connector_streaming_configuration(client, voice_connector_id, input, options \\ [])

Adds a streaming configuration for the specified Amazon Chime Voice Connector.

The streaming configuration specifies whether media streaming is enabled for sending to Amazon Kinesis. It also sets the retention period, in hours, for the Amazon Kinesis data.

put_voice_connector_termination(client, voice_connector_id, input, options \\ [])

Adds termination settings for the specified Amazon Chime Voice Connector.

If emergency calling is configured for the Amazon Chime Voice Connector, it must be deleted prior to turning off termination settings.

put_voice_connector_termination_credentials(client, voice_connector_id, input, options \\ [])

Adds termination SIP credentials for the specified Amazon Chime Voice Connector.

redact_conversation_message(client, account_id, conversation_id, message_id, input, options \\ [])

Redacts the specified message from the specified Amazon Chime conversation.

redact_room_message(client, account_id, message_id, room_id, input, options \\ [])

Redacts the specified message from the specified Amazon Chime chat room.

regenerate_security_token(client, account_id, bot_id, input, options \\ [])

Regenerates the security token for a bot.

reset_personal_p_i_n(client, account_id, user_id, input, options \\ [])

Resets the personal meeting PIN for the specified user on an Amazon Chime account.

Returns the User object with the updated personal meeting PIN.

restore_phone_number(client, phone_number_id, input, options \\ [])

Moves a phone number from the Deletion queue back into the phone number Inventory.

search_available_phone_numbers(client, area_code \\ nil, city \\ nil, country \\ nil, max_results \\ nil, next_token \\ nil, state \\ nil, toll_free_prefix \\ nil, options \\ [])

Searches phone numbers that can be ordered.

tag_attendee(client, attendee_id, meeting_id, input, options \\ [])

Applies the specified tags to the specified Amazon Chime SDK attendee.

tag_meeting(client, meeting_id, input, options \\ [])

Applies the specified tags to the specified Amazon Chime SDK meeting.

tag_resource(client, input, options \\ [])

Applies the specified tags to the specified Amazon Chime SDK meeting resource.

untag_attendee(client, attendee_id, meeting_id, input, options \\ [])

Untags the specified tags from the specified Amazon Chime SDK attendee.

untag_meeting(client, meeting_id, input, options \\ [])

Untags the specified tags from the specified Amazon Chime SDK meeting.

untag_resource(client, input, options \\ [])

Untags the specified tags from the specified Amazon Chime SDK meeting resource.

update_account(client, account_id, input, options \\ [])

Updates account details for the specified Amazon Chime account.

Currently, only account name updates are supported for this action.

update_account_settings(client, account_id, input, options \\ [])

Updates the settings for the specified Amazon Chime account.

You can update settings for remote control of shared screens, or for the dial-out option. For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

update_bot(client, account_id, bot_id, input, options \\ [])

Updates the status of the specified bot, such as starting or stopping the bot from running in your Amazon Chime Enterprise account.

update_global_settings(client, input, options \\ [])

Updates global settings for the administrator's AWS account, such as Amazon Chime Business Calling and Amazon Chime Voice Connector settings.

update_phone_number(client, phone_number_id, input, options \\ [])

Updates phone number details, such as product type or calling name, for the specified phone number ID.

You can update one phone number detail at a time. For example, you can update either the product type or the calling name in one action.

For toll-free numbers, you must use the Amazon Chime Voice Connector product type.

Updates to outbound calling names can take up to 72 hours to complete. Pending updates to outbound calling names must be complete before you can request another update.

update_phone_number_settings(client, input, options \\ [])

Updates the phone number settings for the administrator's AWS account, such as the default outbound calling name.

You can update the default outbound calling name once every seven days. Outbound calling names can take up to 72 hours to update.

update_proxy_session(client, proxy_session_id, voice_connector_id, input, options \\ [])

Updates the specified proxy session details, such as voice or SMS capabilities.

update_room(client, account_id, room_id, input, options \\ [])

Updates room details, such as the room name, for a room in an Amazon Chime Enterprise account.

update_room_membership(client, account_id, member_id, room_id, input, options \\ [])

Updates room membership details, such as the member role, for a room in an Amazon Chime Enterprise account.

The member role designates whether the member is a chat room administrator or a general chat room member. The member role can be updated only for user IDs.

update_user(client, account_id, user_id, input, options \\ [])

Updates user details for a specified user ID.

Currently, only LicenseType updates are supported for this action.

update_user_settings(client, account_id, user_id, input, options \\ [])

Updates the settings for the specified user, such as phone number settings.

update_voice_connector(client, voice_connector_id, input, options \\ [])

Updates details for the specified Amazon Chime Voice Connector.

update_voice_connector_group(client, voice_connector_group_id, input, options \\ [])

Updates details for the specified Amazon Chime Voice Connector group, such as the name and Amazon Chime Voice Connector priority ranking.