aws-elixir v0.6.0 AWS.Cloud9

AWS Cloud9

AWS Cloud9 is a collection of tools that you can use to code, build, run, test, debug, and release software in the cloud.

For more information about AWS Cloud9, see the AWS Cloud9 User Guide.

AWS Cloud9 supports these operations:

CreateEnvironmentEC2: Creates an AWS Cloud9 development environment, launches an Amazon EC2 instance, and then connects from the instance to the environment.

CreateEnvironmentMembership: Adds an environment member to an environment.

DeleteEnvironment: Deletes an environment. If an Amazon EC2 instance is connected to the environment, also terminates the instance.

DeleteEnvironmentMembership: Deletes an environment member from an environment.

DescribeEnvironmentMemberships: Gets information about environment members for an environment.

DescribeEnvironments: Gets information about environments.

DescribeEnvironmentStatus: Gets status information for an environment.

ListEnvironments: Gets a list of environment identifiers.

ListTagsForResource: Gets the tags for an environment.

TagResource: Adds tags to an environment.

UntagResource: Removes tags from an environment.

UpdateEnvironment: Changes the settings of an existing environment.

UpdateEnvironmentMembership: Changes the settings of an existing environment member for an environment.

Functions

create_environment_e_c2(client, input, options \\ [])

Creates an AWS Cloud9 development environment, launches an Amazon Elastic Compute Cloud (Amazon EC2) instance, and then connects from the instance to the environment.

create_environment_membership(client, input, options \\ [])

Adds an environment member to an AWS Cloud9 development environment.

delete_environment(client, input, options \\ [])

Deletes an AWS Cloud9 development environment.

If an Amazon EC2 instance is connected to the environment, also terminates the instance.

delete_environment_membership(client, input, options \\ [])

Deletes an environment member from an AWS Cloud9 development environment.

describe_environment_memberships(client, input, options \\ [])

Gets information about environment members for an AWS Cloud9 development environment.

describe_environment_status(client, input, options \\ [])

Gets status information for an AWS Cloud9 development environment.

describe_environments(client, input, options \\ [])

Gets information about AWS Cloud9 development environments.

list_environments(client, input, options \\ [])

Gets a list of AWS Cloud9 development environment identifiers.

list_tags_for_resource(client, input, options \\ [])

Gets a list of the tags associated with an AWS Cloud9 development environment.

tag_resource(client, input, options \\ [])

Adds tags to an AWS Cloud9 development environment.

Tags that you add to an AWS Cloud9 environment by using this method will NOT be automatically propagated to underlying resources.

untag_resource(client, input, options \\ [])

Removes tags from an AWS Cloud9 development environment.

update_environment(client, input, options \\ [])

Changes the settings of an existing AWS Cloud9 development environment.

update_environment_membership(client, input, options \\ [])

Changes the settings of an existing environment member for an AWS Cloud9 development environment.