aws-elixir v0.6.0 AWS.CloudDirectory

Amazon Cloud Directory

Amazon Cloud Directory is a component of the AWS Directory Service that simplifies the development and management of cloud-scale web, mobile, and IoT applications.

This guide describes the Cloud Directory operations that you can call programmatically and includes detailed information on data types and errors. For information about Cloud Directory features, see AWS Directory Service and the Amazon Cloud Directory Developer Guide.

Functions

add_facet_to_object(client, input, options \\ [])

Adds a new Facet to an object.

An object can have more than one facet applied on it.

apply_schema(client, input, options \\ [])

Copies the input published schema, at the specified version, into the Directory with the same name and version as that of the published schema.

attach_object(client, input, options \\ [])

Attaches an existing object to another object.

An object can be accessed in two ways:

Using the path

Using ObjectIdentifier

attach_policy(client, input, options \\ [])

Attaches a policy object to a regular object.

An object can have a limited number of attached policies.

attach_to_index(client, input, options \\ [])

Attaches the specified object to the specified index.

attach_typed_link(client, input, options \\ [])

Attaches a typed link to a specified source and target object.

For more information, see Typed Links.

batch_read(client, input, options \\ [])

Performs all the read operations in a batch.

batch_write(client, input, options \\ [])

Performs all the write operations in a batch.

Either all the operations succeed or none.

create_directory(client, input, options \\ [])

Creates a Directory by copying the published schema into the directory.

A directory cannot be created without a schema.

You can also quickly create a directory using a managed schema, called the QuickStartSchema. For more information, see Managed Schema in the Amazon Cloud Directory Developer Guide.

create_facet(client, input, options \\ [])

Creates a new Facet in a schema.

Facet creation is allowed only in development or applied schemas.

create_index(client, input, options \\ [])

Creates an index object.

See Indexing and search for more information.

create_object(client, input, options \\ [])

Creates an object in a Directory.

Additionally attaches the object to a parent, if a parent reference and LinkName is specified. An object is simply a collection of Facet attributes. You can also use this API call to create a policy object, if the facet from which you create the object is a policy facet.

create_schema(client, input, options \\ [])

Creates a new schema in a development state.

A schema can exist in three phases:

Development: This is a mutable phase of the schema. All new schemas are in the development phase. Once the schema is finalized, it can be published.

Published: Published schemas are immutable and have a version associated with them.

Applied: Applied schemas are mutable in a way that allows you to add new schema facets. You can also add new, nonrequired attributes to existing schema facets. You can apply only published schemas to directories.

create_typed_link_facet(client, input, options \\ [])

Creates a TypedLinkFacet.

For more information, see Typed Links.

delete_directory(client, input, options \\ [])

Deletes a directory.

Only disabled directories can be deleted. A deleted directory cannot be undone. Exercise extreme caution when deleting directories.

delete_facet(client, input, options \\ [])

Deletes a given Facet.

All attributes and Rules that are associated with the facet will be deleted. Only development schema facets are allowed deletion.

delete_object(client, input, options \\ [])

Deletes an object and its associated attributes.

Only objects with no children and no parents can be deleted. The maximum number of attributes that can be deleted during an object deletion is 30. For more information, see Amazon Cloud Directory Limits.

delete_schema(client, input, options \\ [])

Deletes a given schema.

Schemas in a development and published state can only be deleted.

delete_typed_link_facet(client, input, options \\ [])

Deletes a TypedLinkFacet.

For more information, see Typed Links.

detach_from_index(client, input, options \\ [])

Detaches the specified object from the specified index.

detach_object(client, input, options \\ [])

Detaches a given object from the parent object.

The object that is to be detached from the parent is specified by the link name.

detach_policy(client, input, options \\ [])

Detaches a policy from an object.

detach_typed_link(client, input, options \\ [])

Detaches a typed link from a specified source and target object.

For more information, see Typed Links.

disable_directory(client, input, options \\ [])

Disables the specified directory.

Disabled directories cannot be read or written to. Only enabled directories can be disabled. Disabled directories may be reenabled.

enable_directory(client, input, options \\ [])

Enables the specified directory.

Only disabled directories can be enabled. Once enabled, the directory can then be read and written to.

get_applied_schema_version(client, input, options \\ [])

Returns current applied schema version ARN, including the minor version in use.

get_directory(client, input, options \\ [])

Retrieves metadata about a directory.

get_facet(client, input, options \\ [])

Gets details of the Facet, such as facet name, attributes, Rules, or ObjectType.

You can call this on all kinds of schema facets -- published, development, or applied.

get_link_attributes(client, input, options \\ [])

Retrieves attributes that are associated with a typed link.

get_object_attributes(client, input, options \\ [])

Retrieves attributes within a facet that are associated with an object.

get_object_information(client, input, options \\ [])

Retrieves metadata about an object.

get_schema_as_json(client, input, options \\ [])

Retrieves a JSON representation of the schema.

See JSON Schema Format for more information.

get_typed_link_facet_information(client, input, options \\ [])

Returns the identity attribute order for a specific TypedLinkFacet.

For more information, see Typed Links.

list_applied_schema_arns(client, input, options \\ [])

Lists schema major versions applied to a directory.

If SchemaArn is provided, lists the minor version.

list_attached_indices(client, input, options \\ [])

Lists indices attached to the specified object.

list_development_schema_arns(client, input, options \\ [])

Retrieves each Amazon Resource Name (ARN) of schemas in the development state.

list_directories(client, input, options \\ [])

Lists directories created within an account.

list_facet_attributes(client, input, options \\ [])

Retrieves attributes attached to the facet.

list_facet_names(client, input, options \\ [])

Retrieves the names of facets that exist in a schema.

list_incoming_typed_links(client, input, options \\ [])

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object.

It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

list_index(client, input, options \\ [])

Lists objects attached to the specified index.

list_managed_schema_arns(client, input, options \\ [])

Lists the major version families of each managed schema.

If a major version ARN is provided as SchemaArn, the minor version revisions in that family are listed instead.

list_object_attributes(client, input, options \\ [])

Lists all attributes that are associated with an object.

list_object_children(client, input, options \\ [])

Returns a paginated list of child objects that are associated with a given object.

list_object_parent_paths(client, input, options \\ [])

Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects.

For more information about objects, see Directory Structure.

Use this API to evaluate all parents for an object. The call returns all objects from the root of the directory up to the requested object. The API returns the number of paths based on user-defined MaxResults, in case there are multiple paths to the parent. The order of the paths and nodes returned is consistent among multiple API calls unless the objects are deleted or moved. Paths not leading to the directory root are ignored from the target object.

list_object_parents(client, input, options \\ [])

Lists parent objects that are associated with a given object in pagination fashion.

list_object_policies(client, input, options \\ [])

Returns policies attached to an object in pagination fashion.

list_outgoing_typed_links(client, input, options \\ [])

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object.

It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

list_policy_attachments(client, input, options \\ [])

Returns all of the ObjectIdentifiers to which a given policy is attached.

list_published_schema_arns(client, input, options \\ [])

Lists the major version families of each published schema.

If a major version ARN is provided as SchemaArn, the minor version revisions in that family are listed instead.

list_tags_for_resource(client, input, options \\ [])

Returns tags for a resource.

Tagging is currently supported only for directories with a limit of 50 tags per directory. All 50 tags are returned for a given directory with this API call.

list_typed_link_facet_attributes(client, input, options \\ [])

Returns a paginated list of all attribute definitions for a particular TypedLinkFacet.

For more information, see Typed Links.

list_typed_link_facet_names(client, input, options \\ [])

Returns a paginated list of TypedLink facet names for a particular schema.

For more information, see Typed Links.

lookup_policy(client, input, options \\ [])

Lists all policies from the root of the Directory to the object specified.

If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier, policyId, and policyType. Paths that don't lead to the root from the target object are ignored. For more information, see Policies.

publish_schema(client, input, options \\ [])

Publishes a development schema with a major version and a recommended minor version.

put_schema_from_json(client, input, options \\ [])

Allows a schema to be updated using JSON upload.

Only available for development schemas. See JSON Schema Format for more information.

remove_facet_from_object(client, input, options \\ [])

Removes the specified facet from the specified object.

tag_resource(client, input, options \\ [])

An API operation for adding tags to a resource.

untag_resource(client, input, options \\ [])

An API operation for removing tags from a resource.

update_facet(client, input, options \\ [])

Does the following:

1.

Adds new Attributes, Rules, or ObjectTypes.

Updates existing Attributes, Rules, or ObjectTypes.

Deletes existing Attributes, Rules, or ObjectTypes.

update_link_attributes(client, input, options \\ [])

Updates a given typed link’s attributes.

Attributes to be updated must not contribute to the typed link’s identity, as defined by its IdentityAttributeOrder.

update_object_attributes(client, input, options \\ [])

Updates a given object's attributes.

update_schema(client, input, options \\ [])

Updates the schema name with a new name.

Only development schema names can be updated.

update_typed_link_facet(client, input, options \\ [])

Updates a TypedLinkFacet.

For more information, see Typed Links.

upgrade_applied_schema(client, input, options \\ [])

Upgrades a single directory in-place using the PublishedSchemaArn with schema updates found in MinorVersion.

Backwards-compatible minor version upgrades are instantaneously available for readers on all objects in the directory. Note: This is a synchronous API call and upgrades only one schema on a given directory per call. To upgrade multiple directories from one schema, you would need to call this API on each directory.

upgrade_published_schema(client, input, options \\ [])

Upgrades a published schema under a new minor version revision using the current contents of DevelopmentSchemaArn.