aws-elixir v0.6.0 AWS.CloudFront

Amazon CloudFront

This is the Amazon CloudFront API Reference.

This guide is for developers who need detailed information about CloudFront API actions, data types, and errors. For detailed information about CloudFront features, see the Amazon CloudFront Developer Guide.

Functions

create_cache_policy(client, input, options \\ [])

Creates a cache policy.

After you create a cache policy, you can attach it to one or more cache behaviors. When it’s attached to a cache behavior, the cache policy determines the following:

The values that CloudFront includes in the cache key. These values can include HTTP headers, cookies, and URL query strings. CloudFront uses the cache key to find an object in its cache that it can return to the viewer.

The default, minimum, and maximum time to live (TTL) values that you want objects to stay in the CloudFront cache.

The headers, cookies, and query strings that are included in the cache key are automatically included in requests that CloudFront sends to the origin. CloudFront sends a request when it can’t find an object in its cache that matches the request’s cache key. If you want to send values to the origin but not include them in the cache key, use OriginRequestPolicy.

For more information about cache policies, see Controlling the cache key in the Amazon CloudFront Developer Guide.

create_cloud_front_origin_access_identity(client, input, options \\ [])

Creates a new origin access identity.

If you're using Amazon S3 for your origin, you can use an origin access identity to require users to access your content using a CloudFront URL instead of the Amazon S3 URL. For more information about how to use origin access identities, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

create_distribution(client, input, options \\ [])

Creates a new web distribution.

You create a CloudFront distribution to tell CloudFront where you want content to be delivered from, and the details about how to track and manage content delivery. Send a POST request to the /*CloudFront API version*/distribution/distribution ID resource.

When you update a distribution, there are more required fields than when you create a distribution. When you update your distribution by using UpdateDistribution, follow the steps included in the documentation to get the current configuration and then make your updates. This helps to make sure that you include all of the required fields. To view a summary, see Required Fields for Create Distribution and Update Distribution in the Amazon CloudFront Developer Guide.

create_distribution_with_tags(client, input, options \\ [])

Create a new distribution with tags.

create_field_level_encryption_config(client, input, options \\ [])

Create a new field-level encryption configuration.

create_field_level_encryption_profile(client, input, options \\ [])

Create a field-level encryption profile.

create_invalidation(client, distribution_id, input, options \\ [])

Create a new invalidation.

create_monitoring_subscription(client, distribution_id, input, options \\ [])

Enables additional CloudWatch metrics for the specified CloudFront distribution.

The additional metrics incur an additional cost.

For more information, see Viewing additional CloudFront distribution metrics in the Amazon CloudFront Developer Guide.

create_origin_request_policy(client, input, options \\ [])

Creates an origin request policy.

After you create an origin request policy, you can attach it to one or more cache behaviors. When it’s attached to a cache behavior, the origin request policy determines the values that CloudFront includes in requests that it sends to the origin. Each request that CloudFront sends to the origin includes the following:

The request body and the URL path (without the domain name) from the viewer request.

The headers that CloudFront automatically includes in every origin request, including Host, User-Agent, and X-Amz-Cf-Id.

All HTTP headers, cookies, and URL query strings that are specified in the cache policy or the origin request policy. These can include items from the viewer request and, in the case of headers, additional ones that are added by CloudFront.

CloudFront sends a request when it can’t find a valid object in its cache that matches the request. If you want to send values to the origin and also include them in the cache key, use CachePolicy.

For more information about origin request policies, see Controlling origin requests in the Amazon CloudFront Developer Guide.

create_public_key(client, input, options \\ [])

Add a new public key to CloudFront to use, for example, for field-level encryption.

You can add a maximum of 10 public keys with one AWS account.

create_realtime_log_config(client, input, options \\ [])

Creates a real-time log configuration.

After you create a real-time log configuration, you can attach it to one or more cache behaviors to send real-time log data to the specified Amazon Kinesis data stream.

For more information about real-time log configurations, see Real-time logs in the Amazon CloudFront Developer Guide.

create_streaming_distribution(client, input, options \\ [])

Creates a new RTMP distribution.

An RTMP distribution is similar to a web distribution, but an RTMP distribution streams media files using the Adobe Real-Time Messaging Protocol (RTMP) instead of serving files using HTTP.

To create a new distribution, submit a POST request to the CloudFront API version/distribution resource. The request body must include a document with a StreamingDistributionConfig element. The response echoes the StreamingDistributionConfig element and returns other information about the RTMP distribution.

To get the status of your request, use the GET StreamingDistribution API action. When the value of Enabled is true and the value of Status is Deployed, your distribution is ready. A distribution usually deploys in less than 15 minutes.

For more information about web distributions, see Working with RTMP Distributions in the Amazon CloudFront Developer Guide.

Beginning with the 2012-05-05 version of the CloudFront API, we made substantial changes to the format of the XML document that you include in the request body when you create or update a web distribution or an RTMP distribution, and when you invalidate objects. With previous versions of the API, we discovered that it was too easy to accidentally delete one or more values for an element that accepts multiple values, for example, CNAMEs and trusted signers. Our changes for the 2012-05-05 release are intended to prevent these accidental deletions and to notify you when there's a mismatch between the number of values you say you're specifying in the Quantity element and the number of values specified.

create_streaming_distribution_with_tags(client, input, options \\ [])

Create a new streaming distribution with tags.

delete_cache_policy(client, id, input, options \\ [])

Deletes a cache policy.

You cannot delete a cache policy if it’s attached to a cache behavior. First update your distributions to remove the cache policy from all cache behaviors, then delete the cache policy.

To delete a cache policy, you must provide the policy’s identifier and version. To get these values, you can use ListCachePolicies or GetCachePolicy.

delete_cloud_front_origin_access_identity(client, id, input, options \\ [])

Delete an origin access identity.

delete_distribution(client, id, input, options \\ [])

Delete a distribution.

delete_field_level_encryption_config(client, id, input, options \\ [])

Remove a field-level encryption configuration.

delete_field_level_encryption_profile(client, id, input, options \\ [])

Remove a field-level encryption profile.

delete_monitoring_subscription(client, distribution_id, input, options \\ [])

Disables additional CloudWatch metrics for the specified CloudFront distribution.

delete_origin_request_policy(client, id, input, options \\ [])

Deletes an origin request policy.

You cannot delete an origin request policy if it’s attached to any cache behaviors. First update your distributions to remove the origin request policy from all cache behaviors, then delete the origin request policy.

To delete an origin request policy, you must provide the policy’s identifier and version. To get the identifier, you can use ListOriginRequestPolicies or GetOriginRequestPolicy.

delete_public_key(client, id, input, options \\ [])

Remove a public key you previously added to CloudFront.

delete_realtime_log_config(client, input, options \\ [])

Deletes a real-time log configuration.

You cannot delete a real-time log configuration if it’s attached to a cache behavior. First update your distributions to remove the real-time log configuration from all cache behaviors, then delete the real-time log configuration.

To delete a real-time log configuration, you can provide the configuration’s name or its Amazon Resource Name (ARN). You must provide at least one. If you provide both, CloudFront uses the name to identify the real-time log configuration to delete.

delete_streaming_distribution(client, id, input, options \\ [])

Delete a streaming distribution.

To delete an RTMP distribution using the CloudFront API, perform the following steps.

To delete an RTMP distribution using the CloudFront API:

Disable the RTMP distribution.

Submit a GET Streaming Distribution Config request to get the current configuration and the Etag header for the distribution.

Update the XML document that was returned in the response to your GET Streaming Distribution Config request to change the value of Enabled to false.

Submit a PUT Streaming Distribution Config request to update the configuration for your distribution. In the request body, include the XML document that you updated in Step 3. Then set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Streaming Distribution Config request in Step 2.

Review the response to the PUT Streaming Distribution Config request to confirm that the distribution was successfully disabled.

Submit a GET Streaming Distribution Config request to confirm that your changes have propagated. When propagation is complete, the value of Status is Deployed.

Submit a DELETE Streaming Distribution request. Set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Streaming Distribution Config request in Step 2.

Review the response to your DELETE Streaming Distribution request to confirm that the distribution was successfully deleted.

For information about deleting a distribution using the CloudFront console, see Deleting a Distribution in the Amazon CloudFront Developer Guide.

get_cache_policy(client, id, options \\ [])

Gets a cache policy, including the following metadata:

The policy’s identifier.

The date and time when the policy was last modified.

To get a cache policy, you must provide the policy’s identifier. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

get_cache_policy_config(client, id, options \\ [])

Gets a cache policy configuration.

To get a cache policy configuration, you must provide the policy’s identifier. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

get_cloud_front_origin_access_identity(client, id, options \\ [])

Get the information about an origin access identity.

get_cloud_front_origin_access_identity_config(client, id, options \\ [])

Get the configuration information about an origin access identity.

get_distribution(client, id, options \\ [])

Get the information about a distribution.

get_distribution_config(client, id, options \\ [])

Get the configuration information about a distribution.

get_field_level_encryption(client, id, options \\ [])

Get the field-level encryption configuration information.

get_field_level_encryption_config(client, id, options \\ [])

Get the field-level encryption configuration information.

get_field_level_encryption_profile(client, id, options \\ [])

Get the field-level encryption profile information.

get_field_level_encryption_profile_config(client, id, options \\ [])

Get the field-level encryption profile configuration information.

get_invalidation(client, distribution_id, id, options \\ [])

Get the information about an invalidation.

get_monitoring_subscription(client, distribution_id, options \\ [])

Gets information about whether additional CloudWatch metrics are enabled for the specified CloudFront distribution.

get_origin_request_policy(client, id, options \\ [])

Gets an origin request policy, including the following metadata:

The policy’s identifier.

The date and time when the policy was last modified.

To get an origin request policy, you must provide the policy’s identifier. If the origin request policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the origin request policy is not attached to a cache behavior, you can get the identifier using ListOriginRequestPolicies.

get_origin_request_policy_config(client, id, options \\ [])

Gets an origin request policy configuration.

To get an origin request policy configuration, you must provide the policy’s identifier. If the origin request policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the origin request policy is not attached to a cache behavior, you can get the identifier using ListOriginRequestPolicies.

get_public_key(client, id, options \\ [])

Get the public key information.

get_public_key_config(client, id, options \\ [])

Return public key configuration informaation

get_realtime_log_config(client, input, options \\ [])

Gets a real-time log configuration.

To get a real-time log configuration, you can provide the configuration’s name or its Amazon Resource Name (ARN). You must provide at least one. If you provide both, CloudFront uses the name to identify the real-time log configuration to get.

get_streaming_distribution(client, id, options \\ [])

Gets information about a specified RTMP distribution, including the distribution configuration.

get_streaming_distribution_config(client, id, options \\ [])

Get the configuration information about a streaming distribution.

list_cache_policies(client, marker \\ nil, max_items \\ nil, type \\ nil, options \\ [])

Gets a list of cache policies.

You can optionally apply a filter to return only the managed policies created by AWS, or only the custom policies created in your AWS account.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

list_cloud_front_origin_access_identities(client, marker \\ nil, max_items \\ nil, options \\ [])

Lists origin access identities.

list_distributions(client, marker \\ nil, max_items \\ nil, options \\ [])

List CloudFront distributions.

list_distributions_by_cache_policy_id(client, cache_policy_id, marker \\ nil, max_items \\ nil, options \\ [])

Gets a list of distribution IDs for distributions that have a cache behavior that’s associated with the specified cache policy.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

list_distributions_by_origin_request_policy_id(client, origin_request_policy_id, marker \\ nil, max_items \\ nil, options \\ [])

Gets a list of distribution IDs for distributions that have a cache behavior that’s associated with the specified origin request policy.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

list_distributions_by_realtime_log_config(client, input, options \\ [])

Gets a list of distributions that have a cache behavior that’s associated with the specified real-time log configuration.

You can specify the real-time log configuration by its name or its Amazon Resource Name (ARN). You must provide at least one. If you provide both, CloudFront uses the name to identify the real-time log configuration to list distributions for.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

list_distributions_by_web_a_c_l_id(client, web_a_c_l_id, marker \\ nil, max_items \\ nil, options \\ [])

List the distributions that are associated with a specified AWS WAF web ACL.

list_field_level_encryption_configs(client, marker \\ nil, max_items \\ nil, options \\ [])

List all field-level encryption configurations that have been created in CloudFront for this account.

list_field_level_encryption_profiles(client, marker \\ nil, max_items \\ nil, options \\ [])

Request a list of field-level encryption profiles that have been created in CloudFront for this account.

list_invalidations(client, distribution_id, marker \\ nil, max_items \\ nil, options \\ [])

Lists invalidation batches.

list_origin_request_policies(client, marker \\ nil, max_items \\ nil, type \\ nil, options \\ [])

Gets a list of origin request policies.

You can optionally apply a filter to return only the managed policies created by AWS, or only the custom policies created in your AWS account.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

list_public_keys(client, marker \\ nil, max_items \\ nil, options \\ [])

List all public keys that have been added to CloudFront for this account.

list_realtime_log_configs(client, marker \\ nil, max_items \\ nil, options \\ [])

Gets a list of real-time log configurations.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

list_streaming_distributions(client, marker \\ nil, max_items \\ nil, options \\ [])

List streaming distributions.

list_tags_for_resource(client, resource, options \\ [])

List tags for a CloudFront resource.

tag_resource(client, input, options \\ [])

Add tags to a CloudFront resource.

untag_resource(client, input, options \\ [])

Remove tags from a CloudFront resource.

update_cache_policy(client, id, input, options \\ [])

Updates a cache policy configuration.

When you update a cache policy configuration, all the fields are updated with the values provided in the request. You cannot update some fields independent of others. To update a cache policy configuration:

Use GetCachePolicyConfig to get the current configuration.

Locally modify the fields in the cache policy configuration that you want to update.

Call UpdateCachePolicy by providing the entire cache policy configuration, including the fields that you modified and those that you didn’t.

update_cloud_front_origin_access_identity(client, id, input, options \\ [])

Update an origin access identity.

update_distribution(client, id, input, options \\ [])

Updates the configuration for a web distribution.

When you update a distribution, there are more required fields than when you create a distribution. When you update your distribution by using this API action, follow the steps here to get the current configuration and then make your updates, to make sure that you include all of the required fields. To view a summary, see Required Fields for Create Distribution and Update Distribution in the Amazon CloudFront Developer Guide.

The update process includes getting the current distribution configuration, updating the XML document that is returned to make your changes, and then submitting an UpdateDistribution request to make the updates.

For information about updating a distribution using the CloudFront console instead, see Creating a Distribution in the Amazon CloudFront Developer Guide.

To update a web distribution using the CloudFront API

Submit a GetDistributionConfig request to get the current configuration and an Etag header for the distribution.

If you update the distribution again, you must get a new Etag header.

Update the XML document that was returned in the response to your GetDistributionConfig request to include your changes.

When you edit the XML file, be aware of the following:

You must strip out the ETag parameter that is returned.

Additional fields are required when you update a distribution. There may be fields included in the XML file for features that you haven't configured for your distribution. This is expected and required to successfully update the distribution.

You can't change the value of CallerReference. If you try to change this value, CloudFront returns an IllegalUpdate error.

The new configuration replaces the existing configuration; the values that you specify in an UpdateDistribution request are not merged into your existing configuration. When you add, delete, or replace values in an element that allows multiple values (for example, CNAME), you must specify all of the values that you want to appear in the updated distribution. In addition, you must update the corresponding Quantity element.

Submit an UpdateDistribution request to update the configuration for your distribution:

In the request body, include the XML document that you updated in Step

The request body must include an XML document with a DistributionConfig element.

Set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GetDistributionConfig request in Step 1.

Review the response to the UpdateDistribution request to confirm that the

configuration was successfully updated.

Optional: Submit a GetDistribution request to confirm that your changes have propagated. When propagation is complete, the value of Status is Deployed.

update_field_level_encryption_config(client, id, input, options \\ [])

Update a field-level encryption configuration.

update_field_level_encryption_profile(client, id, input, options \\ [])

Update a field-level encryption profile.

update_origin_request_policy(client, id, input, options \\ [])

Updates an origin request policy configuration.

When you update an origin request policy configuration, all the fields are updated with the values provided in the request. You cannot update some fields independent of others. To update an origin request policy configuration:

Use GetOriginRequestPolicyConfig to get the current configuration.

Locally modify the fields in the origin request policy configuration that you want to update.

Call UpdateOriginRequestPolicy by providing the entire origin request policy configuration, including the fields that you modified and those that you didn’t.

update_public_key(client, id, input, options \\ [])

Update public key information.

Note that the only value you can change is the comment.

update_realtime_log_config(client, input, options \\ [])

Updates a real-time log configuration.

When you update a real-time log configuration, all the parameters are updated with the values provided in the request. You cannot update some parameters independent of others. To update a real-time log configuration:

Call GetRealtimeLogConfig to get the current real-time log configuration.

Locally modify the parameters in the real-time log configuration that you want to update.

Call this API (UpdateRealtimeLogConfig) by providing the entire real-time log configuration, including the parameters that you modified and those that you didn’t.

You cannot update a real-time log configuration’s Name or ARN.

update_streaming_distribution(client, id, input, options \\ [])

Update a streaming distribution.