aws-elixir v0.6.0 AWS.CloudHSMV2

For more information about AWS CloudHSM, see AWS CloudHSM and the AWS CloudHSM User Guide.

Functions

copy_backup_to_region(client, input, options \\ [])

Copy an AWS CloudHSM cluster backup to a different region.

create_cluster(client, input, options \\ [])

Creates a new AWS CloudHSM cluster.

create_hsm(client, input, options \\ [])

Creates a new hardware security module (HSM) in the specified AWS CloudHSM cluster.

delete_backup(client, input, options \\ [])

Deletes a specified AWS CloudHSM backup.

A backup can be restored up to 7 days after the DeleteBackup request is made. For more information on restoring a backup, see RestoreBackup.

delete_cluster(client, input, options \\ [])

Deletes the specified AWS CloudHSM cluster.

Before you can delete a cluster, you must delete all HSMs in the cluster. To see if the cluster contains any HSMs, use DescribeClusters. To delete an HSM, use DeleteHsm.

delete_hsm(client, input, options \\ [])

Deletes the specified HSM.

To specify an HSM, you can use its identifier (ID), the IP address of the HSM's elastic network interface (ENI), or the ID of the HSM's ENI. You need to specify only one of these values. To find these values, use DescribeClusters.

describe_backups(client, input, options \\ [])

Gets information about backups of AWS CloudHSM clusters.

This is a paginated operation, which means that each response might contain only a subset of all the backups. When the response contains only a subset of backups, it includes a NextToken value. Use this value in a subsequent DescribeBackups request to get more backups. When you receive a response with no NextToken (or an empty or null value), that means there are no more backups to get.

describe_clusters(client, input, options \\ [])

Gets information about AWS CloudHSM clusters.

This is a paginated operation, which means that each response might contain only a subset of all the clusters. When the response contains only a subset of clusters, it includes a NextToken value. Use this value in a subsequent DescribeClusters request to get more clusters. When you receive a response with no NextToken (or an empty or null value), that means there are no more clusters to get.

initialize_cluster(client, input, options \\ [])

Claims an AWS CloudHSM cluster by submitting the cluster certificate issued by your issuing certificate authority (CA) and the CA's root certificate.

Before you can claim a cluster, you must sign the cluster's certificate signing request (CSR) with your issuing CA. To get the cluster's CSR, use DescribeClusters.

list_tags(client, input, options \\ [])

Gets a list of tags for the specified AWS CloudHSM cluster.

This is a paginated operation, which means that each response might contain only a subset of all the tags. When the response contains only a subset of tags, it includes a NextToken value. Use this value in a subsequent ListTags request to get more tags. When you receive a response with no NextToken (or an empty or null value), that means there are no more tags to get.

restore_backup(client, input, options \\ [])

Restores a specified AWS CloudHSM backup that is in the PENDING_DELETION state.

For mor information on deleting a backup, see DeleteBackup.

tag_resource(client, input, options \\ [])

Adds or overwrites one or more tags for the specified AWS CloudHSM cluster.

untag_resource(client, input, options \\ [])

Removes the specified tag or tags from the specified AWS CloudHSM cluster.