aws-elixir v0.6.0 AWS.CloudSearch

Amazon CloudSearch Configuration Service

You use the Amazon CloudSearch configuration service to create, configure, and manage search domains.

Configuration service requests are submitted using the AWS Query protocol. AWS Query requests are HTTP or HTTPS requests submitted via HTTP GET or POST with a query parameter named Action.

The endpoint for configuration service requests is region-specific: cloudsearch.region.amazonaws.com. For example, cloudsearch.us-east-1.amazonaws.com. For a current list of supported regions and endpoints, see Regions and Endpoints.

Functions

build_suggesters(client, input, options \\ [])

Indexes the search suggestions.

For more information, see Configuring Suggesters in the Amazon CloudSearch Developer Guide.

create_domain(client, input, options \\ [])

Creates a new search domain.

For more information, see Creating a Search Domain in the Amazon CloudSearch Developer Guide.

define_analysis_scheme(client, input, options \\ [])

Configures an analysis scheme that can be applied to a text or text-array field to define language-specific text processing options.

For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

define_expression(client, input, options \\ [])

Configures an Expression for the search domain.

Used to create new expressions and modify existing ones. If the expression exists, the new configuration replaces the old one. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

define_index_field(client, input, options \\ [])

Configures an IndexField for the search domain.

Used to create new fields and modify existing ones. You must specify the name of the domain you are configuring and an index field configuration. The index field configuration specifies a unique name, the index field type, and the options you want to configure for the field. The options you can specify depend on the IndexFieldType. If the field exists, the new configuration replaces the old one. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

define_suggester(client, input, options \\ [])

Configures a suggester for a domain.

A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

delete_analysis_scheme(client, input, options \\ [])

Deletes an analysis scheme.

For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

delete_domain(client, input, options \\ [])

Permanently deletes a search domain and all of its data.

Once a domain has been deleted, it cannot be recovered. For more information, see Deleting a Search Domain in the Amazon CloudSearch Developer Guide.

delete_expression(client, input, options \\ [])

Removes an Expression from the search domain.

For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

delete_index_field(client, input, options \\ [])

Removes an IndexField from the search domain.

For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

delete_suggester(client, input, options \\ [])

Deletes a suggester.

For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

describe_analysis_schemes(client, input, options \\ [])

Gets the analysis schemes configured for a domain.

An analysis scheme defines language-specific text processing options for a text field. Can be limited to specific analysis schemes by name. By default, shows all analysis schemes and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

describe_availability_options(client, input, options \\ [])

Gets the availability options configured for a domain.

By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

describe_domain_endpoint_options(client, input, options \\ [])

Returns the domain's endpoint options, specifically whether all requests to the domain must arrive over HTTPS.

For more information, see Configuring Domain Endpoint Options in the Amazon CloudSearch Developer Guide.

describe_domains(client, input, options \\ [])

Gets information about the search domains owned by this account.

Can be limited to specific domains. Shows all domains by default. To get the number of searchable documents in a domain, use the console or submit a matchall request to your domain's search endpoint: q=matchall&amp;q.parser=structured&amp;size=0. For more information, see Getting Information about a Search Domain in the Amazon CloudSearch Developer Guide.

describe_expressions(client, input, options \\ [])

Gets the expressions configured for the search domain.

Can be limited to specific expressions by name. By default, shows all expressions and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

describe_index_fields(client, input, options \\ [])

Gets information about the index fields configured for the search domain.

Can be limited to specific fields by name. By default, shows all fields and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Domain Information in the Amazon CloudSearch Developer Guide.

describe_scaling_parameters(client, input, options \\ [])

Gets the scaling parameters configured for a domain.

A domain's scaling parameters specify the desired search instance type and replication count. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

describe_service_access_policies(client, input, options \\ [])

Gets information about the access policies that control access to the domain's document and search endpoints.

By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide.

describe_suggesters(client, input, options \\ [])

Gets the suggesters configured for a domain.

A suggester enables you to display possible matches before users finish typing their queries. Can be limited to specific suggesters by name. By default, shows all suggesters and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

index_documents(client, input, options \\ [])

Tells the search domain to start indexing its documents using the latest indexing options.

This operation must be invoked to activate options whose OptionStatus is RequiresIndexDocuments.

list_domain_names(client, input, options \\ [])

Lists all search domains owned by an account.

update_availability_options(client, input, options \\ [])

Configures the availability options for a domain.

Enabling the Multi-AZ option expands an Amazon CloudSearch domain to an additional Availability Zone in the same Region to increase fault tolerance in the event of a service disruption. Changes to the Multi-AZ option can take about half an hour to become active. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

update_domain_endpoint_options(client, input, options \\ [])

Updates the domain's endpoint options, specifically whether all requests to the domain must arrive over HTTPS.

For more information, see Configuring Domain Endpoint Options in the Amazon CloudSearch Developer Guide.

update_scaling_parameters(client, input, options \\ [])

Configures scaling parameters for a domain.

A domain's scaling parameters specify the desired search instance type and replication count. Amazon CloudSearch will still automatically scale your domain based on the volume of data and traffic, but not below the desired instance type and replication count. If the Multi-AZ option is enabled, these values control the resources used per Availability Zone. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

update_service_access_policies(client, input, options \\ [])

Configures the access rules that control access to the domain's document and search endpoints.

For more information, see Configuring Access for an Amazon CloudSearch Domain.