aws-elixir v0.6.0 AWS.CodeBuild

AWS CodeBuild

AWS CodeBuild is a fully managed build service in the cloud.

AWS CodeBuild compiles your source code, runs unit tests, and produces artifacts that are ready to deploy. AWS CodeBuild eliminates the need to provision, manage, and scale your own build servers. It provides prepackaged build environments for the most popular programming languages and build tools, such as Apache Maven, Gradle, and more. You can also fully customize build environments in AWS CodeBuild to use your own build tools. AWS CodeBuild scales automatically to meet peak build requests. You pay only for the build time you consume. For more information about AWS CodeBuild, see the AWS CodeBuild User Guide.

AWS CodeBuild supports these operations:

BatchDeleteBuilds: Deletes one or more builds.

BatchGetBuilds: Gets information about one or more builds.

BatchGetProjects: Gets information about one or more build projects. A build project defines how AWS CodeBuild runs a build. This includes information such as where to get the source code to build, the build environment to use, the build commands to run, and where to store the build output. A build environment is a representation of operating system, programming language runtime, and tools that AWS CodeBuild uses to run a build. You can add tags to build projects to help manage your resources and costs.

BatchGetReportGroups: Returns an array of report groups.

BatchGetReports: Returns an array of reports.

CreateProject: Creates a build project.

CreateReportGroup: Creates a report group. A report group contains a collection of reports.

CreateWebhook: For an existing AWS CodeBuild build project that has its source code stored in a GitHub or Bitbucket repository, enables AWS CodeBuild to start rebuilding the source code every time a code change is pushed to the repository.

DeleteProject: Deletes a build project.

DeleteReport: Deletes a report.

DeleteReportGroup: Deletes a report group.

DeleteResourcePolicy: Deletes a resource policy that is identified by its resource ARN.

DeleteSourceCredentials: Deletes a set of GitHub, GitHub Enterprise, or Bitbucket source credentials.

DeleteWebhook: For an existing AWS CodeBuild build project that has its source code stored in a GitHub or Bitbucket repository, stops AWS CodeBuild from rebuilding the source code every time a code change is pushed to the repository.

DescribeTestCases: Returns a list of details about test cases for a report.

GetResourcePolicy: Gets a resource policy that is identified by its resource ARN.

ImportSourceCredentials: Imports the source repository credentials for an AWS CodeBuild project that has its source code stored in a GitHub, GitHub Enterprise, or Bitbucket repository.

InvalidateProjectCache: Resets the cache for a project.

ListBuilds: Gets a list of build IDs, with each build ID representing a single build.

ListBuildsForProject: Gets a list of build IDs for the specified build project, with each build ID representing a single build.

ListCuratedEnvironmentImages: Gets information about Docker images that are managed by AWS CodeBuild.

ListProjects: Gets a list of build project names, with each build project name representing a single build project.

ListReportGroups: Gets a list ARNs for the report groups in the current AWS account.

ListReports: Gets a list ARNs for the reports in the current AWS account.

ListReportsForReportGroup: Returns a list of ARNs for the reports that belong to a ReportGroup.

ListSharedProjects: Gets a list of ARNs associated with projects shared with the current AWS account or user.

ListSharedReportGroups: Gets a list of ARNs associated with report groups shared with the current AWS account or user

ListSourceCredentials: Returns a list of SourceCredentialsInfo objects. Each SourceCredentialsInfo object includes the authentication type, token ARN, and type of source provider for one set of credentials.

PutResourcePolicy: Stores a resource policy for the ARN of a Project or ReportGroup object.

StartBuild: Starts running a build.

StopBuild: Attempts to stop running a build.

UpdateProject: Changes the settings of an existing build project.

UpdateReportGroup: Changes a report group.

UpdateWebhook: Changes the settings of an existing webhook.

Functions

batch_delete_builds(client, input, options \\ [])

Deletes one or more builds.

batch_get_build_batches(client, input, options \\ [])

Retrieves information about one or more batch builds.

batch_get_builds(client, input, options \\ [])

Gets information about one or more builds.

batch_get_projects(client, input, options \\ [])

Gets information about one or more build projects.

batch_get_report_groups(client, input, options \\ [])

Returns an array of report groups.

batch_get_reports(client, input, options \\ [])

Returns an array of reports.

create_project(client, input, options \\ [])

Creates a build project.

create_report_group(client, input, options \\ [])

Creates a report group.

A report group contains a collection of reports.

create_webhook(client, input, options \\ [])

For an existing AWS CodeBuild build project that has its source code stored in a GitHub or Bitbucket repository, enables AWS CodeBuild to start rebuilding the source code every time a code change is pushed to the repository.

If you enable webhooks for an AWS CodeBuild project, and the project is used as a build step in AWS CodePipeline, then two identical builds are created for each commit. One build is triggered through webhooks, and one through AWS CodePipeline. Because billing is on a per-build basis, you are billed for both builds. Therefore, if you are using AWS CodePipeline, we recommend that you disable webhooks in AWS CodeBuild. In the AWS CodeBuild console, clear the Webhook box. For more information, see step 5 in Change a Build Project's Settings.

delete_build_batch(client, input, options \\ [])

Deletes a batch build.

delete_project(client, input, options \\ [])

Deletes a build project.

When you delete a project, its builds are not deleted.

delete_report(client, input, options \\ [])

Deletes a report.

delete_report_group(client, input, options \\ [])

Deletes a report group.

Before you delete a report group, you must delete its reports.

delete_resource_policy(client, input, options \\ [])

Deletes a resource policy that is identified by its resource ARN.

delete_source_credentials(client, input, options \\ [])

Deletes a set of GitHub, GitHub Enterprise, or Bitbucket source credentials.

delete_webhook(client, input, options \\ [])

For an existing AWS CodeBuild build project that has its source code stored in a GitHub or Bitbucket repository, stops AWS CodeBuild from rebuilding the source code every time a code change is pushed to the repository.

describe_code_coverages(client, input, options \\ [])

Retrieves one or more code coverage reports.

describe_test_cases(client, input, options \\ [])

Returns a list of details about test cases for a report.

get_resource_policy(client, input, options \\ [])

Gets a resource policy that is identified by its resource ARN.

import_source_credentials(client, input, options \\ [])

Imports the source repository credentials for an AWS CodeBuild project that has its source code stored in a GitHub, GitHub Enterprise, or Bitbucket repository.

invalidate_project_cache(client, input, options \\ [])

Resets the cache for a project.

list_build_batches(client, input, options \\ [])

Retrieves the identifiers of your build batches in the current region.

list_build_batches_for_project(client, input, options \\ [])

Retrieves the identifiers of the build batches for a specific project.

list_builds(client, input, options \\ [])

Gets a list of build IDs, with each build ID representing a single build.

list_builds_for_project(client, input, options \\ [])

Gets a list of build IDs for the specified build project, with each build ID representing a single build.

list_curated_environment_images(client, input, options \\ [])

Gets information about Docker images that are managed by AWS CodeBuild.

list_projects(client, input, options \\ [])

Gets a list of build project names, with each build project name representing a single build project.

list_report_groups(client, input, options \\ [])

Gets a list ARNs for the report groups in the current AWS account.

list_reports(client, input, options \\ [])

Returns a list of ARNs for the reports in the current AWS account.

list_reports_for_report_group(client, input, options \\ [])

Returns a list of ARNs for the reports that belong to a ReportGroup.

list_shared_projects(client, input, options \\ [])

Gets a list of projects that are shared with other AWS accounts or users.

list_shared_report_groups(client, input, options \\ [])

Gets a list of report groups that are shared with other AWS accounts or users.

list_source_credentials(client, input, options \\ [])

Returns a list of SourceCredentialsInfo objects.

put_resource_policy(client, input, options \\ [])

Stores a resource policy for the ARN of a Project or ReportGroup object.

retry_build(client, input, options \\ [])

Restarts a build.

retry_build_batch(client, input, options \\ [])

Restarts a batch build.

start_build(client, input, options \\ [])

Starts running a build.

start_build_batch(client, input, options \\ [])

Starts a batch build for a project.

stop_build(client, input, options \\ [])

Attempts to stop running a build.

stop_build_batch(client, input, options \\ [])

Stops a running batch build.

update_project(client, input, options \\ [])

Changes the settings of a build project.

update_report_group(client, input, options \\ [])

Updates a report group.

update_webhook(client, input, options \\ [])

Updates the webhook associated with an AWS CodeBuild build project.

If you use Bitbucket for your repository, rotateSecret is ignored.