aws-elixir v0.6.0 AWS.CodeDeploy

AWS CodeDeploy

AWS CodeDeploy is a deployment service that automates application deployments to Amazon EC2 instances, on-premises instances running in your own facility, serverless AWS Lambda functions, or applications in an Amazon ECS service.

You can deploy a nearly unlimited variety of application content, such as an updated Lambda function, updated applications in an Amazon ECS service, code, web and configuration files, executables, packages, scripts, multimedia files, and so on. AWS CodeDeploy can deploy application content stored in Amazon S3 buckets, GitHub repositories, or Bitbucket repositories. You do not need to make changes to your existing code before you can use AWS CodeDeploy.

AWS CodeDeploy makes it easier for you to rapidly release new features, helps you avoid downtime during application deployment, and handles the complexity of updating your applications, without many of the risks associated with error-prone manual deployments.

AWS CodeDeploy Components

Use the information in this guide to help you work with the following AWS CodeDeploy components:

Application: A name that uniquely identifies the application you want to deploy. AWS CodeDeploy uses this name, which functions as a container, to ensure the correct combination of revision, deployment configuration, and deployment group are referenced during a deployment.

Deployment group: A set of individual instances, CodeDeploy Lambda deployment configuration settings, or an Amazon ECS service and network details. A Lambda deployment group specifies how to route traffic to a new version of a Lambda function. An Amazon ECS deployment group specifies the service created in Amazon ECS to deploy, a load balancer, and a listener to reroute production traffic to an updated containerized application. An EC2/On-premises deployment group contains individually tagged instances, Amazon EC2 instances in Amazon EC2 Auto Scaling groups, or both. All deployment groups can specify optional trigger, alarm, and rollback settings.

Deployment configuration: A set of deployment rules and deployment success and failure conditions used by AWS CodeDeploy during a deployment.

Deployment: The process and the components used when updating a Lambda function, a containerized application in an Amazon ECS service, or of installing content on one or more instances.

Application revisions: For an AWS Lambda deployment, this is an AppSpec file that specifies the Lambda function to be updated and one or more functions to validate deployment lifecycle events. For an Amazon ECS deployment, this is an AppSpec file that specifies the Amazon ECS task definition, container, and port where production traffic is rerouted. For an EC2/On-premises deployment, this is an archive file that contains source content—source code, webpages, executable files, and deployment scripts—along with an AppSpec file. Revisions are stored in Amazon S3 buckets or GitHub repositories. For Amazon S3, a revision is uniquely identified by its Amazon S3 object key and its ETag, version, or both. For GitHub, a revision is uniquely identified by its commit ID.

This guide also contains information to help you get details about the instances in your deployments, to make on-premises instances available for AWS CodeDeploy deployments, to get details about a Lambda function deployment, and to get details about Amazon ECS service deployments.

AWS CodeDeploy Information Resources

AWS CodeDeploy User Guide * AWS CodeDeploy API Reference Guide

AWS CLI Reference for AWS CodeDeploy

AWS CodeDeploy Developer Forum

Functions

add_tags_to_on_premises_instances(client, input, options \\ [])

Adds tags to on-premises instances.

batch_get_application_revisions(client, input, options \\ [])

Gets information about one or more application revisions.

The maximum number of application revisions that can be returned is 25.

batch_get_applications(client, input, options \\ [])

Gets information about one or more applications.

The maximum number of applications that can be returned is 100.

batch_get_deployment_groups(client, input, options \\ [])

Gets information about one or more deployment groups.

batch_get_deployment_instances(client, input, options \\ [])

This method works, but is deprecated.

Use BatchGetDeploymentTargets instead.

Returns an array of one or more instances associated with a deployment. This method works with EC2/On-premises and AWS Lambda compute platforms. The newer BatchGetDeploymentTargets works with all compute platforms. The maximum number of instances that can be returned is 25.

batch_get_deployment_targets(client, input, options \\ [])

Returns an array of one or more targets associated with a deployment.

This method works with all compute types and should be used instead of the deprecated BatchGetDeploymentInstances. The maximum number of targets that can be returned is 25.

The type of targets returned depends on the deployment's compute platform or deployment method:

EC2/On-premises: Information about EC2 instance targets.

AWS Lambda: Information about Lambda functions targets.

Amazon ECS: Information about Amazon ECS service targets.

CloudFormation: Information about targets of blue/green deployments initiated by a CloudFormation stack update.

batch_get_deployments(client, input, options \\ [])

Gets information about one or more deployments.

The maximum number of deployments that can be returned is 25.

batch_get_on_premises_instances(client, input, options \\ [])

Gets information about one or more on-premises instances.

The maximum number of on-premises instances that can be returned is 25.

continue_deployment(client, input, options \\ [])

For a blue/green deployment, starts the process of rerouting traffic from instances in the original environment to instances in the replacement environment without waiting for a specified wait time to elapse.

(Traffic rerouting, which is achieved by registering instances in the replacement environment with the load balancer, can start as soon as all instances have a status of Ready.)

create_application(client, input, options \\ [])

Creates an application.

create_deployment(client, input, options \\ [])

Deploys an application revision through the specified deployment group.

create_deployment_config(client, input, options \\ [])

Creates a deployment configuration.

create_deployment_group(client, input, options \\ [])

Creates a deployment group to which application revisions are deployed.

delete_application(client, input, options \\ [])

Deletes an application.

delete_deployment_config(client, input, options \\ [])

Deletes a deployment configuration.

A deployment configuration cannot be deleted if it is currently in use. Predefined configurations cannot be deleted.

delete_deployment_group(client, input, options \\ [])

Deletes a deployment group.

delete_git_hub_account_token(client, input, options \\ [])

Deletes a GitHub account connection.

delete_resources_by_external_id(client, input, options \\ [])

Deletes resources linked to an external ID.

deregister_on_premises_instance(client, input, options \\ [])

Deregisters an on-premises instance.

get_application(client, input, options \\ [])

Gets information about an application.

get_application_revision(client, input, options \\ [])

Gets information about an application revision.

get_deployment(client, input, options \\ [])

Gets information about a deployment.

The content property of the appSpecContent object in the returned revision is always null. Use GetApplicationRevision and the sha256 property of the returned appSpecContent object to get the content of the deployment’s AppSpec file.

get_deployment_config(client, input, options \\ [])

Gets information about a deployment configuration.

get_deployment_group(client, input, options \\ [])

Gets information about a deployment group.

get_deployment_instance(client, input, options \\ [])

Gets information about an instance as part of a deployment.

get_deployment_target(client, input, options \\ [])

Returns information about a deployment target.

get_on_premises_instance(client, input, options \\ [])

Gets information about an on-premises instance.

list_application_revisions(client, input, options \\ [])

Lists information about revisions for an application.

list_applications(client, input, options \\ [])

Lists the applications registered with the IAM user or AWS account.

list_deployment_configs(client, input, options \\ [])

Lists the deployment configurations with the IAM user or AWS account.

list_deployment_groups(client, input, options \\ [])

Lists the deployment groups for an application registered with the IAM user or AWS account.

list_deployment_instances(client, input, options \\ [])

The newer BatchGetDeploymentTargets should be used instead because it works with all compute types.

ListDeploymentInstances throws an exception if it is used with a compute platform other than EC2/On-premises or AWS Lambda.

Lists the instance for a deployment associated with the IAM user or AWS account.

list_deployment_targets(client, input, options \\ [])

Returns an array of target IDs that are associated a deployment.

list_deployments(client, input, options \\ [])

Lists the deployments in a deployment group for an application registered with the IAM user or AWS account.

list_git_hub_account_token_names(client, input, options \\ [])

Lists the names of stored connections to GitHub accounts.

list_on_premises_instances(client, input, options \\ [])

Gets a list of names for one or more on-premises instances.

Unless otherwise specified, both registered and deregistered on-premises instance names are listed. To list only registered or deregistered on-premises instance names, use the registration status parameter.

list_tags_for_resource(client, input, options \\ [])

Returns a list of tags for the resource identified by a specified Amazon Resource Name (ARN).

Tags are used to organize and categorize your CodeDeploy resources.

put_lifecycle_event_hook_execution_status(client, input, options \\ [])

Sets the result of a Lambda validation function.

The function validates lifecycle hooks during a deployment that uses the AWS Lambda or Amazon ECS compute platform. For AWS Lambda deployments, the available lifecycle hooks are BeforeAllowTraffic and AfterAllowTraffic. For Amazon ECS deployments, the available lifecycle hooks are BeforeInstall, AfterInstall, AfterAllowTestTraffic, BeforeAllowTraffic, and AfterAllowTraffic. Lambda validation functions return Succeeded or Failed. For more information, see AppSpec 'hooks' Section for an AWS Lambda Deployment and AppSpec 'hooks' Section for an Amazon ECS Deployment.

register_application_revision(client, input, options \\ [])

Registers with AWS CodeDeploy a revision for the specified application.

register_on_premises_instance(client, input, options \\ [])

Registers an on-premises instance.

Only one IAM ARN (an IAM session ARN or IAM user ARN) is supported in the request. You cannot use both.

remove_tags_from_on_premises_instances(client, input, options \\ [])

Removes one or more tags from one or more on-premises instances.

skip_wait_time_for_instance_termination(client, input, options \\ [])

In a blue/green deployment, overrides any specified wait time and starts terminating instances immediately after the traffic routing is complete.

stop_deployment(client, input, options \\ [])

Attempts to stop an ongoing deployment.

tag_resource(client, input, options \\ [])

Associates the list of tags in the input Tags parameter with the resource identified by the ResourceArn input parameter.

untag_resource(client, input, options \\ [])

Disassociates a resource from a list of tags.

The resource is identified by the ResourceArn input parameter. The tags are identified by the list of keys in the TagKeys input parameter.

update_application(client, input, options \\ [])

Changes the name of an application.

update_deployment_group(client, input, options \\ [])

Changes information about a deployment group.