aws-elixir v0.6.0 AWS.CodeGuruProfiler

This section provides documentation for the Amazon CodeGuru Profiler API operations.

Amazon CodeGuru Profiler collects runtime performance data from your live applications, and provides recommendations that can help you fine-tune your application performance. Using machine learning algorithms, CodeGuru Profiler can help you find your most expensive lines of code and suggest ways you can improve efficiency and remove CPU bottlenecks. Amazon CodeGuru Profiler provides different visualizations of profiling data to help you identify what code is running on the CPU, see how much time is consumed, and suggest ways to reduce CPU utilization. Amazon CodeGuru Profiler currently supports applications written in all Java virtual machine (JVM) languages. While CodeGuru Profiler supports both visualizations and recommendations for applications written in Java, it can also generate visualizations and a subset of recommendations for applications written in other JVM languages. For more information, see [What is Amazon CodeGuru Profiler](https://docs.aws.amazon.com/codeguru/latest/profiler-ug/what-is-codeguru-profiler.html) in the *Amazon CodeGuru Profiler User Guide*.

Functions

add_notification_channels(client, profiling_group_name, input, options \\ [])

Add up to 2 anomaly notifications channels for a profiling group.

batch_get_frame_metric_data(client, profiling_group_name, input, options \\ [])

Returns the time series of values for a requested list of frame metrics from a time period.

configure_agent(client, profiling_group_name, input, options \\ [])

Used by profiler agents to report their current state and to receive remote configuration updates.

For example, ConfigureAgent can be used to tell and agent whether to profile or not and for how long to return profiling data.

create_profiling_group(client, input, options \\ [])

Creates a profiling group.

delete_profiling_group(client, profiling_group_name, input, options \\ [])

Deletes a profiling group.

describe_profiling_group(client, profiling_group_name, options \\ [])

Returns a ProfilingGroupDescription object that contains information about the requested profiling group.

get_findings_report_account_summary(client, daily_reports_only \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of FindingsReportSummary objects that contain analysis results for all profiling groups in your AWS account.

get_notification_configuration(client, profiling_group_name, options \\ [])

Get the current configuration for anomaly notifications for a profiling group.

get_policy(client, profiling_group_name, options \\ [])

Returns the JSON-formatted resource-based policy on a profiling group.

get_profile(client, profiling_group_name, end_time \\ nil, max_depth \\ nil, period \\ nil, start_time \\ nil, accept \\ nil, options \\ [])

Gets the aggregated profile of a profiling group for a specified time range.

Amazon CodeGuru Profiler collects posted agent profiles for a profiling group into aggregated profiles.

Because aggregated profiles expire over timeGetProfileis not idempotent. Specify the time range for the requested aggregated profile using 1 or 2 of the following parameters:startTime,endTime,period. The maximum time range allowed is 7 days. If you specify all 3 parameters, an exception is thrown. If you specify onlyperiod, the latest aggregated profile is returned. Aggregated profiles are available with aggregation periods of 5 minutes, 1 hour, and 1 day, aligned to UTC. The aggregation period of an aggregated profile determines how long it is retained. For more information, see [AggregatedProfileTime](https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_AggregatedProfileTime.html). The aggregated profile's aggregation period determines how long it is retained by CodeGuru Profiler. * If the aggregation period is 5 minutes, the aggregated profile is retained for 15 days. * If the aggregation period is 1 hour, the aggregated profile is retained for 60 days. * If the aggregation period is 1 day, the aggregated profile is retained for 3 years. There are two use cases for callingGetProfile. 1. If you want to return an aggregated profile that already exists, use [ListProfileTimes](https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_ListProfileTimes.html) to view the time ranges of existing aggregated profiles. Use them in aGetProfilerequest to return a specific, existing aggregated profile. 2. If you want to return an aggregated profile for a time range that doesn't align with an existing aggregated profile, then CodeGuru Profiler makes a best effort to combine existing aggregated profiles from the requested time range and return them as one aggregated profile. If aggregated profiles do not exist for the full time range requested, then aggregated profiles for a smaller time range are returned. For example, if the requested time range is from 00:00 to 00:20, and the existing aggregated profiles are from 00:15 and 00:25, then the aggregated profiles from 00:15 to 00:20 are returned.

get_recommendations(client, profiling_group_name, end_time, locale \\ nil, start_time, options \\ [])

Returns a list of Recommendation objects that contain recommendations for a profiling group for a given time period.

A list of Anomaly objects that contains details about anomalies detected in the profiling group for the same time period is also returned.

list_findings_reports(client, profiling_group_name, daily_reports_only \\ nil, end_time, max_results \\ nil, next_token \\ nil, start_time, options \\ [])

List the available reports for a given profiling group and time range.

list_profile_times(client, profiling_group_name, end_time, max_results \\ nil, next_token \\ nil, order_by \\ nil, period, start_time, options \\ [])

Lists the start times of the available aggregated profiles of a profiling group for an aggregation period within the specified time range.

list_profiling_groups(client, include_description \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of profiling groups.

The profiling groups are returned as ProfilingGroupDescription objects.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns a list of the tags that are assigned to a specified resource.

post_agent_profile(client, profiling_group_name, input, options \\ [])

Submits profiling data to an aggregated profile of a profiling group.

To get an aggregated profile that is created with this profiling data, use GetProfile .

put_permission(client, action_group, profiling_group_name, input, options \\ [])

Adds permissions to a profiling group's resource-based policy that are provided using an action group.

If a profiling group doesn't have a resource-based policy, one is created for it using the permissions in the action group and the roles and users in the principals parameter.

The one supported action group that can be added isagentPermissionwhich grantsConfigureAgentandPostAgentpermissions. For more information, see [Resource-based policies in CodeGuru Profiler](https://docs.aws.amazon.com/codeguru/latest/profiler-ug/resource-based-policies.html) in the *Amazon CodeGuru Profiler User Guide*, [ConfigureAgent](https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_ConfigureAgent.html), and [PostAgentProfile](https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_PostAgentProfile.html). The first time you callPutPermissionon a profiling group, do not specify arevisionIdbecause it doesn't have a resource-based policy. Subsequent calls must provide arevisionIdto specify which revision of the resource-based policy to add the permissions to. The response contains the profiling group's JSON-formatted resource policy.

remove_notification_channel(client, channel_id, profiling_group_name, input, options \\ [])

Remove one anomaly notifications channel for a profiling group.

remove_permission(client, action_group, profiling_group_name, input, options \\ [])

Removes permissions from a profiling group's resource-based policy that are provided using an action group.

The one supported action group that can be removed is agentPermission which grants ConfigureAgent and PostAgent permissions. For more information, see Resource-based policies in CodeGuru Profiler in the Amazon CodeGuru Profiler User Guide, ConfigureAgent , and PostAgentProfile .

submit_feedback(client, anomaly_instance_id, profiling_group_name, input, options \\ [])

Sends feedback to CodeGuru Profiler about whether the anomaly detected by the analysis is useful or not.

tag_resource(client, resource_arn, input, options \\ [])

Use to assign one or more tags to a resource.

untag_resource(client, resource_arn, input, options \\ [])

Use to remove one or more tags from a resource.

update_profiling_group(client, profiling_group_name, input, options \\ [])

Updates a profiling group.