aws-elixir v0.6.0 AWS.CodeGuruReviewer

This section provides documentation for the Amazon CodeGuru Reviewer API operations.

CodeGuru Reviewer is a service that uses program analysis and machine learning to detect potential defects that are difficult for developers to find and recommends fixes in your Java code.

By proactively detecting and providing recommendations for addressing code defects and implementing best practices, CodeGuru Reviewer improves the overall quality and maintainability of your code base during the code review stage. For more information about CodeGuru Reviewer, see the

Amazon CodeGuru Reviewer User Guide.*

Functions

associate_repository(client, input, options \\ [])

Use to associate an AWS CodeCommit repository or a repostory managed by AWS CodeStar Connections with Amazon CodeGuru Reviewer.

When you associate a repository, CodeGuru Reviewer reviews source code changes in the repository's pull requests and provides automatic recommendations. You can view recommendations using the CodeGuru Reviewer console. For more information, see Recommendations in Amazon CodeGuru Reviewer in the Amazon CodeGuru Reviewer User Guide.

If you associate a CodeCommit repository, it must be in the same AWS Region and AWS account where its CodeGuru Reviewer code reviews are configured.

Bitbucket and GitHub Enterprise Server repositories are managed by AWS CodeStar Connections to connect to CodeGuru Reviewer. For more information, see Connect to a repository source provider in the Amazon CodeGuru Reviewer User Guide.

You cannot use the CodeGuru Reviewer SDK or the AWS CLI to associate a GitHub repository with Amazon CodeGuru Reviewer. To associate a GitHub repository, use the console. For more information, see Getting started with CodeGuru Reviewer in the CodeGuru Reviewer User Guide.

create_code_review(client, input, options \\ [])

Use to create a code review for a repository analysis.

describe_code_review(client, code_review_arn, options \\ [])

Returns the metadata associated with the code review along with its status.

describe_recommendation_feedback(client, code_review_arn, recommendation_id, user_id \\ nil, options \\ [])

Describes the customer feedback for a CodeGuru Reviewer recommendation.

describe_repository_association(client, association_arn, options \\ [])

Returns a RepositoryAssociation object that contains information about the requested repository association.

disassociate_repository(client, association_arn, input, options \\ [])

Removes the association between Amazon CodeGuru Reviewer and a repository.

list_code_reviews(client, max_results \\ nil, next_token \\ nil, provider_types \\ nil, repository_names \\ nil, states \\ nil, type, options \\ [])

Lists all the code reviews that the customer has created in the past 90 days.

list_recommendation_feedback(client, code_review_arn, max_results \\ nil, next_token \\ nil, recommendation_ids \\ nil, user_ids \\ nil, options \\ [])

Returns a list of RecommendationFeedbackSummary objects that contain customer recommendation feedback for all CodeGuru Reviewer users.

list_recommendations(client, code_review_arn, max_results \\ nil, next_token \\ nil, options \\ [])

Returns the list of all recommendations for a completed code review.

list_repository_associations(client, max_results \\ nil, names \\ nil, next_token \\ nil, owners \\ nil, provider_types \\ nil, states \\ nil, options \\ [])

Returns a list of RepositoryAssociationSummary objects that contain summary information about a repository association.

You can filter the returned list by ProviderType , Name , State , and Owner .

put_recommendation_feedback(client, input, options \\ [])

Stores customer feedback for a CodeGuru Reviewer recommendation.

When this API is called again with different reactions the previous feedback is overwritten.