aws-elixir v0.6.0 AWS.CodePipeline

AWS CodePipeline

Overview

This is the AWS CodePipeline API Reference.

This guide provides descriptions of the actions and data types for AWS CodePipeline. Some functionality for your pipeline can only be configured through the API. For more information, see the AWS CodePipeline User Guide.

You can use the AWS CodePipeline API to work with pipelines, stages, actions, and transitions.

Pipelines are models of automated release processes. Each pipeline is uniquely named, and consists of stages, actions, and transitions.

You can work with pipelines by calling:

CreatePipeline, which creates a uniquely named pipeline.

DeletePipeline, which deletes the specified pipeline.

GetPipeline, which returns information about the pipeline structure and pipeline metadata, including the pipeline Amazon Resource Name (ARN).

GetPipelineExecution, which returns information about a specific execution of a pipeline.

GetPipelineState, which returns information about the current state of the stages and actions of a pipeline.

ListActionExecutions, which returns action-level details for past executions. The details include full stage and action-level details, including individual action duration, status, any errors that occurred during the execution, and input and output artifact location details.

ListPipelines, which gets a summary of all of the pipelines associated with your account.

ListPipelineExecutions, which gets a summary of the most recent executions for a pipeline.

StartPipelineExecution, which runs the most recent revision of an artifact through the pipeline.

StopPipelineExecution, which stops the specified pipeline execution from continuing through the pipeline.

UpdatePipeline, which updates a pipeline with edits or changes to the structure of the pipeline.

Pipelines include stages. Each stage contains one or more actions that must complete before the next stage begins. A stage results in success or failure. If a stage fails, the pipeline stops at that stage and remains stopped until either a new version of an artifact appears in the source location, or a user takes action to rerun the most recent artifact through the pipeline. You can call GetPipelineState, which displays the status of a pipeline, including the status of stages in the pipeline, or GetPipeline, which returns the entire structure of the pipeline, including the stages of that pipeline. For more information about the structure of stages and actions, see AWS CodePipeline Pipeline Structure Reference.

Pipeline stages include actions that are categorized into categories such as source or build actions performed in a stage of a pipeline. For example, you can use a source action to import artifacts into a pipeline from a source such as Amazon S3. Like stages, you do not work with actions directly in most cases, but you do define and interact with actions when working with pipeline operations such as CreatePipeline and GetPipelineState. Valid action categories are:

Source

Build

Test

Deploy

Approval

Invoke

Pipelines also include transitions, which allow the transition of artifacts from one stage to the next in a pipeline after the actions in one stage complete.

You can work with transitions by calling:

DisableStageTransition, which prevents artifacts from transitioning to the next stage in a pipeline.

EnableStageTransition, which enables transition of artifacts between stages in a pipeline.

Using the API to integrate with AWS CodePipeline

For third-party integrators or developers who want to create their own integrations with AWS CodePipeline, the expected sequence varies from the standard API user. To integrate with AWS CodePipeline, developers need to work with the following items:

Jobs, which are instances of an action. For example, a job for a source action might import a revision of an artifact from a source.

You can work with jobs by calling:

AcknowledgeJob, which confirms whether a job worker has received the specified job.

GetJobDetails, which returns the details of a job.

PollForJobs, which determines whether there are any jobs to act on.

PutJobFailureResult, which provides details of a job failure.

PutJobSuccessResult, which provides details of a job success.

Third party jobs, which are instances of an action created by a partner action and integrated into AWS CodePipeline. Partner actions are created by members of the AWS Partner Network.

You can work with third party jobs by calling:

AcknowledgeThirdPartyJob, which confirms whether a job worker has received the specified job.

GetThirdPartyJobDetails, which requests the details of a job for a partner action.

PollForThirdPartyJobs, which determines whether there are any jobs to act on.

PutThirdPartyJobFailureResult, which provides details of a job failure.

PutThirdPartyJobSuccessResult, which provides details of a job success.

Functions

acknowledge_job(client, input, options \\ [])

Returns information about a specified job and whether that job has been received by the job worker.

Used for custom actions only.

acknowledge_third_party_job(client, input, options \\ [])

Confirms a job worker has received the specified job.

Used for partner actions only.

create_custom_action_type(client, input, options \\ [])

Creates a new custom action that can be used in all pipelines associated with the AWS account.

Only used for custom actions.

create_pipeline(client, input, options \\ [])

Creates a pipeline.

In the pipeline structure, you must include either artifactStore or artifactStores in your pipeline, but you cannot use both. If you create a cross-region action in your pipeline, you must use artifactStores.

delete_custom_action_type(client, input, options \\ [])

Marks a custom action as deleted.

PollForJobs for the custom action fails after the action is marked for deletion. Used for custom actions only.

To re-create a custom action after it has been deleted you must use a string in the version field that has never been used before. This string can be an incremented version number, for example. To restore a deleted custom action, use a JSON file that is identical to the deleted action, including the original string in the version field.

delete_pipeline(client, input, options \\ [])

Deletes the specified pipeline.

delete_webhook(client, input, options \\ [])

Deletes a previously created webhook by name.

Deleting the webhook stops AWS CodePipeline from starting a pipeline every time an external event occurs. The API returns successfully when trying to delete a webhook that is already deleted. If a deleted webhook is re-created by calling PutWebhook with the same name, it will have a different URL.

deregister_webhook_with_third_party(client, input, options \\ [])

Removes the connection between the webhook that was created by CodePipeline and the external tool with events to be detected.

Currently supported only for webhooks that target an action type of GitHub.

disable_stage_transition(client, input, options \\ [])

Prevents artifacts in a pipeline from transitioning to the next stage in the pipeline.

enable_stage_transition(client, input, options \\ [])

Enables artifacts in a pipeline to transition to a stage in a pipeline.

get_job_details(client, input, options \\ [])

Returns information about a job.

Used for custom actions only.

When this API is called, AWS CodePipeline returns temporary credentials for the S3 bucket used to store artifacts for the pipeline, if the action requires access to that S3 bucket for input or output artifacts. This API also returns any secret values defined for the action.

get_pipeline(client, input, options \\ [])

Returns the metadata, structure, stages, and actions of a pipeline.

Can be used to return the entire structure of a pipeline in JSON format, which can then be modified and used to update the pipeline structure with UpdatePipeline.

get_pipeline_execution(client, input, options \\ [])

Returns information about an execution of a pipeline, including details about artifacts, the pipeline execution ID, and the name, version, and status of the pipeline.

get_pipeline_state(client, input, options \\ [])

Returns information about the state of a pipeline, including the stages and actions.

Values returned in the revisionId and revisionUrl fields indicate the source revision information, such as the commit ID, for the current state.

get_third_party_job_details(client, input, options \\ [])

Requests the details of a job for a third party action.

Used for partner actions only.

When this API is called, AWS CodePipeline returns temporary credentials for the S3 bucket used to store artifacts for the pipeline, if the action requires access to that S3 bucket for input or output artifacts. This API also returns any secret values defined for the action.

list_action_executions(client, input, options \\ [])

Lists the action executions that have occurred in a pipeline.

list_action_types(client, input, options \\ [])

Gets a summary of all AWS CodePipeline action types associated with your account.

list_pipeline_executions(client, input, options \\ [])

Gets a summary of the most recent executions for a pipeline.

list_pipelines(client, input, options \\ [])

Gets a summary of all of the pipelines associated with your account.

list_tags_for_resource(client, input, options \\ [])

Gets the set of key-value pairs (metadata) that are used to manage the resource.

list_webhooks(client, input, options \\ [])

Gets a listing of all the webhooks in this AWS Region for this account.

The output lists all webhooks and includes the webhook URL and ARN and the configuration for each webhook.

poll_for_jobs(client, input, options \\ [])

Returns information about any jobs for AWS CodePipeline to act on.

PollForJobs is valid only for action types with "Custom" in the owner field. If the action type contains "AWS" or "ThirdParty" in the owner field, the PollForJobs action returns an error.

When this API is called, AWS CodePipeline returns temporary credentials for the S3 bucket used to store artifacts for the pipeline, if the action requires access to that S3 bucket for input or output artifacts. This API also returns any secret values defined for the action.

poll_for_third_party_jobs(client, input, options \\ [])

Determines whether there are any third party jobs for a job worker to act on.

Used for partner actions only.

When this API is called, AWS CodePipeline returns temporary credentials for the S3 bucket used to store artifacts for the pipeline, if the action requires access to that S3 bucket for input or output artifacts.

put_action_revision(client, input, options \\ [])

Provides information to AWS CodePipeline about new revisions to a source.

put_approval_result(client, input, options \\ [])

Provides the response to a manual approval request to AWS CodePipeline.

Valid responses include Approved and Rejected.

put_job_failure_result(client, input, options \\ [])

Represents the failure of a job as returned to the pipeline by a job worker.

Used for custom actions only.

put_job_success_result(client, input, options \\ [])

Represents the success of a job as returned to the pipeline by a job worker.

Used for custom actions only.

put_third_party_job_failure_result(client, input, options \\ [])

Represents the failure of a third party job as returned to the pipeline by a job worker.

Used for partner actions only.

put_third_party_job_success_result(client, input, options \\ [])

Represents the success of a third party job as returned to the pipeline by a job worker.

Used for partner actions only.

put_webhook(client, input, options \\ [])

Defines a webhook and returns a unique webhook URL generated by CodePipeline.

This URL can be supplied to third party source hosting providers to call every time there's a code change. When CodePipeline receives a POST request on this URL, the pipeline defined in the webhook is started as long as the POST request satisfied the authentication and filtering requirements supplied when defining the webhook. RegisterWebhookWithThirdParty and DeregisterWebhookWithThirdParty APIs can be used to automatically configure supported third parties to call the generated webhook URL.

register_webhook_with_third_party(client, input, options \\ [])

Configures a connection between the webhook that was created and the external tool with events to be detected.

retry_stage_execution(client, input, options \\ [])

Resumes the pipeline execution by retrying the last failed actions in a stage.

You can retry a stage immediately if any of the actions in the stage fail. When you retry, all actions that are still in progress continue working, and failed actions are triggered again.

start_pipeline_execution(client, input, options \\ [])

Starts the specified pipeline.

Specifically, it begins processing the latest commit to the source location specified as part of the pipeline.

stop_pipeline_execution(client, input, options \\ [])

Stops the specified pipeline execution.

You choose to either stop the pipeline execution by completing in-progress actions without starting subsequent actions, or by abandoning in-progress actions. While completing or abandoning in-progress actions, the pipeline execution is in a Stopping state. After all in-progress actions are completed or abandoned, the pipeline execution is in a Stopped state.

tag_resource(client, input, options \\ [])

Adds to or modifies the tags of the given resource.

Tags are metadata that can be used to manage a resource.

untag_resource(client, input, options \\ [])

Removes tags from an AWS resource.

update_pipeline(client, input, options \\ [])

Updates a specified pipeline with edits or changes to its structure.

Use a JSON file with the pipeline structure and UpdatePipeline to provide the full structure of the pipeline. Updating the pipeline increases the version number of the pipeline by 1.