aws-elixir v0.6.0 AWS.CodeStar

AWS CodeStar

This is the API reference for AWS CodeStar.

This reference provides descriptions of the operations and data types for the AWS CodeStar API along with usage examples.

You can use the AWS CodeStar API to work with:

Projects and their resources, by calling the following:

DeleteProject, which deletes a project.

DescribeProject, which lists the attributes of a project.

ListProjects, which lists all projects associated with your AWS account.

ListResources, which lists the resources associated with a project.

ListTagsForProject, which lists the tags associated with a project.

TagProject, which adds tags to a project.

UntagProject, which removes tags from a project.

UpdateProject, which updates the attributes of a project.

Teams and team members, by calling the following:

AssociateTeamMember, which adds an IAM user to the team for a project.

DisassociateTeamMember, which removes an IAM user from the team for a project.

ListTeamMembers, which lists all the IAM users in the team for a project, including their roles and attributes.

UpdateTeamMember, which updates a team member's attributes in a project.

Users, by calling the following:

CreateUserProfile, which creates a user profile that contains data associated with the user across all projects.

DeleteUserProfile, which deletes all user profile information across all projects.

DescribeUserProfile, which describes the profile of a user.

ListUserProfiles, which lists all user profiles.

UpdateUserProfile, which updates the profile for a user.

Functions

associate_team_member(client, input, options \\ [])

Adds an IAM user to the team for an AWS CodeStar project.

create_project(client, input, options \\ [])

Creates a project, including project resources.

This action creates a project based on a submitted project request. A set of source code files and a toolchain template file can be included with the project request. If these are not provided, an empty project is created.

create_user_profile(client, input, options \\ [])

Creates a profile for a user that includes user preferences, such as the display name and email address assocciated with the user, in AWS CodeStar.

The user profile is not project-specific. Information in the user profile is displayed wherever the user's information appears to other users in AWS CodeStar.

delete_project(client, input, options \\ [])

Deletes a project, including project resources.

Does not delete users associated with the project, but does delete the IAM roles that allowed access to the project.

delete_user_profile(client, input, options \\ [])

Deletes a user profile in AWS CodeStar, including all personal preference data associated with that profile, such as display name and email address.

It does not delete the history of that user, for example the history of commits made by that user.

describe_project(client, input, options \\ [])

Describes a project and its resources.

describe_user_profile(client, input, options \\ [])

Describes a user in AWS CodeStar and the user attributes across all projects.

disassociate_team_member(client, input, options \\ [])

Removes a user from a project.

Removing a user from a project also removes the IAM policies from that user that allowed access to the project and its resources. Disassociating a team member does not remove that user's profile from AWS CodeStar. It does not remove the user from IAM.

list_projects(client, input, options \\ [])

Lists all projects in AWS CodeStar associated with your AWS account.

list_resources(client, input, options \\ [])

Lists resources associated with a project in AWS CodeStar.

list_tags_for_project(client, input, options \\ [])

Gets the tags for a project.

list_team_members(client, input, options \\ [])

Lists all team members associated with a project.

list_user_profiles(client, input, options \\ [])

Lists all the user profiles configured for your AWS account in AWS CodeStar.

tag_project(client, input, options \\ [])

Adds tags to a project.

untag_project(client, input, options \\ [])

Removes tags from a project.

update_project(client, input, options \\ [])

Updates a project in AWS CodeStar.

update_team_member(client, input, options \\ [])

Updates a team member's attributes in an AWS CodeStar project.

For example, you can change a team member's role in the project, or change whether they have remote access to project resources.

update_user_profile(client, input, options \\ [])

Updates a user's profile in AWS CodeStar.

The user profile is not project-specific. Information in the user profile is displayed wherever the user's information appears to other users in AWS CodeStar.