aws-elixir v0.6.0 AWS.Codestarnotifications

This AWS CodeStar Notifications API Reference provides descriptions and usage examples of the operations and data types for the AWS CodeStar Notifications API.

You can use the AWS CodeStar Notifications API to work with the following objects:

Notification rules, by calling the following:

CreateNotificationRule, which creates a notification rule for a resource in your account.

DeleteNotificationRule, which deletes a notification rule.

DescribeNotificationRule, which provides information about a notification rule.

ListNotificationRules, which lists the notification rules associated with your account.

UpdateNotificationRule, which changes the name, events, or targets associated with a notification rule.

Subscribe, which subscribes a target to a notification rule.

Unsubscribe, which removes a target from a notification rule.

Targets, by calling the following:

DeleteTarget, which removes a notification rule target (SNS topic) from a notification rule.

ListTargets, which lists the targets associated with a notification rule.

Events, by calling the following:

ListEventTypes, which lists the event types you can include in a notification rule.

Tags, by calling the following:

ListTagsForResource, which lists the tags already associated with a notification rule in your account.

TagResource, which associates a tag you provide with a notification rule in your account.

UntagResource, which removes a tag from a notification rule in your account.

For information about how to use AWS CodeStar Notifications, see link in the CodeStarNotifications User Guide.

Functions

create_notification_rule(client, input, options \\ [])

Creates a notification rule for a resource.

The rule specifies the events you want notifications about and the targets (such as SNS topics) where you want to receive them.

delete_notification_rule(client, input, options \\ [])

Deletes a notification rule for a resource.

delete_target(client, input, options \\ [])

Deletes a specified target for notifications.

describe_notification_rule(client, input, options \\ [])

Returns information about a specified notification rule.

list_event_types(client, input, options \\ [])

Returns information about the event types available for configuring notifications.

list_notification_rules(client, input, options \\ [])

Returns a list of the notification rules for an AWS account.

list_tags_for_resource(client, input, options \\ [])

Returns a list of the tags associated with a notification rule.

list_targets(client, input, options \\ [])

Returns a list of the notification rule targets for an AWS account.

subscribe(client, input, options \\ [])

Creates an association between a notification rule and an SNS topic so that the associated target can receive notifications when the events described in the rule are triggered.

tag_resource(client, input, options \\ [])

Associates a set of provided tags with a notification rule.

unsubscribe(client, input, options \\ [])

Removes an association between a notification rule and an Amazon SNS topic so that subscribers to that topic stop receiving notifications when the events described in the rule are triggered.

untag_resource(client, input, options \\ [])

Removes the association between one or more provided tags and a notification rule.

update_notification_rule(client, input, options \\ [])

Updates a notification rule for a resource.

You can change the events that trigger the notification rule, the status of the rule, and the targets that receive the notifications.

To add or remove tags for a notification rule, you must use TagResource and UntagResource.