aws-elixir v0.6.0 AWS.CognitoIdentityProvider

Using the Amazon Cognito User Pools API, you can create a user pool to manage directories and users.

You can authenticate a user to obtain tokens related to user identity and access policies.

This API reference provides information about user pools in Amazon Cognito User Pools.

For more information, see the Amazon Cognito Documentation.

Functions

add_custom_attributes(client, input, options \\ [])

Adds additional user attributes to the user pool schema.

admin_add_user_to_group(client, input, options \\ [])

Adds the specified user to the specified group.

Calling this action requires developer credentials.

admin_confirm_sign_up(client, input, options \\ [])

Confirms user registration as an admin without using a confirmation code.

Works on any user.

Calling this action requires developer credentials.

admin_create_user(client, input, options \\ [])

Creates a new user in the specified user pool.

If MessageAction is not set, the default is to send a welcome message via email or phone (SMS).

This message is based on a template that you configured in your call to create or update a user pool. This template includes your custom sign-up instructions and placeholders for user name and temporary password.

Alternatively, you can call AdminCreateUser with “SUPPRESS” for the MessageAction parameter, and Amazon Cognito will not send any email.

In either case, the user will be in the FORCE_CHANGE_PASSWORD state until they sign in and change their password.

AdminCreateUser requires developer credentials.

admin_delete_user(client, input, options \\ [])

Deletes a user as an administrator.

Works on any user.

Calling this action requires developer credentials.

admin_delete_user_attributes(client, input, options \\ [])

Deletes the user attributes in a user pool as an administrator.

Works on any user.

Calling this action requires developer credentials.

admin_disable_provider_for_user(client, input, options \\ [])

Disables the user from signing in with the specified external (SAML or social) identity provider.

If the user to disable is a Cognito User Pools native username + password user, they are not permitted to use their password to sign-in. If the user to disable is a linked external IdP user, any link between that user and an existing user is removed. The next time the external user (no longer attached to the previously linked DestinationUser) signs in, they must create a new user account. See AdminLinkProviderForUser.

This action is enabled only for admin access and requires developer credentials.

The ProviderName must match the value specified when creating an IdP for the pool.

To disable a native username + password user, the ProviderName value must be Cognito and the ProviderAttributeName must be Cognito_Subject, with the ProviderAttributeValue being the name that is used in the user pool for the user.

The ProviderAttributeName must always be Cognito_Subject for social identity providers. The ProviderAttributeValue must always be the exact subject that was used when the user was originally linked as a source user.

For de-linking a SAML identity, there are two scenarios. If the linked identity has not yet been used to sign-in, the ProviderAttributeName and ProviderAttributeValue must be the same values that were used for the SourceUser when the identities were originally linked using AdminLinkProviderForUser call. (If the linking was done with ProviderAttributeName set to Cognito_Subject, the same applies here). However, if the user has already signed in, the ProviderAttributeName must be Cognito_Subject and ProviderAttributeValue must be the subject of the SAML assertion.

admin_disable_user(client, input, options \\ [])

Disables the specified user.

Calling this action requires developer credentials.

admin_enable_user(client, input, options \\ [])

Enables the specified user as an administrator.

Works on any user.

Calling this action requires developer credentials.

admin_forget_device(client, input, options \\ [])

Forgets the device, as an administrator.

Calling this action requires developer credentials.

admin_get_device(client, input, options \\ [])

Gets the device, as an administrator.

Calling this action requires developer credentials.

admin_get_user(client, input, options \\ [])

Gets the specified user by user name in a user pool as an administrator.

Works on any user.

Calling this action requires developer credentials.

admin_initiate_auth(client, input, options \\ [])

Initiates the authentication flow, as an administrator.

Calling this action requires developer credentials.

admin_link_provider_for_user(client, input, options \\ [])

Links an existing user account in a user pool (DestinationUser) to an identity from an external identity provider (SourceUser) based on a specified attribute name and value from the external identity provider.

This allows you to create a link from the existing user account to an external federated user identity that has not yet been used to sign in, so that the federated user identity can be used to sign in as the existing user account.

For example, if there is an existing user with a username and password, this API links that user to a federated user identity, so that when the federated user identity is used, the user signs in as the existing user account.

The maximum number of federated identities linked to a user is 5.

Because this API allows a user with an external federated identity to sign in as an existing user in the user pool, it is critical that it only be used with external identity providers and provider attributes that have been trusted by the application owner.

This action is enabled only for admin access and requires developer credentials.

admin_list_devices(client, input, options \\ [])

Lists devices, as an administrator.

Calling this action requires developer credentials.

admin_list_groups_for_user(client, input, options \\ [])

Lists the groups that the user belongs to.

Calling this action requires developer credentials.

admin_list_user_auth_events(client, input, options \\ [])

Lists a history of user activity and any risks detected as part of Amazon Cognito advanced security.

admin_remove_user_from_group(client, input, options \\ [])

Removes the specified user from the specified group.

Calling this action requires developer credentials.

admin_reset_user_password(client, input, options \\ [])

Resets the specified user's password in a user pool as an administrator.

Works on any user.

When a developer calls this API, the current password is invalidated, so it must be changed. If a user tries to sign in after the API is called, the app will get a PasswordResetRequiredException exception back and should direct the user down the flow to reset the password, which is the same as the forgot password flow. In addition, if the user pool has phone verification selected and a verified phone number exists for the user, or if email verification is selected and a verified email exists for the user, calling this API will also result in sending a message to the end user with the code to change their password.

Calling this action requires developer credentials.

admin_respond_to_auth_challenge(client, input, options \\ [])

Responds to an authentication challenge, as an administrator.

Calling this action requires developer credentials.

admin_set_user_m_f_a_preference(client, input, options \\ [])

Sets the user's multi-factor authentication (MFA) preference, including which MFA options are enabled and if any are preferred.

Only one factor can be set as preferred. The preferred MFA factor will be used to authenticate a user if multiple factors are enabled. If multiple options are enabled and no preference is set, a challenge to choose an MFA option will be returned during sign in.

admin_set_user_password(client, input, options \\ [])

Sets the specified user's password in a user pool as an administrator.

Works on any user.

The password can be temporary or permanent. If it is temporary, the user status will be placed into the FORCE_CHANGE_PASSWORD state. When the user next tries to sign in, the InitiateAuth/AdminInitiateAuth response will contain the NEW_PASSWORD_REQUIRED challenge. If the user does not sign in before it expires, the user will not be able to sign in and their password will need to be reset by an administrator.

Once the user has set a new password, or the password is permanent, the user status will be set to Confirmed.

admin_set_user_settings(client, input, options \\ [])

This action is no longer supported. You can use it to configure only SMS MFA.

You can't use it to configure TOTP software token MFA. To configure either type of MFA, use AdminSetUserMFAPreference instead.

admin_update_auth_event_feedback(client, input, options \\ [])

Provides feedback for an authentication event as to whether it was from a valid user.

This feedback is used for improving the risk evaluation decision for the user pool as part of Amazon Cognito advanced security.

admin_update_device_status(client, input, options \\ [])

Updates the device status as an administrator.

Calling this action requires developer credentials.

admin_update_user_attributes(client, input, options \\ [])

Updates the specified user's attributes, including developer attributes, as an administrator.

Works on any user.

For custom attributes, you must prepend the custom: prefix to the attribute name.

In addition to updating user attributes, this API can also be used to mark phone and email as verified.

Calling this action requires developer credentials.

admin_user_global_sign_out(client, input, options \\ [])

Signs out users from all devices, as an administrator.

It also invalidates all refresh tokens issued to a user. The user's current access and Id tokens remain valid until their expiry. Access and Id tokens expire one hour after they are issued.

Calling this action requires developer credentials.

associate_software_token(client, input, options \\ [])

Returns a unique generated shared secret key code for the user account.

The request takes an access token or a session string, but not both.

change_password(client, input, options \\ [])

Changes the password for a specified user in a user pool.

confirm_device(client, input, options \\ [])

Confirms tracking of the device.

This API call is the call that begins device tracking.

confirm_forgot_password(client, input, options \\ [])

Allows a user to enter a confirmation code to reset a forgotten password.

confirm_sign_up(client, input, options \\ [])

Confirms registration of a user and handles the existing alias from a previous user.

create_group(client, input, options \\ [])

Creates a new group in the specified user pool.

Calling this action requires developer credentials.

create_identity_provider(client, input, options \\ [])

Creates an identity provider for a user pool.

create_resource_server(client, input, options \\ [])

Creates a new OAuth2.0 resource server and defines custom scopes in it.

create_user_import_job(client, input, options \\ [])

Creates the user import job.

create_user_pool(client, input, options \\ [])

Creates a new Amazon Cognito user pool and sets the password policy for the pool.

create_user_pool_client(client, input, options \\ [])

Creates the user pool client.

create_user_pool_domain(client, input, options \\ [])

Creates a new domain for a user pool.

delete_group(client, input, options \\ [])

Deletes a group.

Currently only groups with no members can be deleted.

Calling this action requires developer credentials.

delete_identity_provider(client, input, options \\ [])

Deletes an identity provider for a user pool.

delete_resource_server(client, input, options \\ [])

Deletes a resource server.

delete_user(client, input, options \\ [])

Allows a user to delete himself or herself.

delete_user_attributes(client, input, options \\ [])

Deletes the attributes for a user.

delete_user_pool(client, input, options \\ [])

Deletes the specified Amazon Cognito user pool.

delete_user_pool_client(client, input, options \\ [])

Allows the developer to delete the user pool client.

delete_user_pool_domain(client, input, options \\ [])

Deletes a domain for a user pool.

describe_identity_provider(client, input, options \\ [])

Gets information about a specific identity provider.

describe_resource_server(client, input, options \\ [])

Describes a resource server.

describe_risk_configuration(client, input, options \\ [])

Describes the risk configuration.

describe_user_import_job(client, input, options \\ [])

Describes the user import job.

describe_user_pool(client, input, options \\ [])

Returns the configuration information and metadata of the specified user pool.

describe_user_pool_client(client, input, options \\ [])

Client method for returning the configuration information and metadata of the specified user pool app client.

describe_user_pool_domain(client, input, options \\ [])

Gets information about a domain.

forget_device(client, input, options \\ [])

Forgets the specified device.

forgot_password(client, input, options \\ [])

Calling this API causes a message to be sent to the end user with a confirmation code that is required to change the user's password.

For the Username parameter, you can use the username or user alias. The method used to send the confirmation code is sent according to the specified AccountRecoverySetting. For more information, see Recovering User Accounts in the Amazon Cognito Developer Guide. If neither a verified phone number nor a verified email exists, an InvalidParameterException is thrown. To use the confirmation code for resetting the password, call ConfirmForgotPassword.

get_csv_header(client, input, options \\ [])

Gets the header information for the .csv file to be used as input for the user import job.

get_device(client, input, options \\ [])

Gets the device.

get_group(client, input, options \\ [])

Gets a group.

Calling this action requires developer credentials.

get_identity_provider_by_identifier(client, input, options \\ [])

Gets the specified identity provider.

get_signing_certificate(client, input, options \\ [])

This method takes a user pool ID, and returns the signing certificate.

get_u_i_customization(client, input, options \\ [])

Gets the UI Customization information for a particular app client's app UI, if there is something set.

If nothing is set for the particular client, but there is an existing pool level customization (app clientId will be ALL), then that is returned. If nothing is present, then an empty shape is returned.

get_user(client, input, options \\ [])

Gets the user attributes and metadata for a user.

get_user_attribute_verification_code(client, input, options \\ [])

Gets the user attribute verification code for the specified attribute name.

get_user_pool_mfa_config(client, input, options \\ [])

Gets the user pool multi-factor authentication (MFA) configuration.

global_sign_out(client, input, options \\ [])

Signs out users from all devices.

It also invalidates all refresh tokens issued to a user. The user's current access and Id tokens remain valid until their expiry. Access and Id tokens expire one hour after they are issued.

initiate_auth(client, input, options \\ [])

Initiates the authentication flow.

list_devices(client, input, options \\ [])

Lists the devices.

list_groups(client, input, options \\ [])

Lists the groups associated with a user pool.

Calling this action requires developer credentials.

list_identity_providers(client, input, options \\ [])

Lists information about all identity providers for a user pool.

list_resource_servers(client, input, options \\ [])

Lists the resource servers for a user pool.

list_tags_for_resource(client, input, options \\ [])

Lists the tags that are assigned to an Amazon Cognito user pool.

A tag is a label that you can apply to user pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

You can use this action up to 10 times per second, per account.

list_user_import_jobs(client, input, options \\ [])

Lists the user import jobs.

list_user_pool_clients(client, input, options \\ [])

Lists the clients that have been created for the specified user pool.

list_user_pools(client, input, options \\ [])

Lists the user pools associated with an AWS account.

list_users(client, input, options \\ [])

Lists the users in the Amazon Cognito user pool.

list_users_in_group(client, input, options \\ [])

Lists the users in the specified group.

Calling this action requires developer credentials.

resend_confirmation_code(client, input, options \\ [])

Resends the confirmation (for confirmation of registration) to a specific user in the user pool.

respond_to_auth_challenge(client, input, options \\ [])

Responds to the authentication challenge.

set_risk_configuration(client, input, options \\ [])

Configures actions on detected risks.

To delete the risk configuration for UserPoolId or ClientId, pass null values for all four configuration types.

To enable Amazon Cognito advanced security features, update the user pool to include the UserPoolAddOns keyAdvancedSecurityMode.

set_u_i_customization(client, input, options \\ [])

Sets the UI customization information for a user pool's built-in app UI.

You can specify app UI customization settings for a single client (with a specific clientId) or for all clients (by setting the clientId to ALL). If you specify ALL, the default configuration will be used for every client that has no UI customization set previously. If you specify UI customization settings for a particular client, it will no longer fall back to the ALL configuration.

To use this API, your user pool must have a domain associated with it. Otherwise, there is no place to host the app's pages, and the service will throw an error.

set_user_m_f_a_preference(client, input, options \\ [])

Set the user's multi-factor authentication (MFA) method preference, including which MFA factors are enabled and if any are preferred.

Only one factor can be set as preferred. The preferred MFA factor will be used to authenticate a user if multiple factors are enabled. If multiple options are enabled and no preference is set, a challenge to choose an MFA option will be returned during sign in.

set_user_pool_mfa_config(client, input, options \\ [])

Set the user pool multi-factor authentication (MFA) configuration.

set_user_settings(client, input, options \\ [])

This action is no longer supported. You can use it to configure only SMS MFA.

You can't use it to configure TOTP software token MFA. To configure either type of MFA, use SetUserMFAPreference instead.

sign_up(client, input, options \\ [])

Registers the user in the specified user pool and creates a user name, password, and user attributes.

start_user_import_job(client, input, options \\ [])

Starts the user import.

stop_user_import_job(client, input, options \\ [])

Stops the user import job.

tag_resource(client, input, options \\ [])

Assigns a set of tags to an Amazon Cognito user pool.

A tag is a label that you can use to categorize and manage user pools in different ways, such as by purpose, owner, environment, or other criteria.

Each tag consists of a key and value, both of which you define. A key is a general category for more specific values. For example, if you have two versions of a user pool, one for testing and another for production, you might assign an Environment tag key to both user pools. The value of this key might be Test for one user pool and Production for the other.

Tags are useful for cost tracking and access control. You can activate your tags so that they appear on the Billing and Cost Management console, where you can track the costs associated with your user pools. In an IAM policy, you can constrain permissions for user pools based on specific tags or tag values.

You can use this action up to 5 times per second, per account. A user pool can have as many as 50 tags.

untag_resource(client, input, options \\ [])

Removes the specified tags from an Amazon Cognito user pool.

You can use this action up to 5 times per second, per account

update_auth_event_feedback(client, input, options \\ [])

Provides the feedback for an authentication event whether it was from a valid user or not.

This feedback is used for improving the risk evaluation decision for the user pool as part of Amazon Cognito advanced security.

update_device_status(client, input, options \\ [])

Updates the device status.

update_group(client, input, options \\ [])

Updates the specified group with the specified attributes.

Calling this action requires developer credentials.

If you don't provide a value for an attribute, it will be set to the default value.

update_identity_provider(client, input, options \\ [])

Updates identity provider information for a user pool.

update_resource_server(client, input, options \\ [])

Updates the name and scopes of resource server.

All other fields are read-only.

If you don't provide a value for an attribute, it will be set to the default value.

update_user_attributes(client, input, options \\ [])

Allows a user to update a specific attribute (one at a time).

update_user_pool(client, input, options \\ [])

Updates the specified user pool with the specified attributes.

You can get a list of the current user pool settings using DescribeUserPool.

If you don't provide a value for an attribute, it will be set to the default value.

update_user_pool_client(client, input, options \\ [])

Updates the specified user pool app client with the specified attributes.

You can get a list of the current user pool app client settings using DescribeUserPoolClient.

If you don't provide a value for an attribute, it will be set to the default value.

update_user_pool_domain(client, input, options \\ [])

Updates the Secure Sockets Layer (SSL) certificate for the custom domain for your user pool.

You can use this operation to provide the Amazon Resource Name (ARN) of a new certificate to Amazon Cognito. You cannot use it to change the domain for a user pool.

A custom domain is used to host the Amazon Cognito hosted UI, which provides sign-up and sign-in pages for your application. When you set up a custom domain, you provide a certificate that you manage with AWS Certificate Manager (ACM). When necessary, you can use this operation to change the certificate that you applied to your custom domain.

Usually, this is unnecessary following routine certificate renewal with ACM. When you renew your existing certificate in ACM, the ARN for your certificate remains the same, and your custom domain uses the new certificate automatically.

However, if you replace your existing certificate with a new one, ACM gives the new certificate a new ARN. To apply the new certificate to your custom domain, you must provide this ARN to Amazon Cognito.

When you add your new certificate in ACM, you must choose US East (N. Virginia) as the AWS Region.

After you submit your request, Amazon Cognito requires up to 1 hour to distribute your new certificate to your custom domain.

For more information about adding a custom domain to your user pool, see Using Your Own Domain for the Hosted UI.

verify_software_token(client, input, options \\ [])

Use this API to register a user's entered TOTP code and mark the user's software token MFA status as "verified" if successful.

The request takes an access token or a session string, but not both.

verify_user_attribute(client, input, options \\ [])

Verifies the specified user attributes in the user pool.