aws-elixir v0.6.0 AWS.Comprehend

Amazon Comprehend is an AWS service for gaining insight into the content of documents.

Use these actions to determine the topics contained in your documents, the topics they discuss, the predominant sentiment expressed in them, the predominant language used, and more.

Functions

batch_detect_dominant_language(client, input, options \\ [])

Determines the dominant language of the input text for a batch of documents.

For a list of languages that Amazon Comprehend can detect, see Amazon Comprehend Supported Languages.

batch_detect_entities(client, input, options \\ [])

Inspects the text of a batch of documents for named entities and returns information about them.

For more information about named entities, see how-entities

batch_detect_key_phrases(client, input, options \\ [])

Detects the key noun phrases found in a batch of documents.

batch_detect_sentiment(client, input, options \\ [])

Inspects a batch of documents and returns an inference of the prevailing sentiment, POSITIVE, NEUTRAL, MIXED, or NEGATIVE, in each one.

batch_detect_syntax(client, input, options \\ [])

Inspects the text of a batch of documents for the syntax and part of speech of the words in the document and returns information about them.

For more information, see how-syntax.

classify_document(client, input, options \\ [])

Creates a new document classification request to analyze a single document in real-time, using a previously created and trained custom model and an endpoint.

create_document_classifier(client, input, options \\ [])

Creates a new document classifier that you can use to categorize documents.

To create a classifier, you provide a set of training documents that labeled with the categories that you want to use. After the classifier is trained you can use it to categorize a set of labeled documents into the categories. For more information, see how-document-classification.

create_endpoint(client, input, options \\ [])

Creates a model-specific endpoint for synchronous inference for a previously trained custom model

create_entity_recognizer(client, input, options \\ [])

Creates an entity recognizer using submitted files.

After your CreateEntityRecognizer request is submitted, you can check job status using the API.

delete_document_classifier(client, input, options \\ [])

Deletes a previously created document classifier

Only those classifiers that are in terminated states (IN_ERROR, TRAINED) will be deleted.

If an active inference job is using the model, a ResourceInUseException will be returned.

This is an asynchronous action that puts the classifier into a DELETING state, and it is then removed by a background job. Once removed, the classifier disappears from your account and is no longer available for use.

delete_endpoint(client, input, options \\ [])

Deletes a model-specific endpoint for a previously-trained custom model.

All endpoints must be deleted in order for the model to be deleted.

delete_entity_recognizer(client, input, options \\ [])

Deletes an entity recognizer.

Only those recognizers that are in terminated states (IN_ERROR, TRAINED) will be deleted. If an active inference job is using the model, a ResourceInUseException will be returned.

This is an asynchronous action that puts the recognizer into a DELETING state, and it is then removed by a background job. Once removed, the recognizer disappears from your account and is no longer available for use.

describe_document_classification_job(client, input, options \\ [])

Gets the properties associated with a document classification job.

Use this operation to get the status of a classification job.

describe_document_classifier(client, input, options \\ [])

Gets the properties associated with a document classifier.

describe_dominant_language_detection_job(client, input, options \\ [])

Gets the properties associated with a dominant language detection job.

Use this operation to get the status of a detection job.

describe_endpoint(client, input, options \\ [])

Gets the properties associated with a specific endpoint.

Use this operation to get the status of an endpoint.

describe_entities_detection_job(client, input, options \\ [])

Gets the properties associated with an entities detection job.

Use this operation to get the status of a detection job.

describe_entity_recognizer(client, input, options \\ [])

Provides details about an entity recognizer including status, S3 buckets containing training data, recognizer metadata, metrics, and so on.

describe_key_phrases_detection_job(client, input, options \\ [])

Gets the properties associated with a key phrases detection job.

Use this operation to get the status of a detection job.

describe_pii_entities_detection_job(client, input, options \\ [])

Gets the properties associated with a PII entities detection job.

For example, you can use this operation to get the job status.

describe_sentiment_detection_job(client, input, options \\ [])

Gets the properties associated with a sentiment detection job.

Use this operation to get the status of a detection job.

describe_topics_detection_job(client, input, options \\ [])

Gets the properties associated with a topic detection job.

Use this operation to get the status of a detection job.

detect_dominant_language(client, input, options \\ [])

Determines the dominant language of the input text.

For a list of languages that Amazon Comprehend can detect, see Amazon Comprehend Supported Languages.

detect_entities(client, input, options \\ [])

Inspects text for named entities, and returns information about them.

For more information, about named entities, see how-entities.

detect_key_phrases(client, input, options \\ [])

Detects the key noun phrases found in the text.

detect_pii_entities(client, input, options \\ [])

Inspects the input text for entities that contain personally identifiable information (PII) and returns information about them.

detect_sentiment(client, input, options \\ [])

Inspects text and returns an inference of the prevailing sentiment (POSITIVE, NEUTRAL, MIXED, or NEGATIVE).

detect_syntax(client, input, options \\ [])

Inspects text for syntax and the part of speech of words in the document.

For more information, how-syntax.

list_document_classification_jobs(client, input, options \\ [])

Gets a list of the documentation classification jobs that you have submitted.

list_document_classifiers(client, input, options \\ [])

Gets a list of the document classifiers that you have created.

list_dominant_language_detection_jobs(client, input, options \\ [])

Gets a list of the dominant language detection jobs that you have submitted.

list_endpoints(client, input, options \\ [])

Gets a list of all existing endpoints that you've created.

list_entities_detection_jobs(client, input, options \\ [])

Gets a list of the entity detection jobs that you have submitted.

list_entity_recognizers(client, input, options \\ [])

Gets a list of the properties of all entity recognizers that you created, including recognizers currently in training.

Allows you to filter the list of recognizers based on criteria such as status and submission time. This call returns up to 500 entity recognizers in the list, with a default number of 100 recognizers in the list.

The results of this list are not in any particular order. Please get the list and sort locally if needed.

list_key_phrases_detection_jobs(client, input, options \\ [])

Get a list of key phrase detection jobs that you have submitted.

list_pii_entities_detection_jobs(client, input, options \\ [])

Gets a list of the PII entity detection jobs that you have submitted.

list_sentiment_detection_jobs(client, input, options \\ [])

Gets a list of sentiment detection jobs that you have submitted.

list_tags_for_resource(client, input, options \\ [])

Lists all tags associated with a given Amazon Comprehend resource.

list_topics_detection_jobs(client, input, options \\ [])

Gets a list of the topic detection jobs that you have submitted.

start_document_classification_job(client, input, options \\ [])

Starts an asynchronous document classification job.

Use the operation to track the progress of the job.

start_dominant_language_detection_job(client, input, options \\ [])

Starts an asynchronous dominant language detection job for a collection of documents.

Use the operation to track the status of a job.

start_entities_detection_job(client, input, options \\ [])

Starts an asynchronous entity detection job for a collection of documents.

Use the operation to track the status of a job.

This API can be used for either standard entity detection or custom entity recognition. In order to be used for custom entity recognition, the optional EntityRecognizerArn must be used in order to provide access to the recognizer being used to detect the custom entity.

start_key_phrases_detection_job(client, input, options \\ [])

Starts an asynchronous key phrase detection job for a collection of documents.

Use the operation to track the status of a job.

start_pii_entities_detection_job(client, input, options \\ [])

Starts an asynchronous PII entity detection job for a collection of documents.

start_sentiment_detection_job(client, input, options \\ [])

Starts an asynchronous sentiment detection job for a collection of documents.

use the operation to track the status of a job.

start_topics_detection_job(client, input, options \\ [])

Starts an asynchronous topic detection job.

Use the DescribeTopicDetectionJob operation to track the status of a job.

stop_dominant_language_detection_job(client, input, options \\ [])

Stops a dominant language detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

stop_entities_detection_job(client, input, options \\ [])

Stops an entities detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

stop_key_phrases_detection_job(client, input, options \\ [])

Stops a key phrases detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

stop_pii_entities_detection_job(client, input, options \\ [])

Stops a PII entities detection job in progress.

stop_sentiment_detection_job(client, input, options \\ [])

Stops a sentiment detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is be stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

stop_training_document_classifier(client, input, options \\ [])

Stops a document classifier training job while in progress.

If the training job state is TRAINING, the job is marked for termination and put into the STOP_REQUESTED state. If the training job completes before it can be stopped, it is put into the TRAINED; otherwise the training job is stopped and put into the STOPPED state and the service sends back an HTTP 200 response with an empty HTTP body.

stop_training_entity_recognizer(client, input, options \\ [])

Stops an entity recognizer training job while in progress.

If the training job state is TRAINING, the job is marked for termination and put into the STOP_REQUESTED state. If the training job completes before it can be stopped, it is put into the TRAINED; otherwise the training job is stopped and putted into the STOPPED state and the service sends back an HTTP 200 response with an empty HTTP body.

tag_resource(client, input, options \\ [])

Associates a specific tag with an Amazon Comprehend resource.

A tag is a key-value pair that adds as a metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

untag_resource(client, input, options \\ [])

Removes a specific tag associated with an Amazon Comprehend resource.

update_endpoint(client, input, options \\ [])

Updates information about the specified endpoint.