aws-elixir v0.6.0 AWS.ComprehendMedical

Amazon Comprehend Medical extracts structured information from unstructured clinical text.

Use these actions to gain insight in your documents.

Functions

describe_entities_detection_v2_job(client, input, options \\ [])

Gets the properties associated with a medical entities detection job.

Use this operation to get the status of a detection job.

describe_i_c_d10_c_m_inference_job(client, input, options \\ [])

Gets the properties associated with an InferICD10CM job.

Use this operation to get the status of an inference job.

describe_p_h_i_detection_job(client, input, options \\ [])

Gets the properties associated with a protected health information (PHI) detection job.

Use this operation to get the status of a detection job.

describe_rx_norm_inference_job(client, input, options \\ [])

Gets the properties associated with an InferRxNorm job.

Use this operation to get the status of an inference job.

detect_entities(client, input, options \\ [])

The DetectEntities operation is deprecated.

You should use the DetectEntitiesV2 operation instead.

Inspects the clinical text for a variety of medical entities and returns specific information about them such as entity category, location, and confidence score on that information .

detect_entities_v2(client, input, options \\ [])

Inspects the clinical text for a variety of medical entities and returns specific information about them such as entity category, location, and confidence score on that information.

Amazon Comprehend Medical only detects medical entities in English language texts.

The DetectEntitiesV2 operation replaces the DetectEntities operation. This new action uses a different model for determining the entities in your medical text and changes the way that some entities are returned in the output. You should use the DetectEntitiesV2 operation in all new applications.

The DetectEntitiesV2 operation returns the Acuity and Direction entities as attributes instead of types.

detect_p_h_i(client, input, options \\ [])

Inspects the clinical text for protected health information (PHI) entities and returns the entity category, location, and confidence score for each entity.

Amazon Comprehend Medical only detects entities in English language texts.

infer_i_c_d10_c_m(client, input, options \\ [])

InferICD10CM detects medical conditions as entities listed in a patient record and links those entities to normalized concept identifiers in the ICD-10-CM knowledge base from the Centers for Disease Control.

Amazon Comprehend Medical only detects medical entities in English language texts.

infer_rx_norm(client, input, options \\ [])

InferRxNorm detects medications as entities listed in a patient record and links to the normalized concept identifiers in the RxNorm database from the National Library of Medicine.

Amazon Comprehend Medical only detects medical entities in English language texts.

list_entities_detection_v2_jobs(client, input, options \\ [])

Gets a list of medical entity detection jobs that you have submitted.

list_i_c_d10_c_m_inference_jobs(client, input, options \\ [])

Gets a list of InferICD10CM jobs that you have submitted.

list_p_h_i_detection_jobs(client, input, options \\ [])

Gets a list of protected health information (PHI) detection jobs that you have submitted.

list_rx_norm_inference_jobs(client, input, options \\ [])

Gets a list of InferRxNorm jobs that you have submitted.

start_entities_detection_v2_job(client, input, options \\ [])

Starts an asynchronous medical entity detection job for a collection of documents.

Use the DescribeEntitiesDetectionV2Job operation to track the status of a job.

start_i_c_d10_c_m_inference_job(client, input, options \\ [])

Starts an asynchronous job to detect medical conditions and link them to the ICD-10-CM ontology.

Use the DescribeICD10CMInferenceJob operation to track the status of a job.

start_p_h_i_detection_job(client, input, options \\ [])

Starts an asynchronous job to detect protected health information (PHI).

Use the DescribePHIDetectionJob operation to track the status of a job.

start_rx_norm_inference_job(client, input, options \\ [])

Starts an asynchronous job to detect medication entities and link them to the RxNorm ontology.

Use the DescribeRxNormInferenceJob operation to track the status of a job.

stop_entities_detection_v2_job(client, input, options \\ [])

Stops a medical entities detection job in progress.

stop_i_c_d10_c_m_inference_job(client, input, options \\ [])

Stops an InferICD10CM inference job in progress.

stop_p_h_i_detection_job(client, input, options \\ [])

Stops a protected health information (PHI) detection job in progress.

stop_rx_norm_inference_job(client, input, options \\ [])

Stops an InferRxNorm inference job in progress.