aws-elixir v0.6.0 AWS.Connect

Amazon Connect is a cloud-based contact center solution that makes it easy to set up and manage a customer contact center and provide reliable customer engagement at any scale.

Amazon Connect provides rich metrics and real-time reporting that allow you to optimize contact routing. You can also resolve customer issues more efficiently by putting customers in touch with the right agents.

There are limits to the number of Amazon Connect resources that you can create and limits to the number of requests that you can make per second. For more information, see Amazon Connect Service Quotas in the Amazon Connect Administrator Guide.

To connect programmatically to an AWS service, you use an endpoint. For a list of Amazon Connect endpoints, see Amazon Connect Endpoints.

Working with contact flows? Check out the Amazon Connect Flow language.

Functions

associate_routing_profile_queues(client, instance_id, routing_profile_id, input, options \\ [])

Associates a set of queues with a routing profile.

create_contact_flow(client, instance_id, input, options \\ [])

Creates a contact flow for the specified Amazon Connect instance.

You can also create and update contact flows using the Amazon Connect Flow language.

create_routing_profile(client, instance_id, input, options \\ [])

Creates a new routing profile.

create_user(client, instance_id, input, options \\ [])

Creates a user account for the specified Amazon Connect instance.

For information about how to create user accounts using the Amazon Connect console, see Add Users in the Amazon Connect Administrator Guide.

delete_user(client, instance_id, user_id, input, options \\ [])

Deletes a user account from the specified Amazon Connect instance.

For information about what happens to a user's data when their account is deleted, see Delete Users from Your Amazon Connect Instance in the Amazon Connect Administrator Guide.

describe_contact_flow(client, contact_flow_id, instance_id, options \\ [])

Describes the specified contact flow.

You can also create and update contact flows using the Amazon Connect Flow language.

describe_routing_profile(client, instance_id, routing_profile_id, options \\ [])

Describes the specified routing profile.

describe_user(client, instance_id, user_id, options \\ [])

Describes the specified user account.

You can find the instance ID in the console (it’s the final part of the ARN). The console does not display the user IDs. Instead, list the users and note the IDs provided in the output.

describe_user_hierarchy_group(client, hierarchy_group_id, instance_id, options \\ [])

Describes the specified hierarchy group.

describe_user_hierarchy_structure(client, instance_id, options \\ [])

Describes the hierarchy structure of the specified Amazon Connect instance.

disassociate_routing_profile_queues(client, instance_id, routing_profile_id, input, options \\ [])

Disassociates a set of queues from a routing profile.

get_contact_attributes(client, initial_contact_id, instance_id, options \\ [])

Retrieves the contact attributes for the specified contact.

get_current_metric_data(client, instance_id, input, options \\ [])

Gets the real-time metric data from the specified Amazon Connect instance.

For a description of each metric, see Real-time Metrics Definitions in the Amazon Connect Administrator Guide.

get_federation_token(client, instance_id, options \\ [])

Retrieves a token for federation.

get_metric_data(client, instance_id, input, options \\ [])

Gets historical metric data from the specified Amazon Connect instance.

For a description of each historical metric, see Historical Metrics Definitions in the Amazon Connect Administrator Guide.

list_contact_flows(client, instance_id, contact_flow_types \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Provides information about the contact flows for the specified Amazon Connect instance.

You can also create and update contact flows using the Amazon Connect Flow language.

For more information about contact flows, see Contact Flows in the Amazon Connect Administrator Guide.

list_hours_of_operations(client, instance_id, max_results \\ nil, next_token \\ nil, options \\ [])

Provides information about the hours of operation for the specified Amazon Connect instance.

For more information about hours of operation, see Set the Hours of Operation for a Queue in the Amazon Connect Administrator Guide.

list_phone_numbers(client, instance_id, max_results \\ nil, next_token \\ nil, phone_number_country_codes \\ nil, phone_number_types \\ nil, options \\ [])

Provides information about the phone numbers for the specified Amazon Connect instance.

For more information about phone numbers, see Set Up Phone Numbers for Your Contact Center in the Amazon Connect Administrator Guide.

list_prompts(client, instance_id, max_results \\ nil, next_token \\ nil, options \\ [])

Provides information about the prompts for the specified Amazon Connect instance.

list_queues(client, instance_id, max_results \\ nil, next_token \\ nil, queue_types \\ nil, options \\ [])

Provides information about the queues for the specified Amazon Connect instance.

For more information about queues, see Queues: Standard and Agent in the Amazon Connect Administrator Guide.

list_routing_profile_queues(client, instance_id, routing_profile_id, max_results \\ nil, next_token \\ nil, options \\ [])

List the queues associated with a routing profile.

list_routing_profiles(client, instance_id, max_results \\ nil, next_token \\ nil, options \\ [])

Provides summary information about the routing profiles for the specified Amazon Connect instance.

For more information about routing profiles, see Routing Profiles and Create a Routing Profile in the Amazon Connect Administrator Guide.

list_security_profiles(client, instance_id, max_results \\ nil, next_token \\ nil, options \\ [])

Provides summary information about the security profiles for the specified Amazon Connect instance.

For more information about security profiles, see Security Profiles in the Amazon Connect Administrator Guide.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists the tags for the specified resource.

For sample policies that use tags, see Amazon Connect Identity-Based Policy Examples in the Amazon Connect Administrator Guide.

list_user_hierarchy_groups(client, instance_id, max_results \\ nil, next_token \\ nil, options \\ [])

Provides summary information about the hierarchy groups for the specified Amazon Connect instance.

For more information about agent hierarchies, see Set Up Agent Hierarchies in the Amazon Connect Administrator Guide.

list_users(client, instance_id, max_results \\ nil, next_token \\ nil, options \\ [])

Provides summary information about the users for the specified Amazon Connect instance.

resume_contact_recording(client, input, options \\ [])

When a contact is being recorded, and the recording has been suspended using SuspendContactRecording, this API resumes recording the call.

Only voice recordings are supported at this time.

start_chat_contact(client, input, options \\ [])

Initiates a contact flow to start a new chat for the customer.

Response of this API provides a token required to obtain credentials from the CreateParticipantConnection API in the Amazon Connect Participant Service.

When a new chat contact is successfully created, clients need to subscribe to the participant’s connection for the created chat within 5 minutes. This is achieved by invoking CreateParticipantConnection with WEBSOCKET and CONNECTION_CREDENTIALS.

A 429 error occurs in two situations:

API rate limit is exceeded. API TPS throttling returns a TooManyRequests exception from the API Gateway.

The quota for concurrent active chats is exceeded. Active chat throttling returns a LimitExceededException.

For more information about how chat works, see Chat in the Amazon Connect Administrator Guide.

start_contact_recording(client, input, options \\ [])

This API starts recording the contact when the agent joins the call.

StartContactRecording is a one-time action. For example, if you use StopContactRecording to stop recording an ongoing call, you can't use StartContactRecording to restart it. For scenarios where the recording has started and you want to suspend and resume it, such as when collecting sensitive information (for example, a credit card number), use SuspendContactRecording and ResumeContactRecording.

You can use this API to override the recording behavior configured in the Set recording behavior block.

Only voice recordings are supported at this time.

start_outbound_voice_contact(client, input, options \\ [])

This API places an outbound call to a contact, and then initiates the contact flow.

It performs the actions in the contact flow that's specified (in ContactFlowId).

Agents are not involved in initiating the outbound API (that is, dialing the contact). If the contact flow places an outbound call to a contact, and then puts the contact in queue, that's when the call is routed to the agent, like any other inbound case.

There is a 60 second dialing timeout for this operation. If the call is not connected after 60 seconds, it fails.

UK numbers with a 447 prefix are not allowed by default. Before you can dial these UK mobile numbers, you must submit a service quota increase request. For more information, see Amazon Connect Service Quotas in the Amazon Connect Administrator Guide.

stop_contact(client, input, options \\ [])

Ends the specified contact.

stop_contact_recording(client, input, options \\ [])

When a contact is being recorded, this API stops recording the call.

StopContactRecording is a one-time action. If you use StopContactRecording to stop recording an ongoing call, you can't use StartContactRecording to restart it. For scenarios where the recording has started and you want to suspend it for sensitive information (for example, to collect a credit card number), and then restart it, use SuspendContactRecording and ResumeContactRecording.

Only voice recordings are supported at this time.

suspend_contact_recording(client, input, options \\ [])

When a contact is being recorded, this API suspends recording the call.

For example, you might suspend the call recording while collecting sensitive information, such as a credit card number. Then use ResumeContactRecording to restart recording.

The period of time that the recording is suspended is filled with silence in the final recording.

Only voice recordings are supported at this time.

tag_resource(client, resource_arn, input, options \\ [])

Adds the specified tags to the specified resource.

The supported resource types are users, routing profiles, and contact flows.

For sample policies that use tags, see Amazon Connect Identity-Based Policy Examples in the Amazon Connect Administrator Guide.

untag_resource(client, resource_arn, input, options \\ [])

Removes the specified tags from the specified resource.

update_contact_attributes(client, input, options \\ [])

Creates or updates the contact attributes associated with the specified contact.

You can add or update attributes for both ongoing and completed contacts. For example, you can update the customer's name or the reason the customer called while the call is active, or add notes about steps that the agent took during the call that are displayed to the next agent that takes the call. You can also update attributes for a contact using data from your CRM application and save the data with the contact in Amazon Connect. You could also flag calls for additional analysis, such as legal review or identifying abusive callers.

Contact attributes are available in Amazon Connect for 24 months, and are then deleted.

This operation is also available in the Amazon Connect Flow language. See UpdateContactAttributes.

Important: You cannot use the operation to update attributes for contacts that occurred prior to the release of the API, September 12, 2018. You can update attributes only for contacts that started after the release of the API. If you attempt to update attributes for a contact that occurred prior to the release of the API, a 400 error is returned. This applies also to queued callbacks that were initiated prior to the release of the API but are still active in your instance.

update_contact_flow_content(client, contact_flow_id, instance_id, input, options \\ [])

Updates the specified contact flow.

You can also create and update contact flows using the Amazon Connect Flow language.

update_contact_flow_name(client, contact_flow_id, instance_id, input, options \\ [])

The name of the contact flow.

update_routing_profile_concurrency(client, instance_id, routing_profile_id, input, options \\ [])

Updates the channels that agents can handle in the Contact Control Panel (CCP) for a routing profile.

update_routing_profile_default_outbound_queue(client, instance_id, routing_profile_id, input, options \\ [])

Updates the default outbound queue of a routing profile.

update_routing_profile_name(client, instance_id, routing_profile_id, input, options \\ [])

Updates the name and description of a routing profile.

The request accepts the following data in JSON format. At least Name or Description must be provided.

update_routing_profile_queues(client, instance_id, routing_profile_id, input, options \\ [])

Updates the properties associated with a set of queues for a routing profile.

update_user_hierarchy(client, instance_id, user_id, input, options \\ [])

Assigns the specified hierarchy group to the specified user.

update_user_identity_info(client, instance_id, user_id, input, options \\ [])

Updates the identity information for the specified user.

Someone with the ability to invoke UpdateUserIndentityInfo can change the login credentials of other users by changing their email address. This poses a security risk to your organization. They can change the email address of a user to the attacker's email address, and then reset the password through email. We strongly recommend limiting who has the ability to invoke UpdateUserIndentityInfo. For more information, see Best Practices for Security Profiles in the Amazon Connect Administrator Guide.

update_user_phone_config(client, instance_id, user_id, input, options \\ [])

Updates the phone configuration settings for the specified user.

update_user_routing_profile(client, instance_id, user_id, input, options \\ [])

Assigns the specified routing profile to the specified user.

update_user_security_profiles(client, instance_id, user_id, input, options \\ [])

Assigns the specified security profiles to the specified user.