aws-elixir v0.6.0 AWS.ConnectParticipant

Amazon Connect is a cloud-based contact center solution that makes it easy to set up and manage a customer contact center and provide reliable customer engagement at any scale.

Amazon Connect enables customer contacts through voice or chat.

The APIs described here are used by chat participants, such as agents and customers.

Functions

create_participant_connection(client, input, options \\ [])

Creates the participant's connection.

Note that ParticipantToken is used for invoking this API instead of ConnectionToken.

The participant token is valid for the lifetime of the participant – until the they are part of a contact.

The response URL for WEBSOCKET Type has a connect expiry timeout of 100s. Clients must manually connect to the returned websocket URL and subscribe to the desired topic.

For chat, you need to publish the following on the established websocket connection:

{"topic":"aws/subscribe","content":{"topics":["aws/chat"]}}

Upon websocket URL expiry, as specified in the response ConnectionExpiry parameter, clients need to call this API again to obtain a new websocket URL and perform the same steps as before.

disconnect_participant(client, input, options \\ [])

Disconnects a participant.

Note that ConnectionToken is used for invoking this API instead of ParticipantToken.

get_transcript(client, input, options \\ [])

Retrieves a transcript of the session.

Note that ConnectionToken is used for invoking this API instead of ParticipantToken.

send_event(client, input, options \\ [])

Sends an event.

Note that ConnectionToken is used for invoking this API instead of ParticipantToken.

send_message(client, input, options \\ [])

Sends a message.

Note that ConnectionToken is used for invoking this API instead of ParticipantToken.