aws-elixir v0.6.0 AWS.CostExplorer

The Cost Explorer API enables you to programmatically query your cost and usage data.

You can query for aggregated data such as total monthly costs or total daily usage. You can also query for granular data, such as the number of daily write operations for Amazon DynamoDB database tables in your production environment.

Service Endpoint

The Cost Explorer API provides the following endpoint:

https://ce.us-east-1.amazonaws.com

For information about costs associated with the Cost Explorer API, see AWS Cost Management Pricing.

Functions

create_anomaly_monitor(client, input, options \\ [])

Creates a new cost anomaly detection monitor with the requested type and monitor specification.

create_anomaly_subscription(client, input, options \\ [])

Adds a subscription to a cost anomaly detection monitor.

You can use each subscription to define subscribers with email or SNS notifications. Email subscribers can set a dollar threshold and a time frequency for receiving notifications.

create_cost_category_definition(client, input, options \\ [])

Creates a new Cost Category with the requested name and rules.

delete_anomaly_monitor(client, input, options \\ [])

Deletes a cost anomaly monitor.

delete_anomaly_subscription(client, input, options \\ [])

Deletes a cost anomaly subscription.

delete_cost_category_definition(client, input, options \\ [])

Deletes a Cost Category.

Expenses from this month going forward will no longer be categorized with this Cost Category.

describe_cost_category_definition(client, input, options \\ [])

Returns the name, ARN, rules, definition, and effective dates of a Cost Category that's defined in the account.

You have the option to use EffectiveOn to return a Cost Category that is active on a specific date. If there is no EffectiveOn specified, you’ll see a Cost Category that is effective on the current date. If Cost Category is still effective, EffectiveEnd is omitted in the response.

get_anomalies(client, input, options \\ [])

Retrieves all of the cost anomalies detected on your account, during the time period specified by the DateInterval object.

get_anomaly_monitors(client, input, options \\ [])

Retrieves the cost anomaly monitor definitions for your account.

You can filter using a list of cost anomaly monitor Amazon Resource Names (ARNs).

get_anomaly_subscriptions(client, input, options \\ [])

Retrieves the cost anomaly subscription objects for your account.

You can filter using a list of cost anomaly monitor Amazon Resource Names (ARNs).

get_cost_and_usage(client, input, options \\ [])

Retrieves cost and usage metrics for your account.

You can specify which cost and usage-related metric, such as BlendedCosts or UsageQuantity, that you want the request to return. You can also filter and group your data by various dimensions, such as SERVICE or AZ, in a specific time range. For a complete list of valid dimensions, see the GetDimensionValues operation. Master account in an organization in AWS Organizations have access to all member accounts.

get_cost_and_usage_with_resources(client, input, options \\ [])

Retrieves cost and usage metrics with resources for your account.

You can specify which cost and usage-related metric, such as BlendedCosts or UsageQuantity, that you want the request to return. You can also filter and group your data by various dimensions, such as SERVICE or AZ, in a specific time range. For a complete list of valid dimensions, see the GetDimensionValues operation. Master account in an organization in AWS Organizations have access to all member accounts. This API is currently available for the Amazon Elastic Compute Cloud – Compute service only.

This is an opt-in only feature. You can enable this feature from the Cost Explorer Settings page. For information on how to access the Settings page, see Controlling Access for Cost Explorer in the AWS Billing and Cost Management User Guide.

get_cost_forecast(client, input, options \\ [])

Retrieves a forecast for how much Amazon Web Services predicts that you will spend over the forecast time period that you select, based on your past costs.

get_dimension_values(client, input, options \\ [])

Retrieves all available filter values for a specified filter over a period of time.

You can search the dimension values for an arbitrary string.

get_reservation_coverage(client, input, options \\ [])

Retrieves the reservation coverage for your account.

This enables you to see how much of your Amazon Elastic Compute Cloud, Amazon ElastiCache, Amazon Relational Database Service, or Amazon Redshift usage is covered by a reservation. An organization's master account can see the coverage of the associated member accounts. This supports dimensions, Cost Categories, and nested expressions. For any time period, you can filter data about reservation usage by the following dimensions:

AZ

CACHE_ENGINE

DATABASE_ENGINE

DEPLOYMENT_OPTION

INSTANCE_TYPE

LINKED_ACCOUNT

OPERATING_SYSTEM

PLATFORM

REGION

SERVICE

TAG

TENANCY

To determine valid values for a dimension, use the GetDimensionValues operation.

get_reservation_purchase_recommendation(client, input, options \\ [])

Gets recommendations for which reservations to purchase.

These recommendations could help you reduce your costs. Reservations provide a discounted hourly rate (up to 75%) compared to On-Demand pricing.

AWS generates your recommendations by identifying your On-Demand usage during a specific time period and collecting your usage into categories that are eligible for a reservation. After AWS has these categories, it simulates every combination of reservations in each category of usage to identify the best number of each type of RI to purchase to maximize your estimated savings.

For example, AWS automatically aggregates your Amazon EC2 Linux, shared tenancy, and c4 family usage in the US West (Oregon) Region and recommends that you buy size-flexible regional reservations to apply to the c4 family usage. AWS recommends the smallest size instance in an instance family. This makes it easier to purchase a size-flexible RI. AWS also shows the equal number of normalized units so that you can purchase any instance size that you want. For this example, your RI recommendation would be for c4.large because that is the smallest size instance in the c4 instance family.

get_reservation_utilization(client, input, options \\ [])

Retrieves the reservation utilization for your account.

Master account in an organization have access to member accounts. You can filter data by dimensions in a time period. You can use GetDimensionValues to determine the possible dimension values. Currently, you can group only by SUBSCRIPTION_ID.

get_rightsizing_recommendation(client, input, options \\ [])

Creates recommendations that help you save cost by identifying idle and underutilized Amazon EC2 instances.

Recommendations are generated to either downsize or terminate instances, along with providing savings detail and metrics. For details on calculation and function, see Optimizing Your Cost with Rightsizing Recommendations in the AWS Billing and Cost Management User Guide.

get_savings_plans_coverage(client, input, options \\ [])

Retrieves the Savings Plans covered for your account.

This enables you to see how much of your cost is covered by a Savings Plan. An organization’s master account can see the coverage of the associated member accounts. This supports dimensions, Cost Categories, and nested expressions. For any time period, you can filter data for Savings Plans usage with the following dimensions:

LINKED_ACCOUNT

REGION

SERVICE

INSTANCE_FAMILY

To determine valid values for a dimension, use the GetDimensionValues operation.

get_savings_plans_purchase_recommendation(client, input, options \\ [])

Retrieves your request parameters, Savings Plan Recommendations Summary and Details.

get_savings_plans_utilization(client, input, options \\ [])

Retrieves the Savings Plans utilization for your account across date ranges with daily or monthly granularity.

Master account in an organization have access to member accounts. You can use GetDimensionValues in SAVINGS_PLANS to determine the possible dimension values.

You cannot group by any dimension values for GetSavingsPlansUtilization.

get_savings_plans_utilization_details(client, input, options \\ [])

Retrieves attribute data along with aggregate utilization and savings data for a given time period.

This doesn't support granular or grouped data (daily/monthly) in response. You can't retrieve data by dates in a single response similar to GetSavingsPlanUtilization, but you have the option to make multiple calls to GetSavingsPlanUtilizationDetails by providing individual dates. You can use GetDimensionValues in SAVINGS_PLANS to determine the possible dimension values.

GetSavingsPlanUtilizationDetails internally groups data by SavingsPlansArn.

get_tags(client, input, options \\ [])

Queries for available tag keys and tag values for a specified period.

You can search the tag values for an arbitrary string.

get_usage_forecast(client, input, options \\ [])

Retrieves a forecast for how much Amazon Web Services predicts that you will use over the forecast time period that you select, based on your past usage.

list_cost_category_definitions(client, input, options \\ [])

Returns the name, ARN, NumberOfRules and effective dates of all Cost Categories defined in the account.

You have the option to use EffectiveOn to return a list of Cost Categories that were active on a specific date. If there is no EffectiveOn specified, you’ll see Cost Categories that are effective on the current date. If Cost Category is still effective, EffectiveEnd is omitted in the response. ListCostCategoryDefinitions supports pagination. The request can have a MaxResults range up to 100.

provide_anomaly_feedback(client, input, options \\ [])

Modifies the feedback property of a given cost anomaly.

update_anomaly_monitor(client, input, options \\ [])

Updates an existing cost anomaly monitor.

The changes made are applied going forward, and does not change anomalies detected in the past.

update_anomaly_subscription(client, input, options \\ [])

Updates an existing cost anomaly monitor subscription.

update_cost_category_definition(client, input, options \\ [])

Updates an existing Cost Category.

Changes made to the Cost Category rules will be used to categorize the current month’s expenses and future expenses. This won’t change categorization for the previous months.