aws-elixir v0.6.0 AWS.DLM

Amazon Data Lifecycle Manager

With Amazon Data Lifecycle Manager, you can manage the lifecycle of your AWS resources.

You create lifecycle policies, which are used to automate operations on the specified resources.

Amazon DLM supports Amazon EBS volumes and snapshots. For information about using Amazon DLM with Amazon EBS, see Automating the Amazon EBS Snapshot Lifecycle in the Amazon EC2 User Guide.

Functions

create_lifecycle_policy(client, input, options \\ [])

Creates a policy to manage the lifecycle of the specified AWS resources.

You can create up to 100 lifecycle policies.

delete_lifecycle_policy(client, policy_id, input, options \\ [])

Deletes the specified lifecycle policy and halts the automated operations that the policy specified.

get_lifecycle_policies(client, policy_ids \\ nil, resource_types \\ nil, state \\ nil, tags_to_add \\ nil, target_tags \\ nil, options \\ [])

Gets summary information about all or the specified data lifecycle policies.

To get complete information about a policy, use GetLifecyclePolicy.

get_lifecycle_policy(client, policy_id, options \\ [])

Gets detailed information about the specified lifecycle policy.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists the tags for the specified resource.

tag_resource(client, resource_arn, input, options \\ [])

Adds the specified tags to the specified resource.

untag_resource(client, resource_arn, input, options \\ [])

Removes the specified tags from the specified resource.

update_lifecycle_policy(client, policy_id, input, options \\ [])

Updates the specified lifecycle policy.