aws-elixir v0.6.0 AWS.DataExchange

This is the API reference for AWS Data Exchange.

Functions

cancel_job(client, job_id, input, options \\ [])

This operation cancels a job.

Jobs can be cancelled only when they are in the WAITING state.

create_data_set(client, input, options \\ [])

This operation creates a data set.

create_job(client, input, options \\ [])

This operation creates a job.

create_revision(client, data_set_id, input, options \\ [])

This operation creates a revision for a data set.

delete_asset(client, asset_id, data_set_id, revision_id, input, options \\ [])

This operation deletes an asset.

delete_data_set(client, data_set_id, input, options \\ [])

This operation deletes a data set.

delete_revision(client, data_set_id, revision_id, input, options \\ [])

This operation deletes a revision.

get_asset(client, asset_id, data_set_id, revision_id, options \\ [])

This operation returns information about an asset.

get_data_set(client, data_set_id, options \\ [])

This operation returns information about a data set.

get_job(client, job_id, options \\ [])

This operation returns information about a job.

get_revision(client, data_set_id, revision_id, options \\ [])

This operation returns information about a revision.

list_data_set_revisions(client, data_set_id, max_results \\ nil, next_token \\ nil, options \\ [])

This operation lists a data set's revisions sorted by CreatedAt in descending order.

list_data_sets(client, max_results \\ nil, next_token \\ nil, origin \\ nil, options \\ [])

This operation lists your data sets.

When listing by origin OWNED, results are sorted by CreatedAt in descending order. When listing by origin ENTITLED, there is no order and the maxResults parameter is ignored.

list_jobs(client, data_set_id \\ nil, max_results \\ nil, next_token \\ nil, revision_id \\ nil, options \\ [])

This operation lists your jobs sorted by CreatedAt in descending order.

list_revision_assets(client, data_set_id, revision_id, max_results \\ nil, next_token \\ nil, options \\ [])

This operation lists a revision's assets sorted alphabetically in descending order.

list_tags_for_resource(client, resource_arn, options \\ [])

This operation lists the tags on the resource.

start_job(client, job_id, input, options \\ [])

This operation starts a job.

tag_resource(client, resource_arn, input, options \\ [])

This operation tags a resource.

untag_resource(client, resource_arn, input, options \\ [])

This operation removes one or more tags from a resource.

update_asset(client, asset_id, data_set_id, revision_id, input, options \\ [])

This operation updates an asset.

update_data_set(client, data_set_id, input, options \\ [])

This operation updates a data set.

update_revision(client, data_set_id, revision_id, input, options \\ [])

This operation updates a revision.