aws-elixir v0.6.0 AWS.DataSync

AWS DataSync

AWS DataSync is a managed data transfer service that makes it simpler for you to automate moving data between on-premises storage and Amazon Simple Storage Service (Amazon S3) or Amazon Elastic File System (Amazon EFS).

This API interface reference for AWS DataSync contains documentation for a programming interface that you can use to manage AWS DataSync.

Functions

cancel_task_execution(client, input, options \\ [])

Cancels execution of a task.

When you cancel a task execution, the transfer of some files is abruptly interrupted. The contents of files that are transferred to the destination might be incomplete or inconsistent with the source files. However, if you start a new task execution on the same task and you allow the task execution to complete, file content on the destination is complete and consistent. This applies to other unexpected failures that interrupt a task execution. In all of these cases, AWS DataSync successfully complete the transfer when you start the next task execution.

create_agent(client, input, options \\ [])

Activates an AWS DataSync agent that you have deployed on your host.

The activation process associates your agent with your account. In the activation process, you specify information such as the AWS Region that you want to activate the agent in. You activate the agent in the AWS Region where your target locations (in Amazon S3 or Amazon EFS) reside. Your tasks are created in this AWS Region.

You can activate the agent in a VPC (virtual private cloud) or provide the agent access to a VPC endpoint so you can run tasks without going over the public internet.

You can use an agent for more than one location. If a task uses multiple agents, all of them need to have status AVAILABLE for the task to run. If you use multiple agents for a source location, the status of all the agents must be AVAILABLE for the task to run.

Agents are automatically updated by AWS on a regular basis, using a mechanism that ensures minimal interruption to your tasks.

create_location_efs(client, input, options \\ [])

Creates an endpoint for an Amazon EFS file system.

create_location_fsx_windows(client, input, options \\ [])

Creates an endpoint for an Amazon FSx for Windows file system.

create_location_nfs(client, input, options \\ [])

Defines a file system on a Network File System (NFS) server that can be read from or written to.

create_location_object_storage(client, input, options \\ [])

Creates an endpoint for a self-managed object storage bucket.

For more information about self-managed object storage locations, see create-object-location.

create_location_s3(client, input, options \\ [])

Creates an endpoint for an Amazon S3 bucket.

For more information, see https://docs.aws.amazon.com/datasync/latest/userguide/create-locations-cli.html#create-location-s3-cli in the AWS DataSync User Guide.

create_location_smb(client, input, options \\ [])

Defines a file system on a Server Message Block (SMB) server that can be read from or written to.

create_task(client, input, options \\ [])

Creates a task.

A task is a set of two locations (source and destination) and a set of Options that you use to control the behavior of a task. If you don't specify Options when you create a task, AWS DataSync populates them with service defaults.

When you create a task, it first enters the CREATING state. During CREATING AWS DataSync attempts to mount the on-premises Network File System (NFS) location. The task transitions to the AVAILABLE state without waiting for the AWS location to become mounted. If required, AWS DataSync mounts the AWS location before each task execution.

If an agent that is associated with a source (NFS) location goes offline, the task transitions to the UNAVAILABLE status. If the status of the task remains in the CREATING status for more than a few minutes, it means that your agent might be having trouble mounting the source NFS file system. Check the task's ErrorCode and ErrorDetail. Mount issues are often caused by either a misconfigured firewall or a mistyped NFS server hostname.

delete_agent(client, input, options \\ [])

Deletes an agent.

To specify which agent to delete, use the Amazon Resource Name (ARN) of the agent in your request. The operation disassociates the agent from your AWS account. However, it doesn't delete the agent virtual machine (VM) from your on-premises environment.

delete_location(client, input, options \\ [])

Deletes the configuration of a location used by AWS DataSync.

delete_task(client, input, options \\ [])

Deletes a task.

describe_agent(client, input, options \\ [])

Returns metadata such as the name, the network interfaces, and the status (that is, whether the agent is running or not) for an agent.

To specify which agent to describe, use the Amazon Resource Name (ARN) of the agent in your request.

describe_location_efs(client, input, options \\ [])

Returns metadata, such as the path information about an Amazon EFS location.

describe_location_fsx_windows(client, input, options \\ [])

Returns metadata, such as the path information about an Amazon FSx for Windows location.

describe_location_nfs(client, input, options \\ [])

Returns metadata, such as the path information, about an NFS location.

describe_location_object_storage(client, input, options \\ [])

Returns metadata about a self-managed object storage server location.

For more information about self-managed object storage locations, see create-object-location.

describe_location_s3(client, input, options \\ [])

Returns metadata, such as bucket name, about an Amazon S3 bucket location.

describe_location_smb(client, input, options \\ [])

Returns metadata, such as the path and user information about an SMB location.

describe_task(client, input, options \\ [])

Returns metadata about a task.

describe_task_execution(client, input, options \\ [])

Returns detailed metadata about a task that is being executed.

list_agents(client, input, options \\ [])

Returns a list of agents owned by an AWS account in the AWS Region specified in the request.

The returned list is ordered by agent Amazon Resource Name (ARN).

By default, this operation returns a maximum of 100 agents. This operation supports pagination that enables you to optionally reduce the number of agents returned in a response.

If you have more agents than are returned in a response (that is, the response returns only a truncated list of your agents), the response contains a marker that you can specify in your next request to fetch the next page of agents.

list_locations(client, input, options \\ [])

Returns a list of source and destination locations.

If you have more locations than are returned in a response (that is, the response returns only a truncated list of your agents), the response contains a token that you can specify in your next request to fetch the next page of locations.

list_tags_for_resource(client, input, options \\ [])

Returns all the tags associated with a specified resource.

list_task_executions(client, input, options \\ [])

Returns a list of executed tasks.

list_tasks(client, input, options \\ [])

Returns a list of all the tasks.

start_task_execution(client, input, options \\ [])

Starts a specific invocation of a task.

A TaskExecution value represents an individual run of a task. Each task can have at most one TaskExecution at a time.

TaskExecution has the following transition phases: INITIALIZING	PREPARING

TRANSFERRING	VERIFYING	SUCCESS/FAILURE.

For detailed information, see the Task Execution section in the Components and Terminology topic in the AWS DataSync User Guide.

tag_resource(client, input, options \\ [])

Applies a key-value pair to an AWS resource.

untag_resource(client, input, options \\ [])

Removes a tag from an AWS resource.

update_agent(client, input, options \\ [])

Updates the name of an agent.

update_task(client, input, options \\ [])

Updates the metadata associated with a task.