aws-elixir v0.6.0 AWS.DatabaseMigration

AWS Database Migration Service

AWS Database Migration Service (AWS DMS) can migrate your data to and from the most widely used commercial and open-source databases such as Oracle, PostgreSQL, Microsoft SQL Server, Amazon Redshift, MariaDB, Amazon Aurora, MySQL, and SAP Adaptive Server Enterprise (ASE).

The service supports homogeneous migrations such as Oracle to Oracle, as well as heterogeneous migrations between different database platforms, such as Oracle to MySQL or SQL Server to PostgreSQL.

For more information about AWS DMS, see What Is AWS Database Migration Service? in the AWS Database Migration User Guide.

Functions

add_tags_to_resource(client, input, options \\ [])

Adds metadata tags to an AWS DMS resource, including replication instance, endpoint, security group, and migration task.

These tags can also be used with cost allocation reporting to track cost associated with DMS resources, or used in a Condition statement in an IAM policy for DMS. For more information, see Tag data type description.

apply_pending_maintenance_action(client, input, options \\ [])

Applies a pending maintenance action to a resource (for example, to a replication instance).

cancel_replication_task_assessment_run(client, input, options \\ [])

Cancels a single premigration assessment run.

This operation prevents any individual assessments from running if they haven't started running. It also attempts to cancel any individual assessments that are currently running.

create_endpoint(client, input, options \\ [])

Creates an endpoint using the provided settings.

create_event_subscription(client, input, options \\ [])

Creates an AWS DMS event notification subscription.

You can specify the type of source (SourceType) you want to be notified of, provide a list of AWS DMS source IDs (SourceIds) that triggers the events, and provide a list of event categories (EventCategories) for events you want to be notified of. If you specify both the SourceType and SourceIds, such as SourceType = replication-instance and SourceIdentifier = my-replinstance, you will be notified of all the replication instance events for the specified source. If you specify a SourceType but don't specify a SourceIdentifier, you receive notice of the events for that source type for all your AWS DMS sources. If you don't specify either SourceType nor SourceIdentifier, you will be notified of events generated from all AWS DMS sources belonging to your customer account.

For more information about AWS DMS events, see Working with Events and Notifications in the AWS Database Migration Service User Guide.

create_replication_instance(client, input, options \\ [])

Creates the replication instance using the specified parameters.

AWS DMS requires that your account have certain roles with appropriate permissions before you can create a replication instance. For information on the required roles, see Creating the IAM Roles to Use With the AWS CLI and AWS DMS API. For information on the required permissions, see IAM Permissions Needed to Use AWS DMS.

create_replication_subnet_group(client, input, options \\ [])

Creates a replication subnet group given a list of the subnet IDs in a VPC.

create_replication_task(client, input, options \\ [])

Creates a replication task using the specified parameters.

delete_certificate(client, input, options \\ [])

Deletes the specified certificate.

delete_connection(client, input, options \\ [])

Deletes the connection between a replication instance and an endpoint.

delete_endpoint(client, input, options \\ [])

Deletes the specified endpoint.

All tasks associated with the endpoint must be deleted before you can delete the endpoint.

delete_event_subscription(client, input, options \\ [])

Deletes an AWS DMS event subscription.

delete_replication_instance(client, input, options \\ [])

Deletes the specified replication instance.

You must delete any migration tasks that are associated with the replication instance before you can delete it.

delete_replication_subnet_group(client, input, options \\ [])

Deletes a subnet group.

delete_replication_task(client, input, options \\ [])

Deletes the specified replication task.

delete_replication_task_assessment_run(client, input, options \\ [])

Deletes the record of a single premigration assessment run.

This operation removes all metadata that AWS DMS maintains about this assessment run. However, the operation leaves untouched all information about this assessment run that is stored in your Amazon S3 bucket.

describe_account_attributes(client, input, options \\ [])

Lists all of the AWS DMS attributes for a customer account.

These attributes include AWS DMS quotas for the account and a unique account identifier in a particular DMS region. DMS quotas include a list of resource quotas supported by the account, such as the number of replication instances allowed. The description for each resource quota, includes the quota name, current usage toward that quota, and the quota's maximum value. DMS uses the unique account identifier to name each artifact used by DMS in the given region.

This command does not take any parameters.

describe_applicable_individual_assessments(client, input, options \\ [])

Provides a list of individual assessments that you can specify for a new premigration assessment run, given one or more parameters.

If you specify an existing migration task, this operation provides the default individual assessments you can specify for that task. Otherwise, the specified parameters model elements of a possible migration task on which to base a premigration assessment run.

To use these migration task modeling parameters, you must specify an existing replication instance, a source database engine, a target database engine, and a migration type. This combination of parameters potentially limits the default individual assessments available for an assessment run created for a corresponding migration task.

If you specify no parameters, this operation provides a list of all possible individual assessments that you can specify for an assessment run. If you specify any one of the task modeling parameters, you must specify all of them or the operation cannot provide a list of individual assessments. The only parameter that you can specify alone is for an existing migration task. The specified task definition then determines the default list of individual assessments that you can specify in an assessment run for the task.

describe_certificates(client, input, options \\ [])

Provides a description of the certificate.

describe_connections(client, input, options \\ [])

Describes the status of the connections that have been made between the replication instance and an endpoint.

Connections are created when you test an endpoint.

describe_endpoint_types(client, input, options \\ [])

Returns information about the type of endpoints available.

describe_endpoints(client, input, options \\ [])

Returns information about the endpoints for your account in the current region.

describe_event_categories(client, input, options \\ [])

Lists categories for all event source types, or, if specified, for a specified source type.

You can see a list of the event categories and source types in Working with Events and Notifications in the AWS Database Migration Service User Guide.

describe_event_subscriptions(client, input, options \\ [])

Lists all the event subscriptions for a customer account.

The description of a subscription includes SubscriptionName, SNSTopicARN, CustomerID, SourceType, SourceID, CreationTime, and Status.

If you specify SubscriptionName, this action lists the description for that subscription.

describe_events(client, input, options \\ [])

Lists events for a given source identifier and source type.

You can also specify a start and end time. For more information on AWS DMS events, see Working with Events and Notifications in the AWS Database Migration User Guide.

describe_orderable_replication_instances(client, input, options \\ [])

Returns information about the replication instance types that can be created in the specified region.

describe_pending_maintenance_actions(client, input, options \\ [])

For internal use only

describe_refresh_schemas_status(client, input, options \\ [])

Returns the status of the RefreshSchemas operation.

describe_replication_instance_task_logs(client, input, options \\ [])

Returns information about the task logs for the specified task.

describe_replication_instances(client, input, options \\ [])

Returns information about replication instances for your account in the current region.

describe_replication_subnet_groups(client, input, options \\ [])

Returns information about the replication subnet groups.

describe_replication_task_assessment_results(client, input, options \\ [])

Returns the task assessment results from Amazon S3.

This action always returns the latest results.

describe_replication_task_assessment_runs(client, input, options \\ [])

Returns a paginated list of premigration assessment runs based on filter settings.

These filter settings can specify a combination of premigration assessment runs, migration tasks, replication instances, and assessment run status values.

This operation doesn't return information about individual assessments. For this information, see the DescribeReplicationTaskIndividualAssessments operation.

describe_replication_task_individual_assessments(client, input, options \\ [])

Returns a paginated list of individual assessments based on filter settings.

These filter settings can specify a combination of premigration assessment runs, migration tasks, and assessment status values.

describe_replication_tasks(client, input, options \\ [])

Returns information about replication tasks for your account in the current region.

describe_schemas(client, input, options \\ [])

Returns information about the schema for the specified endpoint.

describe_table_statistics(client, input, options \\ [])

Returns table statistics on the database migration task, including table name, rows inserted, rows updated, and rows deleted.

Note that the "last updated" column the DMS console only indicates the time that AWS DMS last updated the table statistics record for a table. It does not indicate the time of the last update to the table.

import_certificate(client, input, options \\ [])

Uploads the specified certificate.

list_tags_for_resource(client, input, options \\ [])

Lists all metadata tags attached to an AWS DMS resource, including replication instance, endpoint, security group, and migration task.

For more information, see Tag data type description.

modify_endpoint(client, input, options \\ [])

Modifies the specified endpoint.

modify_event_subscription(client, input, options \\ [])

Modifies an existing AWS DMS event notification subscription.

modify_replication_instance(client, input, options \\ [])

Modifies the replication instance to apply new settings.

You can change one or more parameters by specifying these parameters and the new values in the request.

Some settings are applied during the maintenance window.

modify_replication_subnet_group(client, input, options \\ [])

Modifies the settings for the specified replication subnet group.

modify_replication_task(client, input, options \\ [])

Modifies the specified replication task.

You can't modify the task endpoints. The task must be stopped before you can modify it.

For more information about AWS DMS tasks, see Working with Migration Tasks in the AWS Database Migration Service User Guide.

reboot_replication_instance(client, input, options \\ [])

Reboots a replication instance.

Rebooting results in a momentary outage, until the replication instance becomes available again.

refresh_schemas(client, input, options \\ [])

Populates the schema for the specified endpoint.

This is an asynchronous operation and can take several minutes. You can check the status of this operation by calling the DescribeRefreshSchemasStatus operation.

reload_tables(client, input, options \\ [])

Reloads the target database table with the source data.

remove_tags_from_resource(client, input, options \\ [])

Removes metadata tags from an AWS DMS resource, including replication instance, endpoint, security group, and migration task.

For more information, see Tag data type description.

start_replication_task(client, input, options \\ [])

Starts the replication task.

For more information about AWS DMS tasks, see Working with Migration Tasks in the AWS Database Migration Service User Guide.

start_replication_task_assessment(client, input, options \\ [])

Starts the replication task assessment for unsupported data types in the source database.

start_replication_task_assessment_run(client, input, options \\ [])

Starts a new premigration assessment run for one or more individual assessments of a migration task.

The assessments that you can specify depend on the source and target database engine and the migration type defined for the given task. To run this operation, your migration task must already be created. After you run this operation, you can review the status of each individual assessment. You can also run the migration task manually after the assessment run and its individual assessments complete.

stop_replication_task(client, input, options \\ [])

Stops the replication task.

test_connection(client, input, options \\ [])

Tests the connection between the replication instance and the endpoint.