aws-elixir v0.6.0 AWS.DeviceFarm

Welcome to the AWS Device Farm API documentation, which contains APIs for:

Testing on desktop browsers

Device Farm makes it possible for you to test your web applications on desktop browsers using Selenium.

The APIs for desktop browser testing contain TestGrid in their names. For more information, see Testing Web Applications on Selenium with Device Farm.

Testing on real mobile devices

Device Farm makes it possible for you to test apps on physical phones, tablets, and other devices in the cloud. For more information, see the Device Farm Developer Guide.

Functions

create_device_pool(client, input, options \\ [])

Creates a device pool.

create_instance_profile(client, input, options \\ [])

Creates a profile that can be applied to one or more private fleet device instances.

create_network_profile(client, input, options \\ [])

Creates a network profile.

create_project(client, input, options \\ [])

Creates a project.

create_remote_access_session(client, input, options \\ [])

Specifies and starts a remote access session.

create_test_grid_project(client, input, options \\ [])

Creates a Selenium testing project.

Projects are used to track TestGridSession instances.

create_test_grid_url(client, input, options \\ [])

Creates a signed, short-term URL that can be passed to a Selenium RemoteWebDriver constructor.

create_upload(client, input, options \\ [])

Uploads an app or test scripts.

create_v_p_c_e_configuration(client, input, options \\ [])

Creates a configuration record in Device Farm for your Amazon Virtual Private Cloud (VPC) endpoint.

delete_device_pool(client, input, options \\ [])

Deletes a device pool given the pool ARN.

Does not allow deletion of curated pools owned by the system.

delete_instance_profile(client, input, options \\ [])

Deletes a profile that can be applied to one or more private device instances.

delete_network_profile(client, input, options \\ [])

Deletes a network profile.

delete_project(client, input, options \\ [])

Deletes an AWS Device Farm project, given the project ARN.

Deleting this resource does not stop an in-progress run.

delete_remote_access_session(client, input, options \\ [])

Deletes a completed remote access session and its results.

delete_run(client, input, options \\ [])

Deletes the run, given the run ARN.

Deleting this resource does not stop an in-progress run.

delete_test_grid_project(client, input, options \\ [])

Deletes a Selenium testing project and all content generated under it.

You cannot undo this operation.

You cannot delete a project if it has active sessions.

delete_upload(client, input, options \\ [])

Deletes an upload given the upload ARN.

delete_v_p_c_e_configuration(client, input, options \\ [])

Deletes a configuration for your Amazon Virtual Private Cloud (VPC) endpoint.

get_account_settings(client, input, options \\ [])

Returns the number of unmetered iOS or unmetered Android devices that have been purchased by the account.

get_device(client, input, options \\ [])

Gets information about a unique device type.

get_device_instance(client, input, options \\ [])

Returns information about a device instance that belongs to a private device fleet.

get_device_pool(client, input, options \\ [])

Gets information about a device pool.

get_device_pool_compatibility(client, input, options \\ [])

Gets information about compatibility with a device pool.

get_instance_profile(client, input, options \\ [])

Returns information about the specified instance profile.

get_job(client, input, options \\ [])

Gets information about a job.

get_network_profile(client, input, options \\ [])

Returns information about a network profile.

get_offering_status(client, input, options \\ [])

Gets the current status and future status of all offerings purchased by an AWS account.

The response indicates how many offerings are currently available and the offerings that will be available in the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

get_project(client, input, options \\ [])

Gets information about a project.

get_remote_access_session(client, input, options \\ [])

Returns a link to a currently running remote access session.

get_run(client, input, options \\ [])

Gets information about a run.

get_suite(client, input, options \\ [])

Gets information about a suite.

get_test(client, input, options \\ [])

Gets information about a test.

get_test_grid_project(client, input, options \\ [])

Retrieves information about a Selenium testing project.

get_test_grid_session(client, input, options \\ [])

A session is an instance of a browser created through a RemoteWebDriver with the URL from CreateTestGridUrlResult$url.

You can use the following to look up sessions:

The session ARN (GetTestGridSessionRequest$sessionArn).

The project ARN and a session ID (GetTestGridSessionRequest$projectArn and GetTestGridSessionRequest$sessionId).

get_upload(client, input, options \\ [])

Gets information about an upload.

get_v_p_c_e_configuration(client, input, options \\ [])

Returns information about the configuration settings for your Amazon Virtual Private Cloud (VPC) endpoint.

install_to_remote_access_session(client, input, options \\ [])

Installs an application to the device in a remote access session.

For Android applications, the file must be in .apk format. For iOS applications, the file must be in .ipa format.

list_artifacts(client, input, options \\ [])

Gets information about artifacts.

list_device_instances(client, input, options \\ [])

Returns information about the private device instances associated with one or more AWS accounts.

list_device_pools(client, input, options \\ [])

Gets information about device pools.

list_devices(client, input, options \\ [])

Gets information about unique device types.

list_instance_profiles(client, input, options \\ [])

Returns information about all the instance profiles in an AWS account.

list_jobs(client, input, options \\ [])

Gets information about jobs for a given test run.

list_network_profiles(client, input, options \\ [])

Returns the list of available network profiles.

list_offering_promotions(client, input, options \\ [])

Returns a list of offering promotions.

Each offering promotion record contains the ID and description of the promotion. The API returns a NotEligible error if the caller is not permitted to invoke the operation. Contact aws-devicefarm-support@amazon.com if you must be able to invoke this operation.

list_offering_transactions(client, input, options \\ [])

Returns a list of all historical purchases, renewals, and system renewal transactions for an AWS account.

The list is paginated and ordered by a descending timestamp (most recent transactions are first). The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

list_offerings(client, input, options \\ [])

Returns a list of products or offerings that the user can manage through the API.

Each offering record indicates the recurring price per unit and the frequency for that offering. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

list_projects(client, input, options \\ [])

Gets information about projects.

list_remote_access_sessions(client, input, options \\ [])

Returns a list of all currently running remote access sessions.

list_runs(client, input, options \\ [])

Gets information about runs, given an AWS Device Farm project ARN.

list_samples(client, input, options \\ [])

Gets information about samples, given an AWS Device Farm job ARN.

list_suites(client, input, options \\ [])

Gets information about test suites for a given job.

list_tags_for_resource(client, input, options \\ [])

List the tags for an AWS Device Farm resource.

list_test_grid_projects(client, input, options \\ [])

Gets a list of all Selenium testing projects in your account.

list_test_grid_session_actions(client, input, options \\ [])

Returns a list of the actions taken in a TestGridSession.

list_test_grid_session_artifacts(client, input, options \\ [])

Retrieves a list of artifacts created during the session.

list_test_grid_sessions(client, input, options \\ [])

Retrieves a list of sessions for a TestGridProject.

list_tests(client, input, options \\ [])

Gets information about tests in a given test suite.

list_unique_problems(client, input, options \\ [])

Gets information about unique problems, such as exceptions or crashes.

Unique problems are defined as a single instance of an error across a run, job, or suite. For example, if a call in your application consistently raises an exception (OutOfBoundsException in MyActivity.java:386), ListUniqueProblems returns a single entry instead of many individual entries for that exception.

list_uploads(client, input, options \\ [])

Gets information about uploads, given an AWS Device Farm project ARN.

list_v_p_c_e_configurations(client, input, options \\ [])

Returns information about all Amazon Virtual Private Cloud (VPC) endpoint configurations in the AWS account.

purchase_offering(client, input, options \\ [])

Immediately purchases offerings for an AWS account.

Offerings renew with the latest total purchased quantity for an offering, unless the renewal was overridden. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

renew_offering(client, input, options \\ [])

Explicitly sets the quantity of devices to renew for an offering, starting from the effectiveDate of the next period.

The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

schedule_run(client, input, options \\ [])

Schedules a run.

stop_job(client, input, options \\ [])

Initiates a stop request for the current job.

AWS Device Farm immediately stops the job on the device where tests have not started. You are not billed for this device. On the device where tests have started, setup suite and teardown suite tests run to completion on the device. You are billed for setup, teardown, and any tests that were in progress or already completed.

stop_remote_access_session(client, input, options \\ [])

Ends a specified remote access session.

stop_run(client, input, options \\ [])

Initiates a stop request for the current test run.

AWS Device Farm immediately stops the run on devices where tests have not started. You are not billed for these devices. On devices where tests have started executing, setup suite and teardown suite tests run to completion on those devices. You are billed for setup, teardown, and any tests that were in progress or already completed.

tag_resource(client, input, options \\ [])

Associates the specified tags to a resource with the specified resourceArn.

If existing tags on a resource are not specified in the request parameters, they are not changed. When a resource is deleted, the tags associated with that resource are also deleted.

untag_resource(client, input, options \\ [])

Deletes the specified tags from a resource.

update_device_instance(client, input, options \\ [])

Updates information about a private device instance.

update_device_pool(client, input, options \\ [])

Modifies the name, description, and rules in a device pool given the attributes and the pool ARN.

Rule updates are all-or-nothing, meaning they can only be updated as a whole (or not at all).

update_instance_profile(client, input, options \\ [])

Updates information about an existing private device instance profile.

update_network_profile(client, input, options \\ [])

Updates the network profile.

update_project(client, input, options \\ [])

Modifies the specified project name, given the project ARN and a new name.

update_test_grid_project(client, input, options \\ [])

Change details of a project.

update_upload(client, input, options \\ [])

Updates an uploaded test spec.

update_v_p_c_e_configuration(client, input, options \\ [])

Updates information about an Amazon Virtual Private Cloud (VPC) endpoint configuration.