aws-elixir v0.6.0 AWS.DocDB

Amazon DocumentDB API documentation

Functions

add_tags_to_resource(client, input, options \\ [])

Adds metadata tags to an Amazon DocumentDB resource.

You can use these tags with cost allocation reporting to track costs that are associated with Amazon DocumentDB resources. or in a Condition statement in an AWS Identity and Access Management (IAM) policy for Amazon DocumentDB.

apply_pending_maintenance_action(client, input, options \\ [])

Applies a pending maintenance action to a resource (for example, to an Amazon DocumentDB instance).

copy_d_b_cluster_parameter_group(client, input, options \\ [])

Copies the specified cluster parameter group.

copy_d_b_cluster_snapshot(client, input, options \\ [])

Copies a snapshot of a cluster.

To copy a cluster snapshot from a shared manual cluster snapshot, SourceDBClusterSnapshotIdentifier must be the Amazon Resource Name (ARN) of the shared cluster snapshot. You can only copy a shared DB cluster snapshot, whether encrypted or not, in the same AWS Region.

To cancel the copy operation after it is in progress, delete the target cluster snapshot identified by TargetDBClusterSnapshotIdentifier while that cluster snapshot is in the copying status.

create_d_b_cluster(client, input, options \\ [])

Creates a new Amazon DocumentDB cluster.

create_d_b_cluster_parameter_group(client, input, options \\ [])

Creates a new cluster parameter group.

Parameters in a cluster parameter group apply to all of the instances in a cluster.

A cluster parameter group is initially created with the default parameters for the database engine used by instances in the cluster. In Amazon DocumentDB, you cannot make modifications directly to the default.docdb3.6 cluster parameter group. If your Amazon DocumentDB cluster is using the default cluster parameter group and you want to modify a value in it, you must first create a new parameter group or copy an existing parameter group, modify it, and then apply the modified parameter group to your cluster. For the new cluster parameter group and associated settings to take effect, you must then reboot the instances in the cluster without failover. For more information, see Modifying Amazon DocumentDB Cluster Parameter Groups.

create_d_b_cluster_snapshot(client, input, options \\ [])

Creates a snapshot of a cluster.

create_d_b_instance(client, input, options \\ [])

Creates a new instance.

create_d_b_subnet_group(client, input, options \\ [])

Creates a new subnet group.

subnet groups must contain at least one subnet in at least two Availability Zones in the AWS Region.

delete_d_b_cluster(client, input, options \\ [])

Deletes a previously provisioned cluster.

When you delete a cluster, all automated backups for that cluster are deleted and can't be recovered. Manual DB cluster snapshots of the specified cluster are not deleted.

delete_d_b_cluster_parameter_group(client, input, options \\ [])

Deletes a specified cluster parameter group.

The cluster parameter group to be deleted can't be associated with any clusters.

delete_d_b_cluster_snapshot(client, input, options \\ [])

Deletes a cluster snapshot.

If the snapshot is being copied, the copy operation is terminated.

The cluster snapshot must be in the available state to be deleted.

delete_d_b_instance(client, input, options \\ [])

Deletes a previously provisioned instance.

delete_d_b_subnet_group(client, input, options \\ [])

Deletes a subnet group.

The specified database subnet group must not be associated with any DB instances.

describe_certificates(client, input, options \\ [])

Returns a list of certificate authority (CA) certificates provided by Amazon DocumentDB for this AWS account.

describe_d_b_cluster_parameter_groups(client, input, options \\ [])

Returns a list of DBClusterParameterGroup descriptions.

If a DBClusterParameterGroupName parameter is specified, the list contains only the description of the specified cluster parameter group.

describe_d_b_cluster_parameters(client, input, options \\ [])

Returns the detailed parameter list for a particular cluster parameter group.

describe_d_b_cluster_snapshot_attributes(client, input, options \\ [])

Returns a list of cluster snapshot attribute names and values for a manual DB cluster snapshot.

When you share snapshots with other AWS accounts, DescribeDBClusterSnapshotAttributes returns the restore attribute and a list of IDs for the AWS accounts that are authorized to copy or restore the manual cluster snapshot. If all is included in the list of values for the restore attribute, then the manual cluster snapshot is public and can be copied or restored by all AWS accounts.

describe_d_b_cluster_snapshots(client, input, options \\ [])

Returns information about cluster snapshots.

This API operation supports pagination.

describe_d_b_clusters(client, input, options \\ [])

Returns information about provisioned Amazon DocumentDB clusters.

This API operation supports pagination. For certain management features such as cluster and instance lifecycle management, Amazon DocumentDB leverages operational technology that is shared with Amazon RDS and Amazon Neptune. Use the filterName=engine,Values=docdb filter parameter to return only Amazon DocumentDB clusters.

describe_d_b_engine_versions(client, input, options \\ [])

Returns a list of the available engines.

describe_d_b_instances(client, input, options \\ [])

Returns information about provisioned Amazon DocumentDB instances.

This API supports pagination.

describe_d_b_subnet_groups(client, input, options \\ [])

Returns a list of DBSubnetGroup descriptions.

If a DBSubnetGroupName is specified, the list will contain only the descriptions of the specified DBSubnetGroup.

describe_engine_default_cluster_parameters(client, input, options \\ [])

Returns the default engine and system parameter information for the cluster database engine.

describe_event_categories(client, input, options \\ [])

Displays a list of categories for all event source types, or, if specified, for a specified source type.

describe_events(client, input, options \\ [])

Returns events related to instances, security groups, snapshots, and DB parameter groups for the past 14 days.

You can obtain events specific to a particular DB instance, security group, snapshot, or parameter group by providing the name as a parameter. By default, the events of the past hour are returned.

describe_orderable_d_b_instance_options(client, input, options \\ [])

Returns a list of orderable instance options for the specified engine.

describe_pending_maintenance_actions(client, input, options \\ [])

Returns a list of resources (for example, instances) that have at least one pending maintenance action.

failover_d_b_cluster(client, input, options \\ [])

Forces a failover for a cluster.

A failover for a cluster promotes one of the Amazon DocumentDB replicas (read-only instances) in the cluster to be the primary instance (the cluster writer).

If the primary instance fails, Amazon DocumentDB automatically fails over to an Amazon DocumentDB replica, if one exists. You can force a failover when you want to simulate a failure of a primary instance for testing.

list_tags_for_resource(client, input, options \\ [])

Lists all tags on an Amazon DocumentDB resource.

modify_d_b_cluster(client, input, options \\ [])

Modifies a setting for an Amazon DocumentDB cluster.

You can change one or more database configuration parameters by specifying these parameters and the new values in the request.

modify_d_b_cluster_parameter_group(client, input, options \\ [])

Modifies the parameters of a cluster parameter group.

To modify more than one parameter, submit a list of the following: ParameterName, ParameterValue, and ApplyMethod. A maximum of 20 parameters can be modified in a single request.

Changes to dynamic parameters are applied immediately. Changes to static parameters require a reboot or maintenance window before the change can take effect.

After you create a cluster parameter group, you should wait at least 5 minutes before creating your first cluster that uses that cluster parameter group as the default parameter group. This allows Amazon DocumentDB to fully complete the create action before the parameter group is used as the default for a new cluster. This step is especially important for parameters that are critical when creating the default database for a cluster, such as the character set for the default database defined by the character_set_database parameter.

modify_d_b_cluster_snapshot_attribute(client, input, options \\ [])

Adds an attribute and values to, or removes an attribute and values from, a manual DB cluster snapshot.

To share a manual cluster snapshot with other AWS accounts, specify restore as the AttributeName, and use the ValuesToAdd parameter to add a list of IDs of the AWS accounts that are authorized to restore the manual cluster snapshot. Use the value all to make the manual cluster snapshot public, which means that it can be copied or restored by all AWS accounts. Do not add the all value for any manual DB cluster snapshots that contain private information that you don't want available to all AWS accounts. If a manual cluster snapshot is encrypted, it can be shared, but only by specifying a list of authorized AWS account IDs for the ValuesToAdd parameter. You can't use all as a value for that parameter in this case.

modify_d_b_instance(client, input, options \\ [])

Modifies settings for an instance.

You can change one or more database configuration parameters by specifying these parameters and the new values in the request.

modify_d_b_subnet_group(client, input, options \\ [])

Modifies an existing subnet group.

subnet groups must contain at least one subnet in at least two Availability Zones in the AWS Region.

reboot_d_b_instance(client, input, options \\ [])

You might need to reboot your instance, usually for maintenance reasons.

For example, if you make certain changes, or if you change the cluster parameter group that is associated with the instance, you must reboot the instance for the changes to take effect.

Rebooting an instance restarts the database engine service. Rebooting an instance results in a momentary outage, during which the instance status is set to rebooting.

remove_tags_from_resource(client, input, options \\ [])

Removes metadata tags from an Amazon DocumentDB resource.

reset_d_b_cluster_parameter_group(client, input, options \\ [])

Modifies the parameters of a cluster parameter group to the default value.

To reset specific parameters, submit a list of the following: ParameterName and ApplyMethod. To reset the entire cluster parameter group, specify the DBClusterParameterGroupName and ResetAllParameters parameters.

When you reset the entire group, dynamic parameters are updated immediately and static parameters are set to pending-reboot to take effect on the next DB instance reboot.

restore_d_b_cluster_from_snapshot(client, input, options \\ [])

Creates a new cluster from a snapshot or cluster snapshot.

If a snapshot is specified, the target cluster is created from the source DB snapshot with a default configuration and default security group.

If a cluster snapshot is specified, the target cluster is created from the source cluster restore point with the same configuration as the original source DB cluster, except that the new cluster is created with the default security group.

restore_d_b_cluster_to_point_in_time(client, input, options \\ [])

Restores a cluster to an arbitrary point in time.

Users can restore to any point in time before LatestRestorableTime for up to BackupRetentionPeriod days. The target cluster is created from the source cluster with the same configuration as the original cluster, except that the new cluster is created with the default security group.

start_d_b_cluster(client, input, options \\ [])

Restarts the stopped cluster that is specified by DBClusterIdentifier.

For more information, see Stopping and Starting an Amazon DocumentDB Cluster.

stop_d_b_cluster(client, input, options \\ [])

Stops the running cluster that is specified by DBClusterIdentifier.

The cluster must be in the available state. For more information, see Stopping and Starting an Amazon DocumentDB Cluster.