aws-elixir v0.6.0 AWS.ECR

Amazon Elastic Container Registry

Amazon Elastic Container Registry (Amazon ECR) is a managed container image registry service.

Customers can use the familiar Docker CLI, or their preferred client, to push, pull, and manage images. Amazon ECR provides a secure, scalable, and reliable registry for your Docker or Open Container Initiative (OCI) images. Amazon ECR supports private repositories with resource-based permissions using IAM so that specific users or Amazon EC2 instances can access repositories and images.

Functions

batch_check_layer_availability(client, input, options \\ [])

Checks the availability of one or more image layers in a repository.

When an image is pushed to a repository, each image layer is checked to verify if it has been uploaded before. If it has been uploaded, then the image layer is skipped.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

batch_delete_image(client, input, options \\ [])

Deletes a list of specified images within a repository.

Images are specified with either an imageTag or imageDigest.

You can remove a tag from an image by specifying the image's tag in your request. When you remove the last tag from an image, the image is deleted from your repository.

You can completely delete an image (and all of its tags) by specifying the image's digest in your request.

batch_get_image(client, input, options \\ [])

Gets detailed information for an image.

Images are specified with either an imageTag or imageDigest.

When an image is pulled, the BatchGetImage API is called once to retrieve the image manifest.

complete_layer_upload(client, input, options \\ [])

Informs Amazon ECR that the image layer upload has completed for a specified registry, repository name, and upload ID.

You can optionally provide a sha256 digest of the image layer for data validation purposes.

When an image is pushed, the CompleteLayerUpload API is called once per each new image layer to verify that the upload has completed.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

create_repository(client, input, options \\ [])

Creates a repository.

For more information, see Amazon ECR Repositories in the Amazon Elastic Container Registry User Guide.

delete_lifecycle_policy(client, input, options \\ [])

Deletes the lifecycle policy associated with the specified repository.

delete_repository(client, input, options \\ [])

Deletes a repository.

If the repository contains images, you must either delete all images in the repository or use the force option to delete the repository.

delete_repository_policy(client, input, options \\ [])

Deletes the repository policy associated with the specified repository.

describe_image_scan_findings(client, input, options \\ [])

Returns the scan findings for the specified image.

describe_images(client, input, options \\ [])

Returns metadata about the images in a repository.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

describe_repositories(client, input, options \\ [])

Describes image repositories in a registry.

get_authorization_token(client, input, options \\ [])

Retrieves an authorization token.

An authorization token represents your IAM authentication credentials and can be used to access any Amazon ECR registry that your IAM principal has access to. The authorization token is valid for 12 hours.

The authorizationToken returned is a base64 encoded string that can be decoded and used in a docker login command to authenticate to a registry. The AWS CLI offers an get-login-password command that simplifies the login process. For more information, see Registry Authentication in the Amazon Elastic Container Registry User Guide.

get_download_url_for_layer(client, input, options \\ [])

Retrieves the pre-signed Amazon S3 download URL corresponding to an image layer.

You can only get URLs for image layers that are referenced in an image.

When an image is pulled, the GetDownloadUrlForLayer API is called once per image layer that is not already cached.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

get_lifecycle_policy(client, input, options \\ [])

Retrieves the lifecycle policy for the specified repository.

get_lifecycle_policy_preview(client, input, options \\ [])

Retrieves the results of the lifecycle policy preview request for the specified repository.

get_repository_policy(client, input, options \\ [])

Retrieves the repository policy for the specified repository.

initiate_layer_upload(client, input, options \\ [])

Notifies Amazon ECR that you intend to upload an image layer.

When an image is pushed, the InitiateLayerUpload API is called once per image layer that has not already been uploaded. Whether or not an image layer has been uploaded is determined by the BatchCheckLayerAvailability API action.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

list_images(client, input, options \\ [])

Lists all the image IDs for the specified repository.

You can filter images based on whether or not they are tagged by using the tagStatus filter and specifying either TAGGED, UNTAGGED or ANY. For example, you can filter your results to return only UNTAGGED images and then pipe that result to a BatchDeleteImage operation to delete them. Or, you can filter your results to return only TAGGED images to list all of the tags in your repository.

list_tags_for_resource(client, input, options \\ [])

List the tags for an Amazon ECR resource.

put_image(client, input, options \\ [])

Creates or updates the image manifest and tags associated with an image.

When an image is pushed and all new image layers have been uploaded, the PutImage API is called once to create or update the image manifest and the tags associated with the image.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

put_image_scanning_configuration(client, input, options \\ [])

Updates the image scanning configuration for the specified repository.

put_image_tag_mutability(client, input, options \\ [])

Updates the image tag mutability settings for the specified repository.

For more information, see Image Tag Mutability in the Amazon Elastic Container Registry User Guide.

put_lifecycle_policy(client, input, options \\ [])

Creates or updates the lifecycle policy for the specified repository.

For more information, see Lifecycle Policy Template.

set_repository_policy(client, input, options \\ [])

Applies a repository policy to the specified repository to control access permissions.

For more information, see Amazon ECR Repository Policies in the Amazon Elastic Container Registry User Guide.

start_image_scan(client, input, options \\ [])

Starts an image vulnerability scan.

An image scan can only be started once per day on an individual image. This limit includes if an image was scanned on initial push. For more information, see Image Scanning in the Amazon Elastic Container Registry User Guide.

start_lifecycle_policy_preview(client, input, options \\ [])

Starts a preview of a lifecycle policy for the specified repository.

This allows you to see the results before associating the lifecycle policy with the repository.

tag_resource(client, input, options \\ [])

Adds specified tags to a resource with the specified ARN.

Existing tags on a resource are not changed if they are not specified in the request parameters.

untag_resource(client, input, options \\ [])

Deletes specified tags from a resource.

upload_layer_part(client, input, options \\ [])

Uploads an image layer part to Amazon ECR.

When an image is pushed, each new image layer is uploaded in parts. The maximum size of each image layer part can be 20971520 bytes (or about 20MB). The UploadLayerPart API is called once per each new image layer part.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.