aws-elixir v0.6.0 AWS.ElastiCache

Amazon ElastiCache

Amazon ElastiCache is a web service that makes it easier to set up, operate, and scale a distributed cache in the cloud.

With ElastiCache, customers get all of the benefits of a high-performance, in-memory cache with less of the administrative burden involved in launching and managing a distributed cache. The service makes setup, scaling, and cluster failure handling much simpler than in a self-managed cache deployment.

In addition, through integration with Amazon CloudWatch, customers get enhanced visibility into the key performance statistics associated with their cache and can receive alarms if a part of their cache runs hot.

Functions

add_tags_to_resource(client, input, options \\ [])

Adds up to 50 cost allocation tags to the named resource.

A cost allocation tag is a key-value pair where the key and value are case-sensitive. You can use cost allocation tags to categorize and track your AWS costs.

When you apply tags to your ElastiCache resources, AWS generates a cost allocation report as a comma-separated value (CSV) file with your usage and costs aggregated by your tags. You can apply tags that represent business categories (such as cost centers, application names, or owners) to organize your costs across multiple services. For more information, see Using Cost Allocation Tags in Amazon ElastiCache in the ElastiCache User Guide.

authorize_cache_security_group_ingress(client, input, options \\ [])

Allows network ingress to a cache security group.

Applications using ElastiCache must be running on Amazon EC2, and Amazon EC2 security groups are used as the authorization mechanism.

You cannot authorize ingress from an Amazon EC2 security group in one region to an ElastiCache cluster in another region.

batch_apply_update_action(client, input, options \\ [])

Apply the service update.

For more information on service updates and applying them, see Applying Service Updates.

batch_stop_update_action(client, input, options \\ [])

Stop the service update.

For more information on service updates and stopping them, see Stopping Service Updates.

complete_migration(client, input, options \\ [])

Complete the migration of data.

copy_snapshot(client, input, options \\ [])

Makes a copy of an existing snapshot.

This operation is valid for Redis only.

Users or groups that have permissions to use the CopySnapshot operation can create their own Amazon S3 buckets and copy snapshots to it. To control access to your snapshots, use an IAM policy to control who has the ability to use the CopySnapshot operation. For more information about using IAM to control the use of ElastiCache operations, see Exporting Snapshots and Authentication & Access Control.

You could receive the following error messages.

Error Messages

Error Message: The S3 bucket %s is outside of the region.

Solution: Create an Amazon S3 bucket in the same region as your snapshot. For more information, see Step 1: Create an Amazon S3 Bucket in the ElastiCache User Guide.

Error Message: The S3 bucket %s does not exist.

Solution: Create an Amazon S3 bucket in the same region as your snapshot. For more information, see Step 1: Create an Amazon S3 Bucket in the ElastiCache User Guide.

Error Message: The S3 bucket %s is not owned by the authenticated user.

Solution: Create an Amazon S3 bucket in the same region as your snapshot. For more information, see Step 1: Create an Amazon S3 Bucket in the ElastiCache User Guide.

Error Message: The authenticated user does not have sufficient permissions to perform the desired activity.

Solution: Contact your system administrator to get the needed permissions.

Error Message: The S3 bucket %s already contains an object with key %s.

Solution: Give the TargetSnapshotName a new and unique value. If exporting a snapshot, you could alternatively create a new Amazon S3 bucket and use this same value for TargetSnapshotName.

Error Message: ElastiCache has not been granted READ permissions %s on the S3 Bucket.

Solution: Add List and Read permissions on the bucket. For more information, see Step 2: Grant ElastiCache Access to Your Amazon S3 Bucket in the ElastiCache User Guide.

Error Message: ElastiCache has not been granted WRITE permissions %s on the S3 Bucket.

Solution: Add Upload/Delete permissions on the bucket. For more information, see Step 2: Grant ElastiCache Access to Your Amazon S3 Bucket in the ElastiCache User Guide.

Error Message: ElastiCache has not been granted READ_ACP permissions %s on the S3 Bucket.

Solution: Add View Permissions on the bucket. For more information, see Step 2: Grant ElastiCache Access to Your Amazon S3 Bucket in the ElastiCache User Guide.

create_cache_cluster(client, input, options \\ [])

Creates a cluster.

All nodes in the cluster run the same protocol-compliant cache engine software, either Memcached or Redis.

This operation is not supported for Redis (cluster mode enabled) clusters.

create_cache_parameter_group(client, input, options \\ [])

Creates a new Amazon ElastiCache cache parameter group.

An ElastiCache cache parameter group is a collection of parameters and their values that are applied to all of the nodes in any cluster or replication group using the CacheParameterGroup.

A newly created CacheParameterGroup is an exact duplicate of the default parameter group for the CacheParameterGroupFamily. To customize the newly created CacheParameterGroup you can change the values of specific parameters. For more information, see:

* ModifyCacheParameterGroup in the ElastiCache API Reference.

Parameters and Parameter Groups in the ElastiCache User Guide.

create_cache_security_group(client, input, options \\ [])

Creates a new cache security group.

Use a cache security group to control access to one or more clusters.

Cache security groups are only used when you are creating a cluster outside of an Amazon Virtual Private Cloud (Amazon VPC). If you are creating a cluster inside of a VPC, use a cache subnet group instead. For more information, see CreateCacheSubnetGroup.

create_cache_subnet_group(client, input, options \\ [])

Creates a new cache subnet group.

Use this parameter only when you are creating a cluster in an Amazon Virtual Private Cloud (Amazon VPC).

create_global_replication_group(client, input, options \\ [])

Global Datastore for Redis offers fully managed, fast, reliable and secure cross-region replication.

Using Global Datastore for Redis, you can create cross-region read replica clusters for ElastiCache for Redis to enable low-latency reads and disaster recovery across regions. For more information, see Replication Across Regions Using Global Datastore.

The GlobalReplicationGroupIdSuffix is the name of the Global Datastore.

The PrimaryReplicationGroupId represents the name of the primary cluster that accepts writes and will replicate updates to the secondary cluster.

create_replication_group(client, input, options \\ [])

Creates a Redis (cluster mode disabled) or a Redis (cluster mode enabled) replication group.

This API can be used to create a standalone regional replication group or a secondary replication group associated with a Global Datastore.

A Redis (cluster mode disabled) replication group is a collection of clusters, where one of the clusters is a read/write primary and the others are read-only replicas. Writes to the primary are asynchronously propagated to the replicas.

A Redis (cluster mode enabled) replication group is a collection of 1 to 90 node groups (shards). Each node group (shard) has one read/write primary node and up to 5 read-only replica nodes. Writes to the primary are asynchronously propagated to the replicas. Redis (cluster mode enabled) replication groups partition the data across node groups (shards).

When a Redis (cluster mode disabled) replication group has been successfully created, you can add one or more read replicas to it, up to a total of 5 read replicas. If you need to increase or decrease the number of node groups (console: shards), you can avail yourself of ElastiCache for Redis' scaling. For more information, see Scaling ElastiCache for Redis Clusters in the ElastiCache User Guide.

This operation is valid for Redis only.

create_snapshot(client, input, options \\ [])

Creates a copy of an entire cluster or replication group at a specific moment in time.

This operation is valid for Redis only.

create_user(client, input, options \\ [])

For Redis engine version 6.04 onwards: Creates a Redis user.

For more information, see Using Role Based Access Control (RBAC).

create_user_group(client, input, options \\ [])

For Redis engine version 6.04 onwards: Creates a Redis user group.

For more information, see Using Role Based Access Control (RBAC)

decrease_node_groups_in_global_replication_group(client, input, options \\ [])

Decreases the number of node groups in a Global Datastore

decrease_replica_count(client, input, options \\ [])

Dynamically decreases the number of replicas in a Redis (cluster mode disabled) replication group or the number of replica nodes in one or more node groups (shards) of a Redis (cluster mode enabled) replication group.

This operation is performed with no cluster down time.

delete_cache_cluster(client, input, options \\ [])

Deletes a previously provisioned cluster.

DeleteCacheCluster deletes all associated cache nodes, node endpoints and the cluster itself. When you receive a successful response from this operation, Amazon ElastiCache immediately begins deleting the cluster; you cannot cancel or revert this operation.

This operation is not valid for:

Redis (cluster mode enabled) clusters

A cluster that is the last read replica of a replication group

A node group (shard) that has Multi-AZ mode enabled

A cluster from a Redis (cluster mode enabled) replication group

A cluster that is not in the available state

delete_cache_parameter_group(client, input, options \\ [])

Deletes the specified cache parameter group.

You cannot delete a cache parameter group if it is associated with any cache clusters.

delete_cache_security_group(client, input, options \\ [])

Deletes a cache security group.

You cannot delete a cache security group if it is associated with any clusters.

delete_cache_subnet_group(client, input, options \\ [])

Deletes a cache subnet group.

You cannot delete a cache subnet group if it is associated with any clusters.

delete_global_replication_group(client, input, options \\ [])

Deleting a Global Datastore is a two-step process:

First, you must DisassociateGlobalReplicationGroup to remove the secondary clusters in the Global Datastore.

Once the Global Datastore contains only the primary cluster, you can use DeleteGlobalReplicationGroup API to delete the Global Datastore while retainining the primary cluster using Retain…= true.

Since the Global Datastore has only a primary cluster, you can delete the Global Datastore while retaining the primary by setting RetainPrimaryCluster=true.

When you receive a successful response from this operation, Amazon ElastiCache immediately begins deleting the selected resources; you cannot cancel or revert this operation.

delete_replication_group(client, input, options \\ [])

Deletes an existing replication group.

By default, this operation deletes the entire replication group, including the primary/primaries and all of the read replicas. If the replication group has only one primary, you can optionally delete only the read replicas, while retaining the primary by setting RetainPrimaryCluster=true.

When you receive a successful response from this operation, Amazon ElastiCache immediately begins deleting the selected resources; you cannot cancel or revert this operation.

This operation is valid for Redis only.

delete_snapshot(client, input, options \\ [])

Deletes an existing snapshot.

When you receive a successful response from this operation, ElastiCache immediately begins deleting the snapshot; you cannot cancel or revert this operation.

This operation is valid for Redis only.

delete_user(client, input, options \\ [])

For Redis engine version 6.04 onwards: Deletes a user.

The user will be removed from all user groups and in turn removed from all replication groups. For more information, see Using Role Based Access Control (RBAC).

delete_user_group(client, input, options \\ [])

For Redis engine version 6.04 onwards: Deletes a ser group.

The user group must first be disassociated from the replcation group before it can be deleted. For more information, see Using Role Based Access Control (RBAC).

describe_cache_clusters(client, input, options \\ [])

Returns information about all provisioned clusters if no cluster identifier is specified, or about a specific cache cluster if a cluster identifier is supplied.

By default, abbreviated information about the clusters is returned. You can use the optional ShowCacheNodeInfo flag to retrieve detailed information about the cache nodes associated with the clusters. These details include the DNS address and port for the cache node endpoint.

If the cluster is in the creating state, only cluster-level information is displayed until all of the nodes are successfully provisioned.

If the cluster is in the deleting state, only cluster-level information is displayed.

If cache nodes are currently being added to the cluster, node endpoint information and creation time for the additional nodes are not displayed until they are completely provisioned. When the cluster state is available, the cluster is ready for use.

If cache nodes are currently being removed from the cluster, no endpoint information for the removed nodes is displayed.

describe_cache_engine_versions(client, input, options \\ [])

Returns a list of the available cache engines and their versions.

describe_cache_parameter_groups(client, input, options \\ [])

Returns a list of cache parameter group descriptions.

If a cache parameter group name is specified, the list contains only the descriptions for that group.

describe_cache_parameters(client, input, options \\ [])

Returns the detailed parameter list for a particular cache parameter group.

describe_cache_security_groups(client, input, options \\ [])

Returns a list of cache security group descriptions.

If a cache security group name is specified, the list contains only the description of that group. This applicable only when you have ElastiCache in Classic setup

describe_cache_subnet_groups(client, input, options \\ [])

Returns a list of cache subnet group descriptions.

If a subnet group name is specified, the list contains only the description of that group. This is applicable only when you have ElastiCache in VPC setup. All ElastiCache clusters now launch in VPC by default.

describe_engine_default_parameters(client, input, options \\ [])

Returns the default engine and system parameter information for the specified cache engine.

describe_events(client, input, options \\ [])

Returns events related to clusters, cache security groups, and cache parameter groups.

You can obtain events specific to a particular cluster, cache security group, or cache parameter group by providing the name as a parameter.

By default, only the events occurring within the last hour are returned; however, you can retrieve up to 14 days' worth of events if necessary.

describe_global_replication_groups(client, input, options \\ [])

Returns information about a particular global replication group.

If no identifier is specified, returns information about all Global Datastores.

describe_replication_groups(client, input, options \\ [])

Returns information about a particular replication group.

If no identifier is specified, DescribeReplicationGroups returns information about all replication groups.

This operation is valid for Redis only.

describe_reserved_cache_nodes(client, input, options \\ [])

Returns information about reserved cache nodes for this account, or about a specified reserved cache node.

describe_reserved_cache_nodes_offerings(client, input, options \\ [])

Lists available reserved cache node offerings.

describe_service_updates(client, input, options \\ [])

Returns details of the service updates

describe_snapshots(client, input, options \\ [])

Returns information about cluster or replication group snapshots.

By default, DescribeSnapshots lists all of your snapshots; it can optionally describe a single snapshot, or just the snapshots associated with a particular cache cluster.

This operation is valid for Redis only.

describe_update_actions(client, input, options \\ [])

Returns details of the update actions

describe_user_groups(client, input, options \\ [])

Returns a list of user groups.

describe_users(client, input, options \\ [])

Returns a list of users.

disassociate_global_replication_group(client, input, options \\ [])

Remove a secondary cluster from the Global Datastore using the Global Datastore name.

The secondary cluster will no longer receive updates from the primary cluster, but will remain as a standalone cluster in that AWS region.

failover_global_replication_group(client, input, options \\ [])

Used to failover the primary region to a selected secondary region.

The selected secondary region will become primary, and all other clusters will become secondary.

increase_node_groups_in_global_replication_group(client, input, options \\ [])

Increase the number of node groups in the Global Datastore

increase_replica_count(client, input, options \\ [])

Dynamically increases the number of replics in a Redis (cluster mode disabled) replication group or the number of replica nodes in one or more node groups (shards) of a Redis (cluster mode enabled) replication group.

This operation is performed with no cluster down time.

list_allowed_node_type_modifications(client, input, options \\ [])

Lists all available node types that you can scale your Redis cluster's or replication group's current node type.

When you use the ModifyCacheCluster or ModifyReplicationGroup operations to scale your cluster or replication group, the value of the CacheNodeType parameter must be one of the node types returned by this operation.

list_tags_for_resource(client, input, options \\ [])

Lists all cost allocation tags currently on the named resource.

A cost allocation tag is a key-value pair where the key is case-sensitive and the value is optional. You can use cost allocation tags to categorize and track your AWS costs.

If the cluster is not in the available state, ListTagsForResource returns an error.

You can have a maximum of 50 cost allocation tags on an ElastiCache resource. For more information, see Monitoring Costs with Tags.

modify_cache_cluster(client, input, options \\ [])

Modifies the settings for a cluster.

You can use this operation to change one or more cluster configuration parameters by specifying the parameters and the new values.

modify_cache_parameter_group(client, input, options \\ [])

Modifies the parameters of a cache parameter group.

You can modify up to 20 parameters in a single request by submitting a list parameter name and value pairs.

modify_cache_subnet_group(client, input, options \\ [])

Modifies an existing cache subnet group.

modify_global_replication_group(client, input, options \\ [])

Modifies the settings for a Global Datastore.

modify_replication_group(client, input, options \\ [])

Modifies the settings for a replication group.

Scaling for Amazon ElastiCache for Redis (cluster mode enabled) in the ElastiCache User Guide

* ModifyReplicationGroupShardConfiguration in the ElastiCache API Reference

This operation is valid for Redis only.

modify_replication_group_shard_configuration(client, input, options \\ [])

Modifies a replication group's shards (node groups) by allowing you to add shards, remove shards, or rebalance the keyspaces among exisiting shards.

modify_user(client, input, options \\ [])

Changes user password(s) and/or access string.

modify_user_group(client, input, options \\ [])

Changes the list of users that belong to the user group.

purchase_reserved_cache_nodes_offering(client, input, options \\ [])

Allows you to purchase a reserved cache node offering.

rebalance_slots_in_global_replication_group(client, input, options \\ [])

Redistribute slots to ensure uniform distribution across existing shards in the cluster.

reboot_cache_cluster(client, input, options \\ [])

Reboots some, or all, of the cache nodes within a provisioned cluster.

This operation applies any modified cache parameter groups to the cluster. The reboot operation takes place as soon as possible, and results in a momentary outage to the cluster. During the reboot, the cluster status is set to REBOOTING.

The reboot causes the contents of the cache (for each cache node being rebooted) to be lost.

When the reboot is complete, a cluster event is created.

Rebooting a cluster is currently supported on Memcached and Redis (cluster mode disabled) clusters. Rebooting is not supported on Redis (cluster mode enabled) clusters.

If you make changes to parameters that require a Redis (cluster mode enabled) cluster reboot for the changes to be applied, see Rebooting a Cluster for an alternate process.

remove_tags_from_resource(client, input, options \\ [])

Removes the tags identified by the TagKeys list from the named resource.

reset_cache_parameter_group(client, input, options \\ [])

Modifies the parameters of a cache parameter group to the engine or system default value.

You can reset specific parameters by submitting a list of parameter names. To reset the entire cache parameter group, specify the ResetAllParameters and CacheParameterGroupName parameters.

revoke_cache_security_group_ingress(client, input, options \\ [])

Revokes ingress from a cache security group.

Use this operation to disallow access from an Amazon EC2 security group that had been previously authorized.

start_migration(client, input, options \\ [])

Start the migration of data.

test_failover(client, input, options \\ [])

Represents the input of a TestFailover operation which test automatic failover on a specified node group (called shard in the console) in a replication group (called cluster in the console).

Note the following

A customer can use this operation to test automatic failover on up to 5 shards (called node groups in the ElastiCache API and AWS CLI) in any rolling 24-hour period.

If calling this operation on shards in different clusters (called replication groups in the API and CLI), the calls can be made concurrently.

If calling this operation multiple times on different shards in the same Redis (cluster mode enabled) replication group, the first node replacement must complete before a subsequent call can be made.

To determine whether the node replacement is complete you can check Events using the Amazon ElastiCache console, the AWS CLI, or the ElastiCache API. Look for the following automatic failover related events, listed here in order of occurrance:

Replication group message: Test Failover API called for node group <node-group-id>

Cache cluster message: `Failover from master node

to replica node completed` 3. Replication group message: `Failover from master node to replica node completed` 4. Cache cluster message: `Recovering cache nodes ` 5. Cache cluster message: `Finished recovery for cache nodes ` For more information see: * [Viewing ElastiCache Events](https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/ECEvents.Viewing.html) in the *ElastiCache User Guide* * [DescribeEvents](https://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeEvents.html) in the ElastiCache API Reference Also see, [Testing Multi-AZ ](https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/AutoFailover.html#auto-failover-test) in the *ElastiCache User Guide*.