aws-elixir v0.6.0 AWS.ElasticInference

Elastic Inference public APIs.

Functions

describe_accelerator_offerings(client, input, options \\ [])

Describes the locations in which a given accelerator type or set of types is present in a given region.

describe_accelerator_types(client, options \\ [])

Describes the accelerator types available in a given region, as well as their characteristics, such as memory and throughput.

describe_accelerators(client, input, options \\ [])

Describes information over a provided set of accelerators belonging to an account.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns all tags of an Elastic Inference Accelerator.

tag_resource(client, resource_arn, input, options \\ [])

Adds the specified tags to an Elastic Inference Accelerator.

untag_resource(client, resource_arn, input, options \\ [])

Removes the specified tags from an Elastic Inference Accelerator.