aws-elixir v0.6.0 AWS.Elasticsearch

Amazon Elasticsearch Configuration Service

Use the Amazon Elasticsearch Configuration API to create, configure, and manage Elasticsearch domains.

For sample code that uses the Configuration API, see the Amazon Elasticsearch Service Developer Guide. The guide also contains sample code for sending signed HTTP requests to the Elasticsearch APIs.

The endpoint for configuration service requests is region-specific: es.region.amazonaws.com. For example, es.us-east-1.amazonaws.com. For a current list of supported regions and endpoints, see Regions and Endpoints.

Functions

accept_inbound_cross_cluster_search_connection(client, cross_cluster_search_connection_id, input, options \\ [])

Allows the destination domain owner to accept an inbound cross-cluster search connection request.

add_tags(client, input, options \\ [])

Attaches tags to an existing Elasticsearch domain.

Tags are a set of case-sensitive key value pairs. An Elasticsearch domain may have up to 10 tags. See Tagging Amazon Elasticsearch Service Domains for more information.

associate_package(client, domain_name, package_i_d, input, options \\ [])

Associates a package with an Amazon ES domain.

cancel_elasticsearch_service_software_update(client, input, options \\ [])

Cancels a scheduled service software update for an Amazon ES domain.

You can only perform this operation before the AutomatedUpdateDate and when the UpdateStatus is in the PENDING_UPDATE state.

create_elasticsearch_domain(client, input, options \\ [])

Creates a new Elasticsearch domain.

For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

create_outbound_cross_cluster_search_connection(client, input, options \\ [])

Creates a new cross-cluster search connection from a source domain to a destination domain.

create_package(client, input, options \\ [])

Create a package for use with Amazon ES domains.

delete_elasticsearch_domain(client, domain_name, input, options \\ [])

Permanently deletes the specified Elasticsearch domain and all of its data.

Once a domain is deleted, it cannot be recovered.

delete_elasticsearch_service_role(client, input, options \\ [])

Deletes the service-linked role that Elasticsearch Service uses to manage and maintain VPC domains.

Role deletion will fail if any existing VPC domains use the role. You must delete any such Elasticsearch domains before deleting the role. See Deleting Elasticsearch Service Role in VPC Endpoints for Amazon Elasticsearch Service Domains.

delete_inbound_cross_cluster_search_connection(client, cross_cluster_search_connection_id, input, options \\ [])

Allows the destination domain owner to delete an existing inbound cross-cluster search connection.

delete_outbound_cross_cluster_search_connection(client, cross_cluster_search_connection_id, input, options \\ [])

Allows the source domain owner to delete an existing outbound cross-cluster search connection.

delete_package(client, package_i_d, input, options \\ [])

Delete the package.

describe_elasticsearch_domain(client, domain_name, options \\ [])

Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN.

describe_elasticsearch_domain_config(client, domain_name, options \\ [])

Provides cluster configuration information about the specified Elasticsearch domain, such as the state, creation date, update version, and update date for cluster options.

describe_elasticsearch_domains(client, input, options \\ [])

Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN.

describe_elasticsearch_instance_type_limits(client, elasticsearch_version, instance_type, domain_name \\ nil, options \\ [])

Describe Elasticsearch Limits for a given InstanceType and ElasticsearchVersion.

When modifying existing Domain, specify the DomainName to know what Limits are supported for modifying.

describe_inbound_cross_cluster_search_connections(client, input, options \\ [])

Lists all the inbound cross-cluster search connections for a destination domain.

describe_outbound_cross_cluster_search_connections(client, input, options \\ [])

Lists all the outbound cross-cluster search connections for a source domain.

describe_packages(client, input, options \\ [])

Describes all packages available to Amazon ES.

Includes options for filtering, limiting the number of results, and pagination.

describe_reserved_elasticsearch_instance_offerings(client, max_results \\ nil, next_token \\ nil, reserved_elasticsearch_instance_offering_id \\ nil, options \\ [])

Lists available reserved Elasticsearch instance offerings.

describe_reserved_elasticsearch_instances(client, max_results \\ nil, next_token \\ nil, reserved_elasticsearch_instance_id \\ nil, options \\ [])

Returns information about reserved Elasticsearch instances for this account.

dissociate_package(client, domain_name, package_i_d, input, options \\ [])

Dissociates a package from the Amazon ES domain.

get_compatible_elasticsearch_versions(client, domain_name \\ nil, options \\ [])

Returns a list of upgrade compatible Elastisearch versions.

You can optionally pass a DomainName to get all upgrade compatible Elasticsearch versions for that specific domain.

get_upgrade_history(client, domain_name, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves the complete history of the last 10 upgrades that were performed on the domain.

get_upgrade_status(client, domain_name, options \\ [])

Retrieves the latest status of the last upgrade or upgrade eligibility check that was performed on the domain.

list_domain_names(client, options \\ [])

Returns the name of all Elasticsearch domains owned by the current user's account.

list_domains_for_package(client, package_i_d, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all Amazon ES domains associated with the package.

list_elasticsearch_instance_types(client, elasticsearch_version, domain_name \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

List all Elasticsearch instance types that are supported for given ElasticsearchVersion

list_elasticsearch_versions(client, max_results \\ nil, next_token \\ nil, options \\ [])

List all supported Elasticsearch versions

list_packages_for_domain(client, domain_name, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all packages associated with the Amazon ES domain.

list_tags(client, a_r_n, options \\ [])

Returns all tags for the given Elasticsearch domain.

purchase_reserved_elasticsearch_instance_offering(client, input, options \\ [])

Allows you to purchase reserved Elasticsearch instances.

reject_inbound_cross_cluster_search_connection(client, cross_cluster_search_connection_id, input, options \\ [])

Allows the destination domain owner to reject an inbound cross-cluster search connection request.

remove_tags(client, input, options \\ [])

Removes the specified set of tags from the specified Elasticsearch domain.

start_elasticsearch_service_software_update(client, input, options \\ [])

Schedules a service software update for an Amazon ES domain.

update_elasticsearch_domain_config(client, domain_name, input, options \\ [])

Modifies the cluster configuration of the specified Elasticsearch domain, setting as setting the instance type and the number of instances.

upgrade_elasticsearch_domain(client, input, options \\ [])

Allows you to either upgrade your domain or perform an Upgrade eligibility check to a compatible Elasticsearch version.