aws-elixir v0.6.0 AWS.FMS

AWS Firewall Manager

This is the AWS Firewall Manager API Reference.

This guide is for developers who need detailed information about the AWS Firewall Manager API actions, data types, and errors. For detailed information about AWS Firewall Manager features, see the AWS Firewall Manager Developer Guide.

Functions

associate_admin_account(client, input, options \\ [])

Sets the AWS Firewall Manager administrator account.

AWS Firewall Manager must be associated with the master account of your AWS organization or associated with a member account that has the appropriate permissions. If the account ID that you submit is not an AWS Organizations master account, AWS Firewall Manager will set the appropriate permissions for the given member account.

The account that you associate with AWS Firewall Manager is called the AWS Firewall Manager administrator account.

delete_apps_list(client, input, options \\ [])

Permanently deletes an AWS Firewall Manager applications list.

delete_notification_channel(client, input, options \\ [])

Deletes an AWS Firewall Manager association with the IAM role and the Amazon Simple Notification Service (SNS) topic that is used to record AWS Firewall Manager SNS logs.

delete_policy(client, input, options \\ [])

Permanently deletes an AWS Firewall Manager policy.

delete_protocols_list(client, input, options \\ [])

Permanently deletes an AWS Firewall Manager protocols list.

disassociate_admin_account(client, input, options \\ [])

Disassociates the account that has been set as the AWS Firewall Manager administrator account.

To set a different account as the administrator account, you must submit an AssociateAdminAccount request.

get_admin_account(client, input, options \\ [])

Returns the AWS Organizations master account that is associated with AWS Firewall Manager as the AWS Firewall Manager administrator.

get_apps_list(client, input, options \\ [])

Returns information about the specified AWS Firewall Manager applications list.

get_compliance_detail(client, input, options \\ [])

Returns detailed compliance information about the specified member account.

Details include resources that are in and out of compliance with the specified policy. Resources are considered noncompliant for AWS WAF and Shield Advanced policies if the specified policy has not been applied to them. Resources are considered noncompliant for security group policies if they are in scope of the policy, they violate one or more of the policy rules, and remediation is disabled or not possible.

get_notification_channel(client, input, options \\ [])

Information about the Amazon Simple Notification Service (SNS) topic that is used to record AWS Firewall Manager SNS logs.

get_policy(client, input, options \\ [])

Returns information about the specified AWS Firewall Manager policy.

get_protection_status(client, input, options \\ [])

If you created a Shield Advanced policy, returns policy-level attack summary information in the event of a potential DDoS attack.

Other policy types are currently unsupported.

get_protocols_list(client, input, options \\ [])

Returns information about the specified AWS Firewall Manager protocols list.

get_violation_details(client, input, options \\ [])

Retrieves violations for a resource based on the specified AWS Firewall Manager policy and AWS account.

list_apps_lists(client, input, options \\ [])

Returns an array of AppsListDataSummary objects.

list_compliance_status(client, input, options \\ [])

Returns an array of PolicyComplianceStatus objects.

Use PolicyComplianceStatus to get a summary of which member accounts are protected by the specified policy.

list_member_accounts(client, input, options \\ [])

Returns a MemberAccounts object that lists the member accounts in the administrator's AWS organization.

The ListMemberAccounts must be submitted by the account that is set as the AWS Firewall Manager administrator.

list_policies(client, input, options \\ [])

Returns an array of PolicySummary objects.

list_protocols_lists(client, input, options \\ [])

Returns an array of ProtocolsListDataSummary objects.

list_tags_for_resource(client, input, options \\ [])

Retrieves the list of tags for the specified AWS resource.

put_apps_list(client, input, options \\ [])

Creates an AWS Firewall Manager applications list.

put_notification_channel(client, input, options \\ [])

Designates the IAM role and Amazon Simple Notification Service (SNS) topic that AWS Firewall Manager uses to record SNS logs.

put_policy(client, input, options \\ [])

Creates an AWS Firewall Manager policy.

Firewall Manager provides the following types of policies:

A Shield Advanced policy, which applies Shield Advanced protection to specified accounts and resources

An AWS WAF policy (type WAFV2), which defines rule groups to run first in the corresponding AWS WAF web ACL and rule groups to run last in the web ACL.

An AWS WAF Classic policy (type WAF), which defines a rule group.

A security group policy, which manages VPC security groups across your AWS organization.

Each policy is specific to one of the types. If you want to enforce more than one policy type across accounts, create multiple policies. You can create multiple policies for each type.

You must be subscribed to Shield Advanced to create a Shield Advanced policy. For more information about subscribing to Shield Advanced, see CreateSubscription.

put_protocols_list(client, input, options \\ [])

Creates an AWS Firewall Manager protocols list.

tag_resource(client, input, options \\ [])

Adds one or more tags to an AWS resource.

untag_resource(client, input, options \\ [])

Removes one or more tags from an AWS resource.