aws-elixir v0.6.0 AWS.FraudDetector

This is the Amazon Fraud Detector API Reference.

This guide is for developers who need detailed information about Amazon Fraud Detector API actions, data types, and errors. For more information about Amazon Fraud Detector features, see the Amazon Fraud Detector User Guide.

Functions

batch_create_variable(client, input, options \\ [])

Creates a batch of variables.

batch_get_variable(client, input, options \\ [])

Gets a batch of variables.

create_detector_version(client, input, options \\ [])

Creates a detector version.

The detector version starts in a DRAFT status.

create_model(client, input, options \\ [])

Creates a model using the specified model type.

create_model_version(client, input, options \\ [])

Creates a version of the model using the specified model type and model id.

create_rule(client, input, options \\ [])

Creates a rule for use with the specified detector.

create_variable(client, input, options \\ [])

Creates a variable.

delete_detector(client, input, options \\ [])

Deletes the detector.

Before deleting a detector, you must first delete all detector versions and rule versions associated with the detector.

delete_detector_version(client, input, options \\ [])

Deletes the detector version.

You cannot delete detector versions that are in ACTIVE status.

delete_event(client, input, options \\ [])

Deletes the specified event.

delete_rule(client, input, options \\ [])

Deletes the rule.

You cannot delete a rule if it is used by an ACTIVE or INACTIVE detector version.

describe_detector(client, input, options \\ [])

Gets all versions for a specified detector.

describe_model_versions(client, input, options \\ [])

Gets all of the model versions for the specified model type or for the specified model type and model ID.

You can also get details for a single, specified model version.

get_detector_version(client, input, options \\ [])

Gets a particular detector version.

get_detectors(client, input, options \\ [])

Gets all detectors or a single detector if a detectorId is specified.

This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 5 and 10. To get the next page results, provide the pagination token from the GetDetectorsResponse as part of your request. A null pagination token fetches the records from the beginning.

get_entity_types(client, input, options \\ [])

Gets all entity types or a specific entity type if a name is specified.

This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 5 and 10. To get the next page results, provide the pagination token from the GetEntityTypesResponse as part of your request. A null pagination token fetches the records from the beginning.

get_event_prediction(client, input, options \\ [])

Evaluates an event against a detector version.

If a version ID is not provided, the detector’s (ACTIVE) version is used.

get_event_types(client, input, options \\ [])

Gets all event types or a specific event type if name is provided.

This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 5 and 10. To get the next page results, provide the pagination token from the GetEventTypesResponse as part of your request. A null pagination token fetches the records from the beginning.

get_external_models(client, input, options \\ [])

Gets the details for one or more Amazon SageMaker models that have been imported into the service.

This is a paginated API. If you provide a null maxResults, this actions retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 5 and 10. To get the next page results, provide the pagination token from the GetExternalModelsResult as part of your request. A null pagination token fetches the records from the beginning.

get_k_m_s_encryption_key(client, input, options \\ [])

Gets the encryption key if a Key Management Service (KMS) customer master key (CMK) has been specified to be used to encrypt content in Amazon Fraud Detector.

get_labels(client, input, options \\ [])

Gets all labels or a specific label if name is provided.

This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 50 records per page. If you provide a maxResults, the value must be between 10 and 50. To get the next page results, provide the pagination token from the GetGetLabelsResponse as part of your request. A null pagination token fetches the records from the beginning.

get_model_version(client, input, options \\ [])

Gets the details of the specified model version.

get_models(client, input, options \\ [])

Gets one or more models.

Gets all models for the AWS account if no model type and no model id provided. Gets all models for the AWS account and model type, if the model type is specified but model id is not provided. Gets a specific model if (model type, model id) tuple is specified.

This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 1 and 10. To get the next page results, provide the pagination token from the response as part of your request. A null pagination token fetches the records from the beginning.

get_outcomes(client, input, options \\ [])

Gets one or more outcomes.

This is a paginated API. If you provide a null maxResults, this actions retrieves a maximum of 100 records per page. If you provide a maxResults, the value must be between 50 and 100. To get the next page results, provide the pagination token from the GetOutcomesResult as part of your request. A null pagination token fetches the records from the beginning.

get_rules(client, input, options \\ [])

Get all rules for a detector (paginated) if ruleId and ruleVersion are not specified.

Gets all rules for the detector and the ruleId if present (paginated). Gets a specific rule if both the ruleId and the ruleVersion are specified.

This is a paginated API. Providing null maxResults results in retrieving maximum of 100 records per page. If you provide maxResults the value must be between 50 and 100. To get the next page result, a provide a pagination token from GetRulesResult as part of your request. Null pagination token fetches the records from the beginning.

get_variables(client, input, options \\ [])

Gets all of the variables or the specific variable.

This is a paginated API. Providing null maxSizePerPage results in retrieving maximum of 100 records per page. If you provide maxSizePerPage the value must be between 50 and 100. To get the next page result, a provide a pagination token from GetVariablesResult as part of your request. Null pagination token fetches the records from the beginning.

list_tags_for_resource(client, input, options \\ [])

Lists all tags associated with the resource.

This is a paginated API. To get the next page results, provide the pagination token from the response as part of your request. A null pagination token fetches the records from the beginning.

put_detector(client, input, options \\ [])

Creates or updates a detector.

put_entity_type(client, input, options \\ [])

Creates or updates an entity type.

An entity represents who is performing the event. As part of a fraud prediction, you pass the entity ID to indicate the specific entity who performed the event. An entity type classifies the entity. Example classifications include customer, merchant, or account.

put_event_type(client, input, options \\ [])

Creates or updates an event type.

An event is a business activity that is evaluated for fraud risk. With Amazon Fraud Detector, you generate fraud predictions for events. An event type defines the structure for an event sent to Amazon Fraud Detector. This includes the variables sent as part of the event, the entity performing the event (such as a customer), and the labels that classify the event. Example event types include online payment transactions, account registrations, and authentications.

put_external_model(client, input, options \\ [])

Creates or updates an Amazon SageMaker model endpoint.

You can also use this action to update the configuration of the model endpoint, including the IAM role and/or the mapped variables.

put_k_m_s_encryption_key(client, input, options \\ [])

Specifies the Key Management Service (KMS) customer master key (CMK) to be used to encrypt content in Amazon Fraud Detector.

put_label(client, input, options \\ [])

Creates or updates label.

A label classifies an event as fraudulent or legitimate. Labels are associated with event types and used to train supervised machine learning models in Amazon Fraud Detector.

put_outcome(client, input, options \\ [])

Creates or updates an outcome.

tag_resource(client, input, options \\ [])

Assigns tags to a resource.

untag_resource(client, input, options \\ [])

Removes tags from a resource.

update_detector_version(client, input, options \\ [])

Updates a detector version.

The detector version attributes that you can update include models, external model endpoints, rules, rule execution mode, and description. You can only update a DRAFT detector version.

update_detector_version_metadata(client, input, options \\ [])

Updates the detector version's description.

You can update the metadata for any detector version (DRAFT, ACTIVE, or INACTIVE).

update_detector_version_status(client, input, options \\ [])

Updates the detector version’s status.

You can perform the following promotions or demotions using UpdateDetectorVersionStatus: DRAFT to ACTIVE, ACTIVE to INACTIVE, and INACTIVE to ACTIVE.

update_model(client, input, options \\ [])

Updates a model.

You can update the description attribute using this action.

update_model_version(client, input, options \\ [])

Updates a model version.

Updating a model version retrains an existing model version using updated training data and produces a new minor version of the model. You can update the training data set location and data access role attributes using this action. This action creates and trains a new minor version of the model, for example version 1.01, 1.02, 1.03.

update_model_version_status(client, input, options \\ [])

Updates the status of a model version.

You can perform the following status updates:

Change the TRAINING_COMPLETE status to ACTIVE.

Change ACTIVEto INACTIVE.

update_rule_metadata(client, input, options \\ [])

Updates a rule's metadata.

The description attribute can be updated.

update_rule_version(client, input, options \\ [])

Updates a rule version resulting in a new rule version.

Updates a rule version resulting in a new rule version (version 1, 2, 3 ...).

update_variable(client, input, options \\ [])

Updates a variable.