aws-elixir v0.6.0 AWS.Greengrass

AWS IoT Greengrass seamlessly extends AWS onto physical devices so they can act locally on the data they generate, while still using the cloud for management, analytics, and durable storage.

AWS IoT Greengrass ensures your devices can respond quickly to local events and operate with intermittent connectivity. AWS IoT Greengrass minimizes the cost of transmitting data to the cloud by allowing you to author AWS Lambda functions that execute locally.

Functions

associate_role_to_group(client, group_id, input, options \\ [])

Associates a role with a group.

Your Greengrass core will use the role to access AWS cloud services. The role's permissions should allow Greengrass core Lambda functions to perform actions against the cloud.

associate_service_role_to_account(client, input, options \\ [])

Associates a role with your account.

AWS IoT Greengrass will use the role to access your Lambda functions and AWS IoT resources. This is necessary for deployments to succeed. The role must have at least minimum permissions in the policy ''AWSGreengrassResourceAccessRolePolicy''.

create_connector_definition(client, input, options \\ [])

Creates a connector definition.

You may provide the initial version of the connector definition now or use ''CreateConnectorDefinitionVersion'' at a later time.

create_connector_definition_version(client, connector_definition_id, input, options \\ [])

Creates a version of a connector definition which has already been defined.

create_core_definition(client, input, options \\ [])

Creates a core definition.

You may provide the initial version of the core definition now or use ''CreateCoreDefinitionVersion'' at a later time. Greengrass groups must each contain exactly one Greengrass core.

create_core_definition_version(client, core_definition_id, input, options \\ [])

Creates a version of a core definition that has already been defined.

Greengrass groups must each contain exactly one Greengrass core.

create_deployment(client, group_id, input, options \\ [])

Creates a deployment.

''CreateDeployment'' requests are idempotent with respect to the ''X-Amzn-Client-Token'' token and the request parameters.

create_device_definition(client, input, options \\ [])

Creates a device definition.

You may provide the initial version of the device definition now or use ''CreateDeviceDefinitionVersion'' at a later time.

create_device_definition_version(client, device_definition_id, input, options \\ [])

Creates a version of a device definition that has already been defined.

create_function_definition(client, input, options \\ [])

Creates a Lambda function definition which contains a list of Lambda functions and their configurations to be used in a group.

You can create an initial version of the definition by providing a list of Lambda functions and their configurations now, or use ''CreateFunctionDefinitionVersion'' later.

create_function_definition_version(client, function_definition_id, input, options \\ [])

Creates a version of a Lambda function definition that has already been defined.

create_group(client, input, options \\ [])

Creates a group.

You may provide the initial version of the group or use ''CreateGroupVersion'' at a later time. Tip: You can use the ''gg_group_setup'' package (https://github.com/awslabs/aws-greengrass-group-setup) as a library or command-line application to create and deploy Greengrass groups.

create_group_certificate_authority(client, group_id, input, options \\ [])

Creates a CA for the group.

If a CA already exists, it will rotate the existing CA.

create_group_version(client, group_id, input, options \\ [])

Creates a version of a group which has already been defined.

create_logger_definition(client, input, options \\ [])

Creates a logger definition.

You may provide the initial version of the logger definition now or use ''CreateLoggerDefinitionVersion'' at a later time.

create_logger_definition_version(client, logger_definition_id, input, options \\ [])

Creates a version of a logger definition that has already been defined.

create_resource_definition(client, input, options \\ [])

Creates a resource definition which contains a list of resources to be used in a group.

You can create an initial version of the definition by providing a list of resources now, or use ''CreateResourceDefinitionVersion'' later.

create_resource_definition_version(client, resource_definition_id, input, options \\ [])

Creates a version of a resource definition that has already been defined.

create_software_update_job(client, input, options \\ [])

Creates a software update for a core or group of cores (specified as an IoT thing group.) Use this to update the OTA Agent as well as the Greengrass core software.

It makes use of the IoT Jobs feature which provides additional commands to manage a Greengrass core software update job.

create_subscription_definition(client, input, options \\ [])

Creates a subscription definition.

You may provide the initial version of the subscription definition now or use ''CreateSubscriptionDefinitionVersion'' at a later time.

create_subscription_definition_version(client, subscription_definition_id, input, options \\ [])

Creates a version of a subscription definition which has already been defined.

delete_connector_definition(client, connector_definition_id, input, options \\ [])

Deletes a connector definition.

delete_core_definition(client, core_definition_id, input, options \\ [])

Deletes a core definition.

delete_device_definition(client, device_definition_id, input, options \\ [])

Deletes a device definition.

delete_function_definition(client, function_definition_id, input, options \\ [])

Deletes a Lambda function definition.

delete_group(client, group_id, input, options \\ [])

Deletes a group.

delete_logger_definition(client, logger_definition_id, input, options \\ [])

Deletes a logger definition.

delete_resource_definition(client, resource_definition_id, input, options \\ [])

Deletes a resource definition.

delete_subscription_definition(client, subscription_definition_id, input, options \\ [])

Deletes a subscription definition.

disassociate_role_from_group(client, group_id, input, options \\ [])

Disassociates the role from a group.

disassociate_service_role_from_account(client, input, options \\ [])

Disassociates the service role from your account.

Without a service role, deployments will not work.

get_associated_role(client, group_id, options \\ [])

Retrieves the role associated with a particular group.

get_bulk_deployment_status(client, bulk_deployment_id, options \\ [])

Returns the status of a bulk deployment.

get_connectivity_info(client, thing_name, options \\ [])

Retrieves the connectivity information for a core.

get_connector_definition(client, connector_definition_id, options \\ [])

Retrieves information about a connector definition.

get_connector_definition_version(client, connector_definition_id, connector_definition_version_id, next_token \\ nil, options \\ [])

Retrieves information about a connector definition version, including the connectors that the version contains.

Connectors are prebuilt modules that interact with local infrastructure, device protocols, AWS, and other cloud services.

get_core_definition(client, core_definition_id, options \\ [])

Retrieves information about a core definition version.

get_core_definition_version(client, core_definition_id, core_definition_version_id, options \\ [])

Retrieves information about a core definition version.

get_deployment_status(client, deployment_id, group_id, options \\ [])

Returns the status of a deployment.

get_device_definition(client, device_definition_id, options \\ [])

Retrieves information about a device definition.

get_device_definition_version(client, device_definition_id, device_definition_version_id, next_token \\ nil, options \\ [])

Retrieves information about a device definition version.

get_function_definition(client, function_definition_id, options \\ [])

Retrieves information about a Lambda function definition, including its creation time and latest version.

get_function_definition_version(client, function_definition_id, function_definition_version_id, next_token \\ nil, options \\ [])

Retrieves information about a Lambda function definition version, including which Lambda functions are included in the version and their configurations.

get_group(client, group_id, options \\ [])

Retrieves information about a group.

get_group_certificate_authority(client, certificate_authority_id, group_id, options \\ [])

Retreives the CA associated with a group.

Returns the public key of the CA.

get_group_certificate_configuration(client, group_id, options \\ [])

Retrieves the current configuration for the CA used by the group.

get_group_version(client, group_id, group_version_id, options \\ [])

Retrieves information about a group version.

get_logger_definition(client, logger_definition_id, options \\ [])

Retrieves information about a logger definition.

get_logger_definition_version(client, logger_definition_id, logger_definition_version_id, next_token \\ nil, options \\ [])

Retrieves information about a logger definition version.

get_resource_definition(client, resource_definition_id, options \\ [])

Retrieves information about a resource definition, including its creation time and latest version.

get_resource_definition_version(client, resource_definition_id, resource_definition_version_id, options \\ [])

Retrieves information about a resource definition version, including which resources are included in the version.

get_service_role_for_account(client, options \\ [])

Retrieves the service role that is attached to your account.

get_subscription_definition(client, subscription_definition_id, options \\ [])

Retrieves information about a subscription definition.

get_subscription_definition_version(client, subscription_definition_id, subscription_definition_version_id, next_token \\ nil, options \\ [])

Retrieves information about a subscription definition version.

get_thing_runtime_configuration(client, thing_name, options \\ [])

Get the runtime configuration of a thing.

list_bulk_deployment_detailed_reports(client, bulk_deployment_id, max_results \\ nil, next_token \\ nil, options \\ [])

Gets a paginated list of the deployments that have been started in a bulk deployment operation, and their current deployment status.

list_bulk_deployments(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of bulk deployments.

list_connector_definition_versions(client, connector_definition_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the versions of a connector definition, which are containers for connectors.

Connectors run on the Greengrass core and contain built-in integration with local infrastructure, device protocols, AWS, and other cloud services.

list_connector_definitions(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of connector definitions.

list_core_definition_versions(client, core_definition_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the versions of a core definition.

list_core_definitions(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of core definitions.

list_deployments(client, group_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a history of deployments for the group.

list_device_definition_versions(client, device_definition_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the versions of a device definition.

list_device_definitions(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of device definitions.

list_function_definition_versions(client, function_definition_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the versions of a Lambda function definition.

list_function_definitions(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of Lambda function definitions.

list_group_certificate_authorities(client, group_id, options \\ [])

Retrieves the current CAs for a group.

list_group_versions(client, group_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the versions of a group.

list_groups(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of groups.

list_logger_definition_versions(client, logger_definition_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the versions of a logger definition.

list_logger_definitions(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of logger definitions.

list_resource_definition_versions(client, resource_definition_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the versions of a resource definition.

list_resource_definitions(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of resource definitions.

list_subscription_definition_versions(client, subscription_definition_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the versions of a subscription definition.

list_subscription_definitions(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of subscription definitions.

list_tags_for_resource(client, resource_arn, options \\ [])

Retrieves a list of resource tags for a resource arn.

reset_deployments(client, group_id, input, options \\ [])

Resets a group's deployments.

start_bulk_deployment(client, input, options \\ [])

Deploys multiple groups in one operation.

This action starts the bulk deployment of a specified set of group versions. Each group version deployment will be triggered with an adaptive rate that has a fixed upper limit. We recommend that you include an ''X-Amzn-Client-Token'' token in every ''StartBulkDeployment'' request. These requests are idempotent with respect to the token and the request parameters.

stop_bulk_deployment(client, bulk_deployment_id, input, options \\ [])

Stops the execution of a bulk deployment.

This action returns a status of ''Stopping'' until the deployment is stopped. You cannot start a new bulk deployment while a previous deployment is in the ''Stopping'' state. This action doesn't rollback completed deployments or cancel pending deployments.

tag_resource(client, resource_arn, input, options \\ [])

Adds tags to a Greengrass resource.

Valid resources are 'Group', 'ConnectorDefinition', 'CoreDefinition', 'DeviceDefinition', 'FunctionDefinition', 'LoggerDefinition', 'SubscriptionDefinition', 'ResourceDefinition', and 'BulkDeployment'.

untag_resource(client, resource_arn, input, options \\ [])

Remove resource tags from a Greengrass Resource.

update_connectivity_info(client, thing_name, input, options \\ [])

Updates the connectivity information for the core.

Any devices that belong to the group which has this core will receive this information in order to find the location of the core and connect to it.

update_connector_definition(client, connector_definition_id, input, options \\ [])

Updates a connector definition.

update_core_definition(client, core_definition_id, input, options \\ [])

Updates a core definition.

update_device_definition(client, device_definition_id, input, options \\ [])

Updates a device definition.

update_function_definition(client, function_definition_id, input, options \\ [])

Updates a Lambda function definition.

update_group(client, group_id, input, options \\ [])

Updates a group.

update_group_certificate_configuration(client, group_id, input, options \\ [])

Updates the Certificate expiry time for a group.

update_logger_definition(client, logger_definition_id, input, options \\ [])

Updates a logger definition.

update_resource_definition(client, resource_definition_id, input, options \\ [])

Updates a resource definition.

update_subscription_definition(client, subscription_definition_id, input, options \\ [])

Updates a subscription definition.

update_thing_runtime_configuration(client, thing_name, input, options \\ [])

Updates the runtime configuration of a thing.