aws-elixir v0.6.0 AWS.GroundStation

Welcome to the AWS Ground Station API Reference.

AWS Ground Station is a fully managed service that enables you to control satellite communications, downlink and process satellite data, and scale your satellite operations efficiently and cost-effectively without having to build or manage your own ground station infrastructure.

Functions

cancel_contact(client, contact_id, input, options \\ [])

Cancels a contact with a specified contact ID.

create_config(client, input, options \\ [])

Creates a Config with the specified configData parameters.

Only one type of configData can be specified.

create_dataflow_endpoint_group(client, input, options \\ [])

Creates a DataflowEndpoint group containing the specified list of DataflowEndpoint objects.

The name field in each endpoint is used in your mission profile DataflowEndpointConfig to specify which endpoints to use during a contact.

When a contact uses multiple DataflowEndpointConfig objects, each Config must match a DataflowEndpoint in the same group.

create_mission_profile(client, input, options \\ [])

Creates a mission profile.

dataflowEdges is a list of lists of strings. Each lower level list of strings has two elements: a from ARN and a to ARN.

delete_config(client, config_id, config_type, input, options \\ [])

Deletes a Config.

delete_dataflow_endpoint_group(client, dataflow_endpoint_group_id, input, options \\ [])

Deletes a dataflow endpoint group.

delete_mission_profile(client, mission_profile_id, input, options \\ [])

Deletes a mission profile.

describe_contact(client, contact_id, options \\ [])

Describes an existing contact.

get_config(client, config_id, config_type, options \\ [])

Returns Config information.

Only one Config response can be returned.

get_dataflow_endpoint_group(client, dataflow_endpoint_group_id, options \\ [])

Returns the dataflow endpoint group.

get_minute_usage(client, input, options \\ [])

Returns the number of minutes used by account.

get_mission_profile(client, mission_profile_id, options \\ [])

Returns a mission profile.

get_satellite(client, satellite_id, options \\ [])

Returns a satellite.

list_configs(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of Config objects.

list_contacts(client, input, options \\ [])

Returns a list of contacts.

If statusList contains AVAILABLE, the request must include groundStation, missionprofileArn, and satelliteArn.

list_dataflow_endpoint_groups(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of DataflowEndpoint groups.

list_ground_stations(client, max_results \\ nil, next_token \\ nil, satellite_id \\ nil, options \\ [])

Returns a list of ground stations.

list_mission_profiles(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of mission profiles.

list_satellites(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of satellites.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns a list of tags for a specified resource.

reserve_contact(client, input, options \\ [])

Reserves a contact using specified parameters.

tag_resource(client, resource_arn, input, options \\ [])

Assigns a tag to a resource.

untag_resource(client, resource_arn, input, options \\ [])

Deassigns a resource tag.

update_config(client, config_id, config_type, input, options \\ [])

Updates the Config used when scheduling contacts.

Updating a Config will not update the execution parameters for existing future contacts scheduled with this Config.

update_mission_profile(client, mission_profile_id, input, options \\ [])

Updates a mission profile.

Updating a mission profile will not update the execution parameters for existing future contacts.