aws-elixir v0.6.0 AWS.GuardDuty

Amazon GuardDuty is a continuous security monitoring service that analyzes and processes the following data sources: VPC Flow Logs, AWS CloudTrail event logs, and DNS logs.

It uses threat intelligence feeds (such as lists of malicious IPs and domains) and machine learning to identify unexpected, potentially unauthorized, and malicious activity within your AWS environment. This can include issues like escalations of privileges, uses of exposed credentials, or communication with malicious IPs, URLs, or domains. For example, GuardDuty can detect compromised EC2 instances that serve malware or mine bitcoin.

GuardDuty also monitors AWS account access behavior for signs of compromise. Some examples of this are unauthorized infrastructure deployments such as EC2 instances deployed in a Region that has never been used, or unusual API calls like a password policy change to reduce password strength.

GuardDuty informs you of the status of your AWS environment by producing security findings that you can view in the GuardDuty console or through Amazon CloudWatch events. For more information, see the Amazon GuardDuty User Guide .

Functions

accept_invitation(client, detector_id, input, options \\ [])

Accepts the invitation to be monitored by a master GuardDuty account.

archive_findings(client, detector_id, input, options \\ [])

Archives GuardDuty findings that are specified by the list of finding IDs.

Only the master account can archive findings. Member accounts don't have permission to archive findings from their accounts.

create_detector(client, input, options \\ [])

Creates a single Amazon GuardDuty detector.

A detector is a resource that represents the GuardDuty service. To start using GuardDuty, you must create a detector in each Region where you enable the service. You can have only one detector per account per Region. All data sources are enabled in a new detector by default.

create_filter(client, detector_id, input, options \\ [])

Creates a filter using the specified finding criteria.

create_i_p_set(client, detector_id, input, options \\ [])

Creates a new IPSet, which is called a trusted IP list in the console user interface.

An IPSet is a list of IP addresses that are trusted for secure communication with AWS infrastructure and applications. GuardDuty doesn't generate findings for IP addresses that are included in IPSets. Only users from the master account can use this operation.

create_members(client, detector_id, input, options \\ [])

Creates member accounts of the current AWS account by specifying a list of AWS account IDs.

This step is a prerequisite for managing the associated member accounts either by invitation or through an organization.

When using Create Members as an organizations delegated administrator this action will enable GuardDuty in the added member accounts, with the exception of the organization master account, which must enable GuardDuty prior to being added as a member.

If you are adding accounts by invitation use this action after GuardDuty has been enabled in potential member accounts and before using Invite Members .

create_publishing_destination(client, detector_id, input, options \\ [])

Creates a publishing destination to export findings to.

The resource to export findings to must exist before you use this operation.

create_sample_findings(client, detector_id, input, options \\ [])

Generates example findings of types specified by the list of finding types.

If 'NULL' is specified for findingTypes, the API generates example findings of all supported finding types.

create_threat_intel_set(client, detector_id, input, options \\ [])

Creates a new ThreatIntelSet.

ThreatIntelSets consist of known malicious IP addresses. GuardDuty generates findings based on ThreatIntelSets. Only users of the master account can use this operation.

decline_invitations(client, input, options \\ [])

Declines invitations sent to the current member account by AWS accounts specified by their account IDs.

delete_detector(client, detector_id, input, options \\ [])

Deletes an Amazon GuardDuty detector that is specified by the detector ID.

delete_filter(client, detector_id, filter_name, input, options \\ [])

Deletes the filter specified by the filter name.

delete_i_p_set(client, detector_id, ip_set_id, input, options \\ [])

Deletes the IPSet specified by the ipSetId.

IPSets are called trusted IP lists in the console user interface.

delete_invitations(client, input, options \\ [])

Deletes invitations sent to the current member account by AWS accounts specified by their account IDs.

delete_members(client, detector_id, input, options \\ [])

Deletes GuardDuty member accounts (to the current GuardDuty master account) specified by the account IDs.

delete_publishing_destination(client, destination_id, detector_id, input, options \\ [])

Deletes the publishing definition with the specified destinationId.

delete_threat_intel_set(client, detector_id, threat_intel_set_id, input, options \\ [])

Deletes the ThreatIntelSet specified by the ThreatIntelSet ID.

describe_organization_configuration(client, detector_id, options \\ [])

Returns information about the account selected as the delegated administrator for GuardDuty.

describe_publishing_destination(client, destination_id, detector_id, options \\ [])

Returns information about the publishing destination specified by the provided destinationId.

disable_organization_admin_account(client, input, options \\ [])

Disables an AWS account within the Organization as the GuardDuty delegated administrator.

disassociate_from_master_account(client, detector_id, input, options \\ [])

Disassociates the current GuardDuty member account from its master account.

disassociate_members(client, detector_id, input, options \\ [])

Disassociates GuardDuty member accounts (to the current GuardDuty master account) specified by the account IDs.

enable_organization_admin_account(client, input, options \\ [])

Enables an AWS account within the organization as the GuardDuty delegated administrator.

get_detector(client, detector_id, options \\ [])

Retrieves an Amazon GuardDuty detector specified by the detectorId.

get_filter(client, detector_id, filter_name, options \\ [])

Returns the details of the filter specified by the filter name.

get_findings(client, detector_id, input, options \\ [])

Describes Amazon GuardDuty findings specified by finding IDs.

get_findings_statistics(client, detector_id, input, options \\ [])

Lists Amazon GuardDuty findings statistics for the specified detector ID.

get_i_p_set(client, detector_id, ip_set_id, options \\ [])

Retrieves the IPSet specified by the ipSetId.

get_invitations_count(client, options \\ [])

Returns the count of all GuardDuty membership invitations that were sent to the current member account except the currently accepted invitation.

get_master_account(client, detector_id, options \\ [])

Provides the details for the GuardDuty master account associated with the current GuardDuty member account.

get_member_detectors(client, detector_id, input, options \\ [])

Describes which data sources are enabled for the member account's detector.

get_members(client, detector_id, input, options \\ [])

Retrieves GuardDuty member accounts (to the current GuardDuty master account) specified by the account IDs.

get_threat_intel_set(client, detector_id, threat_intel_set_id, options \\ [])

Retrieves the ThreatIntelSet that is specified by the ThreatIntelSet ID.

get_usage_statistics(client, detector_id, input, options \\ [])

Lists Amazon GuardDuty usage statistics over the last 30 days for the specified detector ID.

For newly enabled detectors or data sources the cost returned will include only the usage so far under 30 days, this may differ from the cost metrics in the console, which projects usage over 30 days to provide a monthly cost estimate. For more information see Understanding How Usage Costs are Calculated.

invite_members(client, detector_id, input, options \\ [])

Invites other AWS accounts (created as members of the current AWS account by CreateMembers) to enable GuardDuty, and allow the current AWS account to view and manage these accounts' GuardDuty findings on their behalf as the master account.

list_detectors(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists detectorIds of all the existing Amazon GuardDuty detector resources.

list_filters(client, detector_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a paginated list of the current filters.

list_findings(client, detector_id, input, options \\ [])

Lists Amazon GuardDuty findings for the specified detector ID.

list_i_p_sets(client, detector_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the IPSets of the GuardDuty service specified by the detector ID.

If you use this operation from a member account, the IPSets returned are the IPSets from the associated master account.

list_invitations(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all GuardDuty membership invitations that were sent to the current AWS account.

list_members(client, detector_id, max_results \\ nil, next_token \\ nil, only_associated \\ nil, options \\ [])

Lists details about all member accounts for the current GuardDuty master account.

list_organization_admin_accounts(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the accounts configured as GuardDuty delegated administrators.

list_publishing_destinations(client, detector_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of publishing destinations associated with the specified dectectorId.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists tags for a resource.

Tagging is currently supported for detectors, finding filters, IP sets, and threat intel sets, with a limit of 50 tags per resource. When invoked, this operation returns all assigned tags for a given resource.

list_threat_intel_sets(client, detector_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the ThreatIntelSets of the GuardDuty service specified by the detector ID.

If you use this operation from a member account, the ThreatIntelSets associated with the master account are returned.

start_monitoring_members(client, detector_id, input, options \\ [])

Turns on GuardDuty monitoring of the specified member accounts.

Use this operation to restart monitoring of accounts that you stopped monitoring with the StopMonitoringMembers operation.

stop_monitoring_members(client, detector_id, input, options \\ [])

Stops GuardDuty monitoring for the specified member accounts.

Use the StartMonitoringMembers operation to restart monitoring for those accounts.

tag_resource(client, resource_arn, input, options \\ [])

Adds tags to a resource.

unarchive_findings(client, detector_id, input, options \\ [])

Unarchives GuardDuty findings specified by the findingIds.

untag_resource(client, resource_arn, input, options \\ [])

Removes tags from a resource.

update_detector(client, detector_id, input, options \\ [])

Updates the Amazon GuardDuty detector specified by the detectorId.

update_filter(client, detector_id, filter_name, input, options \\ [])

Updates the filter specified by the filter name.

update_findings_feedback(client, detector_id, input, options \\ [])

Marks the specified GuardDuty findings as useful or not useful.

update_i_p_set(client, detector_id, ip_set_id, input, options \\ [])

Updates the IPSet specified by the IPSet ID.

update_member_detectors(client, detector_id, input, options \\ [])

Contains information on member accounts to be updated.

update_organization_configuration(client, detector_id, input, options \\ [])

Updates the delegated administrator account with the values provided.

update_publishing_destination(client, destination_id, detector_id, input, options \\ [])

Updates information about the publishing destination specified by the destinationId.

update_threat_intel_set(client, detector_id, threat_intel_set_id, input, options \\ [])

Updates the ThreatIntelSet specified by the ThreatIntelSet ID.