aws-elixir v0.6.0 AWS.Identitystore

Functions

describe_group(client, input, options \\ [])

Retrieves the group metadata and attributes from GroupId in an identity store.

describe_user(client, input, options \\ [])

Retrieves the user metadata and attributes from UserId in an identity store.

list_groups(client, input, options \\ [])

Lists the attribute name and value of the group that you specified in the search.

We only support DisplayName as a valid filter attribute path currently, and filter is required. This API returns minimum attributes, including GroupId and group DisplayName in the response.

list_users(client, input, options \\ [])

Lists the attribute name and value of the user that you specified in the search.

We only support UserName as a valid filter attribute path currently, and filter is required. This API returns minimum attributes, including UserId and UserName in the response.