aws-elixir v0.6.0 AWS.Imagebuilder

EC2 Image Builder is a fully managed AWS service that makes it easier to automate the creation, management, and deployment of customized, secure, and up-to-date “golden” server images that are pre-installed and pre-configured with software and settings to meet specific IT standards.

Functions

cancel_image_creation(client, input, options \\ [])

CancelImageCreation cancels the creation of Image.

This operation can only be used on images in a non-terminal state.

create_component(client, input, options \\ [])

Creates a new component that can be used to build, validate, test, and assess your image.

create_distribution_configuration(client, input, options \\ [])

Creates a new distribution configuration.

Distribution configurations define and configure the outputs of your pipeline.

create_image(client, input, options \\ [])

Creates a new image.

This request will create a new image along with all of the configured output resources defined in the distribution configuration.

create_image_pipeline(client, input, options \\ [])

Creates a new image pipeline.

Image pipelines enable you to automate the creation and distribution of images.

create_image_recipe(client, input, options \\ [])

Creates a new image recipe.

Image recipes define how images are configured, tested, and assessed.

create_infrastructure_configuration(client, input, options \\ [])

Creates a new infrastructure configuration.

An infrastructure configuration defines the environment in which your image will be built and tested.

delete_component(client, input, options \\ [])

Deletes a component build version.

delete_distribution_configuration(client, input, options \\ [])

Deletes a distribution configuration.

delete_image(client, input, options \\ [])

Deletes an image.

delete_image_pipeline(client, input, options \\ [])

Deletes an image pipeline.

delete_image_recipe(client, input, options \\ [])

Deletes an image recipe.

delete_infrastructure_configuration(client, input, options \\ [])

Deletes an infrastructure configuration.

get_component(client, component_build_version_arn, options \\ [])

Gets a component object.

get_component_policy(client, component_arn, options \\ [])

Gets a component policy.

get_distribution_configuration(client, distribution_configuration_arn, options \\ [])

Gets a distribution configuration.

get_image(client, image_build_version_arn, options \\ [])

Gets an image.

get_image_pipeline(client, image_pipeline_arn, options \\ [])

Gets an image pipeline.

get_image_policy(client, image_arn, options \\ [])

Gets an image policy.

get_image_recipe(client, image_recipe_arn, options \\ [])

Gets an image recipe.

get_image_recipe_policy(client, image_recipe_arn, options \\ [])

Gets an image recipe policy.

get_infrastructure_configuration(client, infrastructure_configuration_arn, options \\ [])

Gets an infrastructure configuration.

import_component(client, input, options \\ [])

Imports a component and transforms its data into a component document.

list_component_build_versions(client, input, options \\ [])

Returns the list of component build versions for the specified semantic version.

list_components(client, input, options \\ [])

Returns the list of component build versions for the specified semantic version.

list_distribution_configurations(client, input, options \\ [])

Returns a list of distribution configurations.

list_image_build_versions(client, input, options \\ [])

Returns a list of image build versions.

list_image_pipeline_images(client, input, options \\ [])

Returns a list of images created by the specified pipeline.

list_image_pipelines(client, input, options \\ [])

Returns a list of image pipelines.

list_image_recipes(client, input, options \\ [])

Returns a list of image recipes.

list_images(client, input, options \\ [])

Returns the list of images that you have access to.

list_infrastructure_configurations(client, input, options \\ [])

Returns a list of infrastructure configurations.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns the list of tags for the specified resource.

put_component_policy(client, input, options \\ [])

Applies a policy to a component.

We recommend that you call the RAM API CreateResourceShare to share resources. If you call the Image Builder API PutComponentPolicy, you must also call the RAM API PromoteResourceShareCreatedFromPolicy in order for the resource to be visible to all principals with whom the resource is shared.

put_image_policy(client, input, options \\ [])

Applies a policy to an image.

We recommend that you call the RAM API CreateResourceShare to share resources. If you call the Image Builder API PutImagePolicy, you must also call the RAM API PromoteResourceShareCreatedFromPolicy in order for the resource to be visible to all principals with whom the resource is shared.

put_image_recipe_policy(client, input, options \\ [])

Applies a policy to an image recipe.

We recommend that you call the RAM API CreateResourceShare to share resources. If you call the Image Builder API PutImageRecipePolicy, you must also call the RAM API PromoteResourceShareCreatedFromPolicy in order for the resource to be visible to all principals with whom the resource is shared.

start_image_pipeline_execution(client, input, options \\ [])

Manually triggers a pipeline to create an image.

tag_resource(client, resource_arn, input, options \\ [])

Adds a tag to a resource.

untag_resource(client, resource_arn, input, options \\ [])

Removes a tag from a resource.

update_distribution_configuration(client, input, options \\ [])

Updates a new distribution configuration.

Distribution configurations define and configure the outputs of your pipeline.

update_image_pipeline(client, input, options \\ [])

Updates a new image pipeline.

Image pipelines enable you to automate the creation and distribution of images.

update_infrastructure_configuration(client, input, options \\ [])

Updates a new infrastructure configuration.

An infrastructure configuration defines the environment in which your image will be built and tested.