aws-elixir v0.6.0 AWS.IoT

AWS IoT

AWS IoT provides secure, bi-directional communication between Internet-connected devices (such as sensors, actuators, embedded devices, or smart appliances) and the AWS cloud.

You can discover your custom IoT-Data endpoint to communicate with, configure rules for data processing and integration with other services, organize resources associated with each device (Registry), configure logging, and create and manage policies and credentials to authenticate devices.

The service endpoints that expose this API are listed in AWS IoT Core Endpoints and Quotas. You must use the endpoint for the region that has the resources you want to access.

The service name used by AWS Signature Version 4 to sign the request is: execute-api.

For more information about how AWS IoT works, see the Developer Guide.

For information about how to use the credentials provider for AWS IoT, see Authorizing Direct Calls to AWS Services.

Functions

accept_certificate_transfer(client, certificate_id, input, options \\ [])

Accepts a pending certificate transfer.

The default state of the certificate is INACTIVE.

To check for pending certificate transfers, call ListCertificates to enumerate your certificates.

add_thing_to_billing_group(client, input, options \\ [])

Adds a thing to a billing group.

add_thing_to_thing_group(client, input, options \\ [])

Adds a thing to a thing group.

associate_targets_with_job(client, job_id, input, options \\ [])

Associates a group with a continuous job.

The following criteria must be met:

The job must have been created with the targetSelection field set to "CONTINUOUS".

The job status must currently be "IN_PROGRESS".

The total number of targets associated with a job must not exceed 100.

attach_policy(client, policy_name, input, options \\ [])

Attaches a policy to the specified target.

attach_principal_policy(client, policy_name, input, options \\ [])

Attaches the specified policy to the specified principal (certificate or other credential).

Note: This API is deprecated. Please use AttachPolicy instead.

attach_security_profile(client, security_profile_name, input, options \\ [])

Associates a Device Defender security profile with a thing group or this account.

Each thing group or account can have up to five security profiles associated with it.

attach_thing_principal(client, thing_name, input, options \\ [])

Attaches the specified principal to the specified thing.

A principal can be X.509 certificates, IAM users, groups, and roles, Amazon Cognito identities or federated identities.

cancel_audit_mitigation_actions_task(client, task_id, input, options \\ [])

Cancels a mitigation action task that is in progress.

If the task is not in progress, an InvalidRequestException occurs.

cancel_audit_task(client, task_id, input, options \\ [])

Cancels an audit that is in progress.

The audit can be either scheduled or on-demand. If the audit is not in progress, an "InvalidRequestException" occurs.

cancel_certificate_transfer(client, certificate_id, input, options \\ [])

Cancels a pending transfer for the specified certificate.

Note Only the transfer source account can use this operation to cancel a transfer. (Transfer destinations can use RejectCertificateTransfer instead.) After transfer, AWS IoT returns the certificate to the source account in the INACTIVE state. After the destination account has accepted the transfer, the transfer cannot be cancelled.

After a certificate transfer is cancelled, the status of the certificate changes from PENDING_TRANSFER to INACTIVE.

cancel_job(client, job_id, input, options \\ [])

Cancels a job.

cancel_job_execution(client, job_id, thing_name, input, options \\ [])

Cancels the execution of a job for a given thing.

clear_default_authorizer(client, input, options \\ [])

Clears the default authorizer.

confirm_topic_rule_destination(client, confirmation_token, options \\ [])

Confirms a topic rule destination.

When you create a rule requiring a destination, AWS IoT sends a confirmation message to the endpoint or base address you specify. The message includes a token which you pass back when calling ConfirmTopicRuleDestination to confirm that you own or have access to the endpoint.

create_audit_suppression(client, input, options \\ [])

Creates a Device Defender audit suppression.

create_authorizer(client, authorizer_name, input, options \\ [])

Creates an authorizer.

create_billing_group(client, billing_group_name, input, options \\ [])

Creates a billing group.

create_certificate_from_csr(client, input, options \\ [])

Creates an X.509 certificate using the specified certificate signing request.

Note: The CSR must include a public key that is either an RSA key with a length of at least 2048 bits or an ECC key from NIST P-256 or NIST P-384 curves.

Note: Reusing the same certificate signing request (CSR) results in a distinct certificate.

You can create multiple certificates in a batch by creating a directory, copying multiple .csr files into that directory, and then specifying that directory on the command line. The following commands show how to create a batch of certificates given a batch of CSRs.

Assuming a set of CSRs are located inside of the directory my-csr-directory:

On Linux and OS X, the command is:

$ ls my-csr-directory/ | xargs -I {} aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/{}

This command lists all of the CSRs in my-csr-directory and pipes each CSR file name to the aws iot create-certificate-from-csr AWS CLI command to create a certificate for the corresponding CSR.

The aws iot create-certificate-from-csr part of the command can also be run in parallel to speed up the certificate creation process:

$ ls my-csr-directory/ | xargs -P 10 -I {} aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/{}

On Windows PowerShell, the command to create certificates for all CSRs in my-csr-directory is:

ls -Name my-csr-directory | %{aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/$_}

On a Windows command prompt, the command to create certificates for all CSRs in my-csr-directory is:

forfiles /p my-csr-directory /c "cmd /c aws iot create-certificate-from-csr --certificate-signing-request file://@path"

create_dimension(client, name, input, options \\ [])

Create a dimension that you can use to limit the scope of a metric used in a security profile for AWS IoT Device Defender.

For example, using a TOPIC_FILTER dimension, you can narrow down the scope of the metric only to MQTT topics whose name match the pattern specified in the dimension.

create_domain_configuration(client, domain_configuration_name, input, options \\ [])

Creates a domain configuration.

The domain configuration feature is in public preview and is subject to change.

create_dynamic_thing_group(client, thing_group_name, input, options \\ [])

Creates a dynamic thing group.

create_job(client, job_id, input, options \\ [])

Creates a job.

create_keys_and_certificate(client, input, options \\ [])

Creates a 2048-bit RSA key pair and issues an X.509 certificate using the issued public key.

You can also call CreateKeysAndCertificate over MQTT from a device, for more information, see Provisioning MQTT API.

Note This is the only time AWS IoT issues the private key for this certificate, so it is important to keep it in a secure location.

create_mitigation_action(client, action_name, input, options \\ [])

Defines an action that can be applied to audit findings by using StartAuditMitigationActionsTask.

Only certain types of mitigation actions can be applied to specific check names. For more information, see Mitigation actions. Each mitigation action can apply only one type of change.

create_o_t_a_update(client, ota_update_id, input, options \\ [])

Creates an AWS IoT OTAUpdate on a target group of things or groups.

create_policy(client, policy_name, input, options \\ [])

Creates an AWS IoT policy.

The created policy is the default version for the policy. This operation creates a policy version with a version identifier of 1 and sets 1 as the policy's default version.

create_policy_version(client, policy_name, input, options \\ [])

Creates a new version of the specified AWS IoT policy.

To update a policy, create a new policy version. A managed policy can have up to five versions. If the policy has five versions, you must use DeletePolicyVersion to delete an existing version before you create a new one.

Optionally, you can set the new version as the policy's default version. The default version is the operative version (that is, the version that is in effect for the certificates to which the policy is attached).

create_provisioning_claim(client, template_name, input, options \\ [])

Creates a provisioning claim.

create_provisioning_template(client, input, options \\ [])

Creates a fleet provisioning template.

create_provisioning_template_version(client, template_name, input, options \\ [])

Creates a new version of a fleet provisioning template.

create_role_alias(client, role_alias, input, options \\ [])

Creates a role alias.

create_scheduled_audit(client, scheduled_audit_name, input, options \\ [])

Creates a scheduled audit that is run at a specified time interval.

create_security_profile(client, security_profile_name, input, options \\ [])

Creates a Device Defender security profile.

create_stream(client, stream_id, input, options \\ [])

Creates a stream for delivering one or more large files in chunks over MQTT.

A stream transports data bytes in chunks or blocks packaged as MQTT messages from a source like S3. You can have one or more files associated with a stream.

create_thing(client, thing_name, input, options \\ [])

Creates a thing record in the registry.

If this call is made multiple times using the same thing name and configuration, the call will succeed. If this call is made with the same thing name but different configuration a ResourceAlreadyExistsException is thrown.

This is a control plane operation. See Authorization for information about authorizing control plane actions.

create_thing_group(client, thing_group_name, input, options \\ [])

Create a thing group.

This is a control plane operation. See Authorization for information about authorizing control plane actions.

create_thing_type(client, thing_type_name, input, options \\ [])

Creates a new thing type.

create_topic_rule(client, rule_name, input, options \\ [])

Creates a rule.

Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.

create_topic_rule_destination(client, input, options \\ [])

Creates a topic rule destination.

The destination must be confirmed prior to use.

delete_account_audit_configuration(client, input, options \\ [])

Restores the default settings for Device Defender audits for this account.

Any configuration data you entered is deleted and all audit checks are reset to disabled.

delete_audit_suppression(client, input, options \\ [])

Deletes a Device Defender audit suppression.

delete_authorizer(client, authorizer_name, input, options \\ [])

Deletes an authorizer.

delete_billing_group(client, billing_group_name, input, options \\ [])

Deletes the billing group.

delete_c_a_certificate(client, certificate_id, input, options \\ [])

Deletes a registered CA certificate.

delete_certificate(client, certificate_id, input, options \\ [])

Deletes the specified certificate.

A certificate cannot be deleted if it has a policy or IoT thing attached to it or if its status is set to ACTIVE. To delete a certificate, first use the DetachPrincipalPolicy API to detach all policies. Next, use the UpdateCertificate API to set the certificate to the INACTIVE status.

delete_dimension(client, name, input, options \\ [])

Removes the specified dimension from your AWS account.

delete_domain_configuration(client, domain_configuration_name, input, options \\ [])

Deletes the specified domain configuration.

The domain configuration feature is in public preview and is subject to change.

delete_dynamic_thing_group(client, thing_group_name, input, options \\ [])

Deletes a dynamic thing group.

delete_job(client, job_id, input, options \\ [])

Deletes a job and its related job executions.

Deleting a job may take time, depending on the number of job executions created for the job and various other factors. While the job is being deleted, the status of the job will be shown as "DELETION_IN_PROGRESS". Attempting to delete or cancel a job whose status is already "DELETION_IN_PROGRESS" will result in an error.

Only 10 jobs may have status "DELETION_IN_PROGRESS" at the same time, or a LimitExceededException will occur.

delete_job_execution(client, execution_number, job_id, thing_name, input, options \\ [])

Deletes a job execution.

delete_mitigation_action(client, action_name, input, options \\ [])

Deletes a defined mitigation action from your AWS account.

delete_o_t_a_update(client, ota_update_id, input, options \\ [])

Delete an OTA update.

delete_policy(client, policy_name, input, options \\ [])

Deletes the specified policy.

A policy cannot be deleted if it has non-default versions or it is attached to any certificate.

To delete a policy, use the DeletePolicyVersion API to delete all non-default versions of the policy; use the DetachPrincipalPolicy API to detach the policy from any certificate; and then use the DeletePolicy API to delete the policy.

When a policy is deleted using DeletePolicy, its default version is deleted with it.

delete_policy_version(client, policy_name, policy_version_id, input, options \\ [])

Deletes the specified version of the specified policy.

You cannot delete the default version of a policy using this API. To delete the default version of a policy, use DeletePolicy. To find out which version of a policy is marked as the default version, use ListPolicyVersions.

delete_provisioning_template(client, template_name, input, options \\ [])

Deletes a fleet provisioning template.

delete_provisioning_template_version(client, template_name, version_id, input, options \\ [])

Deletes a fleet provisioning template version.

delete_registration_code(client, input, options \\ [])

Deletes a CA certificate registration code.

delete_role_alias(client, role_alias, input, options \\ [])

Deletes a role alias

delete_scheduled_audit(client, scheduled_audit_name, input, options \\ [])

Deletes a scheduled audit.

delete_security_profile(client, security_profile_name, input, options \\ [])

Deletes a Device Defender security profile.

delete_stream(client, stream_id, input, options \\ [])

Deletes a stream.

delete_thing(client, thing_name, input, options \\ [])

Deletes the specified thing.

Returns successfully with no error if the deletion is successful or you specify a thing that doesn't exist.

delete_thing_group(client, thing_group_name, input, options \\ [])

Deletes a thing group.

delete_thing_type(client, thing_type_name, input, options \\ [])

Deletes the specified thing type.

You cannot delete a thing type if it has things associated with it. To delete a thing type, first mark it as deprecated by calling DeprecateThingType, then remove any associated things by calling UpdateThing to change the thing type on any associated thing, and finally use DeleteThingType to delete the thing type.

delete_topic_rule(client, rule_name, input, options \\ [])

Deletes the rule.

delete_topic_rule_destination(client, arn, input, options \\ [])

Deletes a topic rule destination.

delete_v2_logging_level(client, input, options \\ [])

Deletes a logging level.

deprecate_thing_type(client, thing_type_name, input, options \\ [])

Deprecates a thing type.

You can not associate new things with deprecated thing type.

describe_account_audit_configuration(client, options \\ [])

Gets information about the Device Defender audit settings for this account.

Settings include how audit notifications are sent and which audit checks are enabled or disabled.

describe_audit_finding(client, finding_id, options \\ [])

Gets information about a single audit finding.

Properties include the reason for noncompliance, the severity of the issue, and when the audit that returned the finding was started.

describe_audit_mitigation_actions_task(client, task_id, options \\ [])

Gets information about an audit mitigation task that is used to apply mitigation actions to a set of audit findings.

Properties include the actions being applied, the audit checks to which they're being applied, the task status, and aggregated task statistics.

describe_audit_suppression(client, input, options \\ [])

Gets information about a Device Defender audit suppression.

describe_audit_task(client, task_id, options \\ [])

Gets information about a Device Defender audit.

describe_authorizer(client, authorizer_name, options \\ [])

Describes an authorizer.

describe_billing_group(client, billing_group_name, options \\ [])

Returns information about a billing group.

describe_c_a_certificate(client, certificate_id, options \\ [])

Describes a registered CA certificate.

describe_certificate(client, certificate_id, options \\ [])

Gets information about the specified certificate.

describe_default_authorizer(client, options \\ [])

Describes the default authorizer.

describe_dimension(client, name, options \\ [])

Provides details about a dimension that is defined in your AWS account.

describe_domain_configuration(client, domain_configuration_name, options \\ [])

Gets summary information about a domain configuration.

The domain configuration feature is in public preview and is subject to change.

describe_endpoint(client, endpoint_type \\ nil, options \\ [])

Returns a unique endpoint specific to the AWS account making the call.

describe_event_configurations(client, options \\ [])

Describes event configurations.

describe_index(client, index_name, options \\ [])

Describes a search index.

describe_job(client, job_id, options \\ [])

Describes a job.

describe_job_execution(client, job_id, thing_name, execution_number \\ nil, options \\ [])

Describes a job execution.

describe_mitigation_action(client, action_name, options \\ [])

Gets information about a mitigation action.

describe_provisioning_template(client, template_name, options \\ [])

Returns information about a fleet provisioning template.

describe_provisioning_template_version(client, template_name, version_id, options \\ [])

Returns information about a fleet provisioning template version.

describe_role_alias(client, role_alias, options \\ [])

Describes a role alias.

describe_scheduled_audit(client, scheduled_audit_name, options \\ [])

Gets information about a scheduled audit.

describe_security_profile(client, security_profile_name, options \\ [])

Gets information about a Device Defender security profile.

describe_stream(client, stream_id, options \\ [])

Gets information about a stream.

describe_thing(client, thing_name, options \\ [])

Gets information about the specified thing.

describe_thing_group(client, thing_group_name, options \\ [])

Describe a thing group.

describe_thing_registration_task(client, task_id, options \\ [])

Describes a bulk thing provisioning task.

describe_thing_type(client, thing_type_name, options \\ [])

Gets information about the specified thing type.

detach_policy(client, policy_name, input, options \\ [])

Detaches a policy from the specified target.

detach_principal_policy(client, policy_name, input, options \\ [])

Removes the specified policy from the specified certificate.

Note: This API is deprecated. Please use DetachPolicy instead.

detach_security_profile(client, security_profile_name, input, options \\ [])

Disassociates a Device Defender security profile from a thing group or from this account.

detach_thing_principal(client, thing_name, input, options \\ [])

Detaches the specified principal from the specified thing.

A principal can be X.509 certificates, IAM users, groups, and roles, Amazon Cognito identities or federated identities.

This call is asynchronous. It might take several seconds for the detachment to propagate.

disable_topic_rule(client, rule_name, input, options \\ [])

Disables the rule.

enable_topic_rule(client, rule_name, input, options \\ [])

Enables the rule.

get_cardinality(client, input, options \\ [])

Returns the approximate count of unique values that match the query.

get_effective_policies(client, input, options \\ [])

Gets a list of the policies that have an effect on the authorization behavior of the specified device when it connects to the AWS IoT device gateway.

get_indexing_configuration(client, options \\ [])

Gets the indexing configuration.

get_job_document(client, job_id, options \\ [])

Gets a job document.

get_logging_options(client, options \\ [])

Gets the logging options.

NOTE: use of this command is not recommended. Use GetV2LoggingOptions instead.

get_o_t_a_update(client, ota_update_id, options \\ [])

Gets an OTA update.

get_percentiles(client, input, options \\ [])

Groups the aggregated values that match the query into percentile groupings.

The default percentile groupings are: 1,5,25,50,75,95,99, although you can specify your own when you call GetPercentiles. This function returns a value for each percentile group specified (or the default percentile groupings). The percentile group "1" contains the aggregated field value that occurs in approximately one percent of the values that match the query. The percentile group "5" contains the aggregated field value that occurs in approximately five percent of the values that match the query, and so on. The result is an approximation, the more values that match the query, the more accurate the percentile values.

get_policy(client, policy_name, options \\ [])

Gets information about the specified policy with the policy document of the default version.

get_policy_version(client, policy_name, policy_version_id, options \\ [])

Gets information about the specified policy version.

get_registration_code(client, options \\ [])

Gets a registration code used to register a CA certificate with AWS IoT.

get_statistics(client, input, options \\ [])

Returns the count, average, sum, minimum, maximum, sum of squares, variance, and standard deviation for the specified aggregated field.

If the aggregation field is of type String, only the count statistic is returned.

get_topic_rule(client, rule_name, options \\ [])

Gets information about the rule.

get_topic_rule_destination(client, arn, options \\ [])

Gets information about a topic rule destination.

get_v2_logging_options(client, options \\ [])

Gets the fine grained logging options.

list_active_violations(client, max_results \\ nil, next_token \\ nil, security_profile_name \\ nil, thing_name \\ nil, options \\ [])

Lists the active violations for a given Device Defender security profile.

list_attached_policies(client, target, input, options \\ [])

Lists the policies attached to the specified thing group.

list_audit_findings(client, input, options \\ [])

Lists the findings (results) of a Device Defender audit or of the audits performed during a specified time period.

(Findings are retained for 90 days.)

list_audit_mitigation_actions_executions(client, action_status \\ nil, finding_id, max_results \\ nil, next_token \\ nil, task_id, options \\ [])

Gets the status of audit mitigation action tasks that were executed.

list_audit_mitigation_actions_tasks(client, audit_task_id \\ nil, end_time, finding_id \\ nil, max_results \\ nil, next_token \\ nil, start_time, task_status \\ nil, options \\ [])

Gets a list of audit mitigation action tasks that match the specified filters.

list_audit_suppressions(client, input, options \\ [])

Lists your Device Defender audit listings.

list_audit_tasks(client, end_time, max_results \\ nil, next_token \\ nil, start_time, task_status \\ nil, task_type \\ nil, options \\ [])

Lists the Device Defender audits that have been performed during a given time period.

list_authorizers(client, ascending_order \\ nil, marker \\ nil, page_size \\ nil, status \\ nil, options \\ [])

Lists the authorizers registered in your account.

list_billing_groups(client, max_results \\ nil, name_prefix_filter \\ nil, next_token \\ nil, options \\ [])

Lists the billing groups you have created.

list_c_a_certificates(client, ascending_order \\ nil, marker \\ nil, page_size \\ nil, options \\ [])

Lists the CA certificates registered for your AWS account.

The results are paginated with a default page size of 25. You can use the returned marker to retrieve additional results.

list_certificates(client, ascending_order \\ nil, marker \\ nil, page_size \\ nil, options \\ [])

Lists the certificates registered in your AWS account.

The results are paginated with a default page size of 25. You can use the returned marker to retrieve additional results.

list_certificates_by_c_a(client, ca_certificate_id, ascending_order \\ nil, marker \\ nil, page_size \\ nil, options \\ [])

List the device certificates signed by the specified CA certificate.

list_dimensions(client, max_results \\ nil, next_token \\ nil, options \\ [])

List the set of dimensions that are defined for your AWS account.

list_domain_configurations(client, marker \\ nil, page_size \\ nil, service_type \\ nil, options \\ [])

Gets a list of domain configurations for the user.

This list is sorted alphabetically by domain configuration name.

The domain configuration feature is in public preview and is subject to change.

list_indices(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the search indices.

list_job_executions_for_job(client, job_id, max_results \\ nil, next_token \\ nil, status \\ nil, options \\ [])

Lists the job executions for a job.

list_job_executions_for_thing(client, thing_name, max_results \\ nil, next_token \\ nil, status \\ nil, options \\ [])

Lists the job executions for the specified thing.

list_jobs(client, max_results \\ nil, next_token \\ nil, status \\ nil, target_selection \\ nil, thing_group_id \\ nil, thing_group_name \\ nil, options \\ [])

Lists jobs.

list_mitigation_actions(client, action_type \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Gets a list of all mitigation actions that match the specified filter criteria.

list_o_t_a_updates(client, max_results \\ nil, next_token \\ nil, ota_update_status \\ nil, options \\ [])

Lists OTA updates.

list_outgoing_certificates(client, ascending_order \\ nil, marker \\ nil, page_size \\ nil, options \\ [])

Lists certificates that are being transferred but not yet accepted.

list_policies(client, ascending_order \\ nil, marker \\ nil, page_size \\ nil, options \\ [])

Lists your policies.

list_policy_principals(client, ascending_order \\ nil, marker \\ nil, page_size \\ nil, policy_name, options \\ [])

Lists the principals associated with the specified policy.

Note: This API is deprecated. Please use ListTargetsForPolicy instead.

list_policy_versions(client, policy_name, options \\ [])

Lists the versions of the specified policy and identifies the default version.

list_principal_policies(client, ascending_order \\ nil, marker \\ nil, page_size \\ nil, principal, options \\ [])

Lists the policies attached to the specified principal.

If you use an Cognito identity, the ID must be in AmazonCognito Identity format.

Note: This API is deprecated. Please use ListAttachedPolicies instead.

list_principal_things(client, max_results \\ nil, next_token \\ nil, principal, options \\ [])

Lists the things associated with the specified principal.

A principal can be X.509 certificates, IAM users, groups, and roles, Amazon Cognito identities or federated identities.

list_provisioning_template_versions(client, template_name, max_results \\ nil, next_token \\ nil, options \\ [])

A list of fleet provisioning template versions.

list_provisioning_templates(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the fleet provisioning templates in your AWS account.

list_role_aliases(client, ascending_order \\ nil, marker \\ nil, page_size \\ nil, options \\ [])

Lists the role aliases registered in your account.

list_scheduled_audits(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all of your scheduled audits.

list_security_profiles(client, dimension_name \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the Device Defender security profiles you have created.

You can use filters to list only those security profiles associated with a thing group or only those associated with your account.

list_security_profiles_for_target(client, max_results \\ nil, next_token \\ nil, recursive \\ nil, security_profile_target_arn, options \\ [])

Lists the Device Defender security profiles attached to a target (thing group).

list_streams(client, ascending_order \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all of the streams in your AWS account.

list_tags_for_resource(client, next_token \\ nil, resource_arn, options \\ [])

Lists the tags (metadata) you have assigned to the resource.

list_targets_for_policy(client, policy_name, input, options \\ [])

List targets for the specified policy.

list_targets_for_security_profile(client, security_profile_name, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the targets (thing groups) associated with a given Device Defender security profile.

list_thing_groups(client, max_results \\ nil, name_prefix_filter \\ nil, next_token \\ nil, parent_group \\ nil, recursive \\ nil, options \\ [])

List the thing groups in your account.

list_thing_groups_for_thing(client, thing_name, max_results \\ nil, next_token \\ nil, options \\ [])

List the thing groups to which the specified thing belongs.

list_thing_principals(client, thing_name, options \\ [])

Lists the principals associated with the specified thing.

A principal can be X.509 certificates, IAM users, groups, and roles, Amazon Cognito identities or federated identities.

list_thing_registration_task_reports(client, task_id, max_results \\ nil, next_token \\ nil, report_type, options \\ [])

Information about the thing registration tasks.

list_thing_registration_tasks(client, max_results \\ nil, next_token \\ nil, status \\ nil, options \\ [])

List bulk thing provisioning tasks.

list_thing_types(client, max_results \\ nil, next_token \\ nil, thing_type_name \\ nil, options \\ [])

Lists the existing thing types.

list_things(client, attribute_name \\ nil, attribute_value \\ nil, max_results \\ nil, next_token \\ nil, thing_type_name \\ nil, options \\ [])

Lists your things.

Use the attributeName and attributeValue parameters to filter your things. For example, calling ListThings with attributeName=Color and attributeValue=Red retrieves all things in the registry that contain an attribute Color with the value Red.

You will not be charged for calling this API if an Access denied error is returned. You will also not be charged if no attributes or pagination token was provided in request and no pagination token and no results were returned.

list_things_in_billing_group(client, billing_group_name, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the things you have added to the given billing group.

list_things_in_thing_group(client, thing_group_name, max_results \\ nil, next_token \\ nil, recursive \\ nil, options \\ [])

Lists the things in the specified group.

list_topic_rule_destinations(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all the topic rule destinations in your AWS account.

list_topic_rules(client, max_results \\ nil, next_token \\ nil, rule_disabled \\ nil, topic \\ nil, options \\ [])

Lists the rules for the specific topic.

list_v2_logging_levels(client, max_results \\ nil, next_token \\ nil, target_type \\ nil, options \\ [])

Lists logging levels.

list_violation_events(client, end_time, max_results \\ nil, next_token \\ nil, security_profile_name \\ nil, start_time, thing_name \\ nil, options \\ [])

Lists the Device Defender security profile violations discovered during the given time period.

You can use filters to limit the results to those alerts issued for a particular security profile, behavior, or thing (device).

register_c_a_certificate(client, input, options \\ [])

Registers a CA certificate with AWS IoT.

This CA certificate can then be used to sign device certificates, which can be then registered with AWS IoT. You can register up to 10 CA certificates per AWS account that have the same subject field. This enables you to have up to 10 certificate authorities sign your device certificates. If you have more than one CA certificate registered, make sure you pass the CA certificate when you register your device certificates with the RegisterCertificate API.

register_certificate(client, input, options \\ [])

Registers a device certificate with AWS IoT.

If you have more than one CA certificate that has the same subject field, you must specify the CA certificate that was used to sign the device certificate being registered.

register_certificate_without_c_a(client, input, options \\ [])

Register a certificate that does not have a certificate authority (CA).

register_thing(client, input, options \\ [])

Provisions a thing in the device registry.

RegisterThing calls other AWS IoT control plane APIs. These calls might exceed your account level AWS IoT Throttling Limits and cause throttle errors. Please contact AWS Customer Support to raise your throttling limits if necessary.

reject_certificate_transfer(client, certificate_id, input, options \\ [])

Rejects a pending certificate transfer.

After AWS IoT rejects a certificate transfer, the certificate status changes from PENDING_TRANSFER to INACTIVE.

To check for pending certificate transfers, call ListCertificates to enumerate your certificates.

This operation can only be called by the transfer destination. After it is called, the certificate will be returned to the source's account in the INACTIVE state.

remove_thing_from_billing_group(client, input, options \\ [])

Removes the given thing from the billing group.

remove_thing_from_thing_group(client, input, options \\ [])

Remove the specified thing from the specified group.

You must specify either a thingGroupArn or a thingGroupName to identify the thing group and either a thingArn or a thingName to identify the thing to remove from the thing group.

replace_topic_rule(client, rule_name, input, options \\ [])

Replaces the rule.

You must specify all parameters for the new rule. Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.

search_index(client, input, options \\ [])

The query search index.

set_default_authorizer(client, input, options \\ [])

Sets the default authorizer.

This will be used if a websocket connection is made without specifying an authorizer.

set_default_policy_version(client, policy_name, policy_version_id, input, options \\ [])

Sets the specified version of the specified policy as the policy's default (operative) version.

This action affects all certificates to which the policy is attached. To list the principals the policy is attached to, use the ListPrincipalPolicy API.

set_logging_options(client, input, options \\ [])

Sets the logging options.

NOTE: use of this command is not recommended. Use SetV2LoggingOptions instead.

set_v2_logging_level(client, input, options \\ [])

Sets the logging level.

set_v2_logging_options(client, input, options \\ [])

Sets the logging options for the V2 logging service.

start_audit_mitigation_actions_task(client, task_id, input, options \\ [])

Starts a task that applies a set of mitigation actions to the specified target.

start_on_demand_audit_task(client, input, options \\ [])

Starts an on-demand Device Defender audit.

start_thing_registration_task(client, input, options \\ [])

Creates a bulk thing provisioning task.

stop_thing_registration_task(client, task_id, input, options \\ [])

Cancels a bulk thing provisioning task.

tag_resource(client, input, options \\ [])

Adds to or modifies the tags of the given resource.

Tags are metadata which can be used to manage a resource.

test_authorization(client, input, options \\ [])

Tests if a specified principal is authorized to perform an AWS IoT action on a specified resource.

Use this to test and debug the authorization behavior of devices that connect to the AWS IoT device gateway.

test_invoke_authorizer(client, authorizer_name, input, options \\ [])

Tests a custom authorization behavior by invoking a specified custom authorizer.

Use this to test and debug the custom authorization behavior of devices that connect to the AWS IoT device gateway.

transfer_certificate(client, certificate_id, input, options \\ [])

Transfers the specified certificate to the specified AWS account.

You can cancel the transfer until it is acknowledged by the recipient.

No notification is sent to the transfer destination's account. It is up to the caller to notify the transfer target.

The certificate being transferred must not be in the ACTIVE state. You can use the UpdateCertificate API to deactivate it.

The certificate must not have any policies attached to it. You can use the DetachPrincipalPolicy API to detach them.

untag_resource(client, input, options \\ [])

Removes the given tags (metadata) from the resource.

update_account_audit_configuration(client, input, options \\ [])

Configures or reconfigures the Device Defender audit settings for this account.

Settings include how audit notifications are sent and which audit checks are enabled or disabled.

update_audit_suppression(client, input, options \\ [])

Updates a Device Defender audit suppression.

update_authorizer(client, authorizer_name, input, options \\ [])

Updates an authorizer.

update_billing_group(client, billing_group_name, input, options \\ [])

Updates information about the billing group.

update_c_a_certificate(client, certificate_id, input, options \\ [])

Updates a registered CA certificate.

update_certificate(client, certificate_id, input, options \\ [])

Updates the status of the specified certificate.

This operation is idempotent.

Certificates must be in the ACTIVE state to authenticate devices that use a certificate to connect to AWS IoT.

Within a few minutes of updating a certificate from the ACTIVE state to any other state, AWS IoT disconnects all devices that used that certificate to connect. Devices cannot use a certificate that is not in the ACTIVE state to reconnect.

update_dimension(client, name, input, options \\ [])

Updates the definition for a dimension.

You cannot change the type of a dimension after it is created (you can delete it and re-create it).

update_domain_configuration(client, domain_configuration_name, input, options \\ [])

Updates values stored in the domain configuration.

Domain configurations for default endpoints can't be updated.

The domain configuration feature is in public preview and is subject to change.

update_dynamic_thing_group(client, thing_group_name, input, options \\ [])

Updates a dynamic thing group.

update_event_configurations(client, input, options \\ [])

Updates the event configurations.

update_indexing_configuration(client, input, options \\ [])

Updates the search configuration.

update_job(client, job_id, input, options \\ [])

Updates supported fields of the specified job.

update_mitigation_action(client, action_name, input, options \\ [])

Updates the definition for the specified mitigation action.

update_provisioning_template(client, template_name, input, options \\ [])

Updates a fleet provisioning template.

update_role_alias(client, role_alias, input, options \\ [])

Updates a role alias.

update_scheduled_audit(client, scheduled_audit_name, input, options \\ [])

Updates a scheduled audit, including which checks are performed and how often the audit takes place.

update_security_profile(client, security_profile_name, input, options \\ [])

Updates a Device Defender security profile.

update_stream(client, stream_id, input, options \\ [])

Updates an existing stream.

The stream version will be incremented by one.

update_thing(client, thing_name, input, options \\ [])

Updates the data for a thing.

update_thing_group(client, thing_group_name, input, options \\ [])

Update a thing group.

update_thing_groups_for_thing(client, input, options \\ [])

Updates the groups to which the thing belongs.

update_topic_rule_destination(client, input, options \\ [])

Updates a topic rule destination.

You use this to change the status, endpoint URL, or confirmation URL of the destination.

validate_security_profile_behaviors(client, input, options \\ [])

Validates a Device Defender security profile behaviors specification.