aws-elixir v0.6.0 AWS.IoT1ClickDevices

Describes all of the AWS IoT 1-Click device-related API operations for the service.

Also provides sample requests, responses, and errors for the supported web services protocols.

Functions

claim_devices_by_claim_code(client, claim_code, input, options \\ [])

Adds device(s) to your account (i.e., claim one or more devices) if and only if you received a claim code with the device(s).

describe_device(client, device_id, options \\ [])

Given a device ID, returns a DescribeDeviceResponse object describing the details of the device.

finalize_device_claim(client, device_id, input, options \\ [])

Given a device ID, finalizes the claim request for the associated device.

Claiming a device consists of initiating a claim, then publishing a device event, and finalizing the claim. For a device of type button, a device event can be published by simply clicking the device.

get_device_methods(client, device_id, options \\ [])

Given a device ID, returns the invokable methods associated with the device.

initiate_device_claim(client, device_id, input, options \\ [])

Given a device ID, initiates a claim request for the associated device.

Claiming a device consists of initiating a claim, then publishing a device event, and finalizing the claim. For a device of type button, a device event can be published by simply clicking the device.

invoke_device_method(client, device_id, input, options \\ [])

Given a device ID, issues a request to invoke a named device method (with possible parameters).

See the "Example POST" code snippet below.

list_device_events(client, device_id, from_time_stamp, max_results \\ nil, next_token \\ nil, to_time_stamp, options \\ [])

Using a device ID, returns a DeviceEventsResponse object containing an array of events for the device.

list_devices(client, device_type \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the 1-Click compatible devices associated with your AWS account.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists the tags associated with the specified resource ARN.

tag_resource(client, resource_arn, input, options \\ [])

Adds or updates the tags associated with the resource ARN.

See AWS IoT 1-Click Service Limits for the maximum number of tags allowed per resource.

unclaim_device(client, device_id, input, options \\ [])

Disassociates a device from your AWS account using its device ID.

untag_resource(client, resource_arn, input, options \\ [])

Using tag keys, deletes the tags (key/value pairs) associated with the specified resource ARN.

update_device_state(client, device_id, input, options \\ [])

Using a Boolean value (true or false), this operation enables or disables the device given a device ID.