aws-elixir v0.6.0 AWS.IoTAnalytics

AWS IoT Analytics allows you to collect large amounts of device data, process messages, and store them.

You can then query the data and run sophisticated analytics on it. AWS IoT Analytics enables advanced data exploration through integration with Jupyter Notebooks and data visualization through integration with Amazon QuickSight.

Traditional analytics and business intelligence tools are designed to process structured data. IoT data often comes from devices that record noisy processes (such as temperature, motion, or sound). As a result the data from these devices can have significant gaps, corrupted messages, and false readings that must be cleaned up before analysis can occur. Also, IoT data is often only meaningful in the context of other data from external sources.

AWS IoT Analytics automates the steps required to analyze data from IoT devices. AWS IoT Analytics filters, transforms, and enriches IoT data before storing it in a time-series data store for analysis. You can set up the service to collect only the data you need from your devices, apply mathematical transforms to process the data, and enrich the data with device-specific metadata such as device type and location before storing it. Then, you can analyze your data by running queries using the built-in SQL query engine, or perform more complex analytics and machine learning inference. AWS IoT Analytics includes pre-built models for common IoT use cases so you can answer questions like which devices are about to fail or which customers are at risk of abandoning their wearable devices.

Functions

batch_put_message(client, input, options \\ [])

Sends messages to a channel.

cancel_pipeline_reprocessing(client, pipeline_name, reprocessing_id, input, options \\ [])

Cancels the reprocessing of data through the pipeline.

create_channel(client, input, options \\ [])

Creates a channel.

A channel collects data from an MQTT topic and archives the raw, unprocessed messages before publishing the data to a pipeline.

create_dataset(client, input, options \\ [])

Creates a data set.

A data set stores data retrieved from a data store by applying a "queryAction" (a SQL query) or a "containerAction" (executing a containerized application). This operation creates the skeleton of a data set. The data set can be populated manually by calling "CreateDatasetContent" or automatically according to a "trigger" you specify.

create_dataset_content(client, dataset_name, input, options \\ [])

Creates the content of a data set by applying a "queryAction" (a SQL query) or a "containerAction" (executing a containerized application).

create_datastore(client, input, options \\ [])

Creates a data store, which is a repository for messages.

create_pipeline(client, input, options \\ [])

Creates a pipeline.

A pipeline consumes messages from a channel and allows you to process the messages before storing them in a data store. You must specify both a channel and a datastore activity and, optionally, as many as 23 additional activities in the pipelineActivities array.

delete_channel(client, channel_name, input, options \\ [])

Deletes the specified channel.

delete_dataset(client, dataset_name, input, options \\ [])

Deletes the specified data set.

You do not have to delete the content of the data set before you perform this operation.

delete_dataset_content(client, dataset_name, input, options \\ [])

Deletes the content of the specified data set.

delete_datastore(client, datastore_name, input, options \\ [])

Deletes the specified data store.

delete_pipeline(client, pipeline_name, input, options \\ [])

Deletes the specified pipeline.

describe_channel(client, channel_name, include_statistics \\ nil, options \\ [])

Retrieves information about a channel.

describe_dataset(client, dataset_name, options \\ [])

Retrieves information about a data set.

describe_datastore(client, datastore_name, include_statistics \\ nil, options \\ [])

Retrieves information about a data store.

describe_logging_options(client, options \\ [])

Retrieves the current settings of the AWS IoT Analytics logging options.

describe_pipeline(client, pipeline_name, options \\ [])

Retrieves information about a pipeline.

get_dataset_content(client, dataset_name, version_id \\ nil, options \\ [])

Retrieves the contents of a data set as pre-signed URIs.

list_channels(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of channels.

list_dataset_contents(client, dataset_name, max_results \\ nil, next_token \\ nil, scheduled_before \\ nil, scheduled_on_or_after \\ nil, options \\ [])

Lists information about data set contents that have been created.

list_datasets(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves information about data sets.

list_datastores(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of data stores.

list_pipelines(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a list of pipelines.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists the tags (metadata) which you have assigned to the resource.

put_logging_options(client, input, options \\ [])

Sets or updates the AWS IoT Analytics logging options.

Note that if you update the value of any loggingOptions field, it takes up to one minute for the change to take effect. Also, if you change the policy attached to the role you specified in the roleArn field (for example, to correct an invalid policy) it takes up to 5 minutes for that change to take effect.

run_pipeline_activity(client, input, options \\ [])

Simulates the results of running a pipeline activity on a message payload.

sample_channel_data(client, channel_name, end_time \\ nil, max_messages \\ nil, start_time \\ nil, options \\ [])

Retrieves a sample of messages from the specified channel ingested during the specified timeframe.

Up to 10 messages can be retrieved.

start_pipeline_reprocessing(client, pipeline_name, input, options \\ [])

Starts the reprocessing of raw message data through the pipeline.

tag_resource(client, input, options \\ [])

Adds to or modifies the tags of the given resource.

Tags are metadata which can be used to manage a resource.

untag_resource(client, input, options \\ [])

Removes the given tags (metadata) from the resource.

update_channel(client, channel_name, input, options \\ [])

Updates the settings of a channel.

update_dataset(client, dataset_name, input, options \\ [])

Updates the settings of a data set.

update_datastore(client, datastore_name, input, options \\ [])

Updates the settings of a data store.

update_pipeline(client, pipeline_name, input, options \\ [])

Updates the settings of a pipeline.

You must specify both a channel and a datastore activity and, optionally, as many as 23 additional activities in the pipelineActivities array.