aws-elixir v0.6.0 AWS.IoTDataPlane

AWS IoT

AWS IoT-Data enables secure, bi-directional communication between Internet-connected things (such as sensors, actuators, embedded devices, or smart appliances) and the AWS cloud.

It implements a broker for applications and things to publish messages over HTTP (Publish) and retrieve, update, and delete shadows. A shadow is a persistent representation of your things and their state in the AWS cloud.

Find the endpoint address for actions in the AWS IoT data plane by running this CLI command:

aws iot describe-endpoint --endpoint-type iot:Data-ATS

The service name used by AWS Signature Version 4 to sign requests is: iotdevicegateway.

Functions

delete_thing_shadow(client, thing_name, input, options \\ [])

Deletes the shadow for the specified thing.

For more information, see DeleteThingShadow in the AWS IoT Developer Guide.

get_thing_shadow(client, thing_name, shadow_name \\ nil, options \\ [])

Gets the shadow for the specified thing.

For more information, see GetThingShadow in the AWS IoT Developer Guide.

list_named_shadows_for_thing(client, thing_name, next_token \\ nil, page_size \\ nil, options \\ [])

Lists the shadows for the specified thing.

publish(client, topic, input, options \\ [])

Publishes state information.

For more information, see HTTP Protocol in the AWS IoT Developer Guide.

update_thing_shadow(client, thing_name, input, options \\ [])

Updates the shadow for the specified thing.

For more information, see UpdateThingShadow in the AWS IoT Developer Guide.