aws-elixir v0.6.0 AWS.IoTEvents

AWS IoT Events monitors your equipment or device fleets for failures or changes in operation, and triggers actions when such events occur.

You can use AWS IoT Events API operations to create, read, update, and delete inputs and detector models, and to list their versions.

Functions

create_detector_model(client, input, options \\ [])

Creates a detector model.

create_input(client, input, options \\ [])

Creates an input.

delete_detector_model(client, detector_model_name, input, options \\ [])

Deletes a detector model.

Any active instances of the detector model are also deleted.

delete_input(client, input_name, input, options \\ [])

Deletes an input.

describe_detector_model(client, detector_model_name, detector_model_version \\ nil, options \\ [])

Describes a detector model.

If the version parameter is not specified, information about the latest version is returned.

describe_input(client, input_name, options \\ [])

Describes an input.

describe_logging_options(client, options \\ [])

Retrieves the current settings of the AWS IoT Events logging options.

list_detector_model_versions(client, detector_model_name, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all the versions of a detector model.

Only the metadata associated with each detector model version is returned.

list_detector_models(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the detector models you have created.

Only the metadata associated with each detector model is returned.

list_inputs(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the inputs you have created.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists the tags (metadata) you have assigned to the resource.

put_logging_options(client, input, options \\ [])

Sets or updates the AWS IoT Events logging options.

If you update the value of any loggingOptions field, it takes up to one minute for the change to take effect. If you change the policy attached to the role you specified in the roleArn field (for example, to correct an invalid policy), it takes up to five minutes for that change to take effect.

tag_resource(client, input, options \\ [])

Adds to or modifies the tags of the given resource.

Tags are metadata that can be used to manage a resource.

untag_resource(client, input, options \\ [])

Removes the given tags (metadata) from the resource.

update_detector_model(client, detector_model_name, input, options \\ [])

Updates a detector model.

Detectors (instances) spawned by the previous version are deleted and then re-created as new inputs arrive.

update_input(client, input_name, input, options \\ [])

Updates an input.