aws-elixir v0.6.0 AWS.IoTEventsData

AWS IoT Events monitors your equipment or device fleets for failures or changes in operation, and triggers actions when such events occur.

AWS IoT Events Data API commands enable you to send inputs to detectors, list detectors, and view or update a detector's status.

Functions

batch_put_message(client, input, options \\ [])

Sends a set of messages to the AWS IoT Events system.

Each message payload is transformed into the input you specify ("inputName") and ingested into any detectors that monitor that input. If multiple messages are sent, the order in which the messages are processed isn't guaranteed. To guarantee ordering, you must send messages one at a time and wait for a successful response.

batch_update_detector(client, input, options \\ [])

Updates the state, variable values, and timer settings of one or more detectors (instances) of a specified detector model.

describe_detector(client, detector_model_name, key_value \\ nil, options \\ [])

Returns information about the specified detector (instance).

list_detectors(client, detector_model_name, max_results \\ nil, next_token \\ nil, state_name \\ nil, options \\ [])

Lists detectors (the instances of a detector model).