aws-elixir v0.6.0 AWS.IoTSecureTunneling

AWS IoT Secure Tunneling

AWS IoT Secure Tunnling enables you to create remote connections to devices deployed in the field.

For more information about how AWS IoT Secure Tunneling works, see the User Guide.

Functions

close_tunnel(client, input, options \\ [])

Closes a tunnel identified by the unique tunnel id.

When a CloseTunnel request is received, we close the WebSocket connections between the client and proxy server so no data can be transmitted.

describe_tunnel(client, input, options \\ [])

Gets information about a tunnel identified by the unique tunnel id.

list_tags_for_resource(client, input, options \\ [])

Lists the tags for the specified resource.

list_tunnels(client, input, options \\ [])

List all tunnels for an AWS account.

Tunnels are listed by creation time in descending order, newer tunnels will be listed before older tunnels.

open_tunnel(client, input, options \\ [])

Creates a new tunnel, and returns two client access tokens for clients to use to connect to the AWS IoT Secure Tunneling proxy server.

.

tag_resource(client, input, options \\ [])

A resource tag.

untag_resource(client, input, options \\ [])

Removes a tag from a resource.