aws-elixir v0.6.0 AWS.IoTSiteWise

Welcome to the AWS IoT SiteWise API Reference.

AWS IoT SiteWise is an AWS service that connects Industrial Internet of Things (IIoT) devices to the power of the AWS Cloud. For more information, see the AWS IoT SiteWise User Guide. For information about AWS IoT SiteWise quotas, see Quotas in the AWS IoT SiteWise User Guide.

Functions

associate_assets(client, asset_id, input, options \\ [])

Associates a child asset with the given parent asset through a hierarchy defined in the parent asset's model.

For more information, see Associating assets in the AWS IoT SiteWise User Guide.

batch_associate_project_assets(client, project_id, input, options \\ [])

Associates a group (batch) of assets with an AWS IoT SiteWise Monitor project.

batch_disassociate_project_assets(client, project_id, input, options \\ [])

Disassociates a group (batch) of assets from an AWS IoT SiteWise Monitor project.

batch_put_asset_property_value(client, input, options \\ [])

Sends a list of asset property values to AWS IoT SiteWise.

Each value is a timestamp-quality-value (TQV) data point. For more information, see Ingesting data using the API in the AWS IoT SiteWise User Guide.

To identify an asset property, you must specify one of the following:

The assetId and propertyId of an asset property.

A propertyAlias, which is a data stream alias (for example, /company/windfarm/3/turbine/7/temperature). To define an asset property's alias, see UpdateAssetProperty.

With respect to Unix epoch time, AWS IoT SiteWise accepts only TQVs that have a timestamp of no more than 15 minutes in the past and no more than 5 minutes in the future. AWS IoT SiteWise rejects timestamps outside of the inclusive range of [-15, +5] minutes and returns a TimestampOutOfRangeException error.

For each asset property, AWS IoT SiteWise overwrites TQVs with duplicate timestamps unless the newer TQV has a different quality. For example, if you store a TQV {T1, GOOD, V1}, then storing {T1, GOOD, V2} replaces the existing TQV.

AWS IoT SiteWise authorizes access to each BatchPutAssetPropertyValue entry individually. For more information, see BatchPutAssetPropertyValue authorization in the AWS IoT SiteWise User Guide.

create_access_policy(client, input, options \\ [])

Creates an access policy that grants the specified identity (AWS SSO user, AWS SSO group, or IAM user) access to the specified AWS IoT SiteWise Monitor portal or project resource.

create_asset(client, input, options \\ [])

Creates an asset from an existing asset model.

For more information, see Creating assets in the AWS IoT SiteWise User Guide.

create_asset_model(client, input, options \\ [])

Creates an asset model from specified property and hierarchy definitions.

You create assets from asset models. With asset models, you can easily create assets of the same type that have standardized definitions. Each asset created from a model inherits the asset model's property and hierarchy definitions. For more information, see Defining asset models in the AWS IoT SiteWise User Guide.

create_dashboard(client, input, options \\ [])

Creates a dashboard in an AWS IoT SiteWise Monitor project.

create_gateway(client, input, options \\ [])

Creates a gateway, which is a virtual or edge device that delivers industrial data streams from local servers to AWS IoT SiteWise.

For more information, see Ingesting data using a gateway in the AWS IoT SiteWise User Guide.

create_portal(client, input, options \\ [])

Creates a portal, which can contain projects and dashboards.

AWS IoT SiteWise Monitor uses AWS SSO or IAM to authenticate portal users and manage user permissions.

Before you can sign in to a new portal, you must add at least one identity to that portal. For more information, see Adding or removing portal administrators in the AWS IoT SiteWise User Guide.

create_presigned_portal_url(client, portal_id, session_duration_seconds \\ nil, options \\ [])

Creates a pre-signed URL to a portal.

Use this operation to create URLs to portals that use AWS Identity and Access Management (IAM) to authenticate users. An IAM user with access to a portal can call this API to get a URL to that portal. The URL contains a session token that lets the IAM user access the portal.

create_project(client, input, options \\ [])

Creates a project in the specified portal.

delete_access_policy(client, access_policy_id, input, options \\ [])

Deletes an access policy that grants the specified identity access to the specified AWS IoT SiteWise Monitor resource.

You can use this operation to revoke access to an AWS IoT SiteWise Monitor resource.

delete_asset(client, asset_id, input, options \\ [])

Deletes an asset.

This action can't be undone. For more information, see Deleting assets and models in the AWS IoT SiteWise User Guide.

You can't delete an asset that's associated to another asset. For more information, see DisassociateAssets.

delete_asset_model(client, asset_model_id, input, options \\ [])

Deletes an asset model.

This action can't be undone. You must delete all assets created from an asset model before you can delete the model. Also, you can't delete an asset model if a parent asset model exists that contains a property formula expression that depends on the asset model that you want to delete. For more information, see Deleting assets and models in the AWS IoT SiteWise User Guide.

delete_dashboard(client, dashboard_id, input, options \\ [])

Deletes a dashboard from AWS IoT SiteWise Monitor.

delete_gateway(client, gateway_id, input, options \\ [])

Deletes a gateway from AWS IoT SiteWise.

When you delete a gateway, some of the gateway's files remain in your gateway's file system.

delete_portal(client, portal_id, input, options \\ [])

Deletes a portal from AWS IoT SiteWise Monitor.

delete_project(client, project_id, input, options \\ [])

Deletes a project from AWS IoT SiteWise Monitor.

describe_access_policy(client, access_policy_id, options \\ [])

Describes an access policy, which specifies an identity's access to an AWS IoT SiteWise Monitor portal or project.

describe_asset(client, asset_id, options \\ [])

Retrieves information about an asset.

describe_asset_model(client, asset_model_id, options \\ [])

Retrieves information about an asset model.

describe_asset_property(client, asset_id, property_id, options \\ [])

Retrieves information about an asset property.

When you call this operation for an attribute property, this response includes the default attribute value that you define in the asset model. If you update the default value in the model, this operation's response includes the new default value.

This operation doesn't return the value of the asset property. To get the value of an asset property, use GetAssetPropertyValue.

describe_dashboard(client, dashboard_id, options \\ [])

Retrieves information about a dashboard.

describe_gateway(client, gateway_id, options \\ [])

Retrieves information about a gateway.

describe_gateway_capability_configuration(client, capability_namespace, gateway_id, options \\ [])

Retrieves information about a gateway capability configuration.

Each gateway capability defines data sources for a gateway. A capability configuration can contain multiple data source configurations. If you define OPC-UA sources for a gateway in the AWS IoT SiteWise console, all of your OPC-UA sources are stored in one capability configuration. To list all capability configurations for a gateway, use DescribeGateway.

describe_logging_options(client, options \\ [])

Retrieves the current AWS IoT SiteWise logging options.

describe_portal(client, portal_id, options \\ [])

Retrieves information about a portal.

describe_project(client, project_id, options \\ [])

Retrieves information about a project.

disassociate_assets(client, asset_id, input, options \\ [])

Disassociates a child asset from the given parent asset through a hierarchy defined in the parent asset's model.

get_asset_property_aggregates(client, aggregate_types, asset_id \\ nil, end_date, max_results \\ nil, next_token \\ nil, property_alias \\ nil, property_id \\ nil, qualities \\ nil, resolution, start_date, time_ordering \\ nil, options \\ [])

Gets aggregated values for an asset property.

For more information, see Querying aggregates in the AWS IoT SiteWise User Guide.

To identify an asset property, you must specify one of the following:

The assetId and propertyId of an asset property.

A propertyAlias, which is a data stream alias (for example, /company/windfarm/3/turbine/7/temperature). To define an asset property's alias, see UpdateAssetProperty.

get_asset_property_value(client, asset_id \\ nil, property_alias \\ nil, property_id \\ nil, options \\ [])

Gets an asset property's current value.

For more information, see Querying current values in the AWS IoT SiteWise User Guide.

To identify an asset property, you must specify one of the following:

The assetId and propertyId of an asset property.

A propertyAlias, which is a data stream alias (for example, /company/windfarm/3/turbine/7/temperature). To define an asset property's alias, see UpdateAssetProperty.

get_asset_property_value_history(client, asset_id \\ nil, end_date \\ nil, max_results \\ nil, next_token \\ nil, property_alias \\ nil, property_id \\ nil, qualities \\ nil, start_date \\ nil, time_ordering \\ nil, options \\ [])

Gets the history of an asset property's values.

For more information, see Querying historical values in the AWS IoT SiteWise User Guide.

To identify an asset property, you must specify one of the following:

The assetId and propertyId of an asset property.

A propertyAlias, which is a data stream alias (for example, /company/windfarm/3/turbine/7/temperature). To define an asset property's alias, see UpdateAssetProperty.

list_access_policies(client, iam_arn \\ nil, identity_id \\ nil, identity_type \\ nil, max_results \\ nil, next_token \\ nil, resource_id \\ nil, resource_type \\ nil, options \\ [])

Retrieves a paginated list of access policies for an identity (an AWS SSO user, an AWS SSO group, or an IAM user) or an AWS IoT SiteWise Monitor resource (a portal or project).

list_asset_models(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a paginated list of summaries of all asset models.

list_assets(client, asset_model_id \\ nil, filter \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a paginated list of asset summaries.

You can use this operation to do the following:

List assets based on a specific asset model.

List top-level assets.

You can't use this operation to list all assets. To retrieve summaries for all of your assets, use ListAssetModels to get all of your asset model IDs. Then, use ListAssets to get all assets for each asset model.

list_associated_assets(client, asset_id, hierarchy_id \\ nil, max_results \\ nil, next_token \\ nil, traversal_direction \\ nil, options \\ [])

Retrieves a paginated list of associated assets.

You can use this operation to do the following:

List child assets associated to a parent asset by a hierarchy that you specify.

List an asset's parent asset.

list_dashboards(client, max_results \\ nil, next_token \\ nil, project_id, options \\ [])

Retrieves a paginated list of dashboards for an AWS IoT SiteWise Monitor project.

list_gateways(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a paginated list of gateways.

list_portals(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a paginated list of AWS IoT SiteWise Monitor portals.

list_project_assets(client, project_id, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a paginated list of assets associated with an AWS IoT SiteWise Monitor project.

list_projects(client, max_results \\ nil, next_token \\ nil, portal_id, options \\ [])

Retrieves a paginated list of projects for an AWS IoT SiteWise Monitor portal.

list_tags_for_resource(client, resource_arn, options \\ [])

Retrieves the list of tags for an AWS IoT SiteWise resource.

put_logging_options(client, input, options \\ [])

Sets logging options for AWS IoT SiteWise.

tag_resource(client, input, options \\ [])

Adds tags to an AWS IoT SiteWise resource.

If a tag already exists for the resource, this operation updates the tag's value.

untag_resource(client, input, options \\ [])

Removes a tag from an AWS IoT SiteWise resource.

update_access_policy(client, access_policy_id, input, options \\ [])

Updates an existing access policy that specifies an identity's access to an AWS IoT SiteWise Monitor portal or project resource.

update_asset(client, asset_id, input, options \\ [])

Updates an asset's name.

For more information, see Updating assets and models in the AWS IoT SiteWise User Guide.

update_asset_model(client, asset_model_id, input, options \\ [])

Updates an asset model and all of the assets that were created from the model.

Each asset created from the model inherits the updated asset model's property and hierarchy definitions. For more information, see Updating assets and models in the AWS IoT SiteWise User Guide.

This operation overwrites the existing model with the provided model. To avoid deleting your asset model's properties or hierarchies, you must include their IDs and definitions in the updated asset model payload. For more information, see DescribeAssetModel.

If you remove a property from an asset model, AWS IoT SiteWise deletes all previous data for that property. If you remove a hierarchy definition from an asset model, AWS IoT SiteWise disassociates every asset associated with that hierarchy. You can't change the type or data type of an existing property.

update_asset_property(client, asset_id, property_id, input, options \\ [])

Updates an asset property's alias and notification state.

This operation overwrites the property's existing alias and notification state. To keep your existing property's alias or notification state, you must include the existing values in the UpdateAssetProperty request. For more information, see DescribeAssetProperty.

update_dashboard(client, dashboard_id, input, options \\ [])

Updates an AWS IoT SiteWise Monitor dashboard.

update_gateway(client, gateway_id, input, options \\ [])

Updates a gateway's name.

update_gateway_capability_configuration(client, gateway_id, input, options \\ [])

Updates a gateway capability configuration or defines a new capability configuration.

Each gateway capability defines data sources for a gateway. A capability configuration can contain multiple data source configurations. If you define OPC-UA sources for a gateway in the AWS IoT SiteWise console, all of your OPC-UA sources are stored in one capability configuration. To list all capability configurations for a gateway, use DescribeGateway.

update_portal(client, portal_id, input, options \\ [])

Updates an AWS IoT SiteWise Monitor portal.

update_project(client, project_id, input, options \\ [])

Updates an AWS IoT SiteWise Monitor project.