aws-elixir v0.6.0 AWS.Ivs

Introduction

The Amazon Interactive Video Service (IVS) API is REST compatible, using a standard HTTP API and an AWS SNS event stream for responses.

JSON is used for both requests and responses, including errors.

The API is an AWS regional service, currently in these regions: us-west-2, us-east-1, and eu-west-1.

All API request parameters and URLs are case sensitive. *

For a summary of notable documentation changes in each release, see Document History.

Service Endpoints

The following are the Amazon IVS service endpoints (all HTTPS):

Region name: US West (Oregon)

Region: us-west-2

Endpoint: ivs.us-west-2.amazonaws.com

Region name: US East (Virginia)

Region: us-east-1

Endpoint: ivs.us-east-1.amazonaws.com

Region name: EU West (Dublin)

Region: eu-west-1

Endpoint: ivs.eu-west-1.amazonaws.com

Allowed Header Values

**Accept:** application/json

**Accept-Encoding:** gzip, deflate

**Content-Type:**application/json

Resources

The following resources contain information about your IVS live stream (see Getting Started with Amazon IVS):

Channel — Stores configuration data related to your live stream. You first create a channel and then use the channel’s stream key to start your live stream. See the Channel endpoints for more information.

Stream key — An identifier assigned by Amazon IVS when you create a channel, which is then used to authorize streaming. See the StreamKey endpoints for more information. Treat the stream key like a secret, since it allows anyone to stream to the channel.

Playback key pair — Video playback may be restricted using playback-authorization tokens, which use public-key encryption. A playback key pair is the public-private pair of keys used to sign and validate the playback-authorization token. See the PlaybackKeyPair endpoints for more information.

Tagging

A tag is a metadata label that you assign to an AWS resource. A tag comprises a key and a value, both set by you. For example, you might set a tag as topic:nature to label a particular video category. See Tagging AWS Resources for more information, including restrictions that apply to tags.

Tags can help you identify and organize your AWS resources. For example, you can use the same tag for different resources to indicate that they are related. You can also use tags to manage access (see Access Tags).

The Amazon IVS API has these tag-related endpoints: TagResource, UntagResource, and ListTagsForResource. The following resources support tagging: Channels, Stream Keys, and Playback Key Pairs.

Channel Endpoints

CreateChannel — Creates a new channel and an associated stream key to start streaming.

GetChannel — Gets the channel configuration for the specified channel ARN (Amazon Resource Name).

BatchGetChannel — Performs GetChannel on multiple ARNs simultaneously.

ListChannels — Gets summary information about all channels in your account, in the AWS region where the API request is processed. This list can be filtered to match a specified string.

UpdateChannel — Updates a channel's configuration. This does not affect an ongoing stream of this channel. You must stop and restart the stream for the changes to take effect.

DeleteChannel — Deletes the specified channel.

StreamKey Endpoints

CreateStreamKey — Creates a stream key, used to initiate a stream, for the specified channel ARN.

GetStreamKey — Gets stream key information for the specified ARN.

BatchGetStreamKey — Performs GetStreamKey on multiple ARNs simultaneously.

ListStreamKeys — Gets summary information about stream keys for the specified channel.

DeleteStreamKey — Deletes the stream key for the specified ARN, so it can no longer be used to stream.

Stream Endpoints

GetStream — Gets information about the active (live) stream on a specified channel.

ListStreams — Gets summary information about live streams in your account, in the AWS region where the API request is processed.

StopStream — Disconnects the incoming RTMPS stream for the specified channel. Can be used in conjunction with DeleteStreamKey to prevent further streaming to a channel.

PutMetadata — Inserts metadata into an RTMPS stream for the specified channel. A maximum of 5 requests per second per channel is allowed, each with a maximum 1KB payload.

PlaybackKeyPair Endpoints

ImportPlaybackKeyPair — Imports the public portion of a new key pair and returns its arn and fingerprint. The privateKey can then be used to generate viewer authorization tokens, to grant viewers access to authorized channels.

GetPlaybackKeyPair — Gets a specified playback authorization key pair and returns the arn and fingerprint. The privateKey held by the caller can be used to generate viewer authorization tokens, to grant viewers access to authorized channels.

ListPlaybackKeyPairs — Gets summary information about playback key pairs.

DeletePlaybackKeyPair — Deletes a specified authorization key pair. This invalidates future viewer tokens generated using the key pair’s privateKey.

AWS Tags Endpoints

TagResource — Adds or updates tags for the AWS resource with the specified ARN.

UntagResource — Removes tags from the resource with the specified ARN.

ListTagsForResource — Gets information about AWS tags for the specified ARN.

Functions

batch_get_channel(client, input, options \\ [])

Performs GetChannel on multiple ARNs simultaneously.

batch_get_stream_key(client, input, options \\ [])

Performs GetStreamKey on multiple ARNs simultaneously.

create_channel(client, input, options \\ [])

Creates a new channel and an associated stream key to start streaming.

create_stream_key(client, input, options \\ [])

Creates a stream key, used to initiate a stream, for the specified channel ARN.

Note that CreateChannel creates a stream key. If you subsequently use CreateStreamKey on the same channel, it will fail because a stream key already exists and there is a limit of 1 stream key per channel. To reset the stream key on a channel, use DeleteStreamKey and then CreateStreamKey.

delete_channel(client, input, options \\ [])

Deletes the specified channel and its associated stream keys.

delete_playback_key_pair(client, input, options \\ [])

Deletes a specified authorization key pair.

This invalidates future viewer tokens generated using the key pair’s privateKey.

delete_stream_key(client, input, options \\ [])

Deletes the stream key for the specified ARN, so it can no longer be used to stream.

get_channel(client, input, options \\ [])

Gets the channel configuration for the specified channel ARN.

See also BatchGetChannel.

get_playback_key_pair(client, input, options \\ [])

Gets a specified playback authorization key pair and returns the arn and fingerprint.

The privateKey held by the caller can be used to generate viewer authorization tokens, to grant viewers access to authorized channels.

get_stream(client, input, options \\ [])

Gets information about the active (live) stream on a specified channel.

get_stream_key(client, input, options \\ [])

Gets stream-key information for a specified ARN.

import_playback_key_pair(client, input, options \\ [])

Imports the public portion of a new key pair and returns its arn and fingerprint.

The privateKey can then be used to generate viewer authorization tokens, to grant viewers access to authorized channels.

list_channels(client, input, options \\ [])

Gets summary information about all channels in your account, in the AWS region where the API request is processed.

This list can be filtered to match a specified string.

list_playback_key_pairs(client, input, options \\ [])

Gets summary information about playback key pairs.

list_stream_keys(client, input, options \\ [])

Gets summary information about stream keys for the specified channel.

list_streams(client, input, options \\ [])

Gets summary information about live streams in your account, in the AWS region where the API request is processed.

list_tags_for_resource(client, resource_arn, options \\ [])

Gets information about AWS tags for the specified ARN.

put_metadata(client, input, options \\ [])

Inserts metadata into an RTMPS stream for the specified channel.

A maximum of 5 requests per second per channel is allowed, each with a maximum 1KB payload.

stop_stream(client, input, options \\ [])

Disconnects the incoming RTMPS stream for the specified channel.

Can be used in conjunction with DeleteStreamKey to prevent further streaming to a channel.

Many streaming client-software libraries automatically reconnect a dropped RTMPS session, so to stop the stream permanently, you may want to first revoke the streamKey attached to the channel.

tag_resource(client, resource_arn, input, options \\ [])

Adds or updates tags for the AWS resource with the specified ARN.

untag_resource(client, resource_arn, input, options \\ [])

Removes tags from the resource with the specified ARN.

update_channel(client, input, options \\ [])

Updates a channel's configuration.

This does not affect an ongoing stream of this channel. You must stop and restart the stream for the changes to take effect.