aws-elixir v0.6.0 AWS.Kafka

The operations for managing an Amazon MSK cluster.

Functions

batch_associate_scram_secret(client, cluster_arn, input, options \\ [])

Associates one or more Scram Secrets with an Amazon MSK cluster.

batch_disassociate_scram_secret(client, cluster_arn, input, options \\ [])

Disassociates one or more Scram Secrets from an Amazon MSK cluster.

create_cluster(client, input, options \\ [])

Creates a new MSK cluster.

create_configuration(client, input, options \\ [])

Creates a new MSK configuration.

delete_cluster(client, cluster_arn, input, options \\ [])

Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in the request.

delete_configuration(client, arn, input, options \\ [])

Deletes the specified MSK configuration.

The configuration must be in the ACTIVE or DELETE_FAILED state.

describe_cluster(client, cluster_arn, options \\ [])

Returns a description of the MSK cluster whose Amazon Resource Name (ARN) is specified in the request.

describe_cluster_operation(client, cluster_operation_arn, options \\ [])

Returns a description of the cluster operation specified by the ARN.

describe_configuration(client, arn, options \\ [])

Returns a description of this MSK configuration.

describe_configuration_revision(client, arn, revision, options \\ [])

Returns a description of this revision of the configuration.

get_bootstrap_brokers(client, cluster_arn, options \\ [])

A list of brokers that a client application can use to bootstrap.

get_compatible_kafka_versions(client, cluster_arn \\ nil, options \\ [])

Gets the Apache Kafka versions to which you can update the MSK cluster.

list_cluster_operations(client, cluster_arn, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of all the operations that have been performed on the specified MSK cluster.

list_clusters(client, cluster_name_filter \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of all the MSK clusters in the current Region.

list_configuration_revisions(client, arn, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of all the revisions of an MSK configuration.

list_configurations(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of all the MSK configurations in this Region.

list_kafka_versions(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of Kafka versions.

list_nodes(client, cluster_arn, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of the broker nodes in the cluster.

list_scram_secrets(client, cluster_arn, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of the Scram Secrets associated with an Amazon MSK cluster.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns a list of the tags associated with the specified resource.

reboot_broker(client, cluster_arn, input, options \\ [])

Executes a reboot on a broker.

tag_resource(client, resource_arn, input, options \\ [])

Adds tags to the specified MSK resource.

untag_resource(client, resource_arn, input, options \\ [])

Removes the tags associated with the keys that are provided in the query.

update_broker_count(client, cluster_arn, input, options \\ [])

Updates the number of broker nodes in the cluster.

You can use this operation to increase the number of brokers in an existing cluster. You can't decrease the number of brokers.

update_broker_storage(client, cluster_arn, input, options \\ [])

Updates the EBS storage associated with MSK brokers.

update_cluster_configuration(client, cluster_arn, input, options \\ [])

Updates the cluster with the configuration that is specified in the request body.

update_cluster_kafka_version(client, cluster_arn, input, options \\ [])

Updates the Apache Kafka version for the cluster.

update_configuration(client, arn, input, options \\ [])

Updates an existing MSK configuration.

The configuration must be in the Active state.

update_monitoring(client, cluster_arn, input, options \\ [])

Updates the monitoring settings for the cluster.

You can use this operation to specify which Apache Kafka metrics you want Amazon MSK to send to Amazon CloudWatch. You can also specify settings for open monitoring with Prometheus.