aws-elixir v0.6.0 AWS.Kendra

Amazon Kendra is a service for indexing large document sets.

Functions

batch_delete_document(client, input, options \\ [])

Removes one or more documents from an index.

The documents must have been added with the BatchPutDocument operation.

The documents are deleted asynchronously. You can see the progress of the deletion by using AWS CloudWatch. Any error messages releated to the processing of the batch are sent to you CloudWatch log.

batch_put_document(client, input, options \\ [])

Adds one or more documents to an index.

The BatchPutDocument operation enables you to ingest inline documents or a set of documents stored in an Amazon S3 bucket. Use this operation to ingest your text and unstructured text into an index, add custom attributes to the documents, and to attach an access control list to the documents added to the index.

The documents are indexed asynchronously. You can see the progress of the batch using AWS CloudWatch. Any error messages related to processing the batch are sent to your AWS CloudWatch log.

create_data_source(client, input, options \\ [])

Creates a data source that you use to with an Amazon Kendra index.

You specify a name, data source connector type and description for your data source. You also specify configuration information such as document metadata (author, source URI, and so on) and user context information.

CreateDataSource is a synchronous operation. The operation returns 200 if the data source was successfully created. Otherwise, an exception is raised.

create_faq(client, input, options \\ [])

Creates an new set of frequently asked question (FAQ) questions and answers.

create_index(client, input, options \\ [])

Creates a new Amazon Kendra index.

Index creation is an asynchronous operation. To determine if index creation has completed, check the Status field returned from a call to . The Status field is set to ACTIVE when the index is ready to use.

Once the index is active you can index your documents using the operation or using one of the supported data sources.

delete_data_source(client, input, options \\ [])

Deletes an Amazon Kendra data source.

An exception is not thrown if the data source is already being deleted. While the data source is being deleted, the Status field returned by a call to the operation is set to DELETING. For more information, see Deleting Data Sources.

delete_faq(client, input, options \\ [])

Removes an FAQ from an index.

delete_index(client, input, options \\ [])

Deletes an existing Amazon Kendra index.

An exception is not thrown if the index is already being deleted. While the index is being deleted, the Status field returned by a call to the DescribeIndex operation is set to DELETING.

describe_data_source(client, input, options \\ [])

Gets information about a Amazon Kendra data source.

describe_faq(client, input, options \\ [])

Gets information about an FAQ list.

describe_index(client, input, options \\ [])

Describes an existing Amazon Kendra index

list_data_source_sync_jobs(client, input, options \\ [])

Gets statistics about synchronizing Amazon Kendra with a data source.

list_data_sources(client, input, options \\ [])

Lists the data sources that you have created.

list_faqs(client, input, options \\ [])

Gets a list of FAQ lists associated with an index.

list_indices(client, input, options \\ [])

Lists the Amazon Kendra indexes that you have created.

list_tags_for_resource(client, input, options \\ [])

Gets a list of tags associated with a specified resource.

Indexes, FAQs, and data sources can have tags associated with them.

query(client, input, options \\ [])

Searches an active index.

Use this API to search your documents using query. The Query operation enables to do faceted search and to filter results based on document attributes.

It also enables you to provide user context that Amazon Kendra uses to enforce document access control in the search results.

Amazon Kendra searches your index for text content and question and answer (FAQ) content. By default the response contains three types of results.

Relevant passages

Matching FAQs

Relevant documents

You can specify that the query return only one type of result using the QueryResultTypeConfig parameter.

Each query returns the 100 most relevant results.

start_data_source_sync_job(client, input, options \\ [])

Starts a synchronization job for a data source.

If a synchronization job is already in progress, Amazon Kendra returns a ResourceInUseException exception.

stop_data_source_sync_job(client, input, options \\ [])

Stops a running synchronization job.

You can't stop a scheduled synchronization job.

submit_feedback(client, input, options \\ [])

Enables you to provide feedback to Amazon Kendra to improve the performance of the service.

tag_resource(client, input, options \\ [])

Adds the specified tag to the specified index, FAQ, or data source resource.

If the tag already exists, the existing value is replaced with the new value.

untag_resource(client, input, options \\ [])

Removes a tag from an index, FAQ, or a data source.

update_data_source(client, input, options \\ [])

Updates an existing Amazon Kendra data source.

update_index(client, input, options \\ [])

Updates an existing Amazon Kendra index.