aws-elixir v0.6.0 AWS.KinesisAnalyticsV2

Amazon Kinesis Data Analytics is a fully managed service that you can use to process and analyze streaming data using Java, SQL, or Scala.

The service enables you to quickly author and run Java, SQL, or Scala code against streaming sources to perform time series analytics, feed real-time dashboards, and create real-time metrics.

Functions

add_application_cloud_watch_logging_option(client, input, options \\ [])

Adds an Amazon CloudWatch log stream to monitor application configuration errors.

add_application_input(client, input, options \\ [])

Adds a streaming source to your SQL-based Kinesis Data Analytics application.

You can add a streaming source when you create an application, or you can use this operation to add a streaming source after you create an application. For more information, see CreateApplication.

Any configuration update, including adding a streaming source using this operation, results in a new version of the application. You can use the DescribeApplication operation to find the current application version.

add_application_input_processing_configuration(client, input, options \\ [])

Adds an InputProcessingConfiguration to a SQL-based Kinesis Data Analytics application.

An input processor pre-processes records on the input stream before the application's SQL code executes. Currently, the only input processor available is AWS Lambda.

add_application_output(client, input, options \\ [])

Adds an external destination to your SQL-based Kinesis Data Analytics application.

If you want Kinesis Data Analytics to deliver data from an in-application stream within your application to an external destination (such as an Kinesis data stream, a Kinesis Data Firehose delivery stream, or an AWS Lambda function), you add the relevant configuration to your application using this operation. You can configure one or more outputs for your application. Each output configuration maps an in-application stream and an external destination.

You can use one of the output configurations to deliver data from your in-application error stream to an external destination so that you can analyze the errors.

Any configuration update, including adding a streaming source using this operation, results in a new version of the application. You can use the DescribeApplication operation to find the current application version.

add_application_reference_data_source(client, input, options \\ [])

Adds a reference data source to an existing SQL-based Kinesis Data Analytics application.

Kinesis Data Analytics reads reference data (that is, an Amazon S3 object) and creates an in-application table within your application. In the request, you provide the source (S3 bucket name and object key name), name of the in-application table to create, and the necessary mapping information that describes how data in an Amazon S3 object maps to columns in the resulting in-application table.

add_application_vpc_configuration(client, input, options \\ [])

Adds a Virtual Private Cloud (VPC) configuration to the application.

Applications can use VPCs to store and access resources securely.

Note the following about VPC configurations for Kinesis Data Analytics applications:

VPC configurations are not supported for SQL applications.

When a VPC is added to a Kinesis Data Analytics application, the application can no longer be accessed from the Internet directly. To enable Internet access to the application, add an Internet gateway to your VPC.

create_application(client, input, options \\ [])

Creates a Kinesis Data Analytics application.

For information about creating a Kinesis Data Analytics application, see Creating an Application.

create_application_snapshot(client, input, options \\ [])

Creates a snapshot of the application's state data.

delete_application(client, input, options \\ [])

Deletes the specified application.

Kinesis Data Analytics halts application execution and deletes the application.

delete_application_cloud_watch_logging_option(client, input, options \\ [])

Deletes an Amazon CloudWatch log stream from an Kinesis Data Analytics application.

delete_application_input_processing_configuration(client, input, options \\ [])

Deletes an InputProcessingConfiguration from an input.

delete_application_output(client, input, options \\ [])

Deletes the output destination configuration from your SQL-based Kinesis Data Analytics application's configuration.

Kinesis Data Analytics will no longer write data from the corresponding in-application stream to the external output destination.

delete_application_reference_data_source(client, input, options \\ [])

Deletes a reference data source configuration from the specified SQL-based Kinesis Data Analytics application's configuration.

If the application is running, Kinesis Data Analytics immediately removes the in-application table that you created using the AddApplicationReferenceDataSource operation.

delete_application_snapshot(client, input, options \\ [])

Deletes a snapshot of application state.

delete_application_vpc_configuration(client, input, options \\ [])

Removes a VPC configuration from a Kinesis Data Analytics application.

describe_application(client, input, options \\ [])

Returns information about a specific Kinesis Data Analytics application.

If you want to retrieve a list of all applications in your account, use the ListApplications operation.

describe_application_snapshot(client, input, options \\ [])

Returns information about a snapshot of application state data.

discover_input_schema(client, input, options \\ [])

Infers a schema for a SQL-based Kinesis Data Analytics application by evaluating sample records on the specified streaming source (Kinesis data stream or Kinesis Data Firehose delivery stream) or Amazon S3 object.

In the response, the operation returns the inferred schema and also the sample records that the operation used to infer the schema.

You can use the inferred schema when configuring a streaming source for your application. When you create an application using the Kinesis Data Analytics console, the console uses this operation to infer a schema and show it in the console user interface.

list_application_snapshots(client, input, options \\ [])

Lists information about the current application snapshots.

list_applications(client, input, options \\ [])

Returns a list of Kinesis Data Analytics applications in your account.

For each application, the response includes the application name, Amazon Resource Name (ARN), and status.

If you want detailed information about a specific application, use DescribeApplication.

list_tags_for_resource(client, input, options \\ [])

Retrieves the list of key-value tags assigned to the application.

For more information, see Using Tagging.

start_application(client, input, options \\ [])

Starts the specified Kinesis Data Analytics application.

After creating an application, you must exclusively call this operation to start your application.

stop_application(client, input, options \\ [])

Stops the application from processing data.

You can stop an application only if it is in the running state. You can use the DescribeApplication operation to find the application state.

tag_resource(client, input, options \\ [])

Adds one or more key-value tags to a Kinesis Data Analytics application.

Note that the maximum number of application tags includes system tags. The maximum number of user-defined application tags is 50. For more information, see Using Tagging.

untag_resource(client, input, options \\ [])

Removes one or more tags from a Kinesis Data Analytics application.

For more information, see Using Tagging.

update_application(client, input, options \\ [])

Updates an existing Kinesis Data Analytics application.

Using this operation, you can update application code, input configuration, and output configuration.

Kinesis Data Analytics updates the ApplicationVersionId each time you update your application.

You cannot update the RuntimeEnvironment of an existing application. If you need to update an application's RuntimeEnvironment, you must delete the application and create it again.