aws-elixir v0.6.0 AWS.LakeFormation

AWS Lake Formation

Defines the public endpoint for the AWS Lake Formation service.

Functions

batch_grant_permissions(client, input, options \\ [])

Batch operation to grant permissions to the principal.

batch_revoke_permissions(client, input, options \\ [])

Batch operation to revoke permissions from the principal.

deregister_resource(client, input, options \\ [])

Deregisters the resource as managed by the Data Catalog.

When you deregister a path, Lake Formation removes the path from the inline policy attached to your service-linked role.

describe_resource(client, input, options \\ [])

Retrieves the current data access role for the given resource registered in AWS Lake Formation.

get_data_lake_settings(client, input, options \\ [])

Retrieves the list of the data lake administrators of a Lake Formation-managed data lake.

get_effective_permissions_for_path(client, input, options \\ [])

Returns the Lake Formation permissions for a specified table or database resource located at a path in Amazon S3.

GetEffectivePermissionsForPath will not return databases and tables if the catalog is encrypted.

grant_permissions(client, input, options \\ [])

Grants permissions to the principal to access metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3.

For information about permissions, see Security and Access Control to Metadata and Data.

list_permissions(client, input, options \\ [])

Returns a list of the principal permissions on the resource, filtered by the permissions of the caller.

For example, if you are granted an ALTER permission, you are able to see only the principal permissions for ALTER.

This operation returns only those permissions that have been explicitly granted.

For information about permissions, see Security and Access Control to Metadata and Data.

list_resources(client, input, options \\ [])

Lists the resources registered to be managed by the Data Catalog.

put_data_lake_settings(client, input, options \\ [])

Sets the list of data lake administrators who have admin privileges on all resources managed by Lake Formation.

For more information on admin privileges, see Granting Lake Formation Permissions.

This API replaces the current list of data lake admins with the new list being passed. To add an admin, fetch the current list and add the new admin to that list and pass that list in this API.

register_resource(client, input, options \\ [])

Registers the resource as managed by the Data Catalog.

To add or update data, Lake Formation needs read/write access to the chosen Amazon S3 path. Choose a role that you know has permission to do this, or choose the AWSServiceRoleForLakeFormationDataAccess service-linked role. When you register the first Amazon S3 path, the service-linked role and a new inline policy are created on your behalf. Lake Formation adds the first path to the inline policy and attaches it to the service-linked role. When you register subsequent paths, Lake Formation adds the path to the existing policy.

The following request registers a new location and gives AWS Lake Formation permission to use the service-linked role to access that location.

ResourceArn = arn:aws:s3:::my-bucket UseServiceLinkedRole = true

If UseServiceLinkedRole is not set to true, you must provide or set the RoleArn:

arn:aws:iam::12345:role/my-data-access-role

revoke_permissions(client, input, options \\ [])

Revokes permissions to the principal to access metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3.

update_resource(client, input, options \\ [])

Updates the data access role used for vending access to the given (registered) resource in AWS Lake Formation.