aws-elixir v0.6.0 AWS.Lambda

AWS Lambda

Overview

This is the AWS Lambda API Reference.

The AWS Lambda Developer Guide provides additional information. For the service overview, see What is AWS Lambda, and for information about how the service works, see AWS Lambda: How it Works in the AWS Lambda Developer Guide.

Functions

add_layer_version_permission(client, layer_name, version_number, input, options \\ [])

Adds permissions to the resource-based policy of a version of an AWS Lambda layer.

Use this action to grant layer usage permission to other accounts. You can grant permission to a single account, all AWS accounts, or all accounts in an organization.

To revoke permission, call RemoveLayerVersionPermission with the statement ID that you specified when you added it.

add_permission(client, function_name, input, options \\ [])

Grants an AWS service or another account permission to use a function.

You can apply the policy at the function level, or specify a qualifier to restrict access to a single version or alias. If you use a qualifier, the invoker must use the full Amazon Resource Name (ARN) of that version or alias to invoke the function.

To grant permission to another account, specify the account ID as the Principal. For AWS services, the principal is a domain-style identifier defined by the service, like s3.amazonaws.com or sns.amazonaws.com. For AWS services, you can also specify the ARN of the associated resource as the SourceArn. If you grant permission to a service principal without specifying the source, other accounts could potentially configure resources in their account to invoke your Lambda function.

This action adds a statement to a resource-based permissions policy for the function. For more information about function policies, see Lambda Function Policies.

create_alias(client, function_name, input, options \\ [])

Creates an alias for a Lambda function version.

Use aliases to provide clients with a function identifier that you can update to invoke a different version.

You can also map an alias to split invocation requests between two versions. Use the RoutingConfig parameter to specify a second version and the percentage of invocation requests that it receives.

create_event_source_mapping(client, input, options \\ [])

Creates a mapping between an event source and an AWS Lambda function.

Lambda reads items from the event source and triggers the function.

For details about each event source type, see the following topics.

Using AWS Lambda with Amazon DynamoDB

Using AWS Lambda with Amazon Kinesis

Using AWS Lambda with Amazon SQS

Using AWS Lambda with Amazon MSK

The following error handling options are only available for stream sources (DynamoDB and Kinesis):

BisectBatchOnFunctionError - If the function returns an error, split the batch in two and retry.

DestinationConfig - Send discarded records to an Amazon SQS queue or Amazon SNS topic.

MaximumRecordAgeInSeconds - Discard records older than the specified age. Default -1 (infinite). Minimum 60. Maximum 604800.

MaximumRetryAttempts - Discard records after the specified number of retries. Default -1 (infinite). Minimum 0. Maximum 10000. When infinite, failed records will be retried until the record expires.

ParallelizationFactor - Process multiple batches from each shard concurrently.

create_function(client, input, options \\ [])

Creates a Lambda function.

To create a function, you need a deployment package and an execution role. The deployment package contains your function code. The execution role grants the function permission to use AWS services, such as Amazon CloudWatch Logs for log streaming and AWS X-Ray for request tracing.

When you create a function, Lambda provisions an instance of the function and its supporting resources. If your function connects to a VPC, this process can take a minute or so. During this time, you can't invoke or modify the function. The State, StateReason, and StateReasonCode fields in the response from GetFunctionConfiguration indicate when the function is ready to invoke. For more information, see Function States.

A function has an unpublished version, and can have published versions and aliases. The unpublished version changes when you update your function's code and configuration. A published version is a snapshot of your function code and configuration that can't be changed. An alias is a named resource that maps to a version, and can be changed to map to a different version. Use the Publish parameter to create version 1 of your function from its initial configuration.

The other parameters let you configure version-specific and function-level settings. You can modify version-specific settings later with UpdateFunctionConfiguration. Function-level settings apply to both the unpublished and published versions of the function, and include tags (TagResource) and per-function concurrency limits (PutFunctionConcurrency).

If another account or an AWS service invokes your function, use AddPermission to grant permission by creating a resource-based IAM policy. You can grant permissions at the function level, on a version, or on an alias.

To invoke your function directly, use Invoke. To invoke your function in response to events in other AWS services, create an event source mapping (CreateEventSourceMapping), or configure a function trigger in the other service. For more information, see Invoking Functions.

delete_alias(client, function_name, name, input, options \\ [])

Deletes a Lambda function alias.

delete_event_source_mapping(client, uuid, input, options \\ [])

Deletes an event source mapping.

You can get the identifier of a mapping from the output of ListEventSourceMappings.

When you delete an event source mapping, it enters a Deleting state and might not be completely deleted for several seconds.

delete_function(client, function_name, input, options \\ [])

Deletes a Lambda function.

To delete a specific function version, use the Qualifier parameter. Otherwise, all versions and aliases are deleted.

To delete Lambda event source mappings that invoke a function, use DeleteEventSourceMapping. For AWS services and resources that invoke your function directly, delete the trigger in the service where you originally configured it.

delete_function_concurrency(client, function_name, input, options \\ [])

Removes a concurrent execution limit from a function.

delete_function_event_invoke_config(client, function_name, input, options \\ [])

Deletes the configuration for asynchronous invocation for a function, version, or alias.

To configure options for asynchronous invocation, use PutFunctionEventInvokeConfig.

delete_layer_version(client, layer_name, version_number, input, options \\ [])

Deletes a version of an AWS Lambda layer.

Deleted versions can no longer be viewed or added to functions. To avoid breaking functions, a copy of the version remains in Lambda until no functions refer to it.

delete_provisioned_concurrency_config(client, function_name, input, options \\ [])

Deletes the provisioned concurrency configuration for a function.

get_account_settings(client, options \\ [])

Retrieves details about your account's limits and usage in an AWS Region.

get_alias(client, function_name, name, options \\ [])

Returns details about a Lambda function alias.

get_event_source_mapping(client, uuid, options \\ [])

Returns details about an event source mapping.

You can get the identifier of a mapping from the output of ListEventSourceMappings.

get_function(client, function_name, qualifier \\ nil, options \\ [])

Returns information about the function or function version, with a link to download the deployment package that's valid for 10 minutes.

If you specify a function version, only details that are specific to that version are returned.

get_function_concurrency(client, function_name, options \\ [])

Returns details about the reserved concurrency configuration for a function.

To set a concurrency limit for a function, use PutFunctionConcurrency.

get_function_configuration(client, function_name, qualifier \\ nil, options \\ [])

Returns the version-specific settings of a Lambda function or version.

The output includes only options that can vary between versions of a function. To modify these settings, use UpdateFunctionConfiguration.

To get all of a function's details, including function-level settings, use GetFunction.

get_function_event_invoke_config(client, function_name, qualifier \\ nil, options \\ [])

Retrieves the configuration for asynchronous invocation for a function, version, or alias.

To configure options for asynchronous invocation, use PutFunctionEventInvokeConfig.

get_layer_version(client, layer_name, version_number, options \\ [])

Returns information about a version of an AWS Lambda layer, with a link to download the layer archive that's valid for 10 minutes.

get_layer_version_by_arn(client, arn, options \\ [])

Returns information about a version of an AWS Lambda layer, with a link to download the layer archive that's valid for 10 minutes.

get_layer_version_policy(client, layer_name, version_number, options \\ [])

Returns the permission policy for a version of an AWS Lambda layer.

For more information, see AddLayerVersionPermission.

get_policy(client, function_name, qualifier \\ nil, options \\ [])

Returns the resource-based IAM policy for a function, version, or alias.

get_provisioned_concurrency_config(client, function_name, qualifier, options \\ [])

Retrieves the provisioned concurrency configuration for a function's alias or version.

invoke(client, function_name, input, options \\ [])

Invokes a Lambda function.

You can invoke a function synchronously (and wait for the response), or asynchronously. To invoke a function asynchronously, set InvocationType to Event.

For synchronous invocation, details about the function response, including errors, are included in the response body and headers. For either invocation type, you can find more information in the execution log and trace. When an error occurs, your function may be invoked multiple times. Retry behavior varies by error type, client, event source, and invocation type. For example, if you invoke a function asynchronously and it returns an error, Lambda executes the function up to two more times. For more information, see Retry Behavior.

For asynchronous invocation, Lambda adds events to a queue before sending them to your function. If your function does not have enough capacity to keep up with the queue, events may be lost. Occasionally, your function may receive the same event multiple times, even if no error occurs. To retain events that were not processed, configure your function with a dead-letter queue.

The status code in the API response doesn't reflect function errors. Error codes are reserved for errors that prevent your function from executing, such as permissions errors, limit errors, or issues with your function's code and configuration. For example, Lambda returns TooManyRequestsException if executing the function would cause you to exceed a concurrency limit at either the account level (ConcurrentInvocationLimitExceeded) or function level (ReservedFunctionConcurrentInvocationLimitExceeded).

For functions with a long timeout, your client might be disconnected during synchronous invocation while it waits for a response. Configure your HTTP client, SDK, firewall, proxy, or operating system to allow for long connections with timeout or keep-alive settings.

This operation requires permission for the lambda:InvokeFunction action.

invoke_async(client, function_name, input, options \\ [])

For asynchronous function invocation, use Invoke.

Invokes a function asynchronously.

list_aliases(client, function_name, function_version \\ nil, marker \\ nil, max_items \\ nil, options \\ [])

Returns a list of aliases for a Lambda function.

list_event_source_mappings(client, event_source_arn \\ nil, function_name \\ nil, marker \\ nil, max_items \\ nil, options \\ [])

Lists event source mappings.

Specify an EventSourceArn to only show event source mappings for a single event source.

list_function_event_invoke_configs(client, function_name, marker \\ nil, max_items \\ nil, options \\ [])

Retrieves a list of configurations for asynchronous invocation for a function.

To configure options for asynchronous invocation, use PutFunctionEventInvokeConfig.

list_functions(client, function_version \\ nil, marker \\ nil, master_region \\ nil, max_items \\ nil, options \\ [])

Returns a list of Lambda functions, with the version-specific configuration of each.

Lambda returns up to 50 functions per call.

Set FunctionVersion to ALL to include all published versions of each function in addition to the unpublished version. To get more information about a function or version, use GetFunction.

list_layer_versions(client, layer_name, compatible_runtime \\ nil, marker \\ nil, max_items \\ nil, options \\ [])

Lists the versions of an AWS Lambda layer.

Versions that have been deleted aren't listed. Specify a runtime identifier to list only versions that indicate that they're compatible with that runtime.

list_layers(client, compatible_runtime \\ nil, marker \\ nil, max_items \\ nil, options \\ [])

Lists AWS Lambda layers and shows information about the latest version of each.

Specify a runtime identifier to list only layers that indicate that they're compatible with that runtime.

list_provisioned_concurrency_configs(client, function_name, marker \\ nil, max_items \\ nil, options \\ [])

Retrieves a list of provisioned concurrency configurations for a function.

list_tags(client, resource, options \\ [])

Returns a function's tags.

You can also view tags with GetFunction.

list_versions_by_function(client, function_name, marker \\ nil, max_items \\ nil, options \\ [])

Returns a list of versions, with the version-specific configuration of each.

Lambda returns up to 50 versions per call.

publish_layer_version(client, layer_name, input, options \\ [])

Creates an AWS Lambda layer from a ZIP archive.

Each time you call PublishLayerVersion with the same layer name, a new version is created.

Add layers to your function with CreateFunction or UpdateFunctionConfiguration.

publish_version(client, function_name, input, options \\ [])

Creates a version from the current code and configuration of a function.

Use versions to create a snapshot of your function code and configuration that doesn't change.

AWS Lambda doesn't publish a version if the function's configuration and code haven't changed since the last version. Use UpdateFunctionCode or UpdateFunctionConfiguration to update the function before publishing a version.

Clients can invoke versions directly or with an alias. To create an alias, use CreateAlias.

put_function_concurrency(client, function_name, input, options \\ [])

Sets the maximum number of simultaneous executions for a function, and reserves capacity for that concurrency level.

Concurrency settings apply to the function as a whole, including all published versions and the unpublished version. Reserving concurrency both ensures that your function has capacity to process the specified number of events simultaneously, and prevents it from scaling beyond that level. Use GetFunction to see the current setting for a function.

Use GetAccountSettings to see your Regional concurrency limit. You can reserve concurrency for as many functions as you like, as long as you leave at least 100 simultaneous executions unreserved for functions that aren't configured with a per-function limit. For more information, see Managing Concurrency.

put_function_event_invoke_config(client, function_name, input, options \\ [])

Configures options for asynchronous invocation on a function, version, or alias.

If a configuration already exists for a function, version, or alias, this operation overwrites it. If you exclude any settings, they are removed. To set one option without affecting existing settings for other options, use UpdateFunctionEventInvokeConfig.

By default, Lambda retries an asynchronous invocation twice if the function returns an error. It retains events in a queue for up to six hours. When an event fails all processing attempts or stays in the asynchronous invocation queue for too long, Lambda discards it. To retain discarded events, configure a dead-letter queue with UpdateFunctionConfiguration.

To send an invocation record to a queue, topic, function, or event bus, specify a destination. You can configure separate destinations for successful invocations (on-success) and events that fail all processing attempts (on-failure). You can configure destinations in addition to or instead of a dead-letter queue.

put_provisioned_concurrency_config(client, function_name, input, options \\ [])

Adds a provisioned concurrency configuration to a function's alias or version.

remove_layer_version_permission(client, layer_name, statement_id, version_number, input, options \\ [])

Removes a statement from the permissions policy for a version of an AWS Lambda layer.

For more information, see AddLayerVersionPermission.

remove_permission(client, function_name, statement_id, input, options \\ [])

Revokes function-use permission from an AWS service or another account.

You can get the ID of the statement from the output of GetPolicy.

tag_resource(client, resource, input, options \\ [])

Adds tags to a function.

untag_resource(client, resource, input, options \\ [])

Removes tags from a function.

update_alias(client, function_name, name, input, options \\ [])

Updates the configuration of a Lambda function alias.

update_event_source_mapping(client, uuid, input, options \\ [])

Updates an event source mapping.

You can change the function that AWS Lambda invokes, or pause invocation and resume later from the same location.

The following error handling options are only available for stream sources (DynamoDB and Kinesis):

BisectBatchOnFunctionError - If the function returns an error, split the batch in two and retry.

DestinationConfig - Send discarded records to an Amazon SQS queue or Amazon SNS topic.

MaximumRecordAgeInSeconds - Discard records older than the specified age. Default -1 (infinite). Minimum 60. Maximum 604800.

MaximumRetryAttempts - Discard records after the specified number of retries. Default -1 (infinite). Minimum 0. Maximum 10000. When infinite, failed records will be retried until the record expires.

ParallelizationFactor - Process multiple batches from each shard concurrently.

update_function_code(client, function_name, input, options \\ [])

Updates a Lambda function's code.

The function's code is locked when you publish a version. You can't modify the code of a published version, only the unpublished version.

update_function_configuration(client, function_name, input, options \\ [])

Modify the version-specific settings of a Lambda function.

When you update a function, Lambda provisions an instance of the function and its supporting resources. If your function connects to a VPC, this process can take a minute. During this time, you can't modify the function, but you can still invoke it. The LastUpdateStatus, LastUpdateStatusReason, and LastUpdateStatusReasonCode fields in the response from GetFunctionConfiguration indicate when the update is complete and the function is processing events with the new configuration. For more information, see Function States.

These settings can vary between versions of a function and are locked when you publish a version. You can't modify the configuration of a published version, only the unpublished version.

To configure function concurrency, use PutFunctionConcurrency. To grant invoke permissions to an account or AWS service, use AddPermission.

update_function_event_invoke_config(client, function_name, input, options \\ [])

Updates the configuration for asynchronous invocation for a function, version, or alias.

To configure options for asynchronous invocation, use PutFunctionEventInvokeConfig.