aws-elixir v0.6.0 AWS.LicenseManager

AWS License Manager

AWS License Manager makes it easier to manage licenses from software vendors across multiple AWS accounts and on-premises servers.

Functions

create_license_configuration(client, input, options \\ [])

Creates a license configuration.

A license configuration is an abstraction of a customer license agreement that can be consumed and enforced by License Manager. Components include specifications for the license type (licensing by instance, socket, CPU, or vCPU), allowed tenancy (shared tenancy, Dedicated Instance, Dedicated Host, or all of these), license affinity to host (how long a license must be associated with a host), and the number of licenses purchased and used.

delete_license_configuration(client, input, options \\ [])

Deletes the specified license configuration.

You cannot delete a license configuration that is in use.

get_license_configuration(client, input, options \\ [])

Gets detailed information about the specified license configuration.

get_service_settings(client, input, options \\ [])

Gets the License Manager settings for the current Region.

list_associations_for_license_configuration(client, input, options \\ [])

Lists the resource associations for the specified license configuration.

Resource associations need not consume licenses from a license configuration. For example, an AMI or a stopped instance might not consume a license (depending on the license rules).

list_failures_for_license_configuration_operations(client, input, options \\ [])

Lists the license configuration operations that failed.

list_license_configurations(client, input, options \\ [])

Lists the license configurations for your account.

list_license_specifications_for_resource(client, input, options \\ [])

Describes the license configurations for the specified resource.

list_resource_inventory(client, input, options \\ [])

Lists resources managed using Systems Manager inventory.

list_tags_for_resource(client, input, options \\ [])

Lists the tags for the specified license configuration.

list_usage_for_license_configuration(client, input, options \\ [])

Lists all license usage records for a license configuration, displaying license consumption details by resource at a selected point in time.

Use this action to audit the current license consumption for any license inventory and configuration.

tag_resource(client, input, options \\ [])

Adds the specified tags to the specified license configuration.

untag_resource(client, input, options \\ [])

Removes the specified tags from the specified license configuration.

update_license_configuration(client, input, options \\ [])

Modifies the attributes of an existing license configuration.

update_license_specifications_for_resource(client, input, options \\ [])

Adds or removes the specified license configurations for the specified AWS resource.

You can update the license specifications of AMIs, instances, and hosts. You cannot update the license specifications for launch templates and AWS CloudFormation templates, as they send license configurations to the operation that creates the resource.

update_service_settings(client, input, options \\ [])

Updates License Manager settings for the current Region.