aws-elixir v0.6.0 AWS.Macie

Amazon Macie Classic

Amazon Macie Classic is a security service that uses machine learning to automatically discover, classify, and protect sensitive data in AWS.

Macie Classic recognizes sensitive data such as personally identifiable information (PII) or intellectual property, and provides you with dashboards and alerts that give visibility into how this data is being accessed or moved. For more information, see the Amazon Macie Classic User Guide.

A new Amazon Macie is now available with significant design improvements and additional features, at a lower price and in most AWS Regions. We encourage you to explore and use the new and improved features, and benefit from the reduced cost. To learn about features and pricing for the new Amazon Macie, see Amazon Macie.

Functions

associate_member_account(client, input, options \\ [])

Associates a specified AWS account with Amazon Macie Classic as a member account.

associate_s3_resources(client, input, options \\ [])

Associates specified S3 resources with Amazon Macie Classic for monitoring and data classification.

If memberAccountId isn't specified, the action associates specified S3 resources with Macie Classic for the current master account. If memberAccountId is specified, the action associates specified S3 resources with Macie Classic for the specified member account.

disassociate_member_account(client, input, options \\ [])

Removes the specified member account from Amazon Macie Classic.

disassociate_s3_resources(client, input, options \\ [])

Removes specified S3 resources from being monitored by Amazon Macie Classic.

If memberAccountId isn't specified, the action removes specified S3 resources from Macie Classic for the current master account. If memberAccountId is specified, the action removes specified S3 resources from Macie Classic for the specified member account.

list_member_accounts(client, input, options \\ [])

Lists all Amazon Macie Classic member accounts for the current Amazon Macie Classic master account.

list_s3_resources(client, input, options \\ [])

Lists all the S3 resources associated with Amazon Macie Classic.

If memberAccountId isn't specified, the action lists the S3 resources associated with Amazon Macie Classic for the current master account. If memberAccountId is specified, the action lists the S3 resources associated with Amazon Macie Classic for the specified member account.

update_s3_resources(client, input, options \\ [])

Updates the classification types for the specified S3 resources.

If memberAccountId isn't specified, the action updates the classification types of the S3 resources associated with Amazon Macie Classic for the current master account. If memberAccountId is specified, the action updates the classification types of the S3 resources associated with Amazon Macie Classic for the specified member account.