aws-elixir v0.6.0 AWS.Macie2

Amazon Macie is a fully managed data security and data privacy service that uses machine learning and pattern matching to discover and protect your sensitive data in AWS.

Macie automates the discovery of sensitive data, such as PII and intellectual property, to provide you with insight into the data that your organization stores in AWS. Macie also provides an inventory of your Amazon S3 buckets, which it continually monitors for you. If Macie detects sensitive data or potential data access issues, it generates detailed findings for you to review and act upon as necessary.

Functions

accept_invitation(client, input, options \\ [])

Accepts an Amazon Macie membership invitation that was received from a specific account.

batch_get_custom_data_identifiers(client, input, options \\ [])

Retrieves information about one or more custom data identifiers.

create_classification_job(client, input, options \\ [])

Creates and defines the settings for a classification job.

create_custom_data_identifier(client, input, options \\ [])

Creates and defines the criteria and other settings for a custom data identifier.

create_findings_filter(client, input, options \\ [])

Creates and defines the criteria and other settings for a findings filter.

create_invitations(client, input, options \\ [])

Sends an Amazon Macie membership invitation to one or more accounts.

create_member(client, input, options \\ [])

Associates an account with an Amazon Macie master account.

create_sample_findings(client, input, options \\ [])

Creates sample findings.

decline_invitations(client, input, options \\ [])

Declines Amazon Macie membership invitations that were received from specific accounts.

delete_custom_data_identifier(client, id, input, options \\ [])

Soft deletes a custom data identifier.

delete_findings_filter(client, id, input, options \\ [])

Deletes a findings filter.

delete_invitations(client, input, options \\ [])

Deletes Amazon Macie membership invitations that were received from specific accounts.

delete_member(client, id, input, options \\ [])

Deletes the association between an Amazon Macie master account and an account.

describe_buckets(client, input, options \\ [])

Retrieves (queries) statistical data and other information about one or more S3 buckets that Amazon Macie monitors and analyzes.

describe_classification_job(client, job_id, options \\ [])

Retrieves information about the status and settings for a classification job.

describe_organization_configuration(client, options \\ [])

Retrieves information about the Amazon Macie configuration settings for an AWS organization.

disable_macie(client, input, options \\ [])

Disables an Amazon Macie account and deletes Macie resources for the account.

disable_organization_admin_account(client, input, options \\ [])

Disables an account as a delegated administrator of Amazon Macie for an AWS organization.

disassociate_from_master_account(client, input, options \\ [])

Disassociates a member account from its Amazon Macie master account.

disassociate_member(client, id, input, options \\ [])

Disassociates an Amazon Macie master account from a member account.

enable_macie(client, input, options \\ [])

Enables Amazon Macie and specifies the configuration settings for a Macie account.

enable_organization_admin_account(client, input, options \\ [])

Enables an account as a delegated administrator of Amazon Macie for an AWS organization.

get_bucket_statistics(client, input, options \\ [])

Retrieves (queries) aggregated statistical data for all the S3 buckets that Amazon Macie monitors and analyzes.

get_classification_export_configuration(client, options \\ [])

Retrieves the configuration settings for storing data classification results.

get_custom_data_identifier(client, id, options \\ [])

Retrieves information about the criteria and other settings for a custom data identifier.

get_finding_statistics(client, input, options \\ [])

Retrieves (queries) aggregated statistical data about findings.

get_findings(client, input, options \\ [])

Retrieves information about one or more findings.

get_findings_filter(client, id, options \\ [])

Retrieves information about the criteria and other settings for a findings filter.

get_invitations_count(client, options \\ [])

Retrieves the count of Amazon Macie membership invitations that were received by an account.

get_macie_session(client, options \\ [])

Retrieves information about the current status and configuration settings for an Amazon Macie account.

get_master_account(client, options \\ [])

Retrieves information about the Amazon Macie master account for an account.

get_member(client, id, options \\ [])

Retrieves information about a member account that's associated with an Amazon Macie master account.

get_usage_statistics(client, input, options \\ [])

Retrieves (queries) quotas and aggregated usage data for one or more accounts.

get_usage_totals(client, options \\ [])

Retrieves (queries) aggregated usage data for an account.

list_classification_jobs(client, input, options \\ [])

Retrieves a subset of information about one or more classification jobs.

list_custom_data_identifiers(client, input, options \\ [])

Retrieves a subset of information about all the custom data identifiers for an account.

list_findings(client, input, options \\ [])

Retrieves a subset of information about one or more findings.

list_findings_filters(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves a subset of information about all the findings filters for an account.

list_invitations(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves information about all the Amazon Macie membership invitations that were received by an account.

list_members(client, max_results \\ nil, next_token \\ nil, only_associated \\ nil, options \\ [])

Retrieves information about the accounts that are associated with an Amazon Macie master account.

list_organization_admin_accounts(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieves information about the account that's designated as the delegated administrator of Amazon Macie for an AWS organization.

list_tags_for_resource(client, resource_arn, options \\ [])

Retrieves the tags (keys and values) that are associated with a classification job, custom data identifier, findings filter, or member account.

put_classification_export_configuration(client, input, options \\ [])

Creates or updates the configuration settings for storing data classification results.

tag_resource(client, resource_arn, input, options \\ [])

Adds or updates one or more tags (keys and values) that are associated with a classification job, custom data identifier, findings filter, or member account.

test_custom_data_identifier(client, input, options \\ [])

Tests a custom data identifier.

untag_resource(client, resource_arn, input, options \\ [])

Removes one or more tags (keys and values) from a classification job, custom data identifier, findings filter, or member account.

update_classification_job(client, job_id, input, options \\ [])

Cancels a classification job.

update_findings_filter(client, id, input, options \\ [])

Updates the criteria and other settings for a findings filter.

update_macie_session(client, input, options \\ [])

Suspends or re-enables an Amazon Macie account, or updates the configuration settings for a Macie account.

update_member_session(client, id, input, options \\ [])

Enables an Amazon Macie master account to suspend or re-enable a member account.

update_organization_configuration(client, input, options \\ [])

Updates Amazon Macie configuration settings for an AWS organization.