aws-elixir v0.6.0 AWS.ManagedBlockchain

Amazon Managed Blockchain is a fully managed service for creating and managing blockchain networks using open source frameworks.

Blockchain allows you to build applications where multiple parties can securely and transparently run transactions and share data without the need for a trusted, central authority. Currently, Managed Blockchain supports the Hyperledger Fabric open source framework.

Functions

create_member(client, network_id, input, options \\ [])

Creates a member within a Managed Blockchain network.

create_network(client, input, options \\ [])

Creates a new blockchain network using Amazon Managed Blockchain.

create_node(client, member_id, network_id, input, options \\ [])

Creates a peer node in a member.

create_proposal(client, network_id, input, options \\ [])

Creates a proposal for a change to the network that other members of the network can vote on, for example, a proposal to add a new member to the network.

Any member can create a proposal.

delete_member(client, member_id, network_id, input, options \\ [])

Deletes a member.

Deleting a member removes the member and all associated resources from the network. DeleteMember can only be called for a specified MemberId if the principal performing the action is associated with the AWS account that owns the member. In all other cases, the DeleteMember action is carried out as the result of an approved proposal to remove a member. If MemberId is the last member in a network specified by the last AWS account, the network is deleted also.

delete_node(client, member_id, network_id, node_id, input, options \\ [])

Deletes a peer node from a member that your AWS account owns.

All data on the node is lost and cannot be recovered.

get_member(client, member_id, network_id, options \\ [])

Returns detailed information about a member.

get_network(client, network_id, options \\ [])

Returns detailed information about a network.

get_node(client, member_id, network_id, node_id, options \\ [])

Returns detailed information about a peer node.

get_proposal(client, network_id, proposal_id, options \\ [])

Returns detailed information about a proposal.

list_invitations(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a listing of all invitations for the current AWS account.

list_members(client, network_id, is_owned \\ nil, max_results \\ nil, name \\ nil, next_token \\ nil, status \\ nil, options \\ [])

Returns a listing of the members in a network and properties of their configurations.

list_networks(client, framework \\ nil, max_results \\ nil, name \\ nil, next_token \\ nil, status \\ nil, options \\ [])

Returns information about the networks in which the current AWS account has members.

list_nodes(client, member_id, network_id, max_results \\ nil, next_token \\ nil, status \\ nil, options \\ [])

Returns information about the nodes within a network.

list_proposal_votes(client, network_id, proposal_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns the listing of votes for a specified proposal, including the value of each vote and the unique identifier of the member that cast the vote.

list_proposals(client, network_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a listing of proposals for the network.

reject_invitation(client, invitation_id, input, options \\ [])

Rejects an invitation to join a network.

This action can be called by a principal in an AWS account that has received an invitation to create a member and join a network.

update_member(client, member_id, network_id, input, options \\ [])

Updates a member configuration with new parameters.

update_node(client, member_id, network_id, node_id, input, options \\ [])

Updates a node configuration with new parameters.

vote_on_proposal(client, network_id, proposal_id, input, options \\ [])

Casts a vote for a specified ProposalId on behalf of a member.

The member to vote as, specified by VoterMemberId, must be in the same AWS account as the principal that calls the action.