aws-elixir v0.6.0 AWS.MarketplaceCatalog

Catalog API actions allow you to manage your entities through list, describe, and update capabilities.

An entity can be a product or an offer on AWS Marketplace.

You can automate your entity update process by integrating the AWS Marketplace Catalog API with your AWS Marketplace product build or deployment pipelines. You can also create your own applications on top of the Catalog API to manage your products on AWS Marketplace.

Functions

cancel_change_set(client, input, options \\ [])

Used to cancel an open change request.

Must be sent before the status of the request changes to APPLYING, the final stage of completing your change request. You can describe a change during the 60-day request history retention period for API calls.

describe_change_set(client, catalog, change_set_id, options \\ [])

Provides information about a given change set.

describe_entity(client, catalog, entity_id, options \\ [])

Returns the metadata and content of the entity.

list_change_sets(client, input, options \\ [])

Returns the list of change sets owned by the account being used to make the call.

You can filter this list by providing any combination of entityId, ChangeSetName, and status. If you provide more than one filter, the API operation applies a logical AND between the filters.

You can describe a change during the 60-day request history retention period for API calls.

list_entities(client, input, options \\ [])

Provides the list of entities of a given type.

start_change_set(client, input, options \\ [])

This operation allows you to request changes for your entities.

Within a single ChangeSet, you cannot start the same change type against the same entity multiple times. Additionally, when a ChangeSet is running, all the entities targeted by the different changes are locked until the ChangeSet has completed (either succeeded, cancelled, or failed). If you try to start a ChangeSet containing a change against an entity that is already locked, you will receive a ResourceInUseException.

For example, you cannot start the ChangeSet described in the example below because it contains two changes to execute the same change type (AddRevisions) against the same entity (entity-id@1).