aws-elixir v0.6.0 AWS.MediaConnect

API for AWS Elemental MediaConnect

Functions

add_flow_outputs(client, flow_arn, input, options \\ [])

Adds outputs to an existing flow.

You can create up to 50 outputs per flow.

add_flow_sources(client, flow_arn, input, options \\ [])

Adds Sources to flow

add_flow_vpc_interfaces(client, flow_arn, input, options \\ [])

Adds VPC interfaces to flow

create_flow(client, input, options \\ [])

Creates a new flow.

The request must include one source. The request optionally can include outputs (up to 50) and entitlements (up to 50).

delete_flow(client, flow_arn, input, options \\ [])

Deletes a flow.

Before you can delete a flow, you must stop the flow.

describe_flow(client, flow_arn, options \\ [])

Displays the details of a flow.

The response includes the flow ARN, name, and Availability Zone, as well as details about the source, outputs, and entitlements.

describe_offering(client, offering_arn, options \\ [])

Displays the details of an offering.

The response includes the offering description, duration, outbound bandwidth, price, and Amazon Resource Name (ARN).

describe_reservation(client, reservation_arn, options \\ [])

Displays the details of a reservation.

The response includes the reservation name, state, start date and time, and the details of the offering that make up the rest of the reservation (such as price, duration, and outbound bandwidth).

grant_flow_entitlements(client, flow_arn, input, options \\ [])

Grants entitlements to an existing flow.

list_entitlements(client, max_results \\ nil, next_token \\ nil, options \\ [])

Displays a list of all entitlements that have been granted to this account.

This request returns 20 results per page.

list_flows(client, max_results \\ nil, next_token \\ nil, options \\ [])

Displays a list of flows that are associated with this account.

This request returns a paginated result.

list_offerings(client, max_results \\ nil, next_token \\ nil, options \\ [])

Displays a list of all offerings that are available to this account in the current AWS Region.

If you have an active reservation (which means you've purchased an offering that has already started and hasn't expired yet), your account isn't eligible for other offerings.

list_reservations(client, max_results \\ nil, next_token \\ nil, options \\ [])

Displays a list of all reservations that have been purchased by this account in the current AWS Region.

This list includes all reservations in all states (such as active and expired).

list_tags_for_resource(client, resource_arn, options \\ [])

List all tags on an AWS Elemental MediaConnect resource

purchase_offering(client, offering_arn, input, options \\ [])

Submits a request to purchase an offering.

If you already have an active reservation, you can't purchase another offering.

remove_flow_output(client, flow_arn, output_arn, input, options \\ [])

Removes an output from an existing flow.

This request can be made only on an output that does not have an entitlement associated with it. If the output has an entitlement, you must revoke the entitlement instead. When an entitlement is revoked from a flow, the service automatically removes the associated output.

remove_flow_source(client, flow_arn, source_arn, input, options \\ [])

Removes a source from an existing flow.

This request can be made only if there is more than one source on the flow.

remove_flow_vpc_interface(client, flow_arn, vpc_interface_name, input, options \\ [])

Removes a VPC Interface from an existing flow.

This request can be made only on a VPC interface that does not have a Source or Output associated with it. If the VPC interface is referenced by a Source or Output, you must first delete or update the Source or Output to no longer reference the VPC interface.

revoke_flow_entitlement(client, entitlement_arn, flow_arn, input, options \\ [])

Revokes an entitlement from a flow.

Once an entitlement is revoked, the content becomes unavailable to the subscriber and the associated output is removed.

start_flow(client, flow_arn, input, options \\ [])

Starts a flow.

stop_flow(client, flow_arn, input, options \\ [])

Stops a flow.

tag_resource(client, resource_arn, input, options \\ [])

Associates the specified tags to a resource with the specified resourceArn.

If existing tags on a resource are not specified in the request parameters, they are not changed. When a resource is deleted, the tags associated with that resource are deleted as well.

untag_resource(client, resource_arn, input, options \\ [])

Deletes specified tags from a resource.

update_flow(client, flow_arn, input, options \\ [])

Updates flow

update_flow_entitlement(client, entitlement_arn, flow_arn, input, options \\ [])

You can change an entitlement's description, subscribers, and encryption.

If you change the subscribers, the service will remove the outputs that are are used by the subscribers that are removed.

update_flow_output(client, flow_arn, output_arn, input, options \\ [])

Updates an existing flow output.

update_flow_source(client, flow_arn, source_arn, input, options \\ [])

Updates the source of a flow.