aws-elixir v0.6.0 AWS.MediaConvert

AWS Elemental MediaConvert

Functions

associate_certificate(client, input, options \\ [])

Associates an AWS Certificate Manager (ACM) Amazon Resource Name (ARN) with AWS Elemental MediaConvert.

cancel_job(client, id, input, options \\ [])

Permanently cancel a job.

Once you have canceled a job, you can't start it again.

create_job(client, input, options \\ [])

Create a new transcoding job.

For information about jobs and job settings, see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

create_job_template(client, input, options \\ [])

Create a new job template.

For information about job templates see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

create_preset(client, input, options \\ [])

Create a new preset.

For information about job templates see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

create_queue(client, input, options \\ [])

Create a new transcoding queue.

For information about queues, see Working With Queues in the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html

delete_job_template(client, name, input, options \\ [])

Permanently delete a job template you have created.

delete_preset(client, name, input, options \\ [])

Permanently delete a preset you have created.

delete_queue(client, name, input, options \\ [])

Permanently delete a queue you have created.

describe_endpoints(client, input, options \\ [])

Send an request with an empty body to the regional API endpoint to get your account API endpoint.

disassociate_certificate(client, arn, input, options \\ [])

Removes an association between the Amazon Resource Name (ARN) of an AWS Certificate Manager (ACM) certificate and an AWS Elemental MediaConvert resource.

get_job(client, id, options \\ [])

Retrieve the JSON for a specific completed transcoding job.

get_job_template(client, name, options \\ [])

Retrieve the JSON for a specific job template.

get_preset(client, name, options \\ [])

Retrieve the JSON for a specific preset.

get_queue(client, name, options \\ [])

Retrieve the JSON for a specific queue.

list_job_templates(client, category \\ nil, list_by \\ nil, max_results \\ nil, next_token \\ nil, order \\ nil, options \\ [])

Retrieve a JSON array of up to twenty of your job templates.

This will return the templates themselves, not just a list of them. To retrieve the next twenty templates, use the nextToken string returned with the array

list_jobs(client, max_results \\ nil, next_token \\ nil, order \\ nil, queue \\ nil, status \\ nil, options \\ [])

Retrieve a JSON array of up to twenty of your most recently created jobs.

This array includes in-process, completed, and errored jobs. This will return the jobs themselves, not just a list of the jobs. To retrieve the twenty next most recent jobs, use the nextToken string returned with the array.

list_presets(client, category \\ nil, list_by \\ nil, max_results \\ nil, next_token \\ nil, order \\ nil, options \\ [])

Retrieve a JSON array of up to twenty of your presets.

This will return the presets themselves, not just a list of them. To retrieve the next twenty presets, use the nextToken string returned with the array.

list_queues(client, list_by \\ nil, max_results \\ nil, next_token \\ nil, order \\ nil, options \\ [])

Retrieve a JSON array of up to twenty of your queues.

This will return the queues themselves, not just a list of them. To retrieve the next twenty queues, use the nextToken string returned with the array.

list_tags_for_resource(client, arn, options \\ [])

Retrieve the tags for a MediaConvert resource.

tag_resource(client, input, options \\ [])

Add tags to a MediaConvert queue, preset, or job template.

For information about tagging, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/tagging-resources.html

untag_resource(client, arn, input, options \\ [])

Remove tags from a MediaConvert queue, preset, or job template.

For information about tagging, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/tagging-resources.html

update_job_template(client, name, input, options \\ [])

Modify one of your existing job templates.

update_preset(client, name, input, options \\ [])

Modify one of your existing presets.

update_queue(client, name, input, options \\ [])

Modify one of your existing queues.