aws-elixir v0.6.0 AWS.MediaLive

API for AWS Elemental MediaLive

Functions

batch_delete(client, input, options \\ [])

Starts delete of resources.

batch_start(client, input, options \\ [])

Starts existing resources

batch_stop(client, input, options \\ [])

Stops running resources

batch_update_schedule(client, channel_id, input, options \\ [])

Update a channel schedule

create_channel(client, input, options \\ [])

Creates a new channel

create_input(client, input, options \\ [])

Create an input

create_input_security_group(client, input, options \\ [])

Creates a Input Security Group

create_multiplex(client, input, options \\ [])

Create a new multiplex.

create_multiplex_program(client, multiplex_id, input, options \\ [])

Create a new program in the multiplex.

create_tags(client, resource_arn, input, options \\ [])

Create tags for a resource

delete_channel(client, channel_id, input, options \\ [])

Starts deletion of channel.

The associated outputs are also deleted.

delete_input(client, input_id, input, options \\ [])

Deletes the input end point

delete_input_security_group(client, input_security_group_id, input, options \\ [])

Deletes an Input Security Group

delete_multiplex(client, multiplex_id, input, options \\ [])

Delete a multiplex.

The multiplex must be idle.

delete_multiplex_program(client, multiplex_id, program_name, input, options \\ [])

Delete a program from a multiplex.

delete_reservation(client, reservation_id, input, options \\ [])

Delete an expired reservation.

delete_schedule(client, channel_id, input, options \\ [])

Delete all schedule actions on a channel.

delete_tags(client, resource_arn, input, options \\ [])

Removes tags for a resource

describe_channel(client, channel_id, options \\ [])

Gets details about a channel

describe_input(client, input_id, options \\ [])

Produces details about an input

describe_input_device(client, input_device_id, options \\ [])

Gets the details for the input device

describe_input_device_thumbnail(client, input_device_id, accept, options \\ [])

Get the latest thumbnail data for the input device.

describe_input_security_group(client, input_security_group_id, options \\ [])

Produces a summary of an Input Security Group

describe_multiplex(client, multiplex_id, options \\ [])

Gets details about a multiplex.

describe_multiplex_program(client, multiplex_id, program_name, options \\ [])

Get the details for a program in a multiplex.

describe_offering(client, offering_id, options \\ [])

Get details for an offering.

describe_reservation(client, reservation_id, options \\ [])

Get details for a reservation.

describe_schedule(client, channel_id, max_results \\ nil, next_token \\ nil, options \\ [])

Get a channel schedule

list_channels(client, max_results \\ nil, next_token \\ nil, options \\ [])

Produces list of channels that have been created

list_input_devices(client, max_results \\ nil, next_token \\ nil, options \\ [])

List input devices

list_input_security_groups(client, max_results \\ nil, next_token \\ nil, options \\ [])

Produces a list of Input Security Groups for an account

list_inputs(client, max_results \\ nil, next_token \\ nil, options \\ [])

Produces list of inputs that have been created

list_multiplex_programs(client, multiplex_id, max_results \\ nil, next_token \\ nil, options \\ [])

List the programs that currently exist for a specific multiplex.

list_multiplexes(client, max_results \\ nil, next_token \\ nil, options \\ [])

Retrieve a list of the existing multiplexes.

list_offerings(client, channel_class \\ nil, channel_configuration \\ nil, codec \\ nil, duration \\ nil, max_results \\ nil, maximum_bitrate \\ nil, maximum_framerate \\ nil, next_token \\ nil, resolution \\ nil, resource_type \\ nil, special_feature \\ nil, video_quality \\ nil, options \\ [])

List offerings available for purchase.

list_reservations(client, channel_class \\ nil, codec \\ nil, max_results \\ nil, maximum_bitrate \\ nil, maximum_framerate \\ nil, next_token \\ nil, resolution \\ nil, resource_type \\ nil, special_feature \\ nil, video_quality \\ nil, options \\ [])

List purchased reservations.

list_tags_for_resource(client, resource_arn, options \\ [])

Produces list of tags that have been created for a resource

purchase_offering(client, offering_id, input, options \\ [])

Purchase an offering and create a reservation.

start_channel(client, channel_id, input, options \\ [])

Starts an existing channel

start_multiplex(client, multiplex_id, input, options \\ [])

Start (run) the multiplex.

Starting the multiplex does not start the channels. You must explicitly start each channel.

stop_channel(client, channel_id, input, options \\ [])

Stops a running channel

stop_multiplex(client, multiplex_id, input, options \\ [])

Stops a running multiplex.

If the multiplex isn't running, this action has no effect.

update_channel(client, channel_id, input, options \\ [])

Updates a channel.

update_channel_class(client, channel_id, input, options \\ [])

Changes the class of the channel.

update_input(client, input_id, input, options \\ [])

Updates an input.

update_input_device(client, input_device_id, input, options \\ [])

Updates the parameters for the input device.

update_input_security_group(client, input_security_group_id, input, options \\ [])

Update an Input Security Group's Whilelists.

update_multiplex(client, multiplex_id, input, options \\ [])

Updates a multiplex.

update_multiplex_program(client, multiplex_id, program_name, input, options \\ [])

Update a program in a multiplex.

update_reservation(client, reservation_id, input, options \\ [])

Update reservation.