aws-elixir v0.6.0 AWS.MediaPackageVod

AWS Elemental MediaPackage VOD

Functions

create_asset(client, input, options \\ [])

Creates a new MediaPackage VOD Asset resource.

create_packaging_configuration(client, input, options \\ [])

Creates a new MediaPackage VOD PackagingConfiguration resource.

create_packaging_group(client, input, options \\ [])

Creates a new MediaPackage VOD PackagingGroup resource.

delete_asset(client, id, input, options \\ [])

Deletes an existing MediaPackage VOD Asset resource.

delete_packaging_configuration(client, id, input, options \\ [])

Deletes a MediaPackage VOD PackagingConfiguration resource.

delete_packaging_group(client, id, input, options \\ [])

Deletes a MediaPackage VOD PackagingGroup resource.

describe_asset(client, id, options \\ [])

Returns a description of a MediaPackage VOD Asset resource.

describe_packaging_configuration(client, id, options \\ [])

Returns a description of a MediaPackage VOD PackagingConfiguration resource.

describe_packaging_group(client, id, options \\ [])

Returns a description of a MediaPackage VOD PackagingGroup resource.

list_assets(client, max_results \\ nil, next_token \\ nil, packaging_group_id \\ nil, options \\ [])

Returns a collection of MediaPackage VOD Asset resources.

list_packaging_configurations(client, max_results \\ nil, next_token \\ nil, packaging_group_id \\ nil, options \\ [])

Returns a collection of MediaPackage VOD PackagingConfiguration resources.

list_packaging_groups(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a collection of MediaPackage VOD PackagingGroup resources.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns a list of the tags assigned to the specified resource.

tag_resource(client, resource_arn, input, options \\ [])

Adds tags to the specified resource.

You can specify one or more tags to add.

untag_resource(client, resource_arn, input, options \\ [])

Removes tags from the specified resource.

You can specify one or more tags to remove.

update_packaging_group(client, id, input, options \\ [])

Updates a specific packaging group.

You can't change the id attribute or any other system-generated attributes.