aws-elixir v0.6.0 AWS.MediaStoreData

An AWS Elemental MediaStore asset is an object, similar to an object in the Amazon S3 service.

Objects are the fundamental entities that are stored in AWS Elemental MediaStore.

Functions

delete_object(client, path, input, options \\ [])

Deletes an object at the specified path.

describe_object(client, path, input, options \\ [])

Gets the headers for an object at the specified path.

get_object(client, path, range \\ nil, options \\ [])

Downloads the object at the specified path.

If the object’s upload availability is set to streaming, AWS Elemental MediaStore downloads the object even if it’s still uploading the object.

list_items(client, max_results \\ nil, next_token \\ nil, path \\ nil, options \\ [])

Provides a list of metadata entries about folders and objects in the specified folder.

put_object(client, path, input, options \\ [])

Uploads an object to the specified path.

Object sizes are limited to 25 MB for standard upload availability and 10 MB for streaming upload availability.