aws-elixir v0.6.0 AWS.MediaTailor

Use the AWS Elemental MediaTailor SDK to configure scalable ad insertion for your live and VOD content.

With AWS Elemental MediaTailor, you can serve targeted ads to viewers while maintaining broadcast quality in over-the-top (OTT) video applications. For information about using the service, including detailed information about the settings covered in this guide, see the AWS Elemental MediaTailor User Guide.

Through the SDK, you manage AWS Elemental MediaTailor configurations the same as you do through the console. For example, you specify ad insertion behavior and mapping information for the origin server and the ad decision server (ADS).

Functions

delete_playback_configuration(client, name, input, options \\ [])

Deletes the playback configuration for the specified name.

get_playback_configuration(client, name, options \\ [])

Returns the playback configuration for the specified name.

list_playback_configurations(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of the playback configurations defined in AWS Elemental MediaTailor.

You can specify a maximum number of configurations to return at a time. The default maximum is 50. Results are returned in pagefuls. If MediaTailor has more configurations than the specified maximum, it provides parameters in the response that you can use to retrieve the next pageful.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns a list of the tags assigned to the specified playback configuration resource.

put_playback_configuration(client, input, options \\ [])

Adds a new playback configuration to AWS Elemental MediaTailor.

tag_resource(client, resource_arn, input, options \\ [])

Adds tags to the specified playback configuration resource.

You can specify one or more tags to add.

untag_resource(client, resource_arn, input, options \\ [])

Removes tags from the specified playback configuration resource.

You can specify one or more tags to remove.