aws-elixir v0.6.0 AWS.Mobile

AWS Mobile Service provides mobile app and website developers with capabilities required to configure AWS resources and bootstrap their developer desktop projects with the necessary SDKs, constants, tools and samples to make use of those resources.

Functions

create_project(client, input, options \\ [])

Creates an AWS Mobile Hub project.

delete_project(client, project_id, input, options \\ [])

Delets a project in AWS Mobile Hub.

describe_bundle(client, bundle_id, options \\ [])

Get the bundle details for the requested bundle id.

describe_project(client, project_id, sync_from_resources \\ nil, options \\ [])

Gets details about a project in AWS Mobile Hub.

export_bundle(client, bundle_id, input, options \\ [])

Generates customized software development kit (SDK) and or tool packages used to integrate mobile web or mobile app clients with backend AWS resources.

export_project(client, project_id, input, options \\ [])

Exports project configuration to a snapshot which can be downloaded and shared.

Note that mobile app push credentials are encrypted in exported projects, so they can only be shared successfully within the same AWS account.

list_bundles(client, max_results \\ nil, next_token \\ nil, options \\ [])

List all available bundles.

list_projects(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists projects in AWS Mobile Hub.

update_project(client, input, options \\ [])

Update an existing project.