aws-elixir v0.6.0 AWS.Mq

Amazon MQ is a managed message broker service for Apache ActiveMQ that makes it easy to set up and operate message brokers in the cloud.

A message broker allows software applications and components to communicate using various programming languages, operating systems, and formal messaging protocols.

Functions

create_broker(client, input, options \\ [])

Creates a broker.

Note: This API is asynchronous.

create_configuration(client, input, options \\ [])

Creates a new configuration for the specified configuration name.

Amazon MQ uses the default configuration (the engine type and version).

create_tags(client, resource_arn, input, options \\ [])

Add a tag to a resource.

create_user(client, broker_id, username, input, options \\ [])

Creates an ActiveMQ user.

delete_broker(client, broker_id, input, options \\ [])

Deletes a broker.

Note: This API is asynchronous.

delete_tags(client, resource_arn, input, options \\ [])

Removes a tag from a resource.

delete_user(client, broker_id, username, input, options \\ [])

Deletes an ActiveMQ user.

describe_broker(client, broker_id, options \\ [])

Returns information about the specified broker.

describe_broker_engine_types(client, engine_type \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Describe available engine types and versions.

describe_broker_instance_options(client, engine_type \\ nil, host_instance_type \\ nil, max_results \\ nil, next_token \\ nil, storage_type \\ nil, options \\ [])

Describe available broker instance options.

describe_configuration(client, configuration_id, options \\ [])

Returns information about the specified configuration.

describe_configuration_revision(client, configuration_id, configuration_revision, options \\ [])

Returns the specified configuration revision for the specified configuration.

describe_user(client, broker_id, username, options \\ [])

Returns information about an ActiveMQ user.

list_brokers(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of all brokers.

list_configuration_revisions(client, configuration_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of all revisions for the specified configuration.

list_configurations(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of all configurations.

list_tags(client, resource_arn, options \\ [])

Lists tags for a resource.

list_users(client, broker_id, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of all ActiveMQ users.

reboot_broker(client, broker_id, input, options \\ [])

Reboots a broker.

Note: This API is asynchronous.

update_broker(client, broker_id, input, options \\ [])

Adds a pending configuration change to a broker.

update_configuration(client, configuration_id, input, options \\ [])

Updates the specified configuration.

update_user(client, broker_id, username, input, options \\ [])

Updates the information for an ActiveMQ user.