aws-elixir v0.6.0 AWS.NetworkManager

Transit Gateway Network Manager (Network Manager) enables you to create a global network, in which you can monitor your AWS and on-premises networks that are built around transit gateways.

Functions

associate_customer_gateway(client, global_network_id, input, options \\ [])

Associates a customer gateway with a device and optionally, with a link.

If you specify a link, it must be associated with the specified device.

You can only associate customer gateways that are connected to a VPN attachment on a transit gateway. The transit gateway must be registered in your global network. When you register a transit gateway, customer gateways that are connected to the transit gateway are automatically included in the global network. To list customer gateways that are connected to a transit gateway, use the DescribeVpnConnections EC2 API and filter by transit-gateway-id.

You cannot associate a customer gateway with more than one device and link.

associate_link(client, global_network_id, input, options \\ [])

Associates a link to a device.

A device can be associated to multiple links and a link can be associated to multiple devices. The device and link must be in the same global network and the same site.

create_device(client, global_network_id, input, options \\ [])

Creates a new device in a global network.

If you specify both a site ID and a location, the location of the site is used for visualization in the Network Manager console.

create_global_network(client, input, options \\ [])

Creates a new, empty global network.

create_link(client, global_network_id, input, options \\ [])

Creates a new link for a specified site.

create_site(client, global_network_id, input, options \\ [])

Creates a new site in a global network.

delete_device(client, device_id, global_network_id, input, options \\ [])

Deletes an existing device.

You must first disassociate the device from any links and customer gateways.

delete_global_network(client, global_network_id, input, options \\ [])

Deletes an existing global network.

You must first delete all global network objects (devices, links, and sites) and deregister all transit gateways.

delete_link(client, global_network_id, link_id, input, options \\ [])

Deletes an existing link.

You must first disassociate the link from any devices and customer gateways.

delete_site(client, global_network_id, site_id, input, options \\ [])

Deletes an existing site.

The site cannot be associated with any device or link.

deregister_transit_gateway(client, global_network_id, transit_gateway_arn, input, options \\ [])

Deregisters a transit gateway from your global network.

This action does not delete your transit gateway, or modify any of its attachments. This action removes any customer gateway associations.

describe_global_networks(client, global_network_ids \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Describes one or more global networks.

By default, all global networks are described. To describe the objects in your global network, you must use the appropriate Get* action. For example, to list the transit gateways in your global network, use GetTransitGatewayRegistrations.

disassociate_customer_gateway(client, customer_gateway_arn, global_network_id, input, options \\ [])

Disassociates a customer gateway from a device and a link.

disassociate_link(client, global_network_id, input, options \\ [])

Disassociates an existing device from a link.

You must first disassociate any customer gateways that are associated with the link.

get_customer_gateway_associations(client, global_network_id, customer_gateway_arns \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the association information for customer gateways that are associated with devices and links in your global network.

get_devices(client, global_network_id, device_ids \\ nil, max_results \\ nil, next_token \\ nil, site_id \\ nil, options \\ [])

Gets information about one or more of your devices in a global network.

get_link_associations(client, global_network_id, device_id \\ nil, link_id \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Gets the link associations for a device or a link.

Either the device ID or the link ID must be specified.

get_links(client, global_network_id, link_ids \\ nil, max_results \\ nil, next_token \\ nil, provider \\ nil, site_id \\ nil, type \\ nil, options \\ [])

Gets information about one or more links in a specified global network.

If you specify the site ID, you cannot specify the type or provider in the same request. You can specify the type and provider in the same request.

get_sites(client, global_network_id, max_results \\ nil, next_token \\ nil, site_ids \\ nil, options \\ [])

Gets information about one or more of your sites in a global network.

get_transit_gateway_registrations(client, global_network_id, max_results \\ nil, next_token \\ nil, transit_gateway_arns \\ nil, options \\ [])

Gets information about the transit gateway registrations in a specified global network.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists the tags for a specified resource.

register_transit_gateway(client, global_network_id, input, options \\ [])

Registers a transit gateway in your global network.

The transit gateway can be in any AWS Region, but it must be owned by the same AWS account that owns the global network. You cannot register a transit gateway in more than one global network.

tag_resource(client, resource_arn, input, options \\ [])

Tags a specified resource.

untag_resource(client, resource_arn, input, options \\ [])

Removes tags from a specified resource.

update_device(client, device_id, global_network_id, input, options \\ [])

Updates the details for an existing device.

To remove information for any of the parameters, specify an empty string.

update_global_network(client, global_network_id, input, options \\ [])

Updates an existing global network.

To remove information for any of the parameters, specify an empty string.

update_link(client, global_network_id, link_id, input, options \\ [])

Updates the details for an existing link.

To remove information for any of the parameters, specify an empty string.

update_site(client, global_network_id, site_id, input, options \\ [])

Updates the information for an existing site.

To remove information for any of the parameters, specify an empty string.