aws-elixir v0.6.0 AWS.Outposts

AWS Outposts is a fully-managed service that extends AWS infrastructure, APIs, and tools to customer premises.

By providing local access to AWS-managed infrastructure, AWS Outposts enables customers to build and run applications on premises using the same programming interfaces as in AWS Regions, while using local compute and storage resources for lower latency and local data processing needs.

Functions

create_outpost(client, input, options \\ [])

Creates an Outpost.

delete_outpost(client, outpost_id, input, options \\ [])

Deletes the Outpost.

delete_site(client, site_id, input, options \\ [])

Deletes the site.

get_outpost(client, outpost_id, options \\ [])

Gets information about the specified Outpost.

get_outpost_instance_types(client, outpost_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the instance types for the specified Outpost.

list_outposts(client, max_results \\ nil, next_token \\ nil, options \\ [])

List the Outposts for your AWS account.

list_sites(client, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the sites for the specified AWS account.