aws-elixir v0.6.0 AWS.PI

AWS Performance Insights enables you to monitor and explore different dimensions of database load based on data captured from a running RDS instance.

The guide provides detailed information about Performance Insights data types, parameters and errors. For more information about Performance Insights capabilities see Using Amazon RDS Performance Insights in the Amazon RDS User Guide.

The AWS Performance Insights API provides visibility into the performance of your RDS instance, when Performance Insights is enabled for supported engine types. While Amazon CloudWatch provides the authoritative source for AWS service vended monitoring metrics, AWS Performance Insights offers a domain-specific view of database load measured as Average Active Sessions and provided to API consumers as a 2-dimensional time-series dataset. The time dimension of the data provides DB load data for each time point in the queried time range, and each time point decomposes overall load in relation to the requested dimensions, such as SQL, Wait-event, User or Host, measured at that time point.

Functions

describe_dimension_keys(client, input, options \\ [])

For a specific time period, retrieve the top N dimension keys for a metric.

get_resource_metrics(client, input, options \\ [])

Retrieve Performance Insights metrics for a set of data sources, over a time period.

You can provide specific dimension groups and dimensions, and provide aggregation and filtering criteria for each group.