aws-elixir v0.6.0 AWS.PinpointSMSVoice

Pinpoint SMS and Voice Messaging public facing APIs

Functions

create_configuration_set(client, input, options \\ [])

Create a new configuration set.

After you create the configuration set, you can add one or more event destinations to it.

create_configuration_set_event_destination(client, configuration_set_name, input, options \\ [])

Create a new event destination in a configuration set.

delete_configuration_set(client, configuration_set_name, input, options \\ [])

Deletes an existing configuration set.

delete_configuration_set_event_destination(client, configuration_set_name, event_destination_name, input, options \\ [])

Deletes an event destination in a configuration set.

get_configuration_set_event_destinations(client, configuration_set_name, options \\ [])

Obtain information about an event destination, including the types of events it reports, the Amazon Resource Name (ARN) of the destination, and the name of the event destination.

list_configuration_sets(client, next_token \\ nil, page_size \\ nil, options \\ [])

List all of the configuration sets associated with your Amazon Pinpoint account in the current region.

send_voice_message(client, input, options \\ [])

Create a new voice message and send it to a recipient's phone number.

update_configuration_set_event_destination(client, configuration_set_name, event_destination_name, input, options \\ [])

Update an event destination in a configuration set.

An event destination is a location that you publish information about your voice calls to. For example, you can log an event to an Amazon CloudWatch destination when a call fails.