aws-elixir v0.6.0 AWS.QLDB

The control plane for Amazon QLDB

Functions

cancel_journal_kinesis_stream(client, ledger_name, stream_id, input, options \\ [])

Ends a given Amazon QLDB journal stream.

Before a stream can be canceled, its current status must be ACTIVE.

You can't restart a stream after you cancel it. Canceled QLDB stream resources are subject to a 7-day retention period, so they are automatically deleted after this limit expires.

create_ledger(client, input, options \\ [])

Creates a new ledger in your AWS account.

delete_ledger(client, name, input, options \\ [])

Deletes a ledger and all of its contents.

This action is irreversible.

If deletion protection is enabled, you must first disable it before you can delete the ledger using the QLDB API or the AWS Command Line Interface (AWS CLI). You can disable it by calling the UpdateLedger operation to set the flag to false. The QLDB console disables deletion protection for you when you use it to delete a ledger.

describe_journal_kinesis_stream(client, ledger_name, stream_id, options \\ [])

Returns detailed information about a given Amazon QLDB journal stream.

The output includes the Amazon Resource Name (ARN), stream name, current status, creation time, and the parameters of your original stream creation request.

describe_journal_s3_export(client, export_id, name, options \\ [])

Returns information about a journal export job, including the ledger name, export ID, when it was created, current status, and its start and end time export parameters.

This action does not return any expired export jobs. For more information, see Export Job Expiration in the Amazon QLDB Developer Guide.

If the export job with the given ExportId doesn't exist, then throws ResourceNotFoundException.

If the ledger with the given Name doesn't exist, then throws ResourceNotFoundException.

describe_ledger(client, name, options \\ [])

Returns information about a ledger, including its state and when it was created.

export_journal_to_s3(client, name, input, options \\ [])

Exports journal contents within a date and time range from a ledger into a specified Amazon Simple Storage Service (Amazon S3) bucket.

The data is written as files in Amazon Ion format.

If the ledger with the given Name doesn't exist, then throws ResourceNotFoundException.

If the ledger with the given Name is in CREATING status, then throws ResourcePreconditionNotMetException.

You can initiate up to two concurrent journal export requests for each ledger. Beyond this limit, journal export requests throw LimitExceededException.

get_block(client, name, input, options \\ [])

Returns a block object at a specified address in a journal.

Also returns a proof of the specified block for verification if DigestTipAddress is provided.

For information about the data contents in a block, see Journal contents in the Amazon QLDB Developer Guide.

If the specified ledger doesn't exist or is in DELETING status, then throws ResourceNotFoundException.

If the specified ledger is in CREATING status, then throws ResourcePreconditionNotMetException.

If no block exists with the specified address, then throws InvalidParameterException.

get_digest(client, name, input, options \\ [])

Returns the digest of a ledger at the latest committed block in the journal.

The response includes a 256-bit hash value and a block address.

get_revision(client, name, input, options \\ [])

Returns a revision data object for a specified document ID and block address.

Also returns a proof of the specified revision for verification if DigestTipAddress is provided.

list_journal_kinesis_streams_for_ledger(client, ledger_name, max_results \\ nil, next_token \\ nil, options \\ [])

Returns an array of all Amazon QLDB journal stream descriptors for a given ledger.

The output of each stream descriptor includes the same details that are returned by DescribeJournalKinesisStream.

This action returns a maximum of MaxResults items. It is paginated so that you can retrieve all the items by calling ListJournalKinesisStreamsForLedger multiple times.

list_journal_s3_exports(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns an array of journal export job descriptions for all ledgers that are associated with the current AWS account and Region.

This action returns a maximum of MaxResults items, and is paginated so that you can retrieve all the items by calling ListJournalS3Exports multiple times.

This action does not return any expired export jobs. For more information, see Export Job Expiration in the Amazon QLDB Developer Guide.

list_journal_s3_exports_for_ledger(client, name, max_results \\ nil, next_token \\ nil, options \\ [])

Returns an array of journal export job descriptions for a specified ledger.

This action returns a maximum of MaxResults items, and is paginated so that you can retrieve all the items by calling ListJournalS3ExportsForLedger multiple times.

This action does not return any expired export jobs. For more information, see Export Job Expiration in the Amazon QLDB Developer Guide.

list_ledgers(client, max_results \\ nil, next_token \\ nil, options \\ [])

Returns an array of ledger summaries that are associated with the current AWS account and Region.

This action returns a maximum of 100 items and is paginated so that you can retrieve all the items by calling ListLedgers multiple times.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns all tags for a specified Amazon QLDB resource.

stream_journal_to_kinesis(client, ledger_name, input, options \\ [])

Creates a journal stream for a given Amazon QLDB ledger.

The stream captures every document revision that is committed to the ledger's journal and delivers the data to a specified Amazon Kinesis Data Streams resource.

tag_resource(client, resource_arn, input, options \\ [])

Adds one or more tags to a specified Amazon QLDB resource.

A resource can have up to 50 tags. If you try to create more than 50 tags for a resource, your request fails and returns an error.

untag_resource(client, resource_arn, input, options \\ [])

Removes one or more tags from a specified Amazon QLDB resource.

You can specify up to 50 tag keys to remove.

update_ledger(client, name, input, options \\ [])

Updates properties on a ledger.