aws-elixir v0.6.0 AWS.QuickSight

Amazon QuickSight API Reference

Amazon QuickSight is a fully managed, serverless business intelligence service for the AWS Cloud that makes it easy to extend data and insights to every user in your organization.

This API reference contains documentation for a programming interface that you can use to manage Amazon QuickSight.

Functions

cancel_ingestion(client, aws_account_id, data_set_id, ingestion_id, input, options \\ [])

Cancels an ongoing ingestion of data into SPICE.

create_account_customization(client, aws_account_id, input, options \\ [])

Creates Amazon QuickSight customizations the current AWS Region.

Currently, you can add a custom default theme by using the CreateAccountCustomization or UpdateAccountCustomization API operation. To further customize QuickSight by removing QuickSight sample assets and videos for all new users, see Customizing QuickSight in the Amazon QuickSight User Guide.

You can create customizations for your AWS account or, if you specify a namespace, for a QuickSight namespace instead. Customizations that apply to a namespace always override customizations that apply to an AWS account. To find out which customizations apply, use the DescribeAccountCustomization API operation.

Before you use the CreateAccountCustomization API operation to add a theme as the namespace default, make sure that you first share the theme with the namespace. If you don't share it with the namespace, the theme isn't visible to your users even if you make it the default theme. To check if the theme is shared, view the current permissions by using the DescribeThemePermissions API operation. To share the theme, grant permissions by using the UpdateThemePermissions API operation.

create_analysis(client, analysis_id, aws_account_id, input, options \\ [])

Creates an analysis in Amazon QuickSight.

create_dashboard(client, aws_account_id, dashboard_id, input, options \\ [])

Creates a dashboard from a template.

To first create a template, see the CreateTemplate API operation.

A dashboard is an entity in QuickSight that identifies QuickSight reports, created from analyses. You can share QuickSight dashboards. With the right permissions, you can create scheduled email reports from them. If you have the correct permissions, you can create a dashboard from a template that exists in a different AWS account.

create_data_set(client, aws_account_id, input, options \\ [])

Creates a dataset.

create_data_source(client, aws_account_id, input, options \\ [])

Creates a data source.

create_group(client, aws_account_id, namespace, input, options \\ [])

Creates an Amazon QuickSight group.

The permissions resource is arn:aws:quicksight:us-east-1:*<relevant-aws-account-id>*:group/default/*<group-name>*.

The response is a group object.

create_group_membership(client, aws_account_id, group_name, member_name, namespace, input, options \\ [])

Adds an Amazon QuickSight user to an Amazon QuickSight group.

create_i_a_m_policy_assignment(client, aws_account_id, namespace, input, options \\ [])

Creates an assignment with one specified IAM policy, identified by its Amazon Resource Name (ARN).

This policy will be assigned to specified groups or users of Amazon QuickSight. The users and groups need to be in the same namespace.

create_ingestion(client, aws_account_id, data_set_id, ingestion_id, input, options \\ [])

Creates and starts a new SPICE ingestion on a dataset

Any ingestions operating on tagged datasets inherit the same tags automatically for use in access control.

For an example, see How do I create an IAM policy to control access to Amazon EC2 resources using tags? in the AWS Knowledge Center. Tags are visible on the tagged dataset, but not on the ingestion resource.

create_namespace(client, aws_account_id, input, options \\ [])

(Enterprise edition only) Creates a new namespace for you to use with Amazon QuickSight.

A namespace allows you to isolate the QuickSight users and groups that are registered for that namespace. Users that access the namespace can share assets only with other users or groups in the same namespace. They can't see users and groups in other namespaces. You can create a namespace after your AWS account is subscribed to QuickSight. The namespace must be unique within the AWS account. By default, there is a limit of 100 namespaces per AWS account. To increase your limit, create a ticket with AWS Support.

create_template(client, aws_account_id, template_id, input, options \\ [])

Creates a template from an existing QuickSight analysis or template.

You can use the resulting template to create a dashboard.

A template is an entity in QuickSight that encapsulates the metadata required to create an analysis and that you can use to create s dashboard. A template adds a layer of abstraction by using placeholders to replace the dataset associated with the analysis. You can use templates to create dashboards by replacing dataset placeholders with datasets that follow the same schema that was used to create the source analysis and template.

create_template_alias(client, alias_name, aws_account_id, template_id, input, options \\ [])

Creates a template alias for a template.

create_theme(client, aws_account_id, theme_id, input, options \\ [])

Creates a theme.

A theme is set of configuration options for color and layout. Themes apply to analyses and dashboards. For more information, see Using Themes in Amazon QuickSight in the Amazon QuickSight User Guide.

create_theme_alias(client, alias_name, aws_account_id, theme_id, input, options \\ [])

Creates a theme alias for a theme.

delete_account_customization(client, aws_account_id, input, options \\ [])

Deletes all Amazon QuickSight customizations in this AWS Region for the specified AWS account and QuickSight namespace.

delete_analysis(client, analysis_id, aws_account_id, input, options \\ [])

Deletes an analysis from Amazon QuickSight.

You can optionally include a recovery window during which you can restore the analysis. If you don't specify a recovery window value, the operation defaults to 30 days. QuickSight attaches a DeletionTime stamp to the response that specifies the end of the recovery window. At the end of the recovery window, QuickSight deletes the analysis permanently.

At any time before recovery window ends, you can use the RestoreAnalysis API operation to remove the DeletionTime stamp and cancel the deletion of the analysis. The analysis remains visible in the API until it's deleted, so you can describe it but you can't make a template from it.

An analysis that's scheduled for deletion isn't accessible in the QuickSight console. To access it in the console, restore it. Deleting an analysis doesn't delete the dashboards that you publish from it.

delete_dashboard(client, aws_account_id, dashboard_id, input, options \\ [])

Deletes a dashboard.

delete_data_set(client, aws_account_id, data_set_id, input, options \\ [])

Deletes a dataset.

delete_data_source(client, aws_account_id, data_source_id, input, options \\ [])

Deletes the data source permanently.

This operation breaks all the datasets that reference the deleted data source.

delete_group(client, aws_account_id, group_name, namespace, input, options \\ [])

Removes a user group from Amazon QuickSight.

delete_group_membership(client, aws_account_id, group_name, member_name, namespace, input, options \\ [])

Removes a user from a group so that the user is no longer a member of the group.

delete_i_a_m_policy_assignment(client, assignment_name, aws_account_id, namespace, input, options \\ [])

Deletes an existing IAM policy assignment.

delete_namespace(client, aws_account_id, namespace, input, options \\ [])

Deletes a namespace and the users and groups that are associated with the namespace.

This is an asynchronous process. Assets including dashboards, analyses, datasets and data sources are not deleted. To delete these assets, you use the API operations for the relevant asset.

delete_template(client, aws_account_id, template_id, input, options \\ [])

Deletes a template.

delete_template_alias(client, alias_name, aws_account_id, template_id, input, options \\ [])

Deletes the item that the specified template alias points to.

If you provide a specific alias, you delete the version of the template that the alias points to.

delete_theme(client, aws_account_id, theme_id, input, options \\ [])

Deletes a theme.

delete_theme_alias(client, alias_name, aws_account_id, theme_id, input, options \\ [])

Deletes the version of the theme that the specified theme alias points to.

If you provide a specific alias, you delete the version of the theme that the alias points to.

delete_user(client, aws_account_id, namespace, user_name, input, options \\ [])

Deletes the Amazon QuickSight user that is associated with the identity of the AWS Identity and Access Management (IAM) user or role that's making the call.

The IAM user isn't deleted as a result of this call.

delete_user_by_principal_id(client, aws_account_id, namespace, principal_id, input, options \\ [])

Deletes a user identified by its principal ID.

describe_account_customization(client, aws_account_id, namespace \\ nil, resolved \\ nil, options \\ [])

Describes the customizations associated with the provided AWS account and Amazon QuickSight namespace in an AWS Region.

The QuickSight console evaluates which customizations to apply by running this API operation with the Resolved flag included.

To determine what customizations display when you run this command, it can help to visualize the relationship of the entities involved.

AWS Account - The AWS account exists at the top of the hierarchy. It has the potential to use all of the AWS Regions and AWS Services. When you subscribe to QuickSight, you choose one AWS Region to use as your home Region. That's where your free SPICE capacity is located. You can use QuickSight in any supported AWS Region.

AWS Region - In each AWS Region where you sign in to QuickSight at least once, QuickSight acts as a separate instance of the same service. If you have a user directory, it resides in us-east-1, which is the US East (N. Virginia). Generally speaking, these users have access to QuickSight in any AWS Region, unless they are constrained to a namespace.

To run the command in a different AWS Region, you change your Region settings. If you're using the AWS CLI, you can use one of the following options:

* Use [command line options](https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-options.html).

* Use [named profiles](https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-profiles.html).

* Run `aws configure` to change your default AWS Region.

Use Enter to key the same settings for your keys. For more information, see Configuring the AWS CLI.

Namespace - A QuickSight namespace is a partition that contains users and assets (data sources, datasets, dashboards, and so on). To access assets that are in a specific namespace, users and groups must also be part of the same namespace. People who share a namespace are completely isolated from users and assets in other namespaces, even if they are in the same AWS account and AWS Region.

Applied customizations - Within an AWS Region, a set of QuickSight customizations can apply to an AWS account or to a namespace. Settings that you apply to a namespace override settings that you apply to an AWS account. All settings are isolated to a single AWS Region. To apply them in other AWS Regions, run the CreateAccountCustomization command in each AWS Region where you want to apply the same customizations.

describe_account_settings(client, aws_account_id, options \\ [])

Describes the settings that were used when your QuickSight subscription was first created in this AWS account.

describe_analysis(client, analysis_id, aws_account_id, options \\ [])

Provides a summary of the metadata for an analysis.

describe_analysis_permissions(client, analysis_id, aws_account_id, options \\ [])

Provides the read and write permissions for an analysis.

describe_dashboard(client, aws_account_id, dashboard_id, alias_name \\ nil, version_number \\ nil, options \\ [])

Provides a summary for a dashboard.

describe_dashboard_permissions(client, aws_account_id, dashboard_id, options \\ [])

Describes read and write permissions for a dashboard.

describe_data_set(client, aws_account_id, data_set_id, options \\ [])

Describes a dataset.

describe_data_set_permissions(client, aws_account_id, data_set_id, options \\ [])

Describes the permissions on a dataset.

The permissions resource is arn:aws:quicksight:region:aws-account-id:dataset/data-set-id.

describe_data_source(client, aws_account_id, data_source_id, options \\ [])

Describes a data source.

describe_data_source_permissions(client, aws_account_id, data_source_id, options \\ [])

Describes the resource permissions for a data source.

describe_group(client, aws_account_id, group_name, namespace, options \\ [])

Returns an Amazon QuickSight group's description and Amazon Resource Name (ARN).

describe_i_a_m_policy_assignment(client, assignment_name, aws_account_id, namespace, options \\ [])

Describes an existing IAM policy assignment, as specified by the assignment name.

describe_ingestion(client, aws_account_id, data_set_id, ingestion_id, options \\ [])

Describes a SPICE ingestion.

describe_namespace(client, aws_account_id, namespace, options \\ [])

Describes the current namespace.

describe_template(client, aws_account_id, template_id, alias_name \\ nil, version_number \\ nil, options \\ [])

Describes a template's metadata.

describe_template_alias(client, alias_name, aws_account_id, template_id, options \\ [])

Describes the template alias for a template.

describe_template_permissions(client, aws_account_id, template_id, options \\ [])

Describes read and write permissions on a template.

describe_theme(client, aws_account_id, theme_id, alias_name \\ nil, version_number \\ nil, options \\ [])

Describes a theme.

describe_theme_alias(client, alias_name, aws_account_id, theme_id, options \\ [])

Describes the alias for a theme.

describe_theme_permissions(client, aws_account_id, theme_id, options \\ [])

Describes the read and write permissions for a theme.

describe_user(client, aws_account_id, namespace, user_name, options \\ [])

Returns information about a user, given the user name.

get_dashboard_embed_url(client, aws_account_id, dashboard_id, identity_type, reset_disabled \\ nil, session_lifetime_in_minutes \\ nil, undo_redo_disabled \\ nil, user_arn \\ nil, options \\ [])

Generates a session URL and authorization code that you can use to embed an Amazon QuickSight read-only dashboard in your web server code.

Before you use this command, make sure that you have configured the dashboards and permissions.

Currently, you can use GetDashboardEmbedURL only from the server, not from the user's browser. The following rules apply to the combination of URL and authorization code:

They must be used together.

They can be used one time only.

They are valid for 5 minutes after you run this command.

The resulting user session is valid for 10 hours.

For more information, see Embedding Amazon QuickSight in the Amazon QuickSight User Guide .

get_session_embed_url(client, aws_account_id, entry_point \\ nil, session_lifetime_in_minutes \\ nil, user_arn \\ nil, options \\ [])

Generates a session URL and authorization code that you can use to embed the Amazon QuickSight console in your web server code.

Use GetSessionEmbedUrl where you want to provide an authoring portal that allows users to create data sources, datasets, analyses, and dashboards. The users who access an embedded QuickSight console need belong to the author or admin security cohort. If you want to restrict permissions to some of these features, add a custom permissions profile to the user with the UpdateUser API operation. Use RegisterUser API operation to add a new user with a custom permission profile attached. For more information, see the following sections in the Amazon QuickSight User Guide:

Embedding the Amazon QuickSight Console

Customizing Access to the Amazon QuickSight Console

list_analyses(client, aws_account_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists Amazon QuickSight analyses that exist in the specified AWS account.

list_dashboard_versions(client, aws_account_id, dashboard_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all the versions of the dashboards in the QuickSight subscription.

list_dashboards(client, aws_account_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists dashboards in an AWS account.

list_data_sets(client, aws_account_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all of the datasets belonging to the current AWS account in an AWS Region.

The permissions resource is arn:aws:quicksight:region:aws-account-id:dataset/*.

list_data_sources(client, aws_account_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists data sources in current AWS Region that belong to this AWS account.

list_group_memberships(client, aws_account_id, group_name, namespace, max_results \\ nil, next_token \\ nil, options \\ [])

Lists member users in a group.

list_groups(client, aws_account_id, namespace, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all user groups in Amazon QuickSight.

list_i_a_m_policy_assignments(client, aws_account_id, namespace, max_results \\ nil, next_token \\ nil, options \\ [])

Lists IAM policy assignments in the current Amazon QuickSight account.

list_i_a_m_policy_assignments_for_user(client, aws_account_id, namespace, user_name, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all the IAM policy assignments, including the Amazon Resource Names (ARNs) for the IAM policies assigned to the specified user and group or groups that the user belongs to.

list_ingestions(client, aws_account_id, data_set_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the history of SPICE ingestions for a dataset.

list_namespaces(client, aws_account_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the namespaces for the specified AWS account.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists the tags assigned to a resource.

list_template_aliases(client, aws_account_id, template_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all the aliases of a template.

list_template_versions(client, aws_account_id, template_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all the versions of the templates in the current Amazon QuickSight account.

list_templates(client, aws_account_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all the templates in the current Amazon QuickSight account.

list_theme_aliases(client, aws_account_id, theme_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all the aliases of a theme.

list_theme_versions(client, aws_account_id, theme_id, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all the versions of the themes in the current AWS account.

list_themes(client, aws_account_id, max_results \\ nil, next_token \\ nil, type \\ nil, options \\ [])

Lists all the themes in the current AWS account.

list_user_groups(client, aws_account_id, namespace, user_name, max_results \\ nil, next_token \\ nil, options \\ [])

Lists the Amazon QuickSight groups that an Amazon QuickSight user is a member of.

list_users(client, aws_account_id, namespace, max_results \\ nil, next_token \\ nil, options \\ [])

Returns a list of all of the Amazon QuickSight users belonging to this account.

register_user(client, aws_account_id, namespace, input, options \\ [])

Creates an Amazon QuickSight user, whose identity is associated with the AWS Identity and Access Management (IAM) identity or role specified in the request.

restore_analysis(client, analysis_id, aws_account_id, input, options \\ [])

Restores an analysis.

search_analyses(client, aws_account_id, input, options \\ [])

Searches for analyses that belong to the user specified in the filter.

search_dashboards(client, aws_account_id, input, options \\ [])

Searches for dashboards that belong to a user.

tag_resource(client, resource_arn, input, options \\ [])

Assigns one or more tags (key-value pairs) to the specified QuickSight resource.

Tags can help you organize and categorize your resources. You can also use them to scope user permissions, by granting a user permission to access or change only resources with certain tag values. You can use the TagResource operation with a resource that already has tags. If you specify a new tag key for the resource, this tag is appended to the list of tags associated with the resource. If you specify a tag key that is already associated with the resource, the new tag value that you specify replaces the previous value for that tag.

You can associate as many as 50 tags with a resource. QuickSight supports tagging on data set, data source, dashboard, and template.

Tagging for QuickSight works in a similar way to tagging for other AWS services, except for the following:

You can't use tags to track AWS costs for QuickSight. This restriction is because QuickSight costs are based on users and SPICE capacity, which aren't taggable resources.

QuickSight doesn't currently support the Tag Editor for AWS Resource Groups.

untag_resource(client, resource_arn, input, options \\ [])

Removes a tag or tags from a resource.

update_account_customization(client, aws_account_id, input, options \\ [])

Updates Amazon QuickSight customizations the current AWS Region.

Currently, the only customization you can use is a theme.

You can use customizations for your AWS account or, if you specify a namespace, for a QuickSight namespace instead. Customizations that apply to a namespace override customizations that apply to an AWS account. To find out which customizations apply, use the DescribeAccountCustomization API operation.

update_account_settings(client, aws_account_id, input, options \\ [])

Updates the Amazon QuickSight settings in your AWS account.

update_analysis(client, analysis_id, aws_account_id, input, options \\ [])

Updates an analysis in Amazon QuickSight

update_analysis_permissions(client, analysis_id, aws_account_id, input, options \\ [])

Updates the read and write permissions for an analysis.

update_dashboard(client, aws_account_id, dashboard_id, input, options \\ [])

Updates a dashboard in an AWS account.

update_dashboard_permissions(client, aws_account_id, dashboard_id, input, options \\ [])

Updates read and write permissions on a dashboard.

update_dashboard_published_version(client, aws_account_id, dashboard_id, version_number, input, options \\ [])

Updates the published version of a dashboard.

update_data_set(client, aws_account_id, data_set_id, input, options \\ [])

Updates a dataset.

update_data_set_permissions(client, aws_account_id, data_set_id, input, options \\ [])

Updates the permissions on a dataset.

The permissions resource is arn:aws:quicksight:region:aws-account-id:dataset/data-set-id.

update_data_source(client, aws_account_id, data_source_id, input, options \\ [])

Updates a data source.

update_data_source_permissions(client, aws_account_id, data_source_id, input, options \\ [])

Updates the permissions to a data source.

update_group(client, aws_account_id, group_name, namespace, input, options \\ [])

Changes a group description.

update_i_a_m_policy_assignment(client, assignment_name, aws_account_id, namespace, input, options \\ [])

Updates an existing IAM policy assignment.

This operation updates only the optional parameter or parameters that are specified in the request.

update_template(client, aws_account_id, template_id, input, options \\ [])

Updates a template from an existing Amazon QuickSight analysis or another template.

update_template_alias(client, alias_name, aws_account_id, template_id, input, options \\ [])

Updates the template alias of a template.

update_template_permissions(client, aws_account_id, template_id, input, options \\ [])

Updates the resource permissions for a template.

update_theme(client, aws_account_id, theme_id, input, options \\ [])

Updates a theme.

update_theme_alias(client, alias_name, aws_account_id, theme_id, input, options \\ [])

Updates an alias of a theme.

update_theme_permissions(client, aws_account_id, theme_id, input, options \\ [])

Updates the resource permissions for a theme.

Permissions apply to the action to grant or revoke permissions on, for example "quicksight:DescribeTheme".

Theme permissions apply in groupings. Valid groupings include the following for the three levels of permissions, which are user, owner, or no permissions:

User

"quicksight:DescribeTheme"

"quicksight:DescribeThemeAlias"

"quicksight:ListThemeAliases"

"quicksight:ListThemeVersions"

Owner

"quicksight:DescribeTheme"

"quicksight:DescribeThemeAlias"

"quicksight:ListThemeAliases"

"quicksight:ListThemeVersions"

"quicksight:DeleteTheme"

"quicksight:UpdateTheme"

"quicksight:CreateThemeAlias"

"quicksight:DeleteThemeAlias"

"quicksight:UpdateThemeAlias"

"quicksight:UpdateThemePermissions"

"quicksight:DescribeThemePermissions"

To specify no permissions, omit the permissions list.

update_user(client, aws_account_id, namespace, user_name, input, options \\ [])

Updates an Amazon QuickSight user.