aws-elixir v0.6.0 AWS.RAM

Use AWS Resource Access Manager to share AWS resources between AWS accounts.

To share a resource, you create a resource share, associate the resource with the resource share, and specify the principals that can access the resources associated with the resource share. The following principals are supported: AWS accounts, organizational units (OU) from AWS Organizations, and organizations from AWS Organizations.

For more information, see the AWS Resource Access Manager User Guide.

Functions

accept_resource_share_invitation(client, input, options \\ [])

Accepts an invitation to a resource share from another AWS account.

associate_resource_share(client, input, options \\ [])

Associates the specified resource share with the specified principals and resources.

associate_resource_share_permission(client, input, options \\ [])

Associates a permission with a resource share.

create_resource_share(client, input, options \\ [])

Creates a resource share.

delete_resource_share(client, input, options \\ [])

Deletes the specified resource share.

disassociate_resource_share(client, input, options \\ [])

Disassociates the specified principals or resources from the specified resource share.

disassociate_resource_share_permission(client, input, options \\ [])

Disassociates an AWS RAM permission from a resource share.

enable_sharing_with_aws_organization(client, input, options \\ [])

Enables resource sharing within your AWS Organization.

The caller must be the master account for the AWS Organization.

get_permission(client, input, options \\ [])

Gets the contents of an AWS RAM permission in JSON format.

get_resource_policies(client, input, options \\ [])

Gets the policies for the specified resources that you own and have shared.

get_resource_share_associations(client, input, options \\ [])

Gets the resources or principals for the resource shares that you own.

get_resource_share_invitations(client, input, options \\ [])

Gets the invitations for resource sharing that you've received.

get_resource_shares(client, input, options \\ [])

Gets the resource shares that you own or the resource shares that are shared with you.

list_pending_invitation_resources(client, input, options \\ [])

Lists the resources in a resource share that is shared with you but that the invitation is still pending for.

list_permissions(client, input, options \\ [])

Lists the AWS RAM permissions.

list_principals(client, input, options \\ [])

Lists the principals that you have shared resources with or that have shared resources with you.

list_resource_share_permissions(client, input, options \\ [])

Lists the AWS RAM permissions that are associated with a resource share.

list_resource_types(client, input, options \\ [])

Lists the shareable resource types supported by AWS RAM.

list_resources(client, input, options \\ [])

Lists the resources that you added to a resource shares or the resources that are shared with you.

promote_resource_share_created_from_policy(client, input, options \\ [])

Resource shares that were created by attaching a policy to a resource are visible only to the resource share owner, and the resource share cannot be modified in AWS RAM.

Use this API action to promote the resource share. When you promote the resource share, it becomes:

Visible to all principals that it is shared with.

Modifiable in AWS RAM.

reject_resource_share_invitation(client, input, options \\ [])

Rejects an invitation to a resource share from another AWS account.

tag_resource(client, input, options \\ [])

Adds the specified tags to the specified resource share that you own.

untag_resource(client, input, options \\ [])

Removes the specified tags from the specified resource share that you own.

update_resource_share(client, input, options \\ [])

Updates the specified resource share that you own.