aws-elixir v0.6.0 AWS.RedshiftData

You can use the Amazon Redshift Data API to run queries on Amazon Redshift tables.

You can run individual SQL statements, which are committed if the statement succeeds.

Functions

cancel_statement(client, input, options \\ [])

Cancels a running query.

To be canceled, a query must be running.

describe_statement(client, input, options \\ [])

Describes the details about a specific instance when a query was run by the Amazon Redshift Data API.

The information includes when the query started, when it finished, the query status, the number of rows returned, and the SQL statement.

describe_table(client, input, options \\ [])

Describes the detailed information about a table from metadata in the cluster.

The information includes its columns. A token is returned to page through the column list. Depending on the authorization method, use one of the following combinations of request parameters:

AWS Secrets Manager - specify the Amazon Resource Name (ARN) of the secret and the cluster identifier that matches the cluster in the secret.

Temporary credentials - specify the cluster identifier, the database name, and the database user name. Permission to call the redshift:GetClusterCredentials operation is required to use this method.

execute_statement(client, input, options \\ [])

Runs an SQL statement, which can be data manipulation language (DML) or data definition language (DDL).

This statement must be a single SQL statement. Depending on the authorization method, use one of the following combinations of request parameters:

AWS Secrets Manager - specify the Amazon Resource Name (ARN) of the secret and the cluster identifier that matches the cluster in the secret.

Temporary credentials - specify the cluster identifier, the database name, and the database user name. Permission to call the redshift:GetClusterCredentials operation is required to use this method.

get_statement_result(client, input, options \\ [])

Fetches the temporarily cached result of an SQL statement.

A token is returned to page through the statement results.

list_databases(client, input, options \\ [])

List the databases in a cluster.

A token is returned to page through the database list. Depending on the authorization method, use one of the following combinations of request parameters:

AWS Secrets Manager - specify the Amazon Resource Name (ARN) of the secret and the cluster identifier that matches the cluster in the secret.

Temporary credentials - specify the cluster identifier, the database name, and the database user name. Permission to call the redshift:GetClusterCredentials operation is required to use this method.

list_schemas(client, input, options \\ [])

Lists the schemas in a database.

A token is returned to page through the schema list. Depending on the authorization method, use one of the following combinations of request parameters:

AWS Secrets Manager - specify the Amazon Resource Name (ARN) of the secret and the cluster identifier that matches the cluster in the secret.

Temporary credentials - specify the cluster identifier, the database name, and the database user name. Permission to call the redshift:GetClusterCredentials operation is required to use this method.

list_statements(client, input, options \\ [])

List of SQL statements.

By default, only finished statements are shown. A token is returned to page through the statement list.

list_tables(client, input, options \\ [])

List the tables in a database.

If neither SchemaPattern nor TablePattern are specified, then all tables in the database are returned. A token is returned to page through the table list. Depending on the authorization method, use one of the following combinations of request parameters:

AWS Secrets Manager - specify the Amazon Resource Name (ARN) of the secret and the cluster identifier that matches the cluster in the secret.

Temporary credentials - specify the cluster identifier, the database name, and the database user name. Permission to call the redshift:GetClusterCredentials operation is required to use this method.