aws-elixir v0.6.0 AWS.Request

Functions

add_headers(list, headers)

Include additions only if they do not already exist in the provided list.

build_params(mapping, params)

Build request params (header or querystring) based on a list of key-value pairs representing the mapping from the input name (key) to the param name (value) and a map with the input params.

sign_v4(client, method, url, headers, body)

Generate headers with an AWS signature version 4 for the specified request.

sign_v4(client, now, method, url, headers, body)

Generate headers with an AWS signature version 4 for the specified request using the specified time.

sign_v4_query(client, method, url, headers, body)

Generate headers with an AWS signature version 4 for the specified request that can be transformed into a query string.

sign_v4_query(client, now, method, url, headers, body)

Generate headers with an AWS signature version 4 for the specified request using the specified time that can be transformed into a query string.