aws-elixir v0.6.0 AWS.Request.Internal

Functions

add_authorization_header(headers, authorization)

Add an Authorization header with an AWS4-HMAC-SHA256 signature to the list of headers.

add_content_hash(headers, body)

Add an X-Amz-Content-SHA256 header which is the hash of the payload. This header is required for S3 when using the v4 signature. Adding it in requests for all services does not cause any issues.

add_date_header(headers, date)

Add an X-Amz-Date header with a long date value in YYMMDDTHHMMSSZ format to a list of headers.

add_security_token(headers, client)

Add an X-Amz-Security-Token if credentials configurations are configured for it

authorization(access_key_id, credential_scope, signed_headers, signature)

Generate an AWS4-HMAC-SHA256 authorization signature.

canonical_header(arg)

Strip leading and trailing whitespace around name and value, convert name to lowercase, and add a trailing newline.

canonical_headers(headers)

Convert a list of headers to canonical header format. Leading and trailing whitespace around header names and values is stripped, header names are lowercased, and headers are newline-joined in alphabetical order (with a trailing newline).

canonical_request(method, url, headers, body)

Process and merge request values into a canonical request for AWS signature version 4.

credential_scope(short_date, region, service)

Generate a credential scope from a short date in YYMMDD format, a region identifier and a service identifier.

signed_header(arg)

Strip leading and trailing whitespace around Name and convert it to lowercase.

signed_headers(headers)

Convert a list of headers to canonicals signed header format. Leading and trailing whitespace around names is stripped, header names are lowercased, and header names are semicolon-joined in alphabetical order.

signing_key(secret_access_key, short_date, region, service)

Generate a signing key from a secret access key, a short date in YYMMDD format, a region identifier and a service identifier.

split_url(url)

Strip the query string from the URL, if one if present, and return the URL and query string as separate values.

string_to_sign(long_date, credential_scope, hashed_canonical_request)

Generate the text to sign from a long date in YYMMDDTHHMMSSZ format, a credential scope and a hashed canonical request.