aws-elixir v0.6.0 AWS.ResourceGroups

AWS Resource Groups

AWS Resource Groups lets you organize AWS resources such as Amazon EC2 instances, Amazon Relational Database Service databases, and Amazon S3 buckets into groups using criteria that you define as tags.

A resource group is a collection of resources that match the resource types specified in a query, and share one or more tags or portions of tags. You can create a group of resources based on their roles in your cloud infrastructure, lifecycle stages, regions, application layers, or virtually any criteria. Resource Groups enable you to automate management tasks, such as those in AWS Systems Manager Automation documents, on tag-related resources in AWS Systems Manager. Groups of tagged resources also let you quickly view a custom console in AWS Systems Manager that shows AWS Config compliance and other monitoring data about member resources.

To create a resource group, build a resource query, and specify tags that identify the criteria that members of the group have in common. Tags are key-value pairs.

For more information about Resource Groups, see the AWS Resource Groups User Guide.

AWS Resource Groups uses a REST-compliant API that you can use to perform the following types of operations.

Create, Read, Update, and Delete (CRUD) operations on resource groups and resource query entities

Applying, editing, and removing tags from resource groups

Resolving resource group member ARNs so they can be returned as search results

Getting data about resources that are members of a group

Searching AWS resources based on a resource query

Functions

create_group(client, input, options \\ [])

Creates a resource group with the specified name and description.

You can optionally include a resource query, or a service configuration.

delete_group(client, input, options \\ [])

Deletes the specified resource group.

Deleting a resource group does not delete any resources that are members of the group; it only deletes the group structure.

get_group(client, input, options \\ [])

Returns information about a specified resource group.

get_group_configuration(client, input, options \\ [])

Returns the service configuration associated with the specified resource group.

AWS Resource Groups supports configurations for the following resource group types:

AWS::EC2::CapacityReservationPool - Amazon EC2 capacity reservation pools. For more information, see Working with capacity reservation groups in the EC2 Users Guide.

get_group_query(client, input, options \\ [])

Retrieves the resource query associated with the specified resource group.

get_tags(client, arn, options \\ [])

Returns a list of tags that are associated with a resource group, specified by an ARN.

group_resources(client, input, options \\ [])

Adds the specified resources to the specified group.

list_group_resources(client, input, options \\ [])

Returns a list of ARNs of the resources that are members of a specified resource group.

list_groups(client, input, options \\ [])

Returns a list of existing resource groups in your account.

search_resources(client, input, options \\ [])

Returns a list of AWS resource identifiers that matches the specified query.

The query uses the same format as a resource query in a CreateGroup or UpdateGroupQuery operation.

tag(client, arn, input, options \\ [])

Adds tags to a resource group with the specified ARN.

Existing tags on a resource group are not changed if they are not specified in the request parameters.

Do not store personally identifiable information (PII) or other confidential or sensitive information in tags. We use tags to provide you with billing and administration services. Tags are not intended to be used for private or sensitive data.

ungroup_resources(client, input, options \\ [])

Removes the specified resources from the specified group.

untag(client, arn, input, options \\ [])

Deletes tags from a specified resource group.

update_group(client, input, options \\ [])

Updates the description for an existing group.

You cannot update the name of a resource group.

update_group_query(client, input, options \\ [])

Updates the resource query of a group.