aws-elixir v0.6.0 AWS.RoboMaker

This section provides documentation for the AWS RoboMaker API operations.

Functions

batch_delete_worlds(client, input, options \\ [])

Deletes one or more worlds in a batch operation.

batch_describe_simulation_job(client, input, options \\ [])

Describes one or more simulation jobs.

cancel_deployment_job(client, input, options \\ [])

Cancels the specified deployment job.

cancel_simulation_job(client, input, options \\ [])

Cancels the specified simulation job.

cancel_simulation_job_batch(client, input, options \\ [])

Cancels a simulation job batch.

When you cancel a simulation job batch, you are also cancelling all of the active simulation jobs created as part of the batch.

cancel_world_export_job(client, input, options \\ [])

Cancels the specified export job.

cancel_world_generation_job(client, input, options \\ [])

Cancels the specified world generator job.

create_deployment_job(client, input, options \\ [])

Deploys a specific version of a robot application to robots in a fleet.

The robot application must have a numbered applicationVersion for consistency reasons. To create a new version, use CreateRobotApplicationVersion or see Creating a Robot Application Version.

After 90 days, deployment jobs expire and will be deleted. They will no longer be accessible.

create_fleet(client, input, options \\ [])

Creates a fleet, a logical group of robots running the same robot application.

create_robot(client, input, options \\ [])

Creates a robot.

create_robot_application(client, input, options \\ [])

Creates a robot application.

create_robot_application_version(client, input, options \\ [])

Creates a version of a robot application.

create_simulation_application(client, input, options \\ [])

Creates a simulation application.

create_simulation_application_version(client, input, options \\ [])

Creates a simulation application with a specific revision id.

create_simulation_job(client, input, options \\ [])

Creates a simulation job.

After 90 days, simulation jobs expire and will be deleted. They will no longer be accessible.

create_world_export_job(client, input, options \\ [])

Creates a world export job.

create_world_generation_job(client, input, options \\ [])

Creates worlds using the specified template.

create_world_template(client, input, options \\ [])

Creates a world template.

delete_fleet(client, input, options \\ [])

Deletes a fleet.

delete_robot(client, input, options \\ [])

Deletes a robot.

delete_robot_application(client, input, options \\ [])

Deletes a robot application.

delete_simulation_application(client, input, options \\ [])

Deletes a simulation application.

delete_world_template(client, input, options \\ [])

Deletes a world template.

deregister_robot(client, input, options \\ [])

Deregisters a robot.

describe_deployment_job(client, input, options \\ [])

Describes a deployment job.

describe_fleet(client, input, options \\ [])

Describes a fleet.

describe_robot(client, input, options \\ [])

Describes a robot.

describe_robot_application(client, input, options \\ [])

Describes a robot application.

describe_simulation_application(client, input, options \\ [])

Describes a simulation application.

describe_simulation_job(client, input, options \\ [])

Describes a simulation job.

describe_simulation_job_batch(client, input, options \\ [])

Describes a simulation job batch.

describe_world(client, input, options \\ [])

Describes a world.

describe_world_export_job(client, input, options \\ [])

Describes a world export job.

describe_world_generation_job(client, input, options \\ [])

Describes a world generation job.

describe_world_template(client, input, options \\ [])

Describes a world template.

get_world_template_body(client, input, options \\ [])

Gets the world template body.

list_deployment_jobs(client, input, options \\ [])

Returns a list of deployment jobs for a fleet.

You can optionally provide filters to retrieve specific deployment jobs.

list_fleets(client, input, options \\ [])

Returns a list of fleets.

You can optionally provide filters to retrieve specific fleets.

list_robot_applications(client, input, options \\ [])

Returns a list of robot application.

You can optionally provide filters to retrieve specific robot applications.

list_robots(client, input, options \\ [])

Returns a list of robots.

You can optionally provide filters to retrieve specific robots.

list_simulation_applications(client, input, options \\ [])

Returns a list of simulation applications.

You can optionally provide filters to retrieve specific simulation applications.

list_simulation_job_batches(client, input, options \\ [])

Returns a list simulation job batches.

You can optionally provide filters to retrieve specific simulation batch jobs.

list_simulation_jobs(client, input, options \\ [])

Returns a list of simulation jobs.

You can optionally provide filters to retrieve specific simulation jobs.

list_tags_for_resource(client, resource_arn, options \\ [])

Lists all tags on a AWS RoboMaker resource.

list_world_export_jobs(client, input, options \\ [])

Lists world export jobs.

list_world_generation_jobs(client, input, options \\ [])

Lists world generator jobs.

list_world_templates(client, input, options \\ [])

Lists world templates.

list_worlds(client, input, options \\ [])

Lists worlds.

register_robot(client, input, options \\ [])

Registers a robot with a fleet.

restart_simulation_job(client, input, options \\ [])

Restarts a running simulation job.

start_simulation_job_batch(client, input, options \\ [])

Starts a new simulation job batch.

The batch is defined using one or more SimulationJobRequest objects.

sync_deployment_job(client, input, options \\ [])

Syncrhonizes robots in a fleet to the latest deployment.

This is helpful if robots were added after a deployment.

tag_resource(client, resource_arn, input, options \\ [])

Adds or edits tags for a AWS RoboMaker resource.

Each tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty strings.

For information about the rules that apply to tag keys and tag values, see User-Defined Tag Restrictions in the AWS Billing and Cost Management User Guide.

untag_resource(client, resource_arn, input, options \\ [])

Removes the specified tags from the specified AWS RoboMaker resource.

To remove a tag, specify the tag key. To change the tag value of an existing tag key, use TagResource .

update_robot_application(client, input, options \\ [])

Updates a robot application.

update_simulation_application(client, input, options \\ [])

Updates a simulation application.

update_world_template(client, input, options \\ [])

Updates a world template.