aws-elixir v0.6.0 AWS.S3Control

AWS S3 Control provides access to Amazon S3 control plane operations.

Functions

create_access_point(client, name, input, options \\ [])

Creates an access point and associates it with the specified bucket.

For more information, see Managing Data Access with Amazon S3 Access Points in the Amazon Simple Storage Service Developer Guide.

Using this action with Amazon S3 on Outposts

This action:

Requires a virtual private cloud (VPC) configuration as S3 on Outposts only supports VPC style access points.

Does not support ACL on S3 on Outposts buckets.

Does not support Public Access on S3 on Outposts buckets.

Does not support object lock for S3 on Outposts buckets.

For more information, see Using Amazon S3 on Outposts in the Amazon Simple Storage Service Developer Guide .

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to CreateAccessPoint:

GetAccessPoint DeleteAccessPoint

* ListAccessPoints

create_bucket(client, bucket, input, options \\ [])

This API operation creates an Amazon S3 on Outposts bucket.

To create an S3 bucket, see Create Bucket in the Amazon Simple Storage Service API.

Creates a new Outposts bucket. By creating the bucket, you become the bucket owner. To create an Outposts bucket, you must have S3 on Outposts. For more information, see Using Amazon S3 on Outposts in Amazon Simple Storage Service Developer Guide.

Not every string is an acceptable bucket name. For information on bucket naming restrictions, see Working with Amazon S3 Buckets.

S3 on Outposts buckets do not support

ACLs. Instead, configure access point policies to manage access to buckets.

Public access.

Object Lock

Bucket Location constraint

For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and outpost-id in your API request, see the Example section below.

The following actions are related to CreateBucket for Amazon S3 on Outposts:

PutObject * GetBucket

DeleteBucket CreateAccessPoint

* PutAccessPointPolicy

create_job(client, input, options \\ [])

S3 Batch Operations performs large-scale Batch Operations on Amazon S3 objects.

Batch Operations can run a single operation or action on lists of Amazon S3 objects that you specify. For more information, see S3 Batch Operations in the Amazon Simple Storage Service Developer Guide.

This operation creates a S3 Batch Operations job.

Related actions include:

DescribeJob ListJobs

UpdateJobPriority UpdateJobStatus

delete_access_point(client, name, input, options \\ [])

Deletes the specified access point.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the ARN, see the Example section below.

The following actions are related to DeleteAccessPoint:

CreateAccessPoint GetAccessPoint

* ListAccessPoints

delete_access_point_policy(client, name, input, options \\ [])

Deletes the access point policy for the specified access point.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to DeleteAccessPointPolicy:

PutAccessPointPolicy GetAccessPointPolicy

delete_bucket(client, bucket, input, options \\ [])

This API operation deletes an Amazon S3 on Outposts bucket.

To delete an S3 bucket, see DeleteBucket in the Amazon Simple Storage Service API.

Deletes the Amazon S3 on Outposts bucket. All objects (including all object versions and delete markers) in the bucket must be deleted before the bucket itself can be deleted. For more information, see Using Amazon S3 on Outposts in Amazon Simple Storage Service Developer Guide.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

Related Resources

CreateBucket GetBucket

* DeleteObject

delete_bucket_lifecycle_configuration(client, bucket, input, options \\ [])

This API action deletes an Amazon S3 on Outposts bucket's lifecycle configuration.

To delete an S3 bucket's lifecycle configuration, see DeleteBucketLifecycle in the Amazon Simple Storage Service API.

Deletes the lifecycle configuration from the specified Outposts bucket. Amazon S3 on Outposts removes all the lifecycle configuration rules in the lifecycle subresource associated with the bucket. Your objects never expire, and Amazon S3 on Outposts no longer automatically deletes any objects on the basis of rules contained in the deleted lifecycle configuration. For more information, see Using Amazon S3 on Outposts in Amazon Simple Storage Service Developer Guide.

To use this operation, you must have permission to perform the s3outposts:DeleteLifecycleConfiguration action. By default, the bucket owner has this permission and the Outposts bucket owner can grant this permission to others.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

For more information about object expiration, see Elements to Describe Lifecycle Actions.

Related actions include:

PutBucketLifecycleConfiguration GetBucketLifecycleConfiguration

delete_bucket_policy(client, bucket, input, options \\ [])

This API operation deletes an Amazon S3 on Outposts bucket policy.

To delete an S3 bucket policy, see DeleteBucketPolicy in the Amazon Simple Storage Service API.

This implementation of the DELETE operation uses the policy subresource to delete the policy of a specified Amazon S3 on Outposts bucket. If you are using an identity other than the root user of the AWS account that owns the bucket, the calling identity must have the s3outposts:DeleteBucketPolicy permissions on the specified Outposts bucket and belong to the bucket owner's account to use this operation. For more information, see Using Amazon S3 on Outposts in Amazon Simple Storage Service Developer Guide.

If you don't have DeleteBucketPolicy permissions, Amazon S3 returns a 403 Access Denied error. If you have the correct permissions, but you're not using an identity that belongs to the bucket owner's account, Amazon S3 returns a 405 Method Not Allowed error.

As a security precaution, the root user of the AWS account that owns a bucket can always use this operation, even if the policy explicitly denies the root user the ability to perform this action.

For more information about bucket policies, see Using Bucket Policies and User Policies.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to DeleteBucketPolicy:

GetBucketPolicy PutBucketPolicy

delete_bucket_tagging(client, bucket, input, options \\ [])

This API operation deletes an Amazon S3 on Outposts bucket's tags.

To delete an S3 bucket tags, see DeleteBucketTagging in the Amazon Simple Storage Service API.

Deletes the tags from the Outposts bucket. For more information, see Using Amazon S3 on Outposts in Amazon Simple Storage Service Developer Guide.

To use this operation, you must have permission to perform the PutBucketTagging action. By default, the bucket owner has this permission and can grant this permission to others.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to DeleteBucketTagging:

GetBucketTagging PutBucketTagging

delete_job_tagging(client, job_id, input, options \\ [])

Removes the entire tag set from the specified S3 Batch Operations job.

To use this operation, you must have permission to perform the s3:DeleteJobTagging action. For more information, see Controlling access and labeling jobs using tags in the Amazon Simple Storage Service Developer Guide.

Related actions include:

CreateJob GetJobTagging

* PutJobTagging

delete_public_access_block(client, input, options \\ [])

Removes the PublicAccessBlock configuration for an AWS account.

For more information, see Using Amazon S3 block public access.

Related actions include:

GetPublicAccessBlock PutPublicAccessBlock

describe_job(client, job_id, account_id, options \\ [])

Retrieves the configuration parameters and status for a Batch Operations job.

For more information, see S3 Batch Operations in the Amazon Simple Storage Service Developer Guide.

Related actions include:

CreateJob ListJobs

UpdateJobPriority UpdateJobStatus

get_access_point(client, name, account_id, options \\ [])

Returns configuration information about the specified access point.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to GetAccessPoint:

CreateAccessPoint DeleteAccessPoint

* ListAccessPoints

get_access_point_policy(client, name, account_id, options \\ [])

Returns the access point policy associated with the specified access point.

The following actions are related to GetAccessPointPolicy:

PutAccessPointPolicy DeleteAccessPointPolicy

get_access_point_policy_status(client, name, account_id, options \\ [])

Indicates whether the specified access point currently has a policy that allows public access.

For more information about public access through access points, see Managing Data Access with Amazon S3 Access Points in the Amazon Simple Storage Service Developer Guide.

get_bucket(client, bucket, account_id, options \\ [])

Gets an Amazon S3 on Outposts bucket.

For more information, see Using Amazon S3 on Outposts in the Amazon Simple Storage Service Developer Guide.

The following actions are related to GetBucket for Amazon S3 on Outposts:

PutObject * CreateBucket

* DeleteBucket

get_bucket_lifecycle_configuration(client, bucket, account_id, options \\ [])

This API operation gets an Amazon S3 on Outposts bucket's lifecycle configuration.

To get an S3 bucket's lifecycle configuration, see GetBucketLifecycleConfiguration in the Amazon Simple Storage Service API.

Returns the lifecycle configuration information set on the Outposts bucket. For more information, see Using Amazon S3 on Outposts and for information about lifecycle configuration, see Object Lifecycle Management in Amazon Simple Storage Service Developer Guide.

To use this operation, you must have permission to perform the s3outposts:GetLifecycleConfiguration action. The Outposts bucket owner has this permission, by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

GetBucketLifecycleConfiguration has the following special error:

Error code: NoSuchLifecycleConfiguration

Description: The lifecycle configuration does not exist.

HTTP Status Code: 404 Not Found

SOAP Fault Code Prefix: Client

The following actions are related to GetBucketLifecycleConfiguration:

PutBucketLifecycleConfiguration DeleteBucketLifecycleConfiguration

get_bucket_policy(client, bucket, account_id, options \\ [])

This API action gets a bucket policy for an Amazon S3 on Outposts bucket.

To get a policy for an S3 bucket, see GetBucketPolicy in the Amazon Simple Storage Service API.

Returns the policy of a specified Outposts bucket. For more information, see Using Amazon S3 on Outposts in the Amazon Simple Storage Service Developer Guide.

If you are using an identity other than the root user of the AWS account that owns the bucket, the calling identity must have the GetBucketPolicy permissions on the specified bucket and belong to the bucket owner's account in order to use this operation.

If you don't have s3outposts:GetBucketPolicy permissions, Amazon S3 returns a 403 Access Denied error. If you have the correct permissions, but you're not using an identity that belongs to the bucket owner's account, Amazon S3 returns a 405 Method Not Allowed error.

As a security precaution, the root user of the AWS account that owns a bucket can always use this operation, even if the policy explicitly denies the root user the ability to perform this action.

For more information about bucket policies, see Using Bucket Policies and User Policies.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to GetBucketPolicy:

GetObject * PutBucketPolicy

* DeleteBucketPolicy

get_bucket_tagging(client, bucket, account_id, options \\ [])

This API operation gets an Amazon S3 on Outposts bucket's tags.

To get an S3 bucket tags, see GetBucketTagging in the Amazon Simple Storage Service API.

Returns the tag set associated with the Outposts bucket. For more information, see Using Amazon S3 on Outposts in the Amazon Simple Storage Service Developer Guide.

To use this operation, you must have permission to perform the GetBucketTagging action. By default, the bucket owner has this permission and can grant this permission to others.

GetBucketTagging has the following special error:

Error code: NoSuchTagSetError

Description: There is no tag set associated with the bucket.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to GetBucketTagging:

PutBucketTagging DeleteBucketTagging

get_job_tagging(client, job_id, account_id, options \\ [])

Returns the tags on an S3 Batch Operations job.

To use this operation, you must have permission to perform the s3:GetJobTagging action. For more information, see Controlling access and labeling jobs using tags in the Amazon Simple Storage Service Developer Guide.

Related actions include:

CreateJob PutJobTagging

* DeleteJobTagging

get_public_access_block(client, account_id, options \\ [])

Retrieves the PublicAccessBlock configuration for an AWS account.

For more information, see Using Amazon S3 block public access.

Related actions include:

DeletePublicAccessBlock PutPublicAccessBlock

list_access_points(client, bucket \\ nil, max_results \\ nil, next_token \\ nil, account_id, options \\ [])

Returns a list of the access points currently associated with the specified bucket.

You can retrieve up to 1000 access points per call. If the specified bucket has more than 1,000 access points (or the number specified in maxResults, whichever is less), the response will include a continuation token that you can use to list the additional access points.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to ListAccessPoints:

CreateAccessPoint DeleteAccessPoint

* GetAccessPoint

list_jobs(client, job_statuses \\ nil, max_results \\ nil, next_token \\ nil, account_id, options \\ [])

Lists current S3 Batch Operations jobs and jobs that have ended within the last 30 days for the AWS account making the request.

For more information, see S3 Batch Operations in the Amazon Simple Storage Service Developer Guide.

Related actions include:

CreateJob DescribeJob

UpdateJobPriority UpdateJobStatus

list_regional_buckets(client, max_results \\ nil, next_token \\ nil, account_id, outpost_id \\ nil, options \\ [])

Returns a list of all Outposts buckets in an Outposts that are owned by the authenticated sender of the request.

For more information, see Using Amazon S3 on Outposts in the Amazon Simple Storage Service Developer Guide.

For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and outpost-id in your API request, see the Example section below.

put_access_point_policy(client, name, input, options \\ [])

Associates an access policy with the specified access point.

Each access point can have only one policy, so a request made to this API replaces any existing policy associated with the specified access point.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to PutAccessPointPolicy:

GetAccessPointPolicy DeleteAccessPointPolicy

put_bucket_lifecycle_configuration(client, bucket, input, options \\ [])

This API action puts a lifecycle configuration to an Amazon S3 on Outposts bucket.

To put a lifecycle configuration to an S3 bucket, see PutBucketLifecycleConfiguration in the Amazon Simple Storage Service API.

Creates a new lifecycle configuration for the Outposts bucket or replaces an existing lifecycle configuration. Outposts buckets can only support a lifecycle that deletes objects after a certain period of time. For more information, see Managing Lifecycle Permissions for Amazon S3 on Outposts.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to PutBucketLifecycleConfiguration:

GetBucketLifecycleConfiguration DeleteBucketLifecycleConfiguration

put_bucket_policy(client, bucket, input, options \\ [])

This API action puts a bucket policy to an Amazon S3 on Outposts bucket.

To put a policy on an S3 bucket, see PutBucketPolicy in the Amazon Simple Storage Service API.

Applies an Amazon S3 bucket policy to an Outposts bucket. For more information, see Using Amazon S3 on Outposts in the Amazon Simple Storage Service Developer Guide.

If you are using an identity other than the root user of the AWS account that owns the Outposts bucket, the calling identity must have the PutBucketPolicy permissions on the specified Outposts bucket and belong to the bucket owner's account in order to use this operation.

If you don't have PutBucketPolicy permissions, Amazon S3 returns a 403 Access Denied error. If you have the correct permissions, but you're not using an identity that belongs to the bucket owner's account, Amazon S3 returns a 405 Method Not Allowed error.

As a security precaution, the root user of the AWS account that owns a bucket can always use this operation, even if the policy explicitly denies the root user the ability to perform this action.

For more information about bucket policies, see Using Bucket Policies and User Policies.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to PutBucketPolicy:

GetBucketPolicy DeleteBucketPolicy

put_bucket_tagging(client, bucket, input, options \\ [])

This API action puts tags on an Amazon S3 on Outposts bucket.

To put tags on an S3 bucket, see PutBucketTagging in the Amazon Simple Storage Service API.

Sets the tags for an Outposts bucket. For more information, see Using Amazon S3 on Outposts in the Amazon Simple Storage Service Developer Guide.

Use tags to organize your AWS bill to reflect your own cost structure. To do this, sign up to get your AWS account bill with tag key values included. Then, to see the cost of combined resources, organize your billing information according to resources with the same tag key values. For example, you can tag several resources with a specific application name, and then organize your billing information to see the total cost of that application across several services. For more information, see Cost Allocation and Tagging.

Within a bucket, if you add a tag that has the same key as an existing tag, the new value overwrites the old value. For more information, see Using Cost Allocation in Amazon S3 Bucket Tags.

To use this operation, you must have permissions to perform the s3outposts:PutBucketTagging action. The Outposts bucket owner has this permission by default and can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

PutBucketTagging has the following special errors:

Error code: InvalidTagError

Description: The tag provided was not a valid tag. This error can occur if the tag did not pass input validation. For information about tag restrictions, see User-Defined Tag Restrictions and AWS-Generated Cost Allocation Tag Restrictions.

Error code: MalformedXMLError

Description: The XML provided does not match the schema.

Error code: OperationAbortedError

Description: A conflicting conditional operation is currently in progress against this resource. Try again.

Error code: InternalError

Description: The service was unable to apply the provided tag to the bucket.

All Amazon S3 on Outposts REST API requests for this action require an additional parameter of outpost-id to be passed with the request and an S3 on Outposts endpoint hostname prefix instead of s3-control. For an example of the request syntax for Amazon S3 on Outposts that uses the S3 on Outposts endpoint hostname prefix and the outpost-id derived using the access point ARN, see the Example section below.

The following actions are related to PutBucketTagging:

GetBucketTagging DeleteBucketTagging

put_job_tagging(client, job_id, input, options \\ [])

Sets the supplied tag-set on an S3 Batch Operations job.

A tag is a key-value pair. You can associate S3 Batch Operations tags with any job by sending a PUT request against the tagging subresource that is associated with the job. To modify the existing tag set, you can either replace the existing tag set entirely, or make changes within the existing tag set by retrieving the existing tag set using GetJobTagging, modify that tag set, and use this API action to replace the tag set with the one you modified. For more information, see Controlling access and labeling jobs using tags in the Amazon Simple Storage Service Developer Guide.

If you send this request with an empty tag set, Amazon S3 deletes the existing tag set on the Batch Operations job. If you use this method, you are charged for a Tier 1 Request (PUT). For more information, see Amazon S3 pricing. * For deleting existing tags for your Batch Operations job, a DeleteJobTagging request is preferred because it achieves the same result without incurring charges.

A few things to consider about using tags:

Amazon S3 limits the maximum number of tags to 50 tags per job.

You can associate up to 50 tags with a job as long as they have unique tag keys.

A tag key can be up to 128 Unicode characters in length, and tag values can be up to 256 Unicode characters in length.

The key and values are case sensitive.

For tagging-related restrictions related to characters and encodings, see User-Defined Tag Restrictions in the AWS Billing and Cost Management User Guide.

To use this operation, you must have permission to perform the s3:PutJobTagging action.

Related actions include:

CreatJob GetJobTagging

* DeleteJobTagging

put_public_access_block(client, input, options \\ [])

Creates or modifies the PublicAccessBlock configuration for an AWS account.

For more information, see Using Amazon S3 block public access.

Related actions include:

GetPublicAccessBlock DeletePublicAccessBlock

update_job_priority(client, job_id, input, options \\ [])

Updates an existing S3 Batch Operations job's priority.

For more information, see S3 Batch Operations in the Amazon Simple Storage Service Developer Guide.

Related actions include:

CreateJob ListJobs

DescribeJob UpdateJobStatus

update_job_status(client, job_id, input, options \\ [])

Updates the status for the specified job.

Use this operation to confirm that you want to run a job or to cancel an existing job. For more information, see S3 Batch Operations in the Amazon Simple Storage Service Developer Guide.

Related actions include:

CreateJob ListJobs

DescribeJob UpdateJobStatus