aws-elixir v0.6.0 AWS.S3Outposts

Amazon S3 on Outposts provides access to S3 on Outposts operations.

Functions

create_endpoint(client, input, options \\ [])

S3 on Outposts access points simplify managing data access at scale for shared datasets in Amazon S3 on Outposts.

S3 on Outposts uses endpoints to connect to Outposts buckets so that you can perform actions within your virtual private cloud (VPC).

This action creates an endpoint and associates it with the specified Outpost.

Related actions include:

DeleteEndpoint ListEndpoints

delete_endpoint(client, input, options \\ [])

S3 on Outposts access points simplify managing data access at scale for shared datasets in Amazon S3 on Outposts.

S3 on Outposts uses endpoints to connect to Outposts buckets so that you can perform actions within your virtual private cloud (VPC).

This action deletes an endpoint.

Related actions include:

CreateEndpoint ListEndpoints

list_endpoints(client, max_results \\ nil, next_token \\ nil, options \\ [])

S3 on Outposts access points simplify managing data access at scale for shared datasets in Amazon S3 on Outposts.

S3 on Outposts uses endpoints to connect to Outposts buckets so that you can perform actions within your virtual private cloud (VPC).

This action lists endpoints associated with the Outpost.

Related actions include:

CreateEndpoint DeleteEndpoint