aws-elixir v0.6.0 AWS.SESv2

Amazon SES API v2

Welcome to the Amazon SES API v2 Reference.

This guide provides information about the Amazon SES API v2, including supported operations, data types, parameters, and schemas.

Amazon SES is an AWS service that you can use to send email messages to your customers.

If you're new to Amazon SES API v2, you might find it helpful to also review the Amazon Simple Email Service Developer Guide. The Amazon SES Developer Guide provides information and code samples that demonstrate how to use Amazon SES API v2 features programmatically.

The Amazon SES API v2 is available in several AWS Regions and it provides an endpoint for each of these Regions. For a list of all the Regions and endpoints where the API is currently available, see AWS Service Endpoints in the Amazon Web Services General Reference. To learn more about AWS Regions, see Managing AWS Regions in the Amazon Web Services General Reference.

In each Region, AWS maintains multiple Availability Zones. These Availability Zones are physically isolated from each other, but are united by private, low-latency, high-throughput, and highly redundant network connections. These Availability Zones enable us to provide very high levels of availability and redundancy, while also minimizing latency. To learn more about the number of Availability Zones that are available in each Region, see AWS Global Infrastructure.

Functions

create_configuration_set(client, input, options \\ [])

Create a configuration set.

Configuration sets are groups of rules that you can apply to the emails that you send. You apply a configuration set to an email by specifying the name of the configuration set when you call the Amazon SES API v2. When you apply a configuration set to an email, all of the rules in that configuration set are applied to the email.

create_configuration_set_event_destination(client, configuration_set_name, input, options \\ [])

Create an event destination.

Events include message sends, deliveries, opens, clicks, bounces, and complaints. Event destinations are places that you can send information about these events to. For example, you can send event data to Amazon SNS to receive notifications when you receive bounces or complaints, or you can use Amazon Kinesis Data Firehose to stream data to Amazon S3 for long-term storage.

A single configuration set can include more than one event destination.

create_custom_verification_email_template(client, input, options \\ [])

Creates a new custom verification email template.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

create_dedicated_ip_pool(client, input, options \\ [])

Create a new pool of dedicated IP addresses.

A pool can include one or more dedicated IP addresses that are associated with your AWS account. You can associate a pool with a configuration set. When you send an email that uses that configuration set, the message is sent from one of the addresses in the associated pool.

create_deliverability_test_report(client, input, options \\ [])

Create a new predictive inbox placement test.

Predictive inbox placement tests can help you predict how your messages will be handled by various email providers around the world. When you perform a predictive inbox placement test, you provide a sample message that contains the content that you plan to send to your customers. Amazon SES then sends that message to special email addresses spread across several major email providers. After about 24 hours, the test is complete, and you can use the GetDeliverabilityTestReport operation to view the results of the test.

create_email_identity(client, input, options \\ [])

Starts the process of verifying an email identity.

An identity is an email address or domain that you use when you send email. Before you can use an identity to send email, you first have to verify it. By verifying an identity, you demonstrate that you're the owner of the identity, and that you've given Amazon SES API v2 permission to send email from the identity.

When you verify an email address, Amazon SES sends an email to the address. Your email address is verified as soon as you follow the link in the verification email.

When you verify a domain without specifying the DkimSigningAttributes object, this operation provides a set of DKIM tokens. You can convert these tokens into CNAME records, which you then add to the DNS configuration for your domain. Your domain is verified when Amazon SES detects these records in the DNS configuration for your domain. This verification method is known as Easy DKIM.

Alternatively, you can perform the verification process by providing your own public-private key pair. This verification method is known as Bring Your Own DKIM (BYODKIM). To use BYODKIM, your call to the CreateEmailIdentity operation has to include the DkimSigningAttributes object. When you specify this object, you provide a selector (a component of the DNS record name that identifies the public key that you want to use for DKIM authentication) and a private key.

create_email_identity_policy(client, email_identity, policy_name, input, options \\ [])

Creates the specified sending authorization policy for the given identity (an email address or a domain).

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

create_email_template(client, input, options \\ [])

Creates an email template.

Email templates enable you to send personalized email to one or more destinations in a single API operation. For more information, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

create_import_job(client, input, options \\ [])

Creates an import job for a data destination.

delete_configuration_set(client, configuration_set_name, input, options \\ [])

Delete an existing configuration set.

Configuration sets are groups of rules that you can apply to the emails you send. You apply a configuration set to an email by including a reference to the configuration set in the headers of the email. When you apply a configuration set to an email, all of the rules in that configuration set are applied to the email.

delete_configuration_set_event_destination(client, configuration_set_name, event_destination_name, input, options \\ [])

Delete an event destination.

Events include message sends, deliveries, opens, clicks, bounces, and complaints. Event destinations are places that you can send information about these events to. For example, you can send event data to Amazon SNS to receive notifications when you receive bounces or complaints, or you can use Amazon Kinesis Data Firehose to stream data to Amazon S3 for long-term storage.

delete_custom_verification_email_template(client, template_name, input, options \\ [])

Deletes an existing custom verification email template.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

delete_dedicated_ip_pool(client, pool_name, input, options \\ [])

Delete a dedicated IP pool.

delete_email_identity(client, email_identity, input, options \\ [])

Deletes an email identity.

An identity can be either an email address or a domain name.

delete_email_identity_policy(client, email_identity, policy_name, input, options \\ [])

Deletes the specified sending authorization policy for the given identity (an email address or a domain).

This API returns successfully even if a policy with the specified name does not exist.

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

delete_email_template(client, template_name, input, options \\ [])

Deletes an email template.

You can execute this operation no more than once per second.

delete_suppressed_destination(client, email_address, input, options \\ [])

Removes an email address from the suppression list for your account.

get_account(client, options \\ [])

Obtain information about the email-sending status and capabilities of your Amazon SES account in the current AWS Region.

get_blacklist_reports(client, blacklist_item_names, options \\ [])

Retrieve a list of the blacklists that your dedicated IP addresses appear on.

get_configuration_set(client, configuration_set_name, options \\ [])

Get information about an existing configuration set, including the dedicated IP pool that it's associated with, whether or not it's enabled for sending email, and more.

Configuration sets are groups of rules that you can apply to the emails you send. You apply a configuration set to an email by including a reference to the configuration set in the headers of the email. When you apply a configuration set to an email, all of the rules in that configuration set are applied to the email.

get_configuration_set_event_destinations(client, configuration_set_name, options \\ [])

Retrieve a list of event destinations that are associated with a configuration set.

Events include message sends, deliveries, opens, clicks, bounces, and complaints. Event destinations are places that you can send information about these events to. For example, you can send event data to Amazon SNS to receive notifications when you receive bounces or complaints, or you can use Amazon Kinesis Data Firehose to stream data to Amazon S3 for long-term storage.

get_custom_verification_email_template(client, template_name, options \\ [])

Returns the custom email verification template for the template name you specify.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

get_dedicated_ip(client, ip, options \\ [])

Get information about a dedicated IP address, including the name of the dedicated IP pool that it's associated with, as well information about the automatic warm-up process for the address.

get_dedicated_ips(client, next_token \\ nil, page_size \\ nil, pool_name \\ nil, options \\ [])

List the dedicated IP addresses that are associated with your AWS account.

get_deliverability_dashboard_options(client, options \\ [])

Retrieve information about the status of the Deliverability dashboard for your account.

When the Deliverability dashboard is enabled, you gain access to reputation, deliverability, and other metrics for the domains that you use to send email. You also gain the ability to perform predictive inbox placement tests.

When you use the Deliverability dashboard, you pay a monthly subscription charge, in addition to any other fees that you accrue by using Amazon SES and other AWS services. For more information about the features and cost of a Deliverability dashboard subscription, see Amazon SES Pricing.

get_deliverability_test_report(client, report_id, options \\ [])

Retrieve the results of a predictive inbox placement test.

get_domain_deliverability_campaign(client, campaign_id, options \\ [])

Retrieve all the deliverability data for a specific campaign.

This data is available for a campaign only if the campaign sent email by using a domain that the Deliverability dashboard is enabled for.

get_domain_statistics_report(client, domain, end_date, start_date, options \\ [])

Retrieve inbox placement and engagement rates for the domains that you use to send email.

get_email_identity(client, email_identity, options \\ [])

Provides information about a specific identity, including the identity's verification status, sending authorization policies, its DKIM authentication status, and its custom Mail-From settings.

get_email_identity_policies(client, email_identity, options \\ [])

Returns the requested sending authorization policies for the given identity (an email address or a domain).

The policies are returned as a map of policy names to policy contents. You can retrieve a maximum of 20 policies at a time.

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

get_email_template(client, template_name, options \\ [])

Displays the template object (which includes the subject line, HTML part and text part) for the template you specify.

You can execute this operation no more than once per second.

get_import_job(client, job_id, options \\ [])

Provides information about an import job.

get_suppressed_destination(client, email_address, options \\ [])

Retrieves information about a specific email address that's on the suppression list for your account.

list_configuration_sets(client, next_token \\ nil, page_size \\ nil, options \\ [])

List all of the configuration sets associated with your account in the current region.

Configuration sets are groups of rules that you can apply to the emails you send. You apply a configuration set to an email by including a reference to the configuration set in the headers of the email. When you apply a configuration set to an email, all of the rules in that configuration set are applied to the email.

list_custom_verification_email_templates(client, next_token \\ nil, page_size \\ nil, options \\ [])

Lists the existing custom verification email templates for your account in the current AWS Region.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

list_dedicated_ip_pools(client, next_token \\ nil, page_size \\ nil, options \\ [])

List all of the dedicated IP pools that exist in your AWS account in the current Region.

list_deliverability_test_reports(client, next_token \\ nil, page_size \\ nil, options \\ [])

Show a list of the predictive inbox placement tests that you've performed, regardless of their statuses.

For predictive inbox placement tests that are complete, you can use the GetDeliverabilityTestReport operation to view the results.

list_domain_deliverability_campaigns(client, subscribed_domain, end_date, next_token \\ nil, page_size \\ nil, start_date, options \\ [])

Retrieve deliverability data for all the campaigns that used a specific domain to send email during a specified time range.

This data is available for a domain only if you enabled the Deliverability dashboard for the domain.

list_email_identities(client, next_token \\ nil, page_size \\ nil, options \\ [])

Returns a list of all of the email identities that are associated with your AWS account.

An identity can be either an email address or a domain. This operation returns identities that are verified as well as those that aren't. This operation returns identities that are associated with Amazon SES and Amazon Pinpoint.

list_email_templates(client, next_token \\ nil, page_size \\ nil, options \\ [])

Lists the email templates present in your Amazon SES account in the current AWS Region.

You can execute this operation no more than once per second.

list_import_jobs(client, next_token \\ nil, page_size \\ nil, options \\ [])

Lists all of the import jobs.

list_suppressed_destinations(client, end_date \\ nil, next_token \\ nil, page_size \\ nil, reasons \\ nil, start_date \\ nil, options \\ [])

Retrieves a list of email addresses that are on the suppression list for your account.

list_tags_for_resource(client, resource_arn, options \\ [])

Retrieve a list of the tags (keys and values) that are associated with a specified resource.

A tag is a label that you optionally define and associate with a resource. Each tag consists of a required tag key and an optional associated tag value. A tag key is a general label that acts as a category for more specific tag values. A tag value acts as a descriptor within a tag key.

put_account_dedicated_ip_warmup_attributes(client, input, options \\ [])

Enable or disable the automatic warm-up feature for dedicated IP addresses.

put_account_details(client, input, options \\ [])

Update your Amazon SES account details.

put_account_sending_attributes(client, input, options \\ [])

Enable or disable the ability of your account to send email.

put_account_suppression_attributes(client, input, options \\ [])

Change the settings for the account-level suppression list.

put_configuration_set_delivery_options(client, configuration_set_name, input, options \\ [])

Associate a configuration set with a dedicated IP pool.

You can use dedicated IP pools to create groups of dedicated IP addresses for sending specific types of email.

put_configuration_set_reputation_options(client, configuration_set_name, input, options \\ [])

Enable or disable collection of reputation metrics for emails that you send using a particular configuration set in a specific AWS Region.

put_configuration_set_sending_options(client, configuration_set_name, input, options \\ [])

Enable or disable email sending for messages that use a particular configuration set in a specific AWS Region.

put_configuration_set_suppression_options(client, configuration_set_name, input, options \\ [])

Specify the account suppression list preferences for a configuration set.

put_configuration_set_tracking_options(client, configuration_set_name, input, options \\ [])

Specify a custom domain to use for open and click tracking elements in email that you send.

put_dedicated_ip_in_pool(client, ip, input, options \\ [])

Move a dedicated IP address to an existing dedicated IP pool.

The dedicated IP address that you specify must already exist, and must be associated with your AWS account.

The dedicated IP pool you specify must already exist. You can create a new pool by using the CreateDedicatedIpPool operation.

put_dedicated_ip_warmup_attributes(client, ip, input, options \\ [])

put_deliverability_dashboard_option(client, input, options \\ [])

Enable or disable the Deliverability dashboard.

When you enable the Deliverability dashboard, you gain access to reputation, deliverability, and other metrics for the domains that you use to send email. You also gain the ability to perform predictive inbox placement tests.

When you use the Deliverability dashboard, you pay a monthly subscription charge, in addition to any other fees that you accrue by using Amazon SES and other AWS services. For more information about the features and cost of a Deliverability dashboard subscription, see Amazon SES Pricing.

put_email_identity_dkim_attributes(client, email_identity, input, options \\ [])

Used to enable or disable DKIM authentication for an email identity.

put_email_identity_dkim_signing_attributes(client, email_identity, input, options \\ [])

Used to configure or change the DKIM authentication settings for an email domain identity.

You can use this operation to do any of the following:

Update the signing attributes for an identity that uses Bring Your Own DKIM (BYODKIM).

Change from using no DKIM authentication to using Easy DKIM.

Change from using no DKIM authentication to using BYODKIM.

Change from using Easy DKIM to using BYODKIM.

Change from using BYODKIM to using Easy DKIM.

put_email_identity_feedback_attributes(client, email_identity, input, options \\ [])

Used to enable or disable feedback forwarding for an identity.

This setting determines what happens when an identity is used to send an email that results in a bounce or complaint event.

If the value is true, you receive email notifications when bounce or complaint events occur. These notifications are sent to the address that you specified in the Return-Path header of the original email.

You're required to have a method of tracking bounces and complaints. If you haven't set up another mechanism for receiving bounce or complaint notifications (for example, by setting up an event destination), you receive an email notification when these events occur (even if this setting is disabled).

put_email_identity_mail_from_attributes(client, email_identity, input, options \\ [])

Used to enable or disable the custom Mail-From domain configuration for an email identity.

put_suppressed_destination(client, input, options \\ [])

Adds an email address to the suppression list for your account.

send_bulk_email(client, input, options \\ [])

Composes an email message to multiple destinations.

send_custom_verification_email(client, input, options \\ [])

Adds an email address to the list of identities for your Amazon SES account in the current AWS Region and attempts to verify it.

As a result of executing this operation, a customized verification email is sent to the specified address.

To use this operation, you must first create a custom verification email template. For more information about creating and using custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

send_email(client, input, options \\ [])

Sends an email message.

You can use the Amazon SES API v2 to send two types of messages:

Simple – A standard email message. When you create this type of message, you specify the sender, the recipient, and the message body, and Amazon SES assembles the message for you.

Raw – A raw, MIME-formatted email message. When you send this type of email, you have to specify all of the message headers, as well as the message body. You can use this message type to send messages that contain attachments. The message that you specify has to be a valid MIME message.

Templated – A message that contains personalization tags. When you send this type of email, Amazon SES API v2 automatically replaces the tags with values that you specify.

tag_resource(client, input, options \\ [])

Add one or more tags (keys and values) to a specified resource.

A tag is a label that you optionally define and associate with a resource. Tags can help you categorize and manage resources in different ways, such as by purpose, owner, environment, or other criteria. A resource can have as many as 50 tags.

Each tag consists of a required tag key and an associated tag value, both of which you define. A tag key is a general label that acts as a category for more specific tag values. A tag value acts as a descriptor within a tag key.

test_render_email_template(client, template_name, input, options \\ [])

Creates a preview of the MIME content of an email when provided with a template and a set of replacement data.

You can execute this operation no more than once per second.

untag_resource(client, input, options \\ [])

Remove one or more tags (keys and values) from a specified resource.

update_configuration_set_event_destination(client, configuration_set_name, event_destination_name, input, options \\ [])

Update the configuration of an event destination for a configuration set.

Events include message sends, deliveries, opens, clicks, bounces, and complaints. Event destinations are places that you can send information about these events to. For example, you can send event data to Amazon SNS to receive notifications when you receive bounces or complaints, or you can use Amazon Kinesis Data Firehose to stream data to Amazon S3 for long-term storage.

update_custom_verification_email_template(client, template_name, input, options \\ [])

Updates an existing custom verification email template.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

update_email_identity_policy(client, email_identity, policy_name, input, options \\ [])

Updates the specified sending authorization policy for the given identity (an email address or a domain).

This API returns successfully even if a policy with the specified name does not exist.

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

update_email_template(client, template_name, input, options \\ [])

Updates an email template.

Email templates enable you to send personalized email to one or more destinations in a single API operation. For more information, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.