aws-elixir v0.6.0 AWS.SMS

AWS Server Migration Service

AWS Server Migration Service (AWS SMS) makes it easier and faster for you to migrate your on-premises workloads to AWS.

To learn more about AWS SMS, see the following resources:

AWS Server Migration Service product page

AWS Server Migration Service User Guide

Functions

create_app(client, input, options \\ [])

Creates an application.

An application consists of one or more server groups. Each server group contain one or more servers.

create_replication_job(client, input, options \\ [])

Creates a replication job.

The replication job schedules periodic replication runs to replicate your server to AWS. Each replication run creates an Amazon Machine Image (AMI).

delete_app(client, input, options \\ [])

Deletes the specified application.

Optionally deletes the launched stack associated with the application and all AWS SMS replication jobs for servers in the application.

delete_app_launch_configuration(client, input, options \\ [])

Deletes the launch configuration for the specified application.

delete_app_replication_configuration(client, input, options \\ [])

Deletes the replication configuration for the specified application.

delete_app_validation_configuration(client, input, options \\ [])

Deletes the validation configuration for the specified application.

delete_replication_job(client, input, options \\ [])

Deletes the specified replication job.

After you delete a replication job, there are no further replication runs. AWS deletes the contents of the Amazon S3 bucket used to store AWS SMS artifacts. The AMIs created by the replication runs are not deleted.

delete_server_catalog(client, input, options \\ [])

Deletes all servers from your server catalog.

disassociate_connector(client, input, options \\ [])

Disassociates the specified connector from AWS SMS.

After you disassociate a connector, it is no longer available to support replication jobs.

generate_change_set(client, input, options \\ [])

Generates a target change set for a currently launched stack and writes it to an Amazon S3 object in the customer’s Amazon S3 bucket.

generate_template(client, input, options \\ [])

Generates an AWS CloudFormation template based on the current launch configuration and writes it to an Amazon S3 object in the customer’s Amazon S3 bucket.

get_app(client, input, options \\ [])

Retrieve information about the specified application.

get_app_launch_configuration(client, input, options \\ [])

Retrieves the application launch configuration associated with the specified application.

get_app_replication_configuration(client, input, options \\ [])

Retrieves the application replication configuration associated with the specified application.

get_app_validation_configuration(client, input, options \\ [])

Retrieves information about a configuration for validating an application.

get_app_validation_output(client, input, options \\ [])

Retrieves output from validating an application.

get_connectors(client, input, options \\ [])

Describes the connectors registered with the AWS SMS.

get_replication_jobs(client, input, options \\ [])

Describes the specified replication job or all of your replication jobs.

get_replication_runs(client, input, options \\ [])

Describes the replication runs for the specified replication job.

get_servers(client, input, options \\ [])

Describes the servers in your server catalog.

Before you can describe your servers, you must import them using ImportServerCatalog.

import_app_catalog(client, input, options \\ [])

Allows application import from AWS Migration Hub.

import_server_catalog(client, input, options \\ [])

Gathers a complete list of on-premises servers.

Connectors must be installed and monitoring all servers to import.

This call returns immediately, but might take additional time to retrieve all the servers.

launch_app(client, input, options \\ [])

Launches the specified application as a stack in AWS CloudFormation.

list_apps(client, input, options \\ [])

Retrieves summaries for all applications.

notify_app_validation_output(client, input, options \\ [])

Provides information to AWS SMS about whether application validation is successful.

put_app_launch_configuration(client, input, options \\ [])

Creates or updates the launch configuration for the specified application.

put_app_replication_configuration(client, input, options \\ [])

Creates or updates the replication configuration for the specified application.

put_app_validation_configuration(client, input, options \\ [])

Creates or updates a validation configuration for the specified application.

start_app_replication(client, input, options \\ [])

Starts replicating the specified application by creating replication jobs for each server in the application.

start_on_demand_app_replication(client, input, options \\ [])

Starts an on-demand replication run for the specified application.

start_on_demand_replication_run(client, input, options \\ [])

Starts an on-demand replication run for the specified replication job.

This replication run starts immediately. This replication run is in addition to the ones already scheduled.

There is a limit on the number of on-demand replications runs that you can request in a 24-hour period.

stop_app_replication(client, input, options \\ [])

Stops replicating the specified application by deleting the replication job for each server in the application.

terminate_app(client, input, options \\ [])

Terminates the stack for the specified application.

update_app(client, input, options \\ [])

Updates the specified application.

update_replication_job(client, input, options \\ [])

Updates the specified settings for the specified replication job.