aws-elixir v0.6.0 AWS.SSM

AWS Systems Manager

AWS Systems Manager is a collection of capabilities that helps you automate management tasks such as collecting system inventory, applying operating system (OS) patches, automating the creation of Amazon Machine Images (AMIs), and configuring operating systems (OSs) and applications at scale.

Systems Manager lets you remotely and securely manage the configuration of your managed instances. A managed instance is any Amazon Elastic Compute Cloud instance (EC2 instance), or any on-premises server or virtual machine (VM) in your hybrid environment that has been configured for Systems Manager.

This reference is intended to be used with the AWS Systems Manager User Guide.

To get started, verify prerequisites and configure managed instances. For more information, see Setting up AWS Systems Manager in the AWS Systems Manager User Guide.

For information about other API actions you can perform on EC2 instances, see the Amazon EC2 API Reference. For information about how to use a Query API, see Making API requests.

Functions

add_tags_to_resource(client, input, options \\ [])

Adds or overwrites one or more tags for the specified resource.

Tags are metadata that you can assign to your documents, managed instances, maintenance windows, Parameter Store parameters, and patch baselines. Tags enable you to categorize your resources in different ways, for example, by purpose, owner, or environment. Each tag consists of a key and an optional value, both of which you define. For example, you could define a set of tags for your account's managed instances that helps you track each instance's owner and stack level. For example: Key=Owner and Value=DbAdmin, SysAdmin, or Dev. Or Key=Stack and Value=Production, Pre-Production, or Test.

Each resource can have a maximum of 50 tags.

We recommend that you devise a set of tag keys that meets your needs for each resource type. Using a consistent set of tag keys makes it easier for you to manage your resources. You can search and filter the resources based on the tags you add. Tags don't have any semantic meaning to and are interpreted strictly as a string of characters.

For more information about using tags with EC2 instances, see Tagging your Amazon EC2 resources in the Amazon EC2 User Guide.

cancel_command(client, input, options \\ [])

Attempts to cancel the command specified by the Command ID.

There is no guarantee that the command will be terminated and the underlying process stopped.

cancel_maintenance_window_execution(client, input, options \\ [])

Stops a maintenance window execution that is already in progress and cancels any tasks in the window that have not already starting running.

(Tasks already in progress will continue to completion.)

create_activation(client, input, options \\ [])

Generates an activation code and activation ID you can use to register your on-premises server or virtual machine (VM) with Systems Manager.

Registering these machines with Systems Manager makes it possible to manage them using Systems Manager capabilities. You use the activation code and ID when installing SSM Agent on machines in your hybrid environment. For more information about requirements for managing on-premises instances and VMs using Systems Manager, see Setting up AWS Systems Manager for hybrid environments in the AWS Systems Manager User Guide.

On-premises servers or VMs that are registered with Systems Manager and EC2 instances that you manage with Systems Manager are all called managed instances.

create_association(client, input, options \\ [])

A State Manager association defines the state that you want to maintain on your instances.

For example, an association can specify that anti-virus software must be installed and running on your instances, or that certain ports must be closed. For static targets, the association specifies a schedule for when the configuration is reapplied. For dynamic targets, such as an AWS Resource Group or an AWS Autoscaling Group, State Manager applies the configuration when new instances are added to the group. The association also specifies actions to take when applying the configuration. For example, an association for anti-virus software might run once a day. If the software is not installed, then State Manager installs it. If the software is installed, but the service is not running, then the association might instruct State Manager to start the service.

create_association_batch(client, input, options \\ [])

Associates the specified Systems Manager document with the specified instances or targets.

When you associate a document with one or more instances using instance IDs or tags, SSM Agent running on the instance processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system returns the AssociationAlreadyExists exception.

create_document(client, input, options \\ [])

Creates a Systems Manager (SSM) document.

An SSM document defines the actions that Systems Manager performs on your managed instances. For more information about SSM documents, including information about supported schemas, features, and syntax, see AWS Systems Manager Documents in the AWS Systems Manager User Guide.

create_maintenance_window(client, input, options \\ [])

Creates a new maintenance window.

The value you specify for Duration determines the specific end time for the maintenance window based on the time it begins. No maintenance window tasks are permitted to start after the resulting endtime minus the number of hours you specify for Cutoff. For example, if the maintenance window starts at 3 PM, the duration is three hours, and the value you specify for Cutoff is one hour, no maintenance window tasks can start after 5 PM.

create_ops_item(client, input, options \\ [])

Creates a new OpsItem.

You must have permission in AWS Identity and Access Management (IAM) to create a new OpsItem. For more information, see Getting started with OpsCenter in the AWS Systems Manager User Guide.

Operations engineers and IT professionals use OpsCenter to view, investigate, and remediate operational issues impacting the performance and health of their AWS resources. For more information, see AWS Systems Manager OpsCenter in the AWS Systems Manager User Guide.

create_patch_baseline(client, input, options \\ [])

Creates a patch baseline.

For information about valid key and value pairs in PatchFilters for each supported operating system type, see PatchFilter.

create_resource_data_sync(client, input, options \\ [])

A resource data sync helps you view data from multiple sources in a single location.

Systems Manager offers two types of resource data sync: SyncToDestination and SyncFromSource.

You can configure Systems Manager Inventory to use the SyncToDestination type to synchronize Inventory data from multiple AWS Regions to a single S3 bucket. For more information, see Configuring Resource Data Sync for Inventory in the AWS Systems Manager User Guide.

You can configure Systems Manager Explorer to use the SyncFromSource type to synchronize operational work items (OpsItems) and operational data (OpsData) from multiple AWS Regions to a single S3 bucket. This type can synchronize OpsItems and OpsData from multiple AWS accounts and Regions or EntireOrganization by using AWS Organizations. For more information, see Setting up Systems Manager Explorer to display data from multiple accounts and Regions in the AWS Systems Manager User Guide.

A resource data sync is an asynchronous operation that returns immediately. After a successful initial sync is completed, the system continuously syncs data. To check the status of a sync, use the ListResourceDataSync.

By default, data is not encrypted in Amazon S3. We strongly recommend that you enable encryption in Amazon S3 to ensure secure data storage. We also recommend that you secure access to the Amazon S3 bucket by creating a restrictive bucket policy.

delete_activation(client, input, options \\ [])

Deletes an activation.

You are not required to delete an activation. If you delete an activation, you can no longer use it to register additional managed instances. Deleting an activation does not de-register managed instances. You must manually de-register managed instances.

delete_association(client, input, options \\ [])

Disassociates the specified Systems Manager document from the specified instance.

When you disassociate a document from an instance, it does not change the configuration of the instance. To change the configuration state of an instance after you disassociate a document, you must create a new document with the desired configuration and associate it with the instance.

delete_document(client, input, options \\ [])

Deletes the Systems Manager document and all instance associations to the document.

Before you delete the document, we recommend that you use DeleteAssociation to disassociate all instances that are associated with the document.

delete_inventory(client, input, options \\ [])

Delete a custom inventory type or the data associated with a custom Inventory type.

Deleting a custom inventory type is also referred to as deleting a custom inventory schema.

delete_maintenance_window(client, input, options \\ [])

Deletes a maintenance window.

delete_parameter(client, input, options \\ [])

Delete a parameter from the system.

delete_parameters(client, input, options \\ [])

Delete a list of parameters.

delete_patch_baseline(client, input, options \\ [])

Deletes a patch baseline.

delete_resource_data_sync(client, input, options \\ [])

Deletes a Resource Data Sync configuration.

After the configuration is deleted, changes to data on managed instances are no longer synced to or from the target. Deleting a sync configuration does not delete data.

deregister_managed_instance(client, input, options \\ [])

Removes the server or virtual machine from the list of registered servers.

You can reregister the instance again at any time. If you don't plan to use Run Command on the server, we suggest uninstalling SSM Agent first.

deregister_patch_baseline_for_patch_group(client, input, options \\ [])

Removes a patch group from a patch baseline.

deregister_target_from_maintenance_window(client, input, options \\ [])

Removes a target from a maintenance window.

deregister_task_from_maintenance_window(client, input, options \\ [])

Removes a task from a maintenance window.

describe_activations(client, input, options \\ [])

Describes details about the activation, such as the date and time the activation was created, its expiration date, the IAM role assigned to the instances in the activation, and the number of instances registered by using this activation.

describe_association(client, input, options \\ [])

Describes the association for the specified target or instance.

If you created the association by using the Targets parameter, then you must retrieve the association by using the association ID. If you created the association by specifying an instance ID and a Systems Manager document, then you retrieve the association by specifying the document name and the instance ID.

describe_association_execution_targets(client, input, options \\ [])

Use this API action to view information about a specific execution of a specific association.

describe_association_executions(client, input, options \\ [])

Use this API action to view all executions for a specific association ID.

describe_automation_executions(client, input, options \\ [])

Provides details about all active and terminated Automation executions.

describe_automation_step_executions(client, input, options \\ [])

Information about all active and terminated step executions in an Automation workflow.

describe_available_patches(client, input, options \\ [])

Lists all patches eligible to be included in a patch baseline.

describe_document(client, input, options \\ [])

Describes the specified Systems Manager document.

describe_document_permission(client, input, options \\ [])

Describes the permissions for a Systems Manager document.

If you created the document, you are the owner. If a document is shared, it can either be shared privately (by specifying a user's AWS account ID) or publicly (All).

describe_effective_instance_associations(client, input, options \\ [])

All associations for the instance(s).

describe_effective_patches_for_patch_baseline(client, input, options \\ [])

Retrieves the current effective patches (the patch and the approval state) for the specified patch baseline.

Note that this API applies only to Windows patch baselines.

describe_instance_associations_status(client, input, options \\ [])

The status of the associations for the instance(s).

describe_instance_information(client, input, options \\ [])

Describes one or more of your instances, including information about the operating system platform, the version of SSM Agent installed on the instance, instance status, and so on.

If you specify one or more instance IDs, it returns information for those instances. If you do not specify instance IDs, it returns information for all your instances. If you specify an instance ID that is not valid or an instance that you do not own, you receive an error.

The IamRole field for this API action is the Amazon Identity and Access Management (IAM) role assigned to on-premises instances. This call does not return the IAM role for EC2 instances.

describe_instance_patch_states(client, input, options \\ [])

Retrieves the high-level patch state of one or more instances.

describe_instance_patch_states_for_patch_group(client, input, options \\ [])

Retrieves the high-level patch state for the instances in the specified patch group.

describe_instance_patches(client, input, options \\ [])

Retrieves information about the patches on the specified instance and their state relative to the patch baseline being used for the instance.

describe_inventory_deletions(client, input, options \\ [])

Describes a specific delete inventory operation.

describe_maintenance_window_execution_task_invocations(client, input, options \\ [])

Retrieves the individual task executions (one per target) for a particular task run as part of a maintenance window execution.

describe_maintenance_window_execution_tasks(client, input, options \\ [])

For a given maintenance window execution, lists the tasks that were run.

describe_maintenance_window_executions(client, input, options \\ [])

Lists the executions of a maintenance window.

This includes information about when the maintenance window was scheduled to be active, and information about tasks registered and run with the maintenance window.

describe_maintenance_window_schedule(client, input, options \\ [])

Retrieves information about upcoming executions of a maintenance window.

describe_maintenance_window_targets(client, input, options \\ [])

Lists the targets registered with the maintenance window.

describe_maintenance_window_tasks(client, input, options \\ [])

Lists the tasks in a maintenance window.

describe_maintenance_windows(client, input, options \\ [])

Retrieves the maintenance windows in an AWS account.

describe_maintenance_windows_for_target(client, input, options \\ [])

Retrieves information about the maintenance window targets or tasks that an instance is associated with.

describe_ops_items(client, input, options \\ [])

Query a set of OpsItems.

You must have permission in AWS Identity and Access Management (IAM) to query a list of OpsItems. For more information, see Getting started with OpsCenter in the AWS Systems Manager User Guide.

Operations engineers and IT professionals use OpsCenter to view, investigate, and remediate operational issues impacting the performance and health of their AWS resources. For more information, see AWS Systems Manager OpsCenter in the AWS Systems Manager User Guide.

describe_parameters(client, input, options \\ [])

Get information about a parameter.

Request results are returned on a best-effort basis. If you specify MaxResults in the request, the response includes information up to the limit specified. The number of items returned, however, can be between zero and the value of MaxResults. If the service reaches an internal limit while processing the results, it stops the operation and returns the matching values up to that point and a NextToken. You can specify the NextToken in a subsequent call to get the next set of results.

describe_patch_baselines(client, input, options \\ [])

Lists the patch baselines in your AWS account.

describe_patch_group_state(client, input, options \\ [])

Returns high-level aggregated patch compliance state for a patch group.

describe_patch_groups(client, input, options \\ [])

Lists all patch groups that have been registered with patch baselines.

describe_patch_properties(client, input, options \\ [])

Lists the properties of available patches organized by product, product family, classification, severity, and other properties of available patches.

You can use the reported properties in the filters you specify in requests for actions such as CreatePatchBaseline, UpdatePatchBaseline, DescribeAvailablePatches, and DescribePatchBaselines.

The following section lists the properties that can be used in filters for each major operating system type:

Definitions

AMAZON_LINUX

Valid properties: PRODUCT, CLASSIFICATION, SEVERITY

AMAZON_LINUX_2

Valid properties: PRODUCT, CLASSIFICATION, SEVERITY

CENTOS

Valid properties: PRODUCT, CLASSIFICATION, SEVERITY

DEBIAN

Valid properties: PRODUCT, PRIORITY

ORACLE_LINUX

Valid properties: PRODUCT, CLASSIFICATION, SEVERITY

REDHAT_ENTERPRISE_LINUX

Valid properties: PRODUCT, CLASSIFICATION, SEVERITY

SUSE

Valid properties: PRODUCT, CLASSIFICATION, SEVERITY

UBUNTU

Valid properties: PRODUCT, PRIORITY

WINDOWS

Valid properties: PRODUCT, PRODUCT_FAMILY, CLASSIFICATION, MSRC_SEVERITY

describe_sessions(client, input, options \\ [])

Retrieves a list of all active sessions (both connected and disconnected) or terminated sessions from the past 30 days.

get_automation_execution(client, input, options \\ [])

Get detailed information about a particular Automation execution.

get_calendar_state(client, input, options \\ [])

Gets the state of the AWS Systems Manager Change Calendar at an optional, specified time.

If you specify a time, GetCalendarState returns the state of the calendar at a specific time, and returns the next time that the Change Calendar state will transition. If you do not specify a time, GetCalendarState assumes the current time. Change Calendar entries have two possible states: OPEN or CLOSED.

If you specify more than one calendar in a request, the command returns the status of OPEN only if all calendars in the request are open. If one or more calendars in the request are closed, the status returned is CLOSED.

For more information about Systems Manager Change Calendar, see AWS Systems Manager Change Calendar in the AWS Systems Manager User Guide.

get_command_invocation(client, input, options \\ [])

Returns detailed information about command execution for an invocation or plugin.

get_connection_status(client, input, options \\ [])

Retrieves the Session Manager connection status for an instance to determine whether it is running and ready to receive Session Manager connections.

get_default_patch_baseline(client, input, options \\ [])

Retrieves the default patch baseline.

Note that Systems Manager supports creating multiple default patch baselines. For example, you can create a default patch baseline for each operating system.

If you do not specify an operating system value, the default patch baseline for Windows is returned.

get_deployable_patch_snapshot_for_instance(client, input, options \\ [])

Retrieves the current snapshot for the patch baseline the instance uses.

This API is primarily used by the AWS-RunPatchBaseline Systems Manager document.

get_document(client, input, options \\ [])

Gets the contents of the specified Systems Manager document.

get_inventory(client, input, options \\ [])

Query inventory information.

get_inventory_schema(client, input, options \\ [])

Return a list of inventory type names for the account, or return a list of attribute names for a specific Inventory item type.

get_maintenance_window(client, input, options \\ [])

Retrieves a maintenance window.

get_maintenance_window_execution(client, input, options \\ [])

Retrieves details about a specific a maintenance window execution.

get_maintenance_window_execution_task(client, input, options \\ [])

Retrieves the details about a specific task run as part of a maintenance window execution.

get_maintenance_window_execution_task_invocation(client, input, options \\ [])

Retrieves information about a specific task running on a specific target.

get_maintenance_window_task(client, input, options \\ [])

Lists the tasks in a maintenance window.

get_ops_item(client, input, options \\ [])

Get information about an OpsItem by using the ID.

You must have permission in AWS Identity and Access Management (IAM) to view information about an OpsItem. For more information, see Getting started with OpsCenter in the AWS Systems Manager User Guide.

Operations engineers and IT professionals use OpsCenter to view, investigate, and remediate operational issues impacting the performance and health of their AWS resources. For more information, see AWS Systems Manager OpsCenter in the AWS Systems Manager User Guide.

get_ops_summary(client, input, options \\ [])

View a summary of OpsItems based on specified filters and aggregators.

get_parameter(client, input, options \\ [])

Get information about a parameter by using the parameter name.

Don't confuse this API action with the GetParameters API action.

get_parameter_history(client, input, options \\ [])

Query a list of all parameters used by the AWS account.

get_parameters(client, input, options \\ [])

Get details of a parameter.

Don't confuse this API action with the GetParameter API action.

get_parameters_by_path(client, input, options \\ [])

Retrieve information about one or more parameters in a specific hierarchy.

Request results are returned on a best-effort basis. If you specify MaxResults in the request, the response includes information up to the limit specified. The number of items returned, however, can be between zero and the value of MaxResults. If the service reaches an internal limit while processing the results, it stops the operation and returns the matching values up to that point and a NextToken. You can specify the NextToken in a subsequent call to get the next set of results.

get_patch_baseline(client, input, options \\ [])

Retrieves information about a patch baseline.

get_patch_baseline_for_patch_group(client, input, options \\ [])

Retrieves the patch baseline that should be used for the specified patch group.

get_service_setting(client, input, options \\ [])

ServiceSetting is an account-level setting for an AWS service.

This setting defines how a user interacts with or uses a service or a feature of a service. For example, if an AWS service charges money to the account based on feature or service usage, then the AWS service team might create a default setting of "false". This means the user can't use this feature unless they change the setting to "true" and intentionally opt in for a paid feature.

Services map a SettingId object to a setting value. AWS services teams define the default value for a SettingId. You can't create a new SettingId, but you can overwrite the default value if you have the ssm:UpdateServiceSetting permission for the setting. Use the UpdateServiceSetting API action to change the default setting. Or use the ResetServiceSetting to change the value back to the original value defined by the AWS service team.

Query the current service setting for the account.

label_parameter_version(client, input, options \\ [])

A parameter label is a user-defined alias to help you manage different versions of a parameter.

When you modify a parameter, Systems Manager automatically saves a new version and increments the version number by one. A label can help you remember the purpose of a parameter when there are multiple versions.

Parameter labels have the following requirements and restrictions.

A version of a parameter can have a maximum of 10 labels.

You can't attach the same label to different versions of the same parameter. For example, if version 1 has the label Production, then you can't attach Production to version 2.

You can move a label from one version of a parameter to another.

You can't create a label when you create a new parameter. You must attach a label to a specific version of a parameter.

You can't delete a parameter label. If you no longer want to use a parameter label, then you must move it to a different version of a parameter.

A label can have a maximum of 100 characters.

Labels can contain letters (case sensitive), numbers, periods (.), hyphens (-), or underscores (_).

Labels can't begin with a number, "aws," or "ssm" (not case sensitive). If a label fails to meet these requirements, then the label is not associated with a parameter and the system displays it in the list of InvalidLabels.

list_association_versions(client, input, options \\ [])

Retrieves all versions of an association for a specific association ID.

list_associations(client, input, options \\ [])

Returns all State Manager associations in the current AWS account and Region.

You can limit the results to a specific State Manager association document or instance by specifying a filter.

list_command_invocations(client, input, options \\ [])

An invocation is copy of a command sent to a specific instance.

A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user runs SendCommand against three instances, then a command invocation is created for each requested instance ID. ListCommandInvocations provide status about command execution.

list_commands(client, input, options \\ [])

Lists the commands requested by users of the AWS account.

list_compliance_items(client, input, options \\ [])

For a specified resource ID, this API action returns a list of compliance statuses for different resource types.

Currently, you can only specify one resource ID per call. List results depend on the criteria specified in the filter.

list_compliance_summaries(client, input, options \\ [])

Returns a summary count of compliant and non-compliant resources for a compliance type.

For example, this call can return State Manager associations, patches, or custom compliance types according to the filter criteria that you specify.

list_document_versions(client, input, options \\ [])

List all versions for a document.

list_documents(client, input, options \\ [])

Returns all Systems Manager (SSM) documents in the current AWS account and Region.

You can limit the results of this request by using a filter.

list_inventory_entries(client, input, options \\ [])

A list of inventory items returned by the request.

list_resource_compliance_summaries(client, input, options \\ [])

Returns a resource-level summary count.

The summary includes information about compliant and non-compliant statuses and detailed compliance-item severity counts, according to the filter criteria you specify.

list_resource_data_sync(client, input, options \\ [])

Lists your resource data sync configurations.

Includes information about the last time a sync attempted to start, the last sync status, and the last time a sync successfully completed.

The number of sync configurations might be too large to return using a single call to ListResourceDataSync. You can limit the number of sync configurations returned by using the MaxResults parameter. To determine whether there are more sync configurations to list, check the value of NextToken in the output. If there are more sync configurations to list, you can request them by specifying the NextToken returned in the call to the parameter of a subsequent call.

list_tags_for_resource(client, input, options \\ [])

Returns a list of the tags assigned to the specified resource.

modify_document_permission(client, input, options \\ [])

Shares a Systems Manager document publicly or privately.

If you share a document privately, you must specify the AWS user account IDs for those people who can use the document. If you share a document publicly, you must specify All as the account ID.

put_compliance_items(client, input, options \\ [])

Registers a compliance type and other compliance details on a designated resource.

This action lets you register custom compliance details with a resource. This call overwrites existing compliance information on the resource, so you must provide a full list of compliance items each time that you send the request.

ComplianceType can be one of the following:

ExecutionId: The execution ID when the patch, association, or custom compliance item was applied.

ExecutionType: Specify patch, association, or Custom:string.

ExecutionTime. The time the patch, association, or custom compliance item was applied to the instance.

Id: The patch, association, or custom compliance ID.

Title: A title.

Status: The status of the compliance item. For example, approved for patches, or Failed for associations.

Severity: A patch severity. For example, critical.

DocumentName: A SSM document name. For example, AWS-RunPatchBaseline.

DocumentVersion: An SSM document version number. For example, 4.

Classification: A patch classification. For example, security updates.

PatchBaselineId: A patch baseline ID.

PatchSeverity: A patch severity. For example, Critical.

PatchState: A patch state. For example, InstancesWithFailedPatches.

PatchGroup: The name of a patch group.

InstalledTime: The time the association, patch, or custom compliance item was applied to the resource. Specify the time by using the following format: yyyy-MM-dd'T'HH:mm:ss'Z'

put_inventory(client, input, options \\ [])

Bulk update custom inventory items on one more instance.

The request adds an inventory item, if it doesn't already exist, or updates an inventory item, if it does exist.

put_parameter(client, input, options \\ [])

Add a parameter to the system.

register_default_patch_baseline(client, input, options \\ [])

Defines the default patch baseline for the relevant operating system.

To reset the AWS predefined patch baseline as the default, specify the full patch baseline ARN as the baseline ID value. For example, for CentOS, specify arn:aws:ssm:us-east-2:733109147000:patchbaseline/pb-0574b43a65ea646ed instead of pb-0574b43a65ea646ed.

register_patch_baseline_for_patch_group(client, input, options \\ [])

Registers a patch baseline for a patch group.

register_target_with_maintenance_window(client, input, options \\ [])

Registers a target with a maintenance window.

register_task_with_maintenance_window(client, input, options \\ [])

Adds a new task to a maintenance window.

remove_tags_from_resource(client, input, options \\ [])

Removes tag keys from the specified resource.

reset_service_setting(client, input, options \\ [])

ServiceSetting is an account-level setting for an AWS service.

This setting defines how a user interacts with or uses a service or a feature of a service. For example, if an AWS service charges money to the account based on feature or service usage, then the AWS service team might create a default setting of "false". This means the user can't use this feature unless they change the setting to "true" and intentionally opt in for a paid feature.

Services map a SettingId object to a setting value. AWS services teams define the default value for a SettingId. You can't create a new SettingId, but you can overwrite the default value if you have the ssm:UpdateServiceSetting permission for the setting. Use the GetServiceSetting API action to view the current value. Use the UpdateServiceSetting API action to change the default setting.

Reset the service setting for the account to the default value as provisioned by the AWS service team.

resume_session(client, input, options \\ [])

Reconnects a session to an instance after it has been disconnected.

Connections can be resumed for disconnected sessions, but not terminated sessions.

This command is primarily for use by client machines to automatically reconnect during intermittent network issues. It is not intended for any other use.

send_automation_signal(client, input, options \\ [])

Sends a signal to an Automation execution to change the current behavior or status of the execution.

send_command(client, input, options \\ [])

Runs commands on one or more managed instances.

start_associations_once(client, input, options \\ [])

Use this API action to run an association immediately and only one time.

This action can be helpful when troubleshooting associations.

start_automation_execution(client, input, options \\ [])

Initiates execution of an Automation document.

start_session(client, input, options \\ [])

Initiates a connection to a target (for example, an instance) for a Session Manager session.

Returns a URL and token that can be used to open a WebSocket connection for sending input and receiving outputs.

AWS CLI usage: start-session is an interactive command that requires the Session Manager plugin to be installed on the client machine making the call. For information, see Install the Session Manager plugin for the AWS CLI in the AWS Systems Manager User Guide.

AWS Tools for PowerShell usage: Start-SSMSession is not currently supported by AWS Tools for PowerShell on Windows local machines.

stop_automation_execution(client, input, options \\ [])

Stop an Automation that is currently running.

terminate_session(client, input, options \\ [])

Permanently ends a session and closes the data connection between the Session Manager client and SSM Agent on the instance.

A terminated session cannot be resumed.

update_association(client, input, options \\ [])

Updates an association.

You can update the association name and version, the document version, schedule, parameters, and Amazon S3 output.

In order to call this API action, your IAM user account, group, or role must be configured with permission to call the DescribeAssociation API action. If you don't have permission to call DescribeAssociation, then you receive the following error: An error occurred (AccessDeniedException) when calling the UpdateAssociation operation: User: <user_arn> is not authorized to perform: ssm:DescribeAssociation on resource: <resource_arn>

When you update an association, the association immediately runs against the specified targets.

update_association_status(client, input, options \\ [])

Updates the status of the Systems Manager document associated with the specified instance.

update_document(client, input, options \\ [])

Updates one or more values for an SSM document.

update_document_default_version(client, input, options \\ [])

Set the default version of a document.

update_maintenance_window(client, input, options \\ [])

Updates an existing maintenance window.

Only specified parameters are modified.

The value you specify for Duration determines the specific end time for the maintenance window based on the time it begins. No maintenance window tasks are permitted to start after the resulting endtime minus the number of hours you specify for Cutoff. For example, if the maintenance window starts at 3 PM, the duration is three hours, and the value you specify for Cutoff is one hour, no maintenance window tasks can start after 5 PM.

update_maintenance_window_target(client, input, options \\ [])

Modifies the target of an existing maintenance window.

You can change the following:

Name

Description

Owner

IDs for an ID target

Tags for a Tag target

From any supported tag type to another. The three supported tag types are ID target, Tag target, and resource group. For more information, see Target.

If a parameter is null, then the corresponding field is not modified.

update_maintenance_window_task(client, input, options \\ [])

Modifies a task assigned to a maintenance window.

You can't change the task type, but you can change the following values:

TaskARN. For example, you can change a RUN_COMMAND task from AWS-RunPowerShellScript to AWS-RunShellScript.

ServiceRoleArn

TaskInvocationParameters

Priority

MaxConcurrency

MaxErrors

If the value for a parameter in UpdateMaintenanceWindowTask is null, then the corresponding field is not modified. If you set Replace to true, then all fields required by the RegisterTaskWithMaintenanceWindow action are required for this request. Optional fields that aren't specified are set to null.

When you update a maintenance window task that has options specified in TaskInvocationParameters, you must provide again all the TaskInvocationParameters values that you want to retain. The values you do not specify again are removed. For example, suppose that when you registered a Run Command task, you specified TaskInvocationParameters values for Comment, NotificationConfig, and OutputS3BucketName. If you update the maintenance window task and specify only a different OutputS3BucketName value, the values for Comment and NotificationConfig are removed.

update_managed_instance_role(client, input, options \\ [])

Changes the Amazon Identity and Access Management (IAM) role that is assigned to the on-premises instance or virtual machines (VM).

IAM roles are first assigned to these hybrid instances during the activation process. For more information, see CreateActivation.

update_ops_item(client, input, options \\ [])

Edit or change an OpsItem.

You must have permission in AWS Identity and Access Management (IAM) to update an OpsItem. For more information, see Getting started with OpsCenter in the AWS Systems Manager User Guide.

Operations engineers and IT professionals use OpsCenter to view, investigate, and remediate operational issues impacting the performance and health of their AWS resources. For more information, see AWS Systems Manager OpsCenter in the AWS Systems Manager User Guide.

update_patch_baseline(client, input, options \\ [])

Modifies an existing patch baseline.

Fields not specified in the request are left unchanged.

For information about valid key and value pairs in PatchFilters for each supported operating system type, see PatchFilter.

update_resource_data_sync(client, input, options \\ [])

Update a resource data sync.

After you create a resource data sync for a Region, you can't change the account options for that sync. For example, if you create a sync in the us-east-2 (Ohio) Region and you choose the Include only the current account option, you can't edit that sync later and choose the Include all accounts from my AWS Organizations configuration option. Instead, you must delete the first resource data sync, and create a new one.

This API action only supports a resource data sync that was created with a SyncFromSource SyncType.

update_service_setting(client, input, options \\ [])

ServiceSetting is an account-level setting for an AWS service.

This setting defines how a user interacts with or uses a service or a feature of a service. For example, if an AWS service charges money to the account based on feature or service usage, then the AWS service team might create a default setting of "false". This means the user can't use this feature unless they change the setting to "true" and intentionally opt in for a paid feature.

Services map a SettingId object to a setting value. AWS services teams define the default value for a SettingId. You can't create a new SettingId, but you can overwrite the default value if you have the ssm:UpdateServiceSetting permission for the setting. Use the GetServiceSetting API action to view the current value. Or, use the ResetServiceSetting to change the value back to the original value defined by the AWS service team.

Update the service setting for the account.