aws-elixir v0.6.0 AWS.SSOAdmin

Functions

attach_managed_policy_to_permission_set(client, input, options \\ [])

Attaches an IAM managed policy ARN to a permission set.

If the permission set is already referenced by one or more account assignments, you will need to call ProvisionPermissionSet after this action to apply the corresponding IAM policy updates to all assigned accounts.

create_account_assignment(client, input, options \\ [])

Assigns access to a principal for a specified AWS account using a specified permission set.

The term principal here refers to a user or group that is defined in AWS SSO.

As part of a successful CreateAccountAssignment call, the specified permission set will automatically be provisioned to the account in the form of an IAM policy attached to the SSO-created IAM role. If the permission set is subsequently updated, the corresponding IAM policies attached to roles in your accounts will not be updated automatically. In this case, you will need to call ProvisionPermissionSet to make these updates.

create_permission_set(client, input, options \\ [])

Creates a permission set within a specified SSO instance.

To grant users and groups access to AWS account resources, use CreateAccountAssignment.

delete_account_assignment(client, input, options \\ [])

Deletes a principal's access from a specified AWS account using a specified permission set.

delete_inline_policy_from_permission_set(client, input, options \\ [])

Deletes the inline policy from a specified permission set.

delete_permission_set(client, input, options \\ [])

Deletes the specified permission set.

describe_account_assignment_creation_status(client, input, options \\ [])

Describes the status of the assignment creation request.

describe_account_assignment_deletion_status(client, input, options \\ [])

Describes the status of the assignment deletion request.

describe_permission_set(client, input, options \\ [])

Gets the details of the permission set.

describe_permission_set_provisioning_status(client, input, options \\ [])

Describes the status for the given permission set provisioning request.

detach_managed_policy_from_permission_set(client, input, options \\ [])

Detaches the attached IAM managed policy ARN from the specified permission set.

get_inline_policy_for_permission_set(client, input, options \\ [])

Obtains the inline policy assigned to the permission set.

list_account_assignment_creation_status(client, input, options \\ [])

Lists the status of the AWS account assignment creation requests for a specified SSO instance.

list_account_assignment_deletion_status(client, input, options \\ [])

Lists the status of the AWS account assignment deletion requests for a specified SSO instance.

list_account_assignments(client, input, options \\ [])

Lists the assignee of the specified AWS account with the specified permission set.

list_accounts_for_provisioned_permission_set(client, input, options \\ [])

Lists all the AWS accounts where the specified permission set is provisioned.

list_instances(client, input, options \\ [])

Lists the SSO instances that the caller has access to.

list_managed_policies_in_permission_set(client, input, options \\ [])

Lists the IAM managed policy that is attached to a specified permission set.

list_permission_set_provisioning_status(client, input, options \\ [])

Lists the status of the permission set provisioning requests for a specified SSO instance.

list_permission_sets(client, input, options \\ [])

Lists the PermissionSets in an SSO instance.

list_permission_sets_provisioned_to_account(client, input, options \\ [])

Lists all the permission sets that are provisioned to a specified AWS account.

list_tags_for_resource(client, input, options \\ [])

Lists the tags that are attached to a specified resource.

provision_permission_set(client, input, options \\ [])

The process by which a specified permission set is provisioned to the specified target.

put_inline_policy_to_permission_set(client, input, options \\ [])

Attaches an IAM inline policy to a permission set.

If the permission set is already referenced by one or more account assignments, you will need to call ProvisionPermissionSet after this action to apply the corresponding IAM policy updates to all assigned accounts.

tag_resource(client, input, options \\ [])

Associates a set of tags with a specified resource.

untag_resource(client, input, options \\ [])

Disassociates a set of tags from a specified resource.

update_permission_set(client, input, options \\ [])

Updates an existing permission set.