aws-elixir v0.6.0 AWS.SageMakerA2IRuntime

Amazon Augmented AI is in preview release and is subject to change.

We do not recommend using this product in production environments.

Amazon Augmented AI (Amazon A2I) adds the benefit of human judgment to any machine learning application. When an AI application can't evaluate data with a high degree of confidence, human reviewers can take over. This human review is called a human review workflow. To create and start a human review workflow, you need three resources: a worker task template, a flow definition, and a human loop.

For information about these resources and prerequisites for using Amazon A2I, see Get Started with Amazon Augmented AI in the Amazon SageMaker Developer Guide.

This API reference includes information about API actions and data types that you can use to interact with Amazon A2I programmatically. Use this guide to:

Start a human loop with the StartHumanLoop operation when using Amazon A2I with a custom task type. To learn more about the difference between custom and built-in task types, see Use Task Types . To learn how to start a human loop using this API, see Create and Start a Human Loop for a Custom Task Type in the Amazon SageMaker Developer Guide.

Manage your human loops. You can list all human loops that you have created, describe individual human loops, and stop and delete human loops. To learn more, see Monitor and Manage Your Human Loop in the Amazon SageMaker Developer Guide.

Amazon A2I integrates APIs from various AWS services to create and start human review workflows for those services. To learn how Amazon A2I uses these APIs, see Use APIs in Amazon A2I in the Amazon SageMaker Developer Guide.

Functions

delete_human_loop(client, human_loop_name, input, options \\ [])

Deletes the specified human loop for a flow definition.

describe_human_loop(client, human_loop_name, options \\ [])

Returns information about the specified human loop.

list_human_loops(client, creation_time_after \\ nil, creation_time_before \\ nil, flow_definition_arn, max_results \\ nil, next_token \\ nil, sort_order \\ nil, options \\ [])

Returns information about human loops, given the specified parameters.

If a human loop was deleted, it will not be included.

start_human_loop(client, input, options \\ [])

Starts a human loop, provided that at least one activation condition is met.

stop_human_loop(client, input, options \\ [])

Stops the specified human loop.