aws-elixir v0.6.0 AWS.Savingsplans

Savings Plans are a pricing model that offer significant savings on AWS usage (for example, on Amazon EC2 instances).

You commit to a consistent amount of usage, in USD per hour, for a term of 1 or 3 years, and receive a lower price for that usage. For more information, see the AWS Savings Plans User Guide.

Functions

create_savings_plan(client, input, options \\ [])

Creates a Savings Plan.

delete_queued_savings_plan(client, input, options \\ [])

Deletes the queued purchase for the specified Savings Plan.

describe_savings_plan_rates(client, input, options \\ [])

Describes the specified Savings Plans rates.

describe_savings_plans(client, input, options \\ [])

Describes the specified Savings Plans.

describe_savings_plans_offering_rates(client, input, options \\ [])

Describes the specified Savings Plans offering rates.

describe_savings_plans_offerings(client, input, options \\ [])

Describes the specified Savings Plans offerings.

list_tags_for_resource(client, input, options \\ [])

Lists the tags for the specified resource.

tag_resource(client, input, options \\ [])

Adds the specified tags to the specified resource.

untag_resource(client, input, options \\ [])

Removes the specified tags from the specified resource.