aws-elixir v0.6.0 AWS.Schemas

Amazon EventBridge Schema Registry

Functions

create_discoverer(client, input, options \\ [])

Creates a discoverer.

create_registry(client, registry_name, input, options \\ [])

Creates a registry.

create_schema(client, registry_name, schema_name, input, options \\ [])

Creates a schema definition.

Inactive schemas will be deleted after two years.

delete_discoverer(client, discoverer_id, input, options \\ [])

Deletes a discoverer.

delete_registry(client, registry_name, input, options \\ [])

Deletes a Registry.

delete_resource_policy(client, input, options \\ [])

Delete the resource-based policy attached to the specified registry.

delete_schema(client, registry_name, schema_name, input, options \\ [])

Delete a schema definition.

delete_schema_version(client, registry_name, schema_name, schema_version, input, options \\ [])

Delete the schema version definition

describe_code_binding(client, language, registry_name, schema_name, schema_version \\ nil, options \\ [])

Describe the code binding URI.

describe_discoverer(client, discoverer_id, options \\ [])

Describes the discoverer.

describe_registry(client, registry_name, options \\ [])

Describes the registry.

describe_schema(client, registry_name, schema_name, schema_version \\ nil, options \\ [])

Retrieve the schema definition.

export_schema(client, registry_name, schema_name, schema_version \\ nil, type, options \\ [])

Exports a schema to a different specification.

get_code_binding_source(client, language, registry_name, schema_name, schema_version \\ nil, options \\ [])

Get the code binding source URI.

get_discovered_schema(client, input, options \\ [])

Get the discovered schema that was generated based on sampled events.

get_resource_policy(client, registry_name \\ nil, options \\ [])

Retrieves the resource-based policy attached to a given registry.

list_discoverers(client, discoverer_id_prefix \\ nil, limit \\ nil, next_token \\ nil, source_arn_prefix \\ nil, options \\ [])

List the discoverers.

list_registries(client, limit \\ nil, next_token \\ nil, registry_name_prefix \\ nil, scope \\ nil, options \\ [])

List the registries.

list_schema_versions(client, registry_name, schema_name, limit \\ nil, next_token \\ nil, options \\ [])

Provides a list of the schema versions and related information.

list_schemas(client, registry_name, limit \\ nil, next_token \\ nil, schema_name_prefix \\ nil, options \\ [])

List the schemas.

list_tags_for_resource(client, resource_arn, options \\ [])

Get tags for resource.

put_code_binding(client, language, registry_name, schema_name, input, options \\ [])

Put code binding URI

put_resource_policy(client, input, options \\ [])

The name of the policy.

search_schemas(client, registry_name, keywords, limit \\ nil, next_token \\ nil, options \\ [])

Search the schemas

start_discoverer(client, discoverer_id, input, options \\ [])

Starts the discoverer

stop_discoverer(client, discoverer_id, input, options \\ [])

Stops the discoverer

tag_resource(client, resource_arn, input, options \\ [])

Add tags to a resource.

untag_resource(client, resource_arn, input, options \\ [])

Removes tags from a resource.

update_discoverer(client, discoverer_id, input, options \\ [])

Updates the discoverer

update_registry(client, registry_name, input, options \\ [])

Updates a registry.

update_schema(client, registry_name, schema_name, input, options \\ [])

Updates the schema definition

Inactive schemas will be deleted after two years.