aws-elixir v0.6.0 AWS.ServerlessApplicationRepository

The AWS Serverless Application Repository makes it easy for developers and enterprises to quickly find and deploy serverless applications in the AWS Cloud.

For more information about serverless applications, see Serverless Computing and Applications on the AWS website.

The AWS Serverless Application Repository is deeply integrated with the AWS Lambda console, so that developers of all levels can get started with serverless computing without needing to learn anything new. You can use category keywords to browse for applications such as web and mobile backends, data processing applications, or chatbots. You can also search for applications by name, publisher, or event source. To use an application, you simply choose it, configure any required fields, and deploy it with a few clicks.

You can also easily publish applications, sharing them publicly with the community at large, or privately within your team or across your organization. To publish a serverless application (or app), you can use the AWS Management Console, AWS Command Line Interface (AWS CLI), or AWS SDKs to upload the code. Along with the code, you upload a simple manifest file, also known as the AWS Serverless Application Model (AWS SAM) template. For more information about AWS SAM, see AWS Serverless Application Model (AWS SAM) on the AWS Labs GitHub repository.

The AWS Serverless Application Repository Developer Guide contains more information about the two developer experiences available:

* Consuming Applications – Browse for applications and view information about them, including source code and readme files. Also install, configure, and deploy applications of your choosing.

Publishing Applications – Configure and upload applications to make them available to other developers, and publish new versions of applications.

Functions

create_application(client, input, options \\ [])

Creates an application, optionally including an AWS SAM file to create the first application version in the same call.

create_application_version(client, application_id, semantic_version, input, options \\ [])

Creates an application version.

create_cloud_formation_change_set(client, application_id, input, options \\ [])

Creates an AWS CloudFormation change set for the given application.

create_cloud_formation_template(client, application_id, input, options \\ [])

Creates an AWS CloudFormation template.

delete_application(client, application_id, input, options \\ [])

Deletes the specified application.

get_application(client, application_id, semantic_version \\ nil, options \\ [])

Gets the specified application.

get_application_policy(client, application_id, options \\ [])

Retrieves the policy for the application.

get_cloud_formation_template(client, application_id, template_id, options \\ [])

Gets the specified AWS CloudFormation template.

list_application_dependencies(client, application_id, max_items \\ nil, next_token \\ nil, semantic_version \\ nil, options \\ [])

Retrieves the list of applications nested in the containing application.

list_application_versions(client, application_id, max_items \\ nil, next_token \\ nil, options \\ [])

Lists versions for the specified application.

list_applications(client, max_items \\ nil, next_token \\ nil, options \\ [])

Lists applications owned by the requester.

put_application_policy(client, application_id, input, options \\ [])

Sets the permission policy for an application.

For the list of actions supported for this operation, see Application Permissions

.

unshare_application(client, application_id, input, options \\ [])

Unshares an application from an AWS Organization.

This operation can be called only from the organization's master account.

update_application(client, application_id, input, options \\ [])

Updates the specified application.