aws-elixir v0.6.0 AWS.ServiceCatalog

AWS Service Catalog

AWS Service Catalog enables organizations to create and manage catalogs of IT services that are approved for use on AWS.

To get the most out of this documentation, you should be familiar with the terminology discussed in AWS Service Catalog Concepts.

Functions

accept_portfolio_share(client, input, options \\ [])

Accepts an offer to share the specified portfolio.

associate_budget_with_resource(client, input, options \\ [])

Associates the specified budget with the specified resource.

associate_principal_with_portfolio(client, input, options \\ [])

Associates the specified principal ARN with the specified portfolio.

associate_product_with_portfolio(client, input, options \\ [])

Associates the specified product with the specified portfolio.

A delegated admin is authorized to invoke this command.

associate_service_action_with_provisioning_artifact(client, input, options \\ [])

Associates a self-service action with a provisioning artifact.

associate_tag_option_with_resource(client, input, options \\ [])

Associate the specified TagOption with the specified portfolio or product.

batch_associate_service_action_with_provisioning_artifact(client, input, options \\ [])

Associates multiple self-service actions with provisioning artifacts.

batch_disassociate_service_action_from_provisioning_artifact(client, input, options \\ [])

Disassociates a batch of self-service actions from the specified provisioning artifact.

copy_product(client, input, options \\ [])

Copies the specified source product to the specified target product or a new product.

You can copy a product to the same account or another account. You can copy a product to the same region or another region.

This operation is performed asynchronously. To track the progress of the operation, use DescribeCopyProductStatus.

create_constraint(client, input, options \\ [])

Creates a constraint.

A delegated admin is authorized to invoke this command.

create_portfolio(client, input, options \\ [])

Creates a portfolio.

A delegated admin is authorized to invoke this command.

create_portfolio_share(client, input, options \\ [])

Shares the specified portfolio with the specified account or organization node.

Shares to an organization node can only be created by the management account of an organization or by a delegated administrator. You can share portfolios to an organization, an organizational unit, or a specific account.

Note that if a delegated admin is de-registered, they can no longer create portfolio shares.

AWSOrganizationsAccess must be enabled in order to create a portfolio share to an organization node.

You can't share a shared resource. This includes portfolios that contain a shared product.

create_product(client, input, options \\ [])

Creates a product.

A delegated admin is authorized to invoke this command.

create_provisioned_product_plan(client, input, options \\ [])

Creates a plan.

A plan includes the list of resources to be created (when provisioning a new product) or modified (when updating a provisioned product) when the plan is executed.

You can create one plan per provisioned product. To create a plan for an existing provisioned product, the product status must be AVAILBLE or TAINTED.

To view the resource changes in the change set, use DescribeProvisionedProductPlan. To create or modify the provisioned product, use ExecuteProvisionedProductPlan.

create_provisioning_artifact(client, input, options \\ [])

Creates a provisioning artifact (also known as a version) for the specified product.

You cannot create a provisioning artifact for a product that was shared with you.

create_service_action(client, input, options \\ [])

Creates a self-service action.

create_tag_option(client, input, options \\ [])

Creates a TagOption.

delete_constraint(client, input, options \\ [])

Deletes the specified constraint.

A delegated admin is authorized to invoke this command.

delete_portfolio(client, input, options \\ [])

Deletes the specified portfolio.

You cannot delete a portfolio if it was shared with you or if it has associated products, users, constraints, or shared accounts.

A delegated admin is authorized to invoke this command.

delete_portfolio_share(client, input, options \\ [])

Stops sharing the specified portfolio with the specified account or organization node.

Shares to an organization node can only be deleted by the management account of an organization or by a delegated administrator.

Note that if a delegated admin is de-registered, portfolio shares created from that account are removed.

delete_product(client, input, options \\ [])

Deletes the specified product.

You cannot delete a product if it was shared with you or is associated with a portfolio.

A delegated admin is authorized to invoke this command.

delete_provisioned_product_plan(client, input, options \\ [])

Deletes the specified plan.

delete_provisioning_artifact(client, input, options \\ [])

Deletes the specified provisioning artifact (also known as a version) for the specified product.

You cannot delete a provisioning artifact associated with a product that was shared with you. You cannot delete the last provisioning artifact for a product, because a product must have at least one provisioning artifact.

delete_service_action(client, input, options \\ [])

Deletes a self-service action.

delete_tag_option(client, input, options \\ [])

Deletes the specified TagOption.

You cannot delete a TagOption if it is associated with a product or portfolio.

describe_constraint(client, input, options \\ [])

Gets information about the specified constraint.

describe_copy_product_status(client, input, options \\ [])

Gets the status of the specified copy product operation.

describe_portfolio(client, input, options \\ [])

Gets information about the specified portfolio.

A delegated admin is authorized to invoke this command.

describe_portfolio_share_status(client, input, options \\ [])

Gets the status of the specified portfolio share operation.

This API can only be called by the management account in the organization or by a delegated admin.

describe_product(client, input, options \\ [])

Gets information about the specified product.

describe_product_as_admin(client, input, options \\ [])

Gets information about the specified product.

This operation is run with administrator access.

describe_product_view(client, input, options \\ [])

Gets information about the specified product.

describe_provisioned_product(client, input, options \\ [])

Gets information about the specified provisioned product.

describe_provisioned_product_plan(client, input, options \\ [])

Gets information about the resource changes for the specified plan.

describe_provisioning_artifact(client, input, options \\ [])

Gets information about the specified provisioning artifact (also known as a version) for the specified product.

describe_provisioning_parameters(client, input, options \\ [])

Gets information about the configuration required to provision the specified product using the specified provisioning artifact.

If the output contains a TagOption key with an empty list of values, there is a TagOption conflict for that key. The end user cannot take action to fix the conflict, and launch is not blocked. In subsequent calls to ProvisionProduct, do not include conflicted TagOption keys as tags, or this causes the error "Parameter validation failed: Missing required parameter in Tags[N]:Value". Tag the provisioned product with the value sc-tagoption-conflict-portfolioId-productId.

describe_record(client, input, options \\ [])

Gets information about the specified request operation.

Use this operation after calling a request operation (for example, ProvisionProduct, TerminateProvisionedProduct, or UpdateProvisionedProduct).

If a provisioned product was transferred to a new owner using UpdateProvisionedProductProperties, the new owner will be able to describe all past records for that product. The previous owner will no longer be able to describe the records, but will be able to use ListRecordHistory to see the product's history from when he was the owner.

describe_service_action(client, input, options \\ [])

Describes a self-service action.

describe_service_action_execution_parameters(client, input, options \\ [])

Finds the default parameters for a specific self-service action on a specific provisioned product and returns a map of the results to the user.

describe_tag_option(client, input, options \\ [])

Gets information about the specified TagOption.

disable_a_w_s_organizations_access(client, input, options \\ [])

Disable portfolio sharing through AWS Organizations feature.

This feature will not delete your current shares but it will prevent you from creating new shares throughout your organization. Current shares will not be in sync with your organization structure if it changes after calling this API. This API can only be called by the management account in the organization.

This API can't be invoked if there are active delegated administrators in the organization.

Note that a delegated administrator is not authorized to invoke DisableAWSOrganizationsAccess.

disassociate_budget_from_resource(client, input, options \\ [])

Disassociates the specified budget from the specified resource.

disassociate_principal_from_portfolio(client, input, options \\ [])

Disassociates a previously associated principal ARN from a specified portfolio.

disassociate_product_from_portfolio(client, input, options \\ [])

Disassociates the specified product from the specified portfolio.

A delegated admin is authorized to invoke this command.

disassociate_service_action_from_provisioning_artifact(client, input, options \\ [])

Disassociates the specified self-service action association from the specified provisioning artifact.

disassociate_tag_option_from_resource(client, input, options \\ [])

Disassociates the specified TagOption from the specified resource.

enable_a_w_s_organizations_access(client, input, options \\ [])

Enable portfolio sharing feature through AWS Organizations.

This API will allow Service Catalog to receive updates on your organization in order to sync your shares with the current structure. This API can only be called by the management account in the organization.

By calling this API Service Catalog will make a call to organizations:EnableAWSServiceAccess on your behalf so that your shares can be in sync with any changes in your AWS Organizations structure.

Note that a delegated administrator is not authorized to invoke EnableAWSOrganizationsAccess.

execute_provisioned_product_plan(client, input, options \\ [])

Provisions or modifies a product based on the resource changes for the specified plan.

execute_provisioned_product_service_action(client, input, options \\ [])

Executes a self-service action against a provisioned product.

get_a_w_s_organizations_access_status(client, input, options \\ [])

Get the Access Status for AWS Organization portfolio share feature.

This API can only be called by the management account in the organization or by a delegated admin.

get_provisioned_product_outputs(client, input, options \\ [])

This API takes either a ProvisonedProductId or a ProvisionedProductName, along with a list of one or more output keys, and responds with the key/value pairs of those outputs.

list_accepted_portfolio_shares(client, input, options \\ [])

Lists all portfolios for which sharing was accepted by this account.

list_budgets_for_resource(client, input, options \\ [])

Lists all the budgets associated to the specified resource.

list_constraints_for_portfolio(client, input, options \\ [])

Lists the constraints for the specified portfolio and product.

list_launch_paths(client, input, options \\ [])

Lists the paths to the specified product.

A path is how the user has access to a specified product, and is necessary when provisioning a product. A path also determines the constraints put on the product.

list_organization_portfolio_access(client, input, options \\ [])

Lists the organization nodes that have access to the specified portfolio.

This API can only be called by the management account in the organization or by a delegated admin.

If a delegated admin is de-registered, they can no longer perform this operation.

list_portfolio_access(client, input, options \\ [])

Lists the account IDs that have access to the specified portfolio.

A delegated admin can list the accounts that have access to the shared portfolio. Note that if a delegated admin is de-registered, they can no longer perform this operation.

list_portfolios(client, input, options \\ [])

Lists all portfolios in the catalog.

list_portfolios_for_product(client, input, options \\ [])

Lists all portfolios that the specified product is associated with.

list_principals_for_portfolio(client, input, options \\ [])

Lists all principal ARNs associated with the specified portfolio.

list_provisioned_product_plans(client, input, options \\ [])

Lists the plans for the specified provisioned product or all plans to which the user has access.

list_provisioning_artifacts(client, input, options \\ [])

Lists all provisioning artifacts (also known as versions) for the specified product.

list_provisioning_artifacts_for_service_action(client, input, options \\ [])

Lists all provisioning artifacts (also known as versions) for the specified self-service action.

list_record_history(client, input, options \\ [])

Lists the specified requests or all performed requests.

list_resources_for_tag_option(client, input, options \\ [])

Lists the resources associated with the specified TagOption.

list_service_actions(client, input, options \\ [])

Lists all self-service actions.

list_service_actions_for_provisioning_artifact(client, input, options \\ [])

Returns a paginated list of self-service actions associated with the specified Product ID and Provisioning Artifact ID.

list_stack_instances_for_provisioned_product(client, input, options \\ [])

Returns summary information about stack instances that are associated with the specified CFN_STACKSET type provisioned product.

You can filter for stack instances that are associated with a specific AWS account name or region.

list_tag_options(client, input, options \\ [])

Lists the specified TagOptions or all TagOptions.

provision_product(client, input, options \\ [])

Provisions the specified product.

A provisioned product is a resourced instance of a product. For example, provisioning a product based on a CloudFormation template launches a CloudFormation stack and its underlying resources. You can check the status of this request using DescribeRecord.

If the request contains a tag key with an empty list of values, there is a tag conflict for that key. Do not include conflicted keys as tags, or this causes the error "Parameter validation failed: Missing required parameter in Tags[N]:Value".

reject_portfolio_share(client, input, options \\ [])

Rejects an offer to share the specified portfolio.

scan_provisioned_products(client, input, options \\ [])

Lists the provisioned products that are available (not terminated).

To use additional filtering, see SearchProvisionedProducts.

search_products(client, input, options \\ [])

Gets information about the products to which the caller has access.

search_products_as_admin(client, input, options \\ [])

Gets information about the products for the specified portfolio or all products.

search_provisioned_products(client, input, options \\ [])

Gets information about the provisioned products that meet the specified criteria.

terminate_provisioned_product(client, input, options \\ [])

Terminates the specified provisioned product.

This operation does not delete any records associated with the provisioned product.

You can check the status of this request using DescribeRecord.

update_constraint(client, input, options \\ [])

Updates the specified constraint.

update_portfolio(client, input, options \\ [])

Updates the specified portfolio.

You cannot update a product that was shared with you.

update_product(client, input, options \\ [])

Updates the specified product.

update_provisioned_product(client, input, options \\ [])

Requests updates to the configuration of the specified provisioned product.

If there are tags associated with the object, they cannot be updated or added. Depending on the specific updates requested, this operation can update with no interruption, with some interruption, or replace the provisioned product entirely.

You can check the status of this request using DescribeRecord.

update_provisioned_product_properties(client, input, options \\ [])

Requests updates to the properties of the specified provisioned product.

update_provisioning_artifact(client, input, options \\ [])

Updates the specified provisioning artifact (also known as a version) for the specified product.

You cannot update a provisioning artifact for a product that was shared with you.

update_service_action(client, input, options \\ [])

Updates a self-service action.

update_tag_option(client, input, options \\ [])

Updates the specified TagOption.