aws-elixir v0.6.0 AWS.ServiceQuotas

Service Quotas is a web service that you can use to manage many of your AWS service quotas.

Quotas, also referred to as limits, are the maximum values for a resource, item, or operation. This guide provide descriptions of the Service Quotas actions that you can call from an API. For the Service Quotas user guide, which explains how to use Service Quotas from the console, see What is Service Quotas. AWS provides SDKs that consist of libraries and sample code for programming languages and platforms (Java, Ruby, .NET, iOS, Android, etc...,). The SDKs provide a convenient way to create programmatic access to Service Quotas and AWS. For information about the AWS SDKs, including how to download and install them, see the Tools for Amazon Web Services page.

Functions

associate_service_quota_template(client, input, options \\ [])

Associates the Service Quotas template with your organization so that when new accounts are created in your organization, the template submits increase requests for the specified service quotas.

Use the Service Quotas template to request an increase for any adjustable quota value. After you define the Service Quotas template, use this operation to associate, or enable, the template.

delete_service_quota_increase_request_from_template(client, input, options \\ [])

Removes a service quota increase request from the Service Quotas template.

disassociate_service_quota_template(client, input, options \\ [])

Disables the Service Quotas template.

Once the template is disabled, it does not request quota increases for new accounts in your organization. Disabling the quota template does not apply the quota increase requests from the template.

Related operations

To enable the quota template, call AssociateServiceQuotaTemplate.

To delete a specific service quota from the template, use DeleteServiceQuotaIncreaseRequestFromTemplate.

get_a_w_s_default_service_quota(client, input, options \\ [])

Retrieves the default service quotas values.

The Value returned for each quota is the AWS default value, even if the quotas have been increased..

get_association_for_service_quota_template(client, input, options \\ [])

Retrieves the ServiceQuotaTemplateAssociationStatus value from the service.

Use this action to determine if the Service Quota template is associated, or enabled.

get_requested_service_quota_change(client, input, options \\ [])

Retrieves the details for a particular increase request.

get_service_quota(client, input, options \\ [])

Returns the details for the specified service quota.

This operation provides a different Value than the GetAWSDefaultServiceQuota operation. This operation returns the applied value for each quota. GetAWSDefaultServiceQuota returns the default AWS value for each quota.

get_service_quota_increase_request_from_template(client, input, options \\ [])

Returns the details of the service quota increase request in your template.

list_a_w_s_default_service_quotas(client, input, options \\ [])

Lists all default service quotas for the specified AWS service or all AWS services.

ListAWSDefaultServiceQuotas is similar to ListServiceQuotas except for the Value object. The Value object returned by ListAWSDefaultServiceQuotas is the default value assigned by AWS. This request returns a list of all service quotas for the specified service. The listing of each you'll see the default values are the values that AWS provides for the quotas.

Always check the NextToken response parameter when calling any of the List* operations. These operations can return an unexpected list of results, even when there are more results available. When this happens, the NextToken response parameter contains a value to pass the next call to the same API to request the next part of the list.

list_requested_service_quota_change_history(client, input, options \\ [])

Requests a list of the changes to quotas for a service.

list_requested_service_quota_change_history_by_quota(client, input, options \\ [])

Requests a list of the changes to specific service quotas.

This command provides additional granularity over the ListRequestedServiceQuotaChangeHistory command. Once a quota change request has reached CASE_CLOSED, APPROVED, or DENIED, the history has been kept for 90 days.

list_service_quota_increase_requests_in_template(client, input, options \\ [])

Returns a list of the quota increase requests in the template.

list_service_quotas(client, input, options \\ [])

Lists all service quotas for the specified AWS service.

This request returns a list of the service quotas for the specified service. you'll see the default values are the values that AWS provides for the quotas.

Always check the NextToken response parameter when calling any of the List* operations. These operations can return an unexpected list of results, even when there are more results available. When this happens, the NextToken response parameter contains a value to pass the next call to the same API to request the next part of the list.

list_services(client, input, options \\ [])

Lists the AWS services available in Service Quotas.

Not all AWS services are available in Service Quotas. To list the see the list of the service quotas for a specific service, use ListServiceQuotas.

put_service_quota_increase_request_into_template(client, input, options \\ [])

Defines and adds a quota to the service quota template.

To add a quota to the template, you must provide the ServiceCode, QuotaCode, AwsRegion, and DesiredValue. Once you add a quota to the template, use ListServiceQuotaIncreaseRequestsInTemplate to see the list of quotas in the template.

request_service_quota_increase(client, input, options \\ [])

Retrieves the details of a service quota increase request.

The response to this command provides the details in the RequestedServiceQuotaChange object.