aws-elixir v0.6.0 AWS.Signer

With code signing for IoT, you can sign code that you create for any IoT device that is supported by Amazon Web Services (AWS).

Code signing is available through Amazon FreeRTOS and AWS IoT Device Management, and integrated with AWS Certificate Manager (ACM). In order to sign code, you import a third-party code signing certificate with ACM that is used to sign updates in Amazon FreeRTOS and AWS IoT Device Management. For general information about using code signing, see the Code Signing for IoT Developer Guide.

Functions

cancel_signing_profile(client, profile_name, input, options \\ [])

Changes the state of an ACTIVE signing profile to CANCELED.

A canceled profile is still viewable with the ListSigningProfiles operation, but it cannot perform new signing jobs, and is deleted two years after cancelation.

describe_signing_job(client, job_id, options \\ [])

Returns information about a specific code signing job.

You specify the job by using the jobId value that is returned by the StartSigningJob operation.

get_signing_platform(client, platform_id, options \\ [])

Returns information on a specific signing platform.

get_signing_profile(client, profile_name, options \\ [])

Returns information on a specific signing profile.

list_signing_jobs(client, max_results \\ nil, next_token \\ nil, platform_id \\ nil, requested_by \\ nil, status \\ nil, options \\ [])

Lists all your signing jobs.

You can use the maxResults parameter to limit the number of signing jobs that are returned in the response. If additional jobs remain to be listed, code signing returns a nextToken value. Use this value in subsequent calls to ListSigningJobs to fetch the remaining values. You can continue calling ListSigningJobs with your maxResults parameter and with new values that code signing returns in the nextToken parameter until all of your signing jobs have been returned.

list_signing_platforms(client, category \\ nil, max_results \\ nil, next_token \\ nil, partner \\ nil, target \\ nil, options \\ [])

Lists all signing platforms available in code signing that match the request parameters.

If additional jobs remain to be listed, code signing returns a nextToken value. Use this value in subsequent calls to ListSigningJobs to fetch the remaining values. You can continue calling ListSigningJobs with your maxResults parameter and with new values that code signing returns in the nextToken parameter until all of your signing jobs have been returned.

list_signing_profiles(client, include_canceled \\ nil, max_results \\ nil, next_token \\ nil, options \\ [])

Lists all available signing profiles in your AWS account.

Returns only profiles with an ACTIVE status unless the includeCanceled request field is set to true. If additional jobs remain to be listed, code signing returns a nextToken value. Use this value in subsequent calls to ListSigningJobs to fetch the remaining values. You can continue calling ListSigningJobs with your maxResults parameter and with new values that code signing returns in the nextToken parameter until all of your signing jobs have been returned.

list_tags_for_resource(client, resource_arn, options \\ [])

Returns a list of the tags associated with a signing profile resource.

put_signing_profile(client, profile_name, input, options \\ [])

Creates a signing profile.

A signing profile is a code signing template that can be used to carry out a pre-defined signing job. For more information, see http://docs.aws.amazon.com/signer/latest/developerguide/gs-profile.html

start_signing_job(client, input, options \\ [])

Initiates a signing job to be performed on the code provided.

Signing jobs are viewable by the ListSigningJobs operation for two years after they are performed. Note the following requirements:

You must create an Amazon S3 source bucket. For more information, see Create a Bucket in the Amazon S3 Getting Started Guide.

Your S3 source bucket must be version enabled.

You must create an S3 destination bucket. Code signing uses your S3 destination bucket to write your signed code.

You specify the name of the source and destination buckets when calling the StartSigningJob operation.

You must also specify a request token that identifies your request to code signing.

You can call the DescribeSigningJob and the ListSigningJobs actions after you call StartSigningJob.

For a Java example that shows how to use this action, see http://docs.aws.amazon.com/acm/latest/userguide/

tag_resource(client, resource_arn, input, options \\ [])

Adds one or more tags to a signing profile.

Tags are labels that you can use to identify and organize your AWS resources. Each tag consists of a key and an optional value. To specify the signing profile, use its Amazon Resource Name (ARN). To specify the tag, use a key-value pair.

untag_resource(client, resource_arn, input, options \\ [])

Removes one or more tags from a signing profile.

To remove the tags, specify a list of tag keys.