aws-elixir v0.6.0 AWS.TimestreamWrite

Amazon Timestream is a fast, scalable, fully managed time series database service that makes it easy to store and analyze trillions of time series data points per day.

With Timestream, you can easily store and analyze IoT sensor data to derive insights from your IoT applications. You can analyze industrial telemetry to streamline equipment management and maintenance. You can also store and analyze log data and metrics to improve the performance and availability of your applications. Timestream is built from the ground up to effectively ingest, process, and store time series data. It organizes data to optimize query processing. It automatically scales based on the volume of data ingested and on the query volume to ensure you receive optimal performance while inserting and querying data. As your data grows over time, Timestream’s adaptive query processing engine spans across storage tiers to provide fast analysis while reducing costs.

Functions

create_database(client, input, options \\ [])

Creates a new Timestream database.

If the KMS key is not specified, the database will be encrypted with a Timestream managed KMS key located in your account. Refer to AWS managed KMS keys for more info. Service quotas apply. For more information, see Access Management in the Timestream Developer Guide.

create_table(client, input, options \\ [])

The CreateTable operation adds a new table to an existing database in your account.

In an AWS account, table names must be at least unique within each Region if they are in the same database. You may have identical table names in the same Region if the tables are in seperate databases. While creating the table, you must specify the table name, database name, and the retention properties. Service quotas apply. For more information, see Access Management in the Timestream Developer Guide.

delete_database(client, input, options \\ [])

Deletes a given Timestream database.

This is an irreversible operation. After a database is deleted, the time series data from its tables cannot be recovered.

All tables in the database must be deleted first, or a ValidationException error will be thrown.

delete_table(client, input, options \\ [])

Deletes a given Timestream table.

This is an irreversible operation. After a Timestream database table is deleted, the time series data stored in the table cannot be recovered.

describe_database(client, input, options \\ [])

Returns information about the database, including the database name, time that the database was created, and the total number of tables found within the database.

Service quotas apply. For more information, see Access Management in the Timestream Developer Guide.

describe_endpoints(client, input, options \\ [])

DescribeEndpoints returns a list of available endpoints to make Timestream API calls against.

This API is available through both Write and Query.

Because Timestream’s SDKs are designed to transparently work with the service’s architecture, including the management and mapping of the service endpoints, it is not recommended that you use this API unless:

Your application uses a programming language that does not yet have SDK support

You require better control over the client-side implementation

For detailed information on how to use DescribeEndpoints, see The Endpoint Discovery Pattern and REST APIs.

describe_table(client, input, options \\ [])

Returns information about the table, including the table name, database name, retention duration of the memory store and the magnetic store.

Service quotas apply. For more information, see Access Management in the Timestream Developer Guide.

list_databases(client, input, options \\ [])

Returns a list of your Timestream databases.

Service quotas apply. For more information, see Access Management in the Timestream Developer Guide.

list_tables(client, input, options \\ [])

A list of tables, along with the name, status and retention properties of each table.

list_tags_for_resource(client, input, options \\ [])

List all tags on a Timestream resource.

tag_resource(client, input, options \\ [])

Associate a set of tags with a Timestream resource.

You can then activate these user-defined tags so that they appear on the Billing and Cost Management console for cost allocation tracking.

untag_resource(client, input, options \\ [])

Removes the association of tags from a Timestream resource.

update_database(client, input, options \\ [])

Modifies the KMS key for an existing database.

While updating the database, you must specify the database name and the identifier of the new KMS key to be used (KmsKeyId). If there are any concurrent UpdateDatabase requests, first writer wins.

update_table(client, input, options \\ [])

Modifies the retention duration of the memory store and magnetic store for your Timestream table.

Note that the change in retention duration takes effect immediately. For example, if the retention period of the memory store was initially set to 2 hours and then changed to 24 hours, the memory store will be capable of holding 24 hours of data, but will be populated with 24 hours of data 22 hours after this change was made. Timestream does not retrieve data from the magnetic store to populate the memory store.

Service quotas apply. For more information, see Access Management in the Timestream Developer Guide.

write_records(client, input, options \\ [])

The WriteRecords operation enables you to write your time series data into Timestream.

You can specify a single data point or a batch of data points to be inserted into the system. Timestream offers you with a flexible schema that auto detects the column names and data types for your Timestream tables based on the dimension names and data types of the data points you specify when invoking writes into the database. Timestream support eventual consistency read semantics. This means that when you query data immediately after writing a batch of data into Timestream, the query results might not reflect the results of a recently completed write operation. The results may also include some stale data. If you repeat the query request after a short time, the results should return the latest data. Service quotas apply. For more information, see Access Management in the Timestream Developer Guide.