aws-elixir v0.6.0 AWS.Transcribe

Operations and objects for transcribing speech to text.

Functions

create_language_model(client, input, options \\ [])

Creates a new custom language model.

Use Amazon S3 prefixes to provide the location of your input files. The time it takes to create your model depends on the size of your training data.

create_medical_vocabulary(client, input, options \\ [])

Creates a new custom vocabulary that you can use to change how Amazon Transcribe Medical transcribes your audio file.

create_vocabulary(client, input, options \\ [])

Creates a new custom vocabulary that you can use to change the way Amazon Transcribe handles transcription of an audio file.

create_vocabulary_filter(client, input, options \\ [])

Creates a new vocabulary filter that you can use to filter words, such as profane words, from the output of a transcription job.

delete_language_model(client, input, options \\ [])

Deletes a custom language model using its name.

delete_medical_transcription_job(client, input, options \\ [])

Deletes a transcription job generated by Amazon Transcribe Medical and any related information.

delete_medical_vocabulary(client, input, options \\ [])

Deletes a vocabulary from Amazon Transcribe Medical.

delete_transcription_job(client, input, options \\ [])

Deletes a previously submitted transcription job along with any other generated results such as the transcription, models, and so on.

delete_vocabulary(client, input, options \\ [])

Deletes a vocabulary from Amazon Transcribe.

delete_vocabulary_filter(client, input, options \\ [])

Removes a vocabulary filter.

describe_language_model(client, input, options \\ [])

Gets information about a single custom language model.

Use this information to see details about the language model in your AWS account. You can also see whether the base language model used to create your custom language model has been updated. If Amazon Transcribe has updated the base model, you can create a new custom language model using the updated base model. If the language model wasn't created, you can use this operation to understand why Amazon Transcribe couldn't create it.

get_medical_transcription_job(client, input, options \\ [])

Returns information about a transcription job from Amazon Transcribe Medical.

To see the status of the job, check the TranscriptionJobStatus field. If the status is COMPLETED, the job is finished. You find the results of the completed job in the TranscriptFileUri field.

get_medical_vocabulary(client, input, options \\ [])

Retrieves information about a medical vocabulary.

get_transcription_job(client, input, options \\ [])

Returns information about a transcription job.

To see the status of the job, check the TranscriptionJobStatus field. If the status is COMPLETED, the job is finished and you can find the results at the location specified in the TranscriptFileUri field. If you enable content redaction, the redacted transcript appears in RedactedTranscriptFileUri.

get_vocabulary(client, input, options \\ [])

Gets information about a vocabulary.

get_vocabulary_filter(client, input, options \\ [])

Returns information about a vocabulary filter.

list_language_models(client, input, options \\ [])

Provides more information about the custom language models you've created.

You can use the information in this list to find a specific custom language model. You can then use the operation to get more information about it.

list_medical_transcription_jobs(client, input, options \\ [])

Lists medical transcription jobs with a specified status or substring that matches their names.

list_medical_vocabularies(client, input, options \\ [])

Returns a list of vocabularies that match the specified criteria.

If you don't enter a value in any of the request parameters, returns the entire list of vocabularies.

list_transcription_jobs(client, input, options \\ [])

Lists transcription jobs with the specified status.

list_vocabularies(client, input, options \\ [])

Returns a list of vocabularies that match the specified criteria.

If no criteria are specified, returns the entire list of vocabularies.

list_vocabulary_filters(client, input, options \\ [])

Gets information about vocabulary filters.

start_medical_transcription_job(client, input, options \\ [])

Starts a batch job to transcribe medical speech to text.

start_transcription_job(client, input, options \\ [])

Starts an asynchronous job to transcribe speech to text.

update_medical_vocabulary(client, input, options \\ [])

Updates a vocabulary with new values that you provide in a different text file from the one you used to create the vocabulary.

The UpdateMedicalVocabulary operation overwrites all of the existing information with the values that you provide in the request.

update_vocabulary(client, input, options \\ [])

Updates an existing vocabulary with new values.

The UpdateVocabulary operation overwrites all of the existing information with the values that you provide in the request.

update_vocabulary_filter(client, input, options \\ [])

Updates a vocabulary filter with a new list of filtered words.