aws-elixir v0.6.0 AWS.WorkDocs

The WorkDocs API is designed for the following use cases:

File Migration: File migration applications are supported for users who want to migrate their files from an on-premises or off-premises file system or service.

Users can insert files into a user directory structure, as well as allow for basic metadata changes, such as modifications to the permissions of files.

Security: Support security applications are supported for users who have additional security needs, such as antivirus or data loss prevention. The API actions, along with AWS CloudTrail, allow these applications to detect when changes occur in Amazon WorkDocs. Then, the application can take the necessary actions and replace the target file. If the target file violates the policy, the application can also choose to email the user.

eDiscovery/Analytics: General administrative applications are supported, such as eDiscovery and analytics. These applications can choose to mimic or record the actions in an Amazon WorkDocs site, along with AWS CloudTrail, to replicate data for eDiscovery, backup, or analytical applications.

All Amazon WorkDocs API actions are Amazon authenticated and certificate-signed. They not only require the use of the AWS SDK, but also allow for the exclusive use of IAM users and roles to help facilitate access, trust, and permission policies. By creating a role and allowing an IAM user to access the Amazon WorkDocs site, the IAM user gains full administrative visibility into the entire Amazon WorkDocs site (or as set in the IAM policy). This includes, but is not limited to, the ability to modify file permissions and upload any file to any user. This allows developers to perform the three use cases above, as well as give users the ability to grant access on a selective basis using the IAM model.

Functions

abort_document_version_upload(client, document_id, version_id, input, options \\ [])

Aborts the upload of the specified document version that was previously initiated by InitiateDocumentVersionUpload.

The client should make this call only when it no longer intends to upload the document version, or fails to do so.

activate_user(client, user_id, input, options \\ [])

Activates the specified user.

Only active users can access Amazon WorkDocs.

add_resource_permissions(client, resource_id, input, options \\ [])

Creates a set of permissions for the specified folder or document.

The resource permissions are overwritten if the principals already have different permissions.

create_comment(client, document_id, version_id, input, options \\ [])

Adds a new comment to the specified document version.

create_custom_metadata(client, resource_id, input, options \\ [])

Adds one or more custom properties to the specified resource (a folder, document, or version).

create_folder(client, input, options \\ [])

Creates a folder with the specified name and parent folder.

create_labels(client, resource_id, input, options \\ [])

Adds the specified list of labels to the given resource (a document or folder)

create_notification_subscription(client, organization_id, input, options \\ [])

Configure Amazon WorkDocs to use Amazon SNS notifications.

The endpoint receives a confirmation message, and must confirm the subscription.

For more information, see Subscribe to Notifications in the Amazon WorkDocs Developer Guide.

create_user(client, input, options \\ [])

Creates a user in a Simple AD or Microsoft AD directory.

The status of a newly created user is "ACTIVE". New users can access Amazon WorkDocs.

deactivate_user(client, user_id, input, options \\ [])

Deactivates the specified user, which revokes the user's access to Amazon WorkDocs.

delete_comment(client, comment_id, document_id, version_id, input, options \\ [])

Deletes the specified comment from the document version.

delete_custom_metadata(client, resource_id, input, options \\ [])

Deletes custom metadata from the specified resource.

delete_document(client, document_id, input, options \\ [])

Permanently deletes the specified document and its associated metadata.

delete_folder(client, folder_id, input, options \\ [])

Permanently deletes the specified folder and its contents.

delete_folder_contents(client, folder_id, input, options \\ [])

Deletes the contents of the specified folder.

delete_labels(client, resource_id, input, options \\ [])

Deletes the specified list of labels from a resource.

delete_notification_subscription(client, organization_id, subscription_id, input, options \\ [])

Deletes the specified subscription from the specified organization.

delete_user(client, user_id, input, options \\ [])

Deletes the specified user from a Simple AD or Microsoft AD directory.

describe_activities(client, activity_types \\ nil, end_time \\ nil, include_indirect_activities \\ nil, limit \\ nil, marker \\ nil, organization_id \\ nil, resource_id \\ nil, start_time \\ nil, user_id \\ nil, authentication_token \\ nil, options \\ [])

Describes the user activities in a specified time period.

describe_comments(client, document_id, version_id, limit \\ nil, marker \\ nil, authentication_token \\ nil, options \\ [])

List all the comments for the specified document version.

describe_document_versions(client, document_id, fields \\ nil, include \\ nil, limit \\ nil, marker \\ nil, authentication_token \\ nil, options \\ [])

Retrieves the document versions for the specified document.

By default, only active versions are returned.

describe_folder_contents(client, folder_id, include \\ nil, limit \\ nil, marker \\ nil, order \\ nil, sort \\ nil, type \\ nil, authentication_token \\ nil, options \\ [])

Describes the contents of the specified folder, including its documents and subfolders.

By default, Amazon WorkDocs returns the first 100 active document and folder metadata items. If there are more results, the response includes a marker that you can use to request the next set of results. You can also request initialized documents.

describe_groups(client, limit \\ nil, marker \\ nil, organization_id \\ nil, search_query, authentication_token \\ nil, options \\ [])

Describes the groups specified by the query.

Groups are defined by the underlying Active Directory.

describe_notification_subscriptions(client, organization_id, limit \\ nil, marker \\ nil, options \\ [])

Lists the specified notification subscriptions.

describe_resource_permissions(client, resource_id, limit \\ nil, marker \\ nil, principal_id \\ nil, authentication_token \\ nil, options \\ [])

Describes the permissions of a specified resource.

describe_root_folders(client, limit \\ nil, marker \\ nil, authentication_token, options \\ [])

Describes the current user's special folders; the RootFolder and the RecycleBin.

RootFolder is the root of user's files and folders and RecycleBin is the root of recycled items. This is not a valid action for SigV4 (administrative API) clients.

This action requires an authentication token. To get an authentication token, register an application with Amazon WorkDocs. For more information, see Authentication and Access Control for User Applications in the Amazon WorkDocs Developer Guide.

describe_users(client, fields \\ nil, include \\ nil, limit \\ nil, marker \\ nil, order \\ nil, organization_id \\ nil, query \\ nil, sort \\ nil, user_ids \\ nil, authentication_token \\ nil, options \\ [])

Describes the specified users.

You can describe all users or filter the results (for example, by status or organization).

By default, Amazon WorkDocs returns the first 24 active or pending users. If there are more results, the response includes a marker that you can use to request the next set of results.

get_current_user(client, authentication_token, options \\ [])

Retrieves details of the current user for whom the authentication token was generated.

This is not a valid action for SigV4 (administrative API) clients.

This action requires an authentication token. To get an authentication token, register an application with Amazon WorkDocs. For more information, see Authentication and Access Control for User Applications in the Amazon WorkDocs Developer Guide.

get_document(client, document_id, include_custom_metadata \\ nil, authentication_token \\ nil, options \\ [])

Retrieves details of a document.

get_document_path(client, document_id, fields \\ nil, limit \\ nil, marker \\ nil, authentication_token \\ nil, options \\ [])

Retrieves the path information (the hierarchy from the root folder) for the requested document.

By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested document and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the names of the parent folders.

get_document_version(client, document_id, version_id, fields \\ nil, include_custom_metadata \\ nil, authentication_token \\ nil, options \\ [])

Retrieves version metadata for the specified document.

get_folder(client, folder_id, include_custom_metadata \\ nil, authentication_token \\ nil, options \\ [])

Retrieves the metadata of the specified folder.

get_folder_path(client, folder_id, fields \\ nil, limit \\ nil, marker \\ nil, authentication_token \\ nil, options \\ [])

Retrieves the path information (the hierarchy from the root folder) for the specified folder.

By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested folder and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the parent folder names.

get_resources(client, collection_type \\ nil, limit \\ nil, marker \\ nil, user_id \\ nil, authentication_token \\ nil, options \\ [])

Retrieves a collection of resources, including folders and documents.

The only CollectionType supported is SHARED_WITH_ME.

initiate_document_version_upload(client, input, options \\ [])

Creates a new document object and version object.

The client specifies the parent folder ID and name of the document to upload. The ID is optionally specified when creating a new version of an existing document. This is the first step to upload a document. Next, upload the document to the URL returned from the call, and then call UpdateDocumentVersion.

To cancel the document upload, call AbortDocumentVersionUpload.

remove_all_resource_permissions(client, resource_id, input, options \\ [])

Removes all the permissions from the specified resource.

remove_resource_permission(client, principal_id, resource_id, input, options \\ [])

Removes the permission for the specified principal from the specified resource.

update_document(client, document_id, input, options \\ [])

Updates the specified attributes of a document.

The user must have access to both the document and its parent folder, if applicable.

update_document_version(client, document_id, version_id, input, options \\ [])

Changes the status of the document version to ACTIVE.

Amazon WorkDocs also sets its document container to ACTIVE. This is the last step in a document upload, after the client uploads the document to an S3-presigned URL returned by InitiateDocumentVersionUpload.

update_folder(client, folder_id, input, options \\ [])

Updates the specified attributes of the specified folder.

The user must have access to both the folder and its parent folder, if applicable.

update_user(client, user_id, input, options \\ [])

Updates the specified attributes of the specified user, and grants or revokes administrative privileges to the Amazon WorkDocs site.