aws-elixir v0.6.0 AWS.WorkLink

Amazon WorkLink is a cloud-based service that provides secure access to internal websites and web apps from iOS and Android phones.

In a single step, your users, such as employees, can access internal websites as efficiently as they access any other public website. They enter a URL in their web browser, or choose a link to an internal website in an email. Amazon WorkLink authenticates the user's access and securely renders authorized internal web content in a secure rendering service in the AWS cloud. Amazon WorkLink doesn't download or store any internal web content on mobile devices.

Functions

associate_domain(client, input, options \\ [])

Specifies a domain to be associated to Amazon WorkLink.

associate_website_authorization_provider(client, input, options \\ [])

Associates a website authorization provider with a specified fleet.

This is used to authorize users against associated websites in the company network.

associate_website_certificate_authority(client, input, options \\ [])

Imports the root certificate of a certificate authority (CA) used to obtain TLS certificates used by associated websites within the company network.

create_fleet(client, input, options \\ [])

Creates a fleet.

A fleet consists of resources and the configuration that delivers associated websites to authorized users who download and set up the Amazon WorkLink app.

delete_fleet(client, input, options \\ [])

Deletes a fleet.

Prevents users from accessing previously associated websites.

describe_audit_stream_configuration(client, input, options \\ [])

Describes the configuration for delivering audit streams to the customer account.

describe_company_network_configuration(client, input, options \\ [])

Describes the networking configuration to access the internal websites associated with the specified fleet.

describe_device(client, input, options \\ [])

Provides information about a user's device.

describe_device_policy_configuration(client, input, options \\ [])

Describes the device policy configuration for the specified fleet.

describe_domain(client, input, options \\ [])

Provides information about the domain.

describe_fleet_metadata(client, input, options \\ [])

Provides basic information for the specified fleet, excluding identity provider, networking, and device configuration details.

describe_identity_provider_configuration(client, input, options \\ [])

Describes the identity provider configuration of the specified fleet.

describe_website_certificate_authority(client, input, options \\ [])

Provides information about the certificate authority.

disassociate_domain(client, input, options \\ [])

Disassociates a domain from Amazon WorkLink.

End users lose the ability to access the domain with Amazon WorkLink.

disassociate_website_authorization_provider(client, input, options \\ [])

Disassociates a website authorization provider from a specified fleet.

After the disassociation, users can't load any associated websites that require this authorization provider.

disassociate_website_certificate_authority(client, input, options \\ [])

Removes a certificate authority (CA).

list_devices(client, input, options \\ [])

Retrieves a list of devices registered with the specified fleet.

list_domains(client, input, options \\ [])

Retrieves a list of domains associated to a specified fleet.

list_fleets(client, input, options \\ [])

Retrieves a list of fleets for the current account and Region.

list_tags_for_resource(client, resource_arn, options \\ [])

Retrieves a list of tags for the specified resource.

list_website_authorization_providers(client, input, options \\ [])

Retrieves a list of website authorization providers associated with a specified fleet.

list_website_certificate_authorities(client, input, options \\ [])

Retrieves a list of certificate authorities added for the current account and Region.

restore_domain_access(client, input, options \\ [])

Moves a domain to ACTIVE status if it was in the INACTIVE status.

revoke_domain_access(client, input, options \\ [])

Moves a domain to INACTIVE status if it was in the ACTIVE status.

sign_out_user(client, input, options \\ [])

Signs the user out from all of their devices.

The user can sign in again if they have valid credentials.

tag_resource(client, resource_arn, input, options \\ [])

Adds or overwrites one or more tags for the specified resource, such as a fleet.

Each tag consists of a key and an optional value. If a resource already has a tag with the same key, this operation updates its value.

untag_resource(client, resource_arn, input, options \\ [])

Removes one or more tags from the specified resource.

update_audit_stream_configuration(client, input, options \\ [])

Updates the audit stream configuration for the fleet.

update_company_network_configuration(client, input, options \\ [])

Updates the company network configuration for the fleet.

update_device_policy_configuration(client, input, options \\ [])

Updates the device policy configuration for the fleet.

update_domain_metadata(client, input, options \\ [])

Updates domain metadata, such as DisplayName.

update_fleet_metadata(client, input, options \\ [])

Updates fleet metadata, such as DisplayName.

update_identity_provider_configuration(client, input, options \\ [])

Updates the identity provider configuration for the fleet.