aws-elixir v0.6.0 AWS.WorkMail

Amazon WorkMail is a secure, managed business email and calendaring service with support for existing desktop and mobile email clients.

You can access your email, contacts, and calendars using Microsoft Outlook, your browser, or other native iOS and Android email applications. You can integrate WorkMail with your existing corporate directory and control both the keys that encrypt your data and the location in which your data is stored.

The WorkMail API is designed for the following scenarios:

Listing and describing organizations

Managing users

Managing groups

Managing resources

All WorkMail API operations are Amazon-authenticated and certificate-signed. They not only require the use of the AWS SDK, but also allow for the exclusive use of AWS Identity and Access Management users and roles to help facilitate access, trust, and permission policies. By creating a role and allowing an IAM user to access the WorkMail site, the IAM user gains full administrative visibility into the entire WorkMail organization (or as set in the IAM policy). This includes, but is not limited to, the ability to create, update, and delete users, groups, and resources. This allows developers to perform the scenarios listed above, as well as give users the ability to grant access on a selective basis using the IAM model.

Functions

associate_delegate_to_resource(client, input, options \\ [])

Adds a member (user or group) to the resource's set of delegates.

associate_member_to_group(client, input, options \\ [])

Adds a member (user or group) to the group's set.

cancel_mailbox_export_job(client, input, options \\ [])

Cancels a mailbox export job.

If the mailbox export job is near completion, it might not be possible to cancel it.

create_alias(client, input, options \\ [])

Adds an alias to the set of a given member (user or group) of Amazon WorkMail.

create_group(client, input, options \\ [])

Creates a group that can be used in Amazon WorkMail by calling the RegisterToWorkMail operation.

create_resource(client, input, options \\ [])

Creates a new Amazon WorkMail resource.

create_user(client, input, options \\ [])

Creates a user who can be used in Amazon WorkMail by calling the RegisterToWorkMail operation.

delete_access_control_rule(client, input, options \\ [])

Deletes an access control rule for the specified WorkMail organization.

delete_alias(client, input, options \\ [])

Remove one or more specified aliases from a set of aliases for a given user.

delete_group(client, input, options \\ [])

Deletes a group from Amazon WorkMail.

delete_mailbox_permissions(client, input, options \\ [])

Deletes permissions granted to a member (user or group).

delete_resource(client, input, options \\ [])

Deletes the specified resource.

delete_retention_policy(client, input, options \\ [])

Deletes the specified retention policy from the specified organization.

delete_user(client, input, options \\ [])

Deletes a user from Amazon WorkMail and all subsequent systems.

Before you can delete a user, the user state must be DISABLED. Use the DescribeUser action to confirm the user state.

Deleting a user is permanent and cannot be undone. WorkMail archives user mailboxes for 30 days before they are permanently removed.

deregister_from_work_mail(client, input, options \\ [])

Mark a user, group, or resource as no longer used in Amazon WorkMail.

This action disassociates the mailbox and schedules it for clean-up. WorkMail keeps mailboxes for 30 days before they are permanently removed. The functionality in the console is Disable.

describe_group(client, input, options \\ [])

Returns the data available for the group.

describe_mailbox_export_job(client, input, options \\ [])

Describes the current status of a mailbox export job.

describe_organization(client, input, options \\ [])

Provides more information regarding a given organization based on its identifier.

describe_resource(client, input, options \\ [])

Returns the data available for the resource.

describe_user(client, input, options \\ [])

Provides information regarding the user.

disassociate_delegate_from_resource(client, input, options \\ [])

Removes a member from the resource's set of delegates.

disassociate_member_from_group(client, input, options \\ [])

Removes a member from a group.

get_access_control_effect(client, input, options \\ [])

Gets the effects of an organization's access control rules as they apply to a specified IPv4 address, access protocol action, or user ID.

get_default_retention_policy(client, input, options \\ [])

Gets the default retention policy details for the specified organization.

get_mailbox_details(client, input, options \\ [])

Requests a user's mailbox details for a specified organization and user.

list_access_control_rules(client, input, options \\ [])

Lists the access control rules for the specified organization.

list_aliases(client, input, options \\ [])

Creates a paginated call to list the aliases associated with a given entity.

list_group_members(client, input, options \\ [])

Returns an overview of the members of a group.

Users and groups can be members of a group.

list_groups(client, input, options \\ [])

Returns summaries of the organization's groups.

list_mailbox_export_jobs(client, input, options \\ [])

Lists the mailbox export jobs started for the specified organization within the last seven days.

list_mailbox_permissions(client, input, options \\ [])

Lists the mailbox permissions associated with a user, group, or resource mailbox.

list_organizations(client, input, options \\ [])

Returns summaries of the customer's organizations.

list_resource_delegates(client, input, options \\ [])

Lists the delegates associated with a resource.

Users and groups can be resource delegates and answer requests on behalf of the resource.

list_resources(client, input, options \\ [])

Returns summaries of the organization's resources.

list_tags_for_resource(client, input, options \\ [])

Lists the tags applied to an Amazon WorkMail organization resource.

list_users(client, input, options \\ [])

Returns summaries of the organization's users.

put_access_control_rule(client, input, options \\ [])

Adds a new access control rule for the specified organization.

The rule allows or denies access to the organization for the specified IPv4 addresses, access protocol actions, and user IDs. Adding a new rule with the same name as an existing rule replaces the older rule.

put_mailbox_permissions(client, input, options \\ [])

Sets permissions for a user, group, or resource.

This replaces any pre-existing permissions.

put_retention_policy(client, input, options \\ [])

Puts a retention policy to the specified organization.

register_to_work_mail(client, input, options \\ [])

Registers an existing and disabled user, group, or resource for Amazon WorkMail use by associating a mailbox and calendaring capabilities.

It performs no change if the user, group, or resource is enabled and fails if the user, group, or resource is deleted. This operation results in the accumulation of costs. For more information, see Pricing. The equivalent console functionality for this operation is Enable.

Users can either be created by calling the CreateUser API operation or they can be synchronized from your directory. For more information, see DeregisterFromWorkMail.

reset_password(client, input, options \\ [])

Allows the administrator to reset the password for a user.

start_mailbox_export_job(client, input, options \\ [])

Starts a mailbox export job to export MIME-format email messages and calendar items from the specified mailbox to the specified Amazon Simple Storage Service (Amazon S3) bucket.

For more information, see Exporting mailbox content in the Amazon WorkMail Administrator Guide.

tag_resource(client, input, options \\ [])

Applies the specified tags to the specified Amazon WorkMail organization resource.

untag_resource(client, input, options \\ [])

Untags the specified tags from the specified Amazon WorkMail organization resource.

update_mailbox_quota(client, input, options \\ [])

Updates a user's current mailbox quota for a specified organization and user.

update_primary_email_address(client, input, options \\ [])

Updates the primary email for a user, group, or resource.

The current email is moved into the list of aliases (or swapped between an existing alias and the current primary email), and the email provided in the input is promoted as the primary.

update_resource(client, input, options \\ [])

Updates data for the resource.

To have the latest information, it must be preceded by a DescribeResource call. The dataset in the request should be the one expected when performing another DescribeResource call.