aws-elixir v0.6.0 AWS.WorkSpaces

Amazon WorkSpaces Service

Amazon WorkSpaces enables you to provision virtual, cloud-based Microsoft Windows and Amazon Linux desktops for your users.

Functions

associate_connection_alias(client, input, options \\ [])

Associates the specified connection alias with the specified directory to enable cross-Region redirection.

For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Before performing this operation, call DescribeConnectionAliases to make sure that the current state of the connection alias is CREATED.

associate_ip_groups(client, input, options \\ [])

Associates the specified IP access control group with the specified directory.

authorize_ip_rules(client, input, options \\ [])

Adds one or more rules to the specified IP access control group.

This action gives users permission to access their WorkSpaces from the CIDR address ranges specified in the rules.

copy_workspace_image(client, input, options \\ [])

Copies the specified image from the specified Region to the current Region.

create_connection_alias(client, input, options \\ [])

Creates the specified connection alias for use with cross-Region redirection.

For more information, see Cross-Region Redirection for Amazon WorkSpaces.

create_ip_group(client, input, options \\ [])

Creates an IP access control group.

An IP access control group provides you with the ability to control the IP addresses from which users are allowed to access their WorkSpaces. To specify the CIDR address ranges, add rules to your IP access control group and then associate the group with your directory. You can add rules when you create the group or at any time using AuthorizeIpRules.

There is a default IP access control group associated with your directory. If you don't associate an IP access control group with your directory, the default group is used. The default group includes a default rule that allows users to access their WorkSpaces from anywhere. You cannot modify the default IP access control group for your directory.

create_tags(client, input, options \\ [])

Creates the specified tags for the specified WorkSpaces resource.

create_workspaces(client, input, options \\ [])

Creates one or more WorkSpaces.

This operation is asynchronous and returns before the WorkSpaces are created.

delete_connection_alias(client, input, options \\ [])

Deletes the specified connection alias.

For more information, see Cross-Region Redirection for Amazon WorkSpaces.

If you will no longer be using a fully qualified domain name (FQDN) as the registration code for your WorkSpaces users, you must take certain precautions to prevent potential security issues. For more information, see Security Considerations if You Stop Using Cross-Region Redirection.

To delete a connection alias that has been shared, the shared account must first disassociate the connection alias from any directories it has been associated with. Then you must unshare the connection alias from the account it has been shared with. You can delete a connection alias only after it is no longer shared with any accounts or associated with any directories.

delete_ip_group(client, input, options \\ [])

Deletes the specified IP access control group.

You cannot delete an IP access control group that is associated with a directory.

delete_tags(client, input, options \\ [])

Deletes the specified tags from the specified WorkSpaces resource.

delete_workspace_image(client, input, options \\ [])

Deletes the specified image from your account.

To delete an image, you must first delete any bundles that are associated with the image and unshare the image if it is shared with other accounts.

deregister_workspace_directory(client, input, options \\ [])

Deregisters the specified directory.

This operation is asynchronous and returns before the WorkSpace directory is deregistered. If any WorkSpaces are registered to this directory, you must remove them before you can deregister the directory.

describe_account(client, input, options \\ [])

Retrieves a list that describes the configuration of Bring Your Own License (BYOL) for the specified account.

describe_account_modifications(client, input, options \\ [])

Retrieves a list that describes modifications to the configuration of Bring Your Own License (BYOL) for the specified account.

describe_client_properties(client, input, options \\ [])

Retrieves a list that describes one or more specified Amazon WorkSpaces clients.

describe_connection_alias_permissions(client, input, options \\ [])

Describes the permissions that the owner of a connection alias has granted to another AWS account for the specified connection alias.

For more information, see Cross-Region Redirection for Amazon WorkSpaces.

describe_connection_aliases(client, input, options \\ [])

Retrieves a list that describes the connection aliases used for cross-Region redirection.

For more information, see Cross-Region Redirection for Amazon WorkSpaces.

describe_ip_groups(client, input, options \\ [])

Describes one or more of your IP access control groups.

describe_tags(client, input, options \\ [])

Describes the specified tags for the specified WorkSpaces resource.

describe_workspace_bundles(client, input, options \\ [])

Retrieves a list that describes the available WorkSpace bundles.

You can filter the results using either bundle ID or owner, but not both.

describe_workspace_directories(client, input, options \\ [])

Describes the available directories that are registered with Amazon WorkSpaces.

describe_workspace_image_permissions(client, input, options \\ [])

Describes the permissions that the owner of an image has granted to other AWS accounts for an image.

describe_workspace_images(client, input, options \\ [])

Retrieves a list that describes one or more specified images, if the image identifiers are provided.

Otherwise, all images in the account are described.

describe_workspace_snapshots(client, input, options \\ [])

Describes the snapshots for the specified WorkSpace.

describe_workspaces(client, input, options \\ [])

Describes the specified WorkSpaces.

You can filter the results by using the bundle identifier, directory identifier, or owner, but you can specify only one filter at a time.

describe_workspaces_connection_status(client, input, options \\ [])

Describes the connection status of the specified WorkSpaces.

disassociate_connection_alias(client, input, options \\ [])

Disassociates a connection alias from a directory.

Disassociating a connection alias disables cross-Region redirection between two directories in different AWS Regions. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Before performing this operation, call DescribeConnectionAliases to make sure that the current state of the connection alias is CREATED.

disassociate_ip_groups(client, input, options \\ [])

Disassociates the specified IP access control group from the specified directory.

import_workspace_image(client, input, options \\ [])

Imports the specified Windows 10 Bring Your Own License (BYOL) image into Amazon WorkSpaces.

The image must be an already licensed Amazon EC2 image that is in your AWS account, and you must own the image. For more information about creating BYOL images, see Bring Your Own Windows Desktop Licenses.

list_available_management_cidr_ranges(client, input, options \\ [])

Retrieves a list of IP address ranges, specified as IPv4 CIDR blocks, that you can use for the network management interface when you enable Bring Your Own License (BYOL).

The management network interface is connected to a secure Amazon WorkSpaces management network. It is used for interactive streaming of the WorkSpace desktop to Amazon WorkSpaces clients, and to allow Amazon WorkSpaces to manage the WorkSpace.

migrate_workspace(client, input, options \\ [])

Migrates a WorkSpace from one operating system or bundle type to another, while retaining the data on the user volume.

The migration process recreates the WorkSpace by using a new root volume from the target bundle image and the user volume from the last available snapshot of the original WorkSpace. During migration, the original D:Users%USERNAME% user profile folder is renamed to D:Users%USERNAME%MMddyyTHHmmss%.NotMigrated. A new D:Users%USERNAME% folder is generated by the new OS. Certain files in the old user profile are moved to the new user profile.

For available migration scenarios, details about what happens during migration, and best practices, see Migrate a WorkSpace.

modify_account(client, input, options \\ [])

Modifies the configuration of Bring Your Own License (BYOL) for the specified account.

modify_client_properties(client, input, options \\ [])

Modifies the properties of the specified Amazon WorkSpaces clients.

modify_selfservice_permissions(client, input, options \\ [])

Modifies the self-service WorkSpace management capabilities for your users.

For more information, see Enable Self-Service WorkSpace Management Capabilities for Your Users.

modify_workspace_access_properties(client, input, options \\ [])

Specifies which devices and operating systems users can use to access their WorkSpaces.

For more information, see Control Device Access.

modify_workspace_creation_properties(client, input, options \\ [])

Modify the default properties used to create WorkSpaces.

modify_workspace_properties(client, input, options \\ [])

Modifies the specified WorkSpace properties.

For important information about how to modify the size of the root and user volumes, see Modify a WorkSpace.

modify_workspace_state(client, input, options \\ [])

Sets the state of the specified WorkSpace.

To maintain a WorkSpace without being interrupted, set the WorkSpace state to ADMIN_MAINTENANCE. WorkSpaces in this state do not respond to requests to reboot, stop, start, rebuild, or restore. An AutoStop WorkSpace in this state is not stopped. Users cannot log into a WorkSpace in the ADMIN_MAINTENANCE state.

reboot_workspaces(client, input, options \\ [])

Reboots the specified WorkSpaces.

You cannot reboot a WorkSpace unless its state is AVAILABLE or UNHEALTHY.

This operation is asynchronous and returns before the WorkSpaces have rebooted.

rebuild_workspaces(client, input, options \\ [])

Rebuilds the specified WorkSpace.

You cannot rebuild a WorkSpace unless its state is AVAILABLE, ERROR, UNHEALTHY, STOPPED, or REBOOTING.

Rebuilding a WorkSpace is a potentially destructive action that can result in the loss of data. For more information, see Rebuild a WorkSpace.

This operation is asynchronous and returns before the WorkSpaces have been completely rebuilt.

register_workspace_directory(client, input, options \\ [])

Registers the specified directory.

This operation is asynchronous and returns before the WorkSpace directory is registered. If this is the first time you are registering a directory, you will need to create the workspaces_DefaultRole role before you can register a directory. For more information, see Creating the workspaces_DefaultRole Role.

restore_workspace(client, input, options \\ [])

Restores the specified WorkSpace to its last known healthy state.

You cannot restore a WorkSpace unless its state is AVAILABLE, ERROR, UNHEALTHY, or STOPPED.

Restoring a WorkSpace is a potentially destructive action that can result in the loss of data. For more information, see Restore a WorkSpace.

This operation is asynchronous and returns before the WorkSpace is completely restored.

revoke_ip_rules(client, input, options \\ [])

Removes one or more rules from the specified IP access control group.

start_workspaces(client, input, options \\ [])

Starts the specified WorkSpaces.

You cannot start a WorkSpace unless it has a running mode of AutoStop and a state of STOPPED.

stop_workspaces(client, input, options \\ [])

Stops the specified WorkSpaces.

You cannot stop a WorkSpace unless it has a running mode of AutoStop and a state of AVAILABLE, IMPAIRED, UNHEALTHY, or ERROR.

terminate_workspaces(client, input, options \\ [])

Terminates the specified WorkSpaces.

Terminating a WorkSpace is a permanent action and cannot be undone. The user's data is destroyed. If you need to archive any user data, contact Amazon Web Services before terminating the WorkSpace.

You can terminate a WorkSpace that is in any state except SUSPENDED.

This operation is asynchronous and returns before the WorkSpaces have been completely terminated.

update_connection_alias_permission(client, input, options \\ [])

Shares or unshares a connection alias with one account by specifying whether that account has permission to associate the connection alias with a directory.

If the association permission is granted, the connection alias is shared with that account. If the association permission is revoked, the connection alias is unshared with the account. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Before performing this operation, call DescribeConnectionAliases to make sure that the current state of the connection alias is CREATED.

To delete a connection alias that has been shared, the shared account must first disassociate the connection alias from any directories it has been associated with. Then you must unshare the connection alias from the account it has been shared with. You can delete a connection alias only after it is no longer shared with any accounts or associated with any directories.

update_rules_of_ip_group(client, input, options \\ [])

Replaces the current rules of the specified IP access control group with the specified rules.

update_workspace_image_permission(client, input, options \\ [])

Shares or unshares an image with one account by specifying whether that account has permission to copy the image.

If the copy image permission is granted, the image is shared with that account. If the copy image permission is revoked, the image is unshared with the account.

To delete an image that has been shared, you must unshare the image before you delete it.

Sharing Bring Your Own License (BYOL) images across AWS accounts isn't supported at this time in the AWS GovCloud (US-West) Region. To share BYOL images across accounts in the AWS GovCloud (US-West) Region, contact AWS Support.