aws-elixir v0.6.0 AWS.Pinpoint

Doc Engage API - Amazon Pinpoint API

Functions

create_app(client, input, options \\ [])

Creates an application.

create_campaign(client, application_id, input, options \\ [])

Creates a new campaign for an application or updates the settings of an existing campaign for an application.

create_email_template(client, template_name, input, options \\ [])

Creates a message template for messages that are sent through the email channel.

create_export_job(client, application_id, input, options \\ [])

Creates an export job for an application.

create_import_job(client, application_id, input, options \\ [])

Creates an import job for an application.

create_journey(client, application_id, input, options \\ [])

Creates a journey for an application.

create_push_template(client, template_name, input, options \\ [])

Creates a message template for messages that are sent through a push notification channel.

create_recommender_configuration(client, input, options \\ [])

Creates an Amazon Pinpoint configuration for a recommender model.

create_segment(client, application_id, input, options \\ [])

Creates a new segment for an application or updates the configuration, dimension, and other settings for an existing segment that's associated with an application.

create_sms_template(client, template_name, input, options \\ [])

Creates a message template for messages that are sent through the SMS channel.

create_voice_template(client, template_name, input, options \\ [])

Creates a message template for messages that are sent through the voice channel.

delete_adm_channel(client, application_id, input, options \\ [])

Disables the ADM channel for an application and deletes any existing settings for the channel.

delete_apns_channel(client, application_id, input, options \\ [])

Disables the APNs channel for an application and deletes any existing settings for the channel.

delete_apns_sandbox_channel(client, application_id, input, options \\ [])

Disables the APNs sandbox channel for an application and deletes any existing settings for the channel.

delete_apns_voip_channel(client, application_id, input, options \\ [])

Disables the APNs VoIP channel for an application and deletes any existing settings for the channel.

delete_apns_voip_sandbox_channel(client, application_id, input, options \\ [])

Disables the APNs VoIP sandbox channel for an application and deletes any existing settings for the channel.

delete_app(client, application_id, input, options \\ [])

Deletes an application.

delete_baidu_channel(client, application_id, input, options \\ [])

Disables the Baidu channel for an application and deletes any existing settings for the channel.

delete_campaign(client, application_id, campaign_id, input, options \\ [])

Deletes a campaign from an application.

delete_email_channel(client, application_id, input, options \\ [])

Disables the email channel for an application and deletes any existing settings for the channel.

delete_email_template(client, template_name, input, options \\ [])

Deletes a message template for messages that were sent through the email channel.

delete_endpoint(client, application_id, endpoint_id, input, options \\ [])

Deletes an endpoint from an application.

delete_event_stream(client, application_id, input, options \\ [])

Deletes the event stream for an application.

delete_gcm_channel(client, application_id, input, options \\ [])

Disables the GCM channel for an application and deletes any existing settings for the channel.

delete_journey(client, application_id, journey_id, input, options \\ [])

Deletes a journey from an application.

delete_push_template(client, template_name, input, options \\ [])

Deletes a message template for messages that were sent through a push notification channel.

delete_recommender_configuration(client, recommender_id, input, options \\ [])

Deletes an Amazon Pinpoint configuration for a recommender model.

delete_segment(client, application_id, segment_id, input, options \\ [])

Deletes a segment from an application.

delete_sms_channel(client, application_id, input, options \\ [])

Disables the SMS channel for an application and deletes any existing settings for the channel.

delete_sms_template(client, template_name, input, options \\ [])

Deletes a message template for messages that were sent through the SMS channel.

delete_user_endpoints(client, application_id, user_id, input, options \\ [])

Deletes all the endpoints that are associated with a specific user ID.

delete_voice_channel(client, application_id, input, options \\ [])

Disables the voice channel for an application and deletes any existing settings for the channel.

delete_voice_template(client, template_name, input, options \\ [])

Deletes a message template for messages that were sent through the voice channel.

get_adm_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the ADM channel for an application.

get_apns_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the APNs channel for an application.

get_apns_sandbox_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the APNs sandbox channel for an application.

get_apns_voip_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the APNs VoIP channel for an application.

get_apns_voip_sandbox_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the APNs VoIP sandbox channel for an application.

get_app(client, application_id, options \\ [])

Retrieves information about an application.

get_application_date_range_kpi(client, application_id, kpi_name, end_time \\ nil, next_token \\ nil, page_size \\ nil, start_time \\ nil, options \\ [])

Retrieves (queries) pre-aggregated data for a standard metric that applies to an application.

get_application_settings(client, application_id, options \\ [])

Retrieves information about the settings for an application.

get_apps(client, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about all the applications that are associated with your Amazon Pinpoint account.

get_baidu_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the Baidu channel for an application.

get_campaign(client, application_id, campaign_id, options \\ [])

Retrieves information about the status, configuration, and other settings for a campaign.

get_campaign_activities(client, application_id, campaign_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about all the activities for a campaign.

get_campaign_date_range_kpi(client, application_id, campaign_id, kpi_name, end_time \\ nil, next_token \\ nil, page_size \\ nil, start_time \\ nil, options \\ [])

Retrieves (queries) pre-aggregated data for a standard metric that applies to a campaign.

get_campaign_version(client, application_id, campaign_id, version, options \\ [])

Retrieves information about the status, configuration, and other settings for a specific version of a campaign.

get_campaign_versions(client, application_id, campaign_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about the status, configuration, and other settings for all versions of a campaign.

get_campaigns(client, application_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about the status, configuration, and other settings for all the campaigns that are associated with an application.

get_channels(client, application_id, options \\ [])

Retrieves information about the history and status of each channel for an application.

get_email_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the email channel for an application.

get_email_template(client, template_name, version \\ nil, options \\ [])

Retrieves the content and settings of a message template for messages that are sent through the email channel.

get_endpoint(client, application_id, endpoint_id, options \\ [])

Retrieves information about the settings and attributes of a specific endpoint for an application.

get_event_stream(client, application_id, options \\ [])

Retrieves information about the event stream settings for an application.

get_export_job(client, application_id, job_id, options \\ [])

Retrieves information about the status and settings of a specific export job for an application.

get_export_jobs(client, application_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about the status and settings of all the export jobs for an application.

get_gcm_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the GCM channel for an application.

get_import_job(client, application_id, job_id, options \\ [])

Retrieves information about the status and settings of a specific import job for an application.

get_import_jobs(client, application_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about the status and settings of all the import jobs for an application.

get_journey(client, application_id, journey_id, options \\ [])

Retrieves information about the status, configuration, and other settings for a journey.

get_journey_date_range_kpi(client, application_id, journey_id, kpi_name, end_time \\ nil, next_token \\ nil, page_size \\ nil, start_time \\ nil, options \\ [])

Retrieves (queries) pre-aggregated data for a standard engagement metric that applies to a journey.

get_journey_execution_activity_metrics(client, application_id, journey_activity_id, journey_id, next_token \\ nil, page_size \\ nil, options \\ [])

Retrieves (queries) pre-aggregated data for a standard execution metric that applies to a journey activity.

get_journey_execution_metrics(client, application_id, journey_id, next_token \\ nil, page_size \\ nil, options \\ [])

Retrieves (queries) pre-aggregated data for a standard execution metric that applies to a journey.

get_push_template(client, template_name, version \\ nil, options \\ [])

Retrieves the content and settings of a message template for messages that are sent through a push notification channel.

get_recommender_configuration(client, recommender_id, options \\ [])

Retrieves information about an Amazon Pinpoint configuration for a recommender model.

get_recommender_configurations(client, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about all the recommender model configurations that are associated with your Amazon Pinpoint account.

get_segment(client, application_id, segment_id, options \\ [])

Retrieves information about the configuration, dimension, and other settings for a specific segment that's associated with an application.

get_segment_export_jobs(client, application_id, segment_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about the status and settings of the export jobs for a segment.

get_segment_import_jobs(client, application_id, segment_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about the status and settings of the import jobs for a segment.

get_segment_version(client, application_id, segment_id, version, options \\ [])

Retrieves information about the configuration, dimension, and other settings for a specific version of a segment that's associated with an application.

get_segment_versions(client, application_id, segment_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about the configuration, dimension, and other settings for all the versions of a specific segment that's associated with an application.

get_segments(client, application_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about the configuration, dimension, and other settings for all the segments that are associated with an application.

get_sms_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the SMS channel for an application.

get_sms_template(client, template_name, version \\ nil, options \\ [])

Retrieves the content and settings of a message template for messages that are sent through the SMS channel.

get_user_endpoints(client, application_id, user_id, options \\ [])

Retrieves information about all the endpoints that are associated with a specific user ID.

get_voice_channel(client, application_id, options \\ [])

Retrieves information about the status and settings of the voice channel for an application.

get_voice_template(client, template_name, version \\ nil, options \\ [])

Retrieves the content and settings of a message template for messages that are sent through the voice channel.

list_journeys(client, application_id, page_size \\ nil, token \\ nil, options \\ [])

Retrieves information about the status, configuration, and other settings for all the journeys that are associated with an application.

list_tags_for_resource(client, resource_arn, options \\ [])

Retrieves all the tags (keys and values) that are associated with an application, campaign, message template, or segment.

list_template_versions(client, template_name, template_type, next_token \\ nil, page_size \\ nil, options \\ [])

Retrieves information about all the versions of a specific message template.

list_templates(client, next_token \\ nil, page_size \\ nil, prefix \\ nil, template_type \\ nil, options \\ [])

Retrieves information about all the message templates that are associated with your Amazon Pinpoint account.

phone_number_validate(client, input, options \\ [])

Retrieves information about a phone number.

put_event_stream(client, application_id, input, options \\ [])

Creates a new event stream for an application or updates the settings of an existing event stream for an application.

put_events(client, application_id, input, options \\ [])

Creates a new event to record for endpoints, or creates or updates endpoint data that existing events are associated with.

remove_attributes(client, application_id, attribute_type, input, options \\ [])

Removes one or more attributes, of the same attribute type, from all the endpoints that are associated with an application.

send_messages(client, application_id, input, options \\ [])

Creates and sends a direct message.

send_users_messages(client, application_id, input, options \\ [])

Creates and sends a message to a list of users.

tag_resource(client, resource_arn, input, options \\ [])

Adds one or more tags (keys and values) to an application, campaign, message template, or segment.

untag_resource(client, resource_arn, input, options \\ [])

Removes one or more tags (keys and values) from an application, campaign, message template, or segment.

update_adm_channel(client, application_id, input, options \\ [])

Enables the ADM channel for an application or updates the status and settings of the ADM channel for an application.

update_apns_channel(client, application_id, input, options \\ [])

Enables the APNs channel for an application or updates the status and settings of the APNs channel for an application.

update_apns_sandbox_channel(client, application_id, input, options \\ [])

Enables the APNs sandbox channel for an application or updates the status and settings of the APNs sandbox channel for an application.

update_apns_voip_channel(client, application_id, input, options \\ [])

Enables the APNs VoIP channel for an application or updates the status and settings of the APNs VoIP channel for an application.

update_apns_voip_sandbox_channel(client, application_id, input, options \\ [])

Enables the APNs VoIP sandbox channel for an application or updates the status and settings of the APNs VoIP sandbox channel for an application.

update_application_settings(client, application_id, input, options \\ [])

Updates the settings for an application.

update_baidu_channel(client, application_id, input, options \\ [])

Enables the Baidu channel for an application or updates the status and settings of the Baidu channel for an application.

update_campaign(client, application_id, campaign_id, input, options \\ [])

Updates the configuration and other settings for a campaign.

update_email_channel(client, application_id, input, options \\ [])

Enables the email channel for an application or updates the status and settings of the email channel for an application.

update_email_template(client, template_name, input, options \\ [])

Updates an existing message template for messages that are sent through the email channel.

update_endpoint(client, application_id, endpoint_id, input, options \\ [])

Creates a new endpoint for an application or updates the settings and attributes of an existing endpoint for an application.

You can also use this operation to define custom attributes for an endpoint. If an update includes one or more values for a custom attribute, Amazon Pinpoint replaces (overwrites) any existing values with the new values.

update_endpoints_batch(client, application_id, input, options \\ [])

Creates a new batch of endpoints for an application or updates the settings and attributes of a batch of existing endpoints for an application.

You can also use this operation to define custom attributes for a batch of endpoints. If an update includes one or more values for a custom attribute, Amazon Pinpoint replaces (overwrites) any existing values with the new values.

update_gcm_channel(client, application_id, input, options \\ [])

Enables the GCM channel for an application or updates the status and settings of the GCM channel for an application.

update_journey(client, application_id, journey_id, input, options \\ [])

Updates the configuration and other settings for a journey.

update_journey_state(client, application_id, journey_id, input, options \\ [])

Cancels (stops) an active journey.

update_push_template(client, template_name, input, options \\ [])

Updates an existing message template for messages that are sent through a push notification channel.

update_recommender_configuration(client, recommender_id, input, options \\ [])

Updates an Amazon Pinpoint configuration for a recommender model.

update_segment(client, application_id, segment_id, input, options \\ [])

Creates a new segment for an application or updates the configuration, dimension, and other settings for an existing segment that's associated with an application.

update_sms_channel(client, application_id, input, options \\ [])

Enables the SMS channel for an application or updates the status and settings of the SMS channel for an application.

update_sms_template(client, template_name, input, options \\ [])

Updates an existing message template for messages that are sent through the SMS channel.

update_template_active_version(client, template_name, template_type, input, options \\ [])

Changes the status of a specific version of a message template to active.

update_voice_channel(client, application_id, input, options \\ [])

Enables the voice channel for an application or updates the status and settings of the voice channel for an application.

update_voice_template(client, template_name, input, options \\ [])

Updates an existing message template for messages that are sent through the voice channel.